AdventOfCode: Make it appear TDD
