+++ pom.xml
@@ -82,6 +82,12 @@
             <artifactId>mockito-core</artifactId>
             <scope>test</scope>
         </dependency>
+        <dependency>
+            <groupId>org.scalatest</groupId>
+            <artifactId>scalatest_2.13</artifactId>
+            <version>3.2.2</version>
+            <scope>test</scope>
+        </dependency>
 
         <!-- Why the F do we need THIS MANY Junit dependencies -->
