--- src/main/scala/io/github/teonistor/ttt/experimental/AdventOfCode.scala
@@ -1,326 +1,0 @@
-package io.github.teonistor.ttt.experimental
-
-import java.lang.Integer.parseInt
-import java.util.regex.Pattern
-
-import scala.collection.immutable.HashMap
-import scala.util.Try
-
-object AdventOfCode {
-  def main(args: Array[String]) {
-    _11_1()
-    _11_2()
-  }
-
-  def _02_1() {
-    val input = "13-17 s: ssssssssssssgsssj\n7-9 p: pnlzhcppvl\n5-6 z: zzbwrv\n7-15 w: wwwwwcqwwwwwwwww\n8-9 h: hhhhhhhhsh\n12-13 l: mtrkqfllrglll\n3-4 t: nntt\n10-14 g: svgggggfhqggghlg\n3-10 x: djxxxpmcxx\n1-10 z: ztzzzgzzzz\n3-4 j: jjjjj\n3-4 c: fbhnsccbc\n2-4 q: thql\n15-18 w: tcwzpwzfwwqftvczbw\n5-7 l: lblwblbmllqnlbl\n3-4 b: bvbfnbb\n4-5 f: ffgzf\n7-9 c: cctcccccc\n5-6 k: szkgsk\n2-5 n: nqvnmgnqnsxfn\n2-10 s: dsfsbsssltss\n12-14 s: ssskrssssssfsxpsqsp\n13-15 j: jjjjjjjjjjjjjjjj\n16-17 l: jllclhllkgcpljzwk\n6-9 f: fsffwffbffbfff\n5-9 s: xtzscxsstcss\n6-9 h: hmnhshhhhqhrthmh\n3-6 x: xxxxxx\n9-11 m: mcmmmktmtmmmmm\n2-6 r: grrprrsmr\n15-16 v: vvvvvvvvvvvvvvwdv\n2-14 j: jmjdbhjjjjjjjxnjj\n5-9 r: rrwrrrrrr\n3-9 b: wbbcbvbtbbbbb\n2-9 r: rrtjlplkrgmrrrrxq\n9-12 s: sssshsmsssss\n2-14 t: qzxxvthtthtthv\n2-11 p: pvlsppwmgcspl\n12-15 s: dsjqsshdssbssstsssss\n3-9 k: tkkkpkkklk\n7-15 s: sssssssssssssgsssss\n6-9 x: xxxsxzxglpwxl\n2-3 x: xxmh\n7-9 g: ggsggggggg\n4-17 t: vtttrrrhtghfjvnqtttj\n1-2 n: jgnpn\n11-12 k: vlkkbkkkkklk\n9-12 v: vvvvvvvvwvvvv\n10-11 j: jjjjljjjjvm\n1-5 x: vvxlx\n3-4 l: flll\n10-11 q: qqqqqqqqqlq\n4-5 d: vdvhmpdtrdhdzfdsgk\n14-16 b: bbqbbbxbhrbpbtmhbbb\n18-19 d: dzpnbxngpgsjtnvtcdd\n2-16 v: jtrhtlfhvkbbflfq\n15-17 q: qqqsqqqqqqjqqqlqm\n5-6 q: tqrhtqqspdqvzqbx\n4-8 b: lbbzbwtv\n10-13 f: hrsnsfflfffsfjff\n10-14 k: kkjkkkkkkkkkkkkkkhk\n4-9 c: ccwcccccc\n5-7 r: grkrrrwjrrgbdtd\n8-12 p: tgspnljpfhtr\n5-6 v: vvvvvvm\n5-7 v: vvvvvvv\n6-8 q: qqqqqdqqq\n2-4 v: xfvqvlp\n3-4 g: gghb\n3-14 z: xflgzzrkzqzxzzxgd\n1-7 g: ldsgswgdgdmjlggb\n5-10 h: rhhmhhnbdhhqhwp\n16-18 m: mmmmmmwmmmmmmmmmmmm\n5-11 j: scssgcjjjjpzjjn\n13-14 s: wznksvkfvfskfs\n3-9 q: qdqqkqtbcnqqqsrklqbf\n7-9 c: cccchcmcvdcczwt\n15-16 g: gggdgggggggggggg\n1-5 k: kkkkkk\n7-10 q: qqqqqqqqqqqqq\n3-12 b: bbbrhbspbrbfbbqbbbb\n3-5 r: gczrgr\n2-6 h: whnlvhsftp\n8-20 h: fhhhsdfhhfzgmhpqhhjh\n15-16 z: sbzzzzzzdzjbwhgb\n4-8 h: jgtbmhwb\n8-9 s: hssksrnhlsdslssvcsbs\n2-6 v: vvvvvvv\n13-14 x: xxxxxxxxxxxxnsbx\n12-13 l: pgllllltpdfpxlll\n2-17 v: vqnncnxkssmhvkzstzgp\n4-5 r: rrrhw\n8-11 p: pppdppphppp\n6-7 t: lrmqtts\n8-12 h: hhlthhhhhhhhhtrzh\n9-10 g: jpgpmpcrwfkz\n12-13 w: wwwwwwwwwwhwwww\n5-13 l: jwtlqglllllxkllls\n2-4 c: ncpcgd\n3-7 l: llslllll\n2-5 n: jdfwnn\n3-5 z: xctxd\n4-6 l: lllrll\n6-7 q: fqqqqqqzqqqqq\n9-12 c: gccwgwqlcxccchchcrcb\n7-12 p: pppgpppppppps\n10-16 g: xrgfggzzghtgxgbggg\n7-11 d: tdddddqdwdgxldj\n11-14 p: pppppbpppppppppp\n1-2 j: jhjj\n8-9 v: pvvvvvvvcgvq\n5-6 n: qcrhnnkktxvdxcp\n7-12 x: lfnxxxdxxxxp\n10-12 s: sssssssssbsgs\n2-3 x: rqhzxmjnjxx\n13-14 w: wwwwvwwwwwwwqhww\n10-11 t: dgttttgttttt\n7-9 w: wdzwwwwjj\n4-10 v: vvsjvvbvvncvvvvxv\n6-8 w: wwwwwpwq\n7-9 j: jtjjjjrjmxg\n5-11 z: zzdgztshstzczw\n1-4 s: ssss\n3-10 p: pqppgrphrpkdmp\n9-12 x: xxxqxxxlxxsx\n3-4 v: nvswzv\n3-9 d: vdjvdcxdkdd\n2-10 v: vvjcvvvzvkb\n11-14 g: gggzggggggjgghg\n4-19 p: pppqppppppppppppppvq\n3-5 n: ngznkgdpn\n13-14 t: tttwtbtttttxxmttt\n3-7 q: qnqqqqqtqqz\n3-4 r: rkgr\n9-12 l: lqkfljzllqtl\n5-8 m: vpmmsnfmtxbcknlmpb\n15-19 s: ssptssnhsksnszsdpqss\n13-14 j: jsjjjjnjjjjjjjj\n6-9 p: ppppppppz\n5-8 r: tqzbnvgsvfrx\n7-9 h: hhhhhhhhp\n8-11 s: ssssssssssss\n3-10 t: ttttttttttttt\n4-5 g: ngtgg\n7-8 n: tnnnnnnqn\n12-16 x: bxxxxxzxxxxxxxbxx\n1-3 z: zzzt\n11-14 x: xxgxxxxxxxwcxfx\n6-11 v: rgkkhpkjvvf\n16-17 n: nfgnnnnnnnnnnnngvn\n3-7 w: wkwwwww\n5-16 g: ggggcggwgggggggggg\n1-3 l: sjvjlkzrpn\n17-19 p: pppppppppppppppptpd\n2-6 b: vjjwbqbdbpwsbw\n4-8 g: gggggjvgd\n2-6 w: pzzpwdm\n3-10 g: ggggcggggggg\n1-3 v: vxvjcvcpslpfn\n9-10 r: qrrwrrjrrr\n5-10 m: qbmdzvqmxnnmmmtk\n2-14 q: vqttpdjhqvkqzqxz\n12-13 r: rrxbrrrrrcrptjrr\n13-15 s: cscssssssssssssss\n7-10 z: zxbrzxbdbhzzzfzz\n2-16 n: skqkxqmqbvbrnspnrgq\n14-18 b: bbxsbbdpbnbbbnbbbpbb\n5-8 b: bmcrdhbbk\n11-13 f: fffffffzfffwffb\n5-6 w: wwbwlww\n12-14 d: gzxclqbnmnxjlzd\n13-19 z: fzzzctzzdsbgzzzzwzpb\n2-12 w: lcwwtwwpwffdjqwms\n9-10 x: hznxnxxxxxxxgkxx\n2-4 g: cwfjtlgf\n11-12 k: bqfkkkklkkkkb\n2-4 l: llllgxk\n5-9 m: cpwmqkzhxs\n8-15 b: bbwlwkblvbwdvrjbgsb\n3-4 f: ffmff\n4-10 c: nqxcqcfcqc\n10-12 d: ddqddddddtdhd\n3-4 n: dghnnqnn\n6-15 s: rsbxbrfmssdknppw\n2-13 m: mmkmmmmqdmkmmm\n8-12 n: nnnnnnnnnnnn\n2-5 d: mdkwd\n14-15 l: hllllllllllllgl\n6-11 g: jgccfnggggpgkt\n4-6 j: jjjtjjjjjj\n9-12 v: hgpjjlvhvhljb\n2-4 g: jfwdgfr\n4-5 f: fhffffchf\n16-17 j: jjjjjjjjjjjjjjjjj\n1-4 t: tcttwtttttrkt\n6-13 x: qmxfxxwxrxdxxxfvgv\n6-7 c: cckpccc\n3-10 m: qwmmpknsgmdmfvrcvw\n2-4 r: rszv\n1-7 q: qqqqqqqq\n1-6 l: lzlmjl\n11-12 q: mqqqqxbqqrqqqdqqbkq\n8-11 h: jtvbhxhhphhhh\n9-11 b: bbbbfbzbzbb\n5-7 l: gljssxddzscvhwrlrb\n3-6 c: fxcbccddxxcn\n3-5 s: ssssz\n17-18 n: jctqdsjchwknkffpnn\n8-10 c: crtwbwcfcqkf\n3-5 r: qrzqc\n7-12 z: zzzzzzzzsfzsz\n13-15 p: pppppppppzrllplp\n11-12 h: nhhhhjhqhhplhhh\n4-5 l: lllll\n4-10 g: gggzgggggg\n15-16 h: qhhhshrhhfhhhhch\n5-7 k: jsrvnjlkq\n16-17 p: pppplppppppppppml\n13-14 m: xqhmjdmrmmmffr\n6-11 c: cccccgccccrccccccc\n9-10 l: llllllllxllt\n8-9 b: bbbpbbbwbdnkbb\n5-9 m: sfmmpmmmzn\n5-15 k: kkkkkkkkkkkkkkkckkkk\n3-4 h: hhhh\n15-19 g: gggwgggrslgxjgfgbfs\n12-13 d: dddddddddddxb\n6-14 k: kkrgdvhscffwwm\n14-16 m: zmmmmqmmmmmmmgmm\n8-9 w: wwwwgwwww\n1-12 x: xxxxxxxxxxxx\n1-5 b: bwrbbfwgh\n1-3 t: ftttttttttttttt\n10-11 m: mwmmmwkmxsf\n8-9 t: tqtttttttdtttt\n7-8 z: hzbkpzfg\n2-3 h: htblhdhh\n3-4 p: pptnlpc\n4-7 l: lqlnqhljt\n12-14 n: nnnnnnnnnnjnnt\n1-4 d: dszd\n12-13 f: fvrffffffffkfff\n9-15 t: bkzstqvtgtwxtzqttd\n6-7 v: vrvvvkgv\n6-9 n: knbpnnjnnsn\n5-8 n: nznkzzqrbn\n9-12 s: ssscssssgsqwsspssp\n4-16 q: slqqqqqhqqqqqqdqc\n1-3 d: gdttdd\n2-6 h: hhqzmh\n4-9 k: lkckkpbzkl\n8-17 b: bbbbkbbpbbbbbbbbn\n1-2 r: bqmrw\n8-9 s: ztssqssbz\n1-13 s: stbqmhspsdgjs\n8-10 z: trnsmlxzmzzzz\n6-8 v: vlxvgzbzvvv\n3-6 c: xddhwtcbpcf\n13-19 g: rjvsgpnfjvzlxxzzgxgm\n5-9 m: qffpmxmfmk\n14-15 c: dcfwcpxcmvpxxtcbmzhc\n6-10 k: jkpbkvghpxksfkkk\n2-3 j: jmgjjjj\n7-12 z: szrznmkjxnzj\n4-9 s: ssssssssss\n2-4 h: hkqvfgpsxlnhtrqr\n6-11 g: gggglrlgggwv\n1-6 v: vrbvlvrvv\n4-10 j: jjjjjjjjjz\n3-12 k: kkjkkkkkfckgkkkkk\n3-15 x: xxrxxxxxxxwxpxg\n5-8 v: vvhvvdzvqfrvgqvt\n5-17 z: zznzzzzzzzzzzzzvf\n4-12 n: fdrncrggjlbgbmwv\n6-17 d: tcddhdbdvtsnrdstd\n1-10 m: mqljmgchlmsx\n4-6 l: fbllbl\n8-9 p: kpmtpphwp\n3-6 j: sjjjjj\n5-6 n: nnnnbs\n7-12 p: ppqppngcphpk\n5-15 x: xxxxgxxxxxxxxxqx\n1-3 p: ppppm\n3-6 w: wwwkwww\n1-4 r: rrrx\n3-4 z: zzznp\n12-14 n: tnzwtmnnnnndnfnnn\n9-10 d: ddfdddtdddd\n4-14 l: hvllsgncmhxhwl\n10-15 j: kxftfsmjvjgslsjpp\n2-10 g: gtgvggqggggxggdg\n1-9 v: cvnxhhffnbcv\n1-14 c: ccccccccccccccc\n3-8 q: fnqmkbhqtbjqvczq\n2-4 v: fvphvgv\n10-11 k: kkkkkkkkkpt\n1-10 d: dddldkddmdd\n4-6 x: xxxxxxxx\n4-11 k: kxkfkkkskkrkmk\n1-6 n: nlmngnkqww\n1-9 j: jjjjjjjjsj\n5-6 b: bbbbbb\n9-11 q: qqnqqmqqqpq\n4-7 r: drrrrsrrz\n10-15 n: nnhncnpsnxtnpnqnnghn\n3-6 d: ddddddd\n4-7 x: bxvxbxjwxgx\n1-2 n: cmpwsbjzghgndj\n2-4 r: przczfjfrbwj\n3-17 k: qbvmqxxffdfpkwxdgxv\n7-12 v: llmjvscvjvpvp\n2-7 z: vzxzxxz\n9-18 r: srdrjcrrrvhrcqrrrr\n7-8 f: rffffvtrfsqff\n5-8 w: pdcwwvmwwwkzwwjwxks\n3-4 c: cccc\n15-19 c: cccccccnccclbnplccqc\n2-13 f: fffffffffffffff\n3-9 t: thrkzdttb\n6-11 p: pppppgppppjp\n5-8 q: qqqrqfqs\n7-11 r: crprgtwwrzp\n2-10 f: qfhgfftfdbfnlffffff\n4-5 h: dhvvlcw\n3-6 s: sffsfws\n2-7 b: jrvxbhbp\n10-11 w: wwwgwwwwwkb\n5-9 d: dddddddddddddkdv\n4-14 b: bbbbbbbbbbbbbbb\n4-8 x: xxxxdxbqxc\n5-9 l: phlkhlllbc\n1-2 s: ssgskqvrsrwdt\n19-20 x: xcxqxvxxxtstnqxhxxxx\n6-7 x: gxxxxgb\n4-5 c: cccccc\n3-10 q: pzzjqjqpfbqws\n4-10 t: ttdttsccttpt\n14-18 c: ccrccccccccccccccrc\n2-5 x: xxzqx\n5-11 c: qbckpccckccccws\n5-6 g: gggggg\n3-5 n: wknbnnlnnq\n9-13 j: jjxjjjnjjjdqjjj\n5-9 s: smsxsrfss\n2-4 v: vxphvvk\n6-7 x: xxxqxxbbx\n1-3 g: zrgtghjwd\n4-5 n: nmntwnn\n6-8 g: ggznggggg\n4-5 k: tkkzs\n9-10 g: hqdhggggqggk\n12-16 w: wwztwmwwspwhmwwz\n3-5 m: mmmvmjmmm\n3-4 j: jjjj\n6-16 z: kzbtzqrkzzcfhvpwjszx\n3-9 f: ffwffffffff\n13-15 b: bbbmsbbbbbbwbbbbwb\n11-14 n: nnnnnnnnnnnnnnv\n8-9 v: bvvflvvbjvbtjqrvsxf\n1-5 g: gjgnggg\n15-16 n: nnnnnnnnnnnnnnwvqnn\n12-13 l: lllhllrlllllllll\n5-6 l: llmlll\n8-9 z: zzqzdzvzzzz\n3-4 d: ntdd\n4-5 b: sbndbkjxldqpzfbzbqx\n2-5 t: bwcfzdlfslsd\n4-5 d: dddddd\n1-3 t: jtpt\n10-12 q: qqqqgqqqqqqmqq\n10-19 m: mmmmmmmmmsmmmmmmmgs\n2-14 x: xxxxxgxvxmxxxbx\n3-13 l: tzljvxspwxdnl\n6-10 p: lhnppkpswrppgppzpppn\n2-6 j: djgwjjhvpwkcgsfbf\n10-11 p: ppppppppprpppppp\n7-12 l: lvgcbnlclvjlk\n6-11 z: zzzzzzzzzzn\n3-5 x: xzvhxnpwbxgxx\n4-6 h: hhlrhnl\n8-10 b: brngbslhqhb\n6-11 s: ssqjxsssdsssss\n2-5 s: dbjdjknspbdldfbjd\n10-16 z: ctzzzzznzcwzztzl\n14-15 j: jjjjjjjjjjjjjjjj\n9-13 d: qvdpwddddddddd\n10-12 c: ccccchcccccq\n3-5 x: xxxxx\n13-14 t: ttstttztttttltt\n3-16 k: qvfjklgzlwfwnbjklqdx\n12-13 g: ggggggtgggggggvg\n11-15 j: vqzjvzvxjhjnjtj\n7-8 t: tttttttg\n6-7 x: xxxxfmn\n10-13 m: mmlmmgvfmmmdmmcmmq\n7-14 w: wwkwwwwwwwmptwww\n17-18 b: bbbbbbbbbbbbbbbbbb\n2-8 c: kbvcnrpx\n16-17 n: nnnnnnnnpnnnnnngl\n10-15 k: kbkjkdkwpkkhkkksklx\n2-12 p: pjhdkvwdltvqrhm\n2-4 x: vxhxxk\n3-5 m: mmfmf\n6-8 n: ngnggfnwnn\n9-16 b: bnbbbtcbbbvbbrbbbbb\n2-8 p: ppphpzppcgptpprhpp\n9-12 x: xxxxxxxxxxxx\n5-7 d: dknbddd\n7-9 r: rrrrrrrrj\n3-6 f: fjmfxk\n3-4 n: jnnnbjckggnpjnddhx\n6-12 d: dznpsfdtdsdndjddvcdd\n7-12 x: xxxxxxhqxxxbxbsx\n1-2 k: kkfkkkk\n4-7 l: lprlqllll\n10-13 x: qxxqkjxxxxxxmz\n3-8 f: rxfffwcff\n4-5 s: cpscszdssc\n5-6 t: tttftt\n2-6 p: kpvtkprdqhxzpxrbfp\n6-12 s: ksnsklwqfssdsbsss\n6-7 l: lzlflfl\n7-9 t: ttktttttt\n8-14 d: hdddzndhsddpfpdxt\n8-10 d: ddnddrddvxbzskddzd\n11-13 v: mvhvrtvvvvzvrvdv\n4-5 g: gggjw\n3-6 m: xhmrpm\n7-13 k: kkkkkkkkkkkkk\n10-13 n: nnnnnnmnngnnnnnn\n7-15 m: kcmtmjtmwjmcsbbms\n13-16 j: jjjjjjjjjjjjpjjlj\n2-4 n: nnmnkn\n12-14 j: jjjjjjjjjjjdjj\n1-2 c: cchc\n9-12 m: mmmqbmlwmmjmzmmml\n3-8 r: rrkkrrrt\n4-5 k: wckknxkpkktmks\n9-10 d: ddddpdddnpx\n1-8 q: qmpmxzfqnqctwpprm\n10-12 q: sgkfxhqjkqqq\n5-11 p: rwpppdlmhpptkbrlp\n1-2 d: lzrbxnlcsqhssgdpvjs\n5-6 c: scccck\n4-5 s: ssshxsssws\n3-11 l: nslwgmxmtblzttxxg\n10-11 z: zzzzzzzzzzz\n7-8 j: bjjjjjqdbjj\n7-8 d: glpfjbpdxd\n2-4 k: wkvwrk\n1-5 c: vtcgcbc\n16-17 x: xxxxxxxxxxxxxxxxm\n2-11 f: ffffffsfffffffffffff\n4-8 m: mrmmmmmmmmmmmm\n2-3 x: xvzxx\n2-4 h: hxpn\n13-17 x: xxxxxxxxxxxrxpxxxx\n4-13 q: wcqqtwwmwqqtqmqtqrqq\n10-12 d: dddtdqdnjxdzdfdd\n13-14 f: jffffsffmffqjff\n6-7 d: sxdvddddwqxd\n5-11 d: ddddfdddddcdddd\n11-14 b: bbbbbbbbbhlbbj\n3-4 p: pfpp\n4-7 j: jjjgwkgwnfffvfzfl\n6-9 l: gknrmjgxtlzcbvkgxb\n4-10 n: nnnnnnnnntnnnnnnxn\n1-4 r: rjrr\n3-4 m: nkmr\n9-10 f: fffffffffff\n2-6 r: rnhtml\n3-12 q: qqkqqqqqqqqkqqqqq\n3-10 r: rrnrrrrrrrrrrrrr\n6-8 m: bmmmdmmmmdmh\n10-12 g: ppdvvxgkhfskfqc\n2-10 v: txpvkgvcbqvqfk\n1-5 p: tqcvbtp\n6-7 z: zzzzzzz\n7-16 z: fdjbbkrzjzdbdblhxb\n2-4 x: xlkz\n4-11 m: glhknzmvqzfmnsbn\n6-8 q: qhmdgqvq\n8-16 t: tttdtttdtttttttbvt\n4-7 v: vptcmgvvjvvvfvvmht\n7-8 t: ttttttngt\n5-15 t: ztgbttrntttdvgtptjdb\n10-12 b: bvndbbbbbbbbbbbbtb\n2-3 t: sbqts\n12-16 w: wwwwwwwwwwwwwnwww\n2-3 m: kmmr\n12-15 k: kkmkrkkkkzkdkrgkk\n5-13 h: hhhhzhhhhhhhzhhhh\n16-17 l: lllllllllllllcllq\n1-3 b: bbjb\n2-5 t: vtjhmf\n4-5 q: vzdqfd\n2-3 n: nnnn\n7-14 j: jpjjjjjjljjjjjjjj\n12-15 j: crsjpjjjqjjfjkv\n10-11 t: ttttttttttttt\n1-2 j: jjbjsjvhbfssznf\n9-13 f: ffmfffffzfffgfff\n15-18 g: sgblrqdftpwzggvgqt\n19-20 r: rrrrrrrrrrrrrrrrrrvj\n1-4 j: mjlt\n3-10 x: jxxxxxxxxgxxx\n6-7 p: ppptppp\n7-8 b: jbmnbmsc\n2-6 f: zshfcbhsh\n2-5 r: rrrqwtrvclf\n3-8 g: chgtwlbg\n11-15 n: knnhnnhnnnpnlknnkhnn\n3-8 z: zkqlzzpmjph\n5-10 b: bbbbhbbbbbbbbbbbbb\n6-15 c: ccccdkrncwrcclc\n6-7 c: ckcccwdhcc\n12-13 t: ntttktttttftttt\n5-8 j: jpjzcjjjjj\n7-8 k: kpkkksbkkkkkr\n9-12 s: sqssssbssxcs\n3-7 p: pbbplhtfpktpnppx\n5-7 n: ncnnnsn\n4-6 s: sssrss\n3-16 v: kvvvvvvczvvvvvvvkww\n3-6 k: kkkpmk\n8-9 h: hhhhhhhmb\n7-9 c: czwcccmtckccc\n5-7 t: tstglltkt\n2-6 c: cccccv\n11-14 z: zzzkzqzzzzzvzzz\n3-4 d: ddch\n4-6 j: jjjjjvj\n5-9 c: cccccccccccccc\n7-9 j: hjkwpjjbjfsq\n4-6 t: tttlcdb\n13-15 z: zzzzzzzzzsqzzgzzzkz\n8-10 t: ptckdjtsptlmzrktwcw\n2-5 n: tnbhnnkvnq\n11-12 m: fqmmmmmqmtmkcmmvnwmm\n7-8 z: zzzzztvl\n6-7 j: jbpjjjjpxmqxcbwsjrjj\n2-8 t: ttkdzfwdtflfswlkntt\n6-11 j: rnjjjjjjjjj\n10-13 f: fwhkbfdfffvfs\n5-8 l: pflplntf\n3-5 r: rrrrr\n16-20 g: gggggggggggggggggggg\n1-12 t: xtnsxbjtttxtt\n7-9 r: zrrrrrkrnrrrrrrrrr\n2-11 n: nrjnnnwnnprnnn\n10-12 z: zzzkxzzhznzdz\n7-13 d: pdcjdddqddlddcdkdxk\n10-15 r: rrrbrrrmvrcphrrqr\n1-3 l: rsll\n7-8 x: xxxqvrxx\n5-14 p: ppcdvppbppppjpbg\n5-11 x: dxdxdhkxxvxqxxxfx\n6-12 r: pvbfrhqrhftqrrxcrr\n4-7 t: zdbtzst\n1-13 k: kkkpkghkqkskk\n5-17 g: gvggqgggggqmgxggfgg\n12-13 l: llllwllllllwf\n2-11 b: bbbbbbbbbbbbb\n7-17 h: hhhhhhhhhhhhhhhhhh\n2-3 w: lwww\n2-3 s: vstcs\n4-6 s: vssssxss\n14-16 r: fmrrrrrrrrrrwwrvrr\n6-7 g: dgwgggp\n12-16 g: wggxgggrggggzgbgggg\n7-9 s: tjhxrscnsscssssnms\n3-12 n: nnnnnnnnjnnlnnnnn\n7-8 b: zncbmbglqbbbbpgm\n12-18 l: lllllllllllvlllllwl\n10-16 q: qqqgqfqqqqqrrqgqq\n1-7 n: nnnsnrnnnp\n10-11 q: qqqqqbqqqqq\n5-8 k: bnsfzkkwfkknkccwqkm\n4-5 c: ccdcccc\n5-9 z: nlsgzzzdz\n2-3 k: xkgxhqkpftx\n10-13 d: ddddddddhdddddddd\n4-6 x: xmjxwxx\n14-16 z: zzzzzzzzzzzzzrzq\n3-14 n: qnznjvlzdnnrgwfr\n7-13 s: sssxsspsssssrz\n4-11 k: pkkkcklbkkkk\n9-11 x: xxxxxxxxxxtx\n4-6 z: zzjzzfz\n3-6 d: zddddddz\n10-20 c: smccccvcpbccbqcxbccq\n13-15 m: mmmmmvmmmjmnnmlpsm\n12-13 j: jjjjjjjjjjjjj\n3-5 m: pmmbm\n7-17 m: mmmmmwcmmdfqrmmhlmm\n3-17 g: gcggnggbgdggggggg\n10-16 b: sqpsqkfbnbqzswbb\n3-4 w: wwwn\n5-6 d: dddddpd\n12-15 m: mmmmmmdmmmqlmnvms\n13-15 j: jjjjjtjjjjjjzjcj\n17-18 t: ttttttttjtttttttttt\n17-18 d: dcddmchdwbqrllsxjdv\n16-17 h: hshghhndnhhfhhhhh\n3-5 g: ggggch\n6-13 c: ccvccpccccccvcccccc\n2-4 h: hbfh\n3-9 l: lllptphml\n2-3 s: xwrqssfzr\n9-13 m: mktmmwmmmmrmmrqmmmcm\n8-11 l: rkvxvrglltlc\n1-8 q: jfqqxngqnqchq\n2-6 x: zxgjxx\n5-14 g: mlbgvdglrgcqjgz\n4-5 t: tqgtv\n2-6 s: ssfnzs\n9-11 m: mxwhmmmmmmzm\n1-5 d: dbhjzd\n2-16 f: fffffffffffkffffff\n6-10 b: dbbbbbbbbbbb\n5-6 x: xxxmxxcx\n7-8 l: lllpllcb\n16-17 d: dzdddddddddddddddd\n3-10 t: tklkqftnwcj\n9-15 g: jccxpmfjgntpptgkkcvt\n1-8 r: sdrwgcvprt\n4-5 v: nhbwx\n17-18 z: zzzzzzzzzzzzzzzzxsz\n5-13 l: cbdklqhnklkmwhpp\n4-5 f: fffff\n1-4 q: qpwq\n9-11 x: vxxgrgxxfvvgx\n3-4 r: rrfd\n3-16 x: sxqxxxxxxxxxxxxwxxx\n5-8 s: ssssssss\n5-13 b: jbbhqqbblbbbbfcb\n17-18 p: ppppppppprpppwppphp\n3-9 p: kdpnpfphppvffpwf\n7-8 r: jvrrrrrnrpqrn\n5-6 m: ncmmdxmm\n10-11 v: vvvvvvvvvdl\n3-8 g: jlsscggg\n1-3 m: gwms\n4-5 f: fpkvnf\n6-15 j: zfdrxjgxtbkbbjjctmsk\n5-6 v: qnnbvvwqvbwqblqd\n2-5 v: kvhvt\n2-6 x: lxrgsxg\n6-8 s: sssxswsjsssxss\n1-6 t: dxwjttnqkt\n13-14 s: sssssssssssvjg\n1-2 t: rtttnf\n9-17 h: hjhhhvhhbhhwhhhhphd\n2-3 m: mmmwcm\n10-12 v: vvvvvvvvvvvv\n6-10 j: flcjtzwhwsnjjjrjrj\n3-8 d: ddmrddddd\n5-8 c: ccccslcrwrc\n13-14 s: hpsqtxvkrsssshh\n11-15 t: hrxtjgtwtlmpqfx\n11-14 l: llbxlsllllqrqlllzlb\n3-4 c: crlcvcxwd\n3-9 s: vjbsxhwwdvshfxstc\n7-8 f: jfhnffbgsfjfwf\n15-16 x: ksxqvfsvnxsgvwgpjzl\n16-19 g: gggggggggggggggdggg\n9-11 h: hhhhhchhzmqhh\n2-6 f: kfdlcfrxftzgq\n2-8 p: ppnppkpwwmgp\n1-2 g: ggfgrg\n1-7 n: nnnnnnqwnnn\n5-6 h: hzjgthqphwhnjh\n10-15 l: xlllllllvlllllll\n3-4 l: lzjb\n4-6 l: ldzlql\n7-12 b: mzlbqzbrqjjbddr\n10-13 t: tttttxtttpttt\n4-6 d: dddddd\n4-5 k: kzvkkrkf\n1-2 h: fqml\n17-18 f: jhtfdfgmnchprbwbfrf\n3-10 s: sjsnsksbdsss\n1-17 j: jgcbjmjjjjjcxsjjjjhj\n15-16 k: tkjkjbkkrkkkklmgkj\n2-13 k: kkkkfkqkkkkkkkkpkk\n4-13 d: ddkkdtddzdnnddjdd\n1-3 v: fxvnxdvxnrbjs\n8-11 s: sssssssssspssssss\n1-8 k: khkkkkkkhk\n8-9 k: kkktmkkzk\n4-9 w: wrqdfwxwnfwwwnmzww\n3-5 l: lgtcvl\n13-18 n: nnnnnnnnnnqnnnvnnnn\n15-17 z: zzzzzzzzzhzzzzqzv\n3-5 q: lxnqsvk\n8-13 g: rzzwdlzgbcmggct\n10-14 d: ddvdbddddnddpd\n6-10 d: dphsszqrvz\n8-16 b: bnbgfbkbbbqlbcbbgqbb\n13-15 d: dddjddmqndddddddvjdd\n8-11 q: tnhgqkqfpjhwqgktq\n1-3 n: nnnbn\n11-12 f: fwfffbfcvflff\n7-8 s: sssglsvssvls\n2-5 z: zzzzzzzzzqjzz\n2-11 m: mhzrwphzxgj\n4-6 k: kkkwkd\n8-15 m: mrmmmmmzmmmmzmgmm\n1-7 w: wwwwwwxwwww\n1-3 l: nllllllll\n4-5 m: zbshj\n3-9 x: xssxqxxxsxkzx\n5-9 d: dnpfhsdmrxbvgxqrs\n7-12 r: vhrgcnxrvksg\n8-10 c: ccccccctzv\n16-17 l: lllllllllllllllhl\n6-15 n: nnnnnnngnnfhwwzsnnn\n6-19 z: zzzzzlzzzzzzzzzzzjv\n7-12 b: bxfbbwsjtbbvbqvbmbpb\n5-14 g: gggcjggggggggmgg\n5-8 d: wdddddkdpd\n4-17 x: xtxxxxxxxxxxxxxxmkx\n4-10 c: cccccccccjcscv\n9-11 v: ttjzfjldvvmswpqt\n5-8 r: tjrwhzgfrkgfq\n5-8 m: cvhsmnzmncsfmbqmm\n7-8 z: krctfpzz\n2-6 w: mlnlswvh\n12-15 t: zrntxzttqlthfdttt\n1-8 r: pkcsjkrrrzxxsfnjw\n5-15 p: xbhpkpnngpqvpcwdppg\n3-4 t: tttt\n8-11 j: wjjjmjjjjjqp\n8-14 h: gjslnklhhkhthxh\n2-6 q: qqlqqq\n3-6 p: pppgppppppppppp\n7-11 l: llhllllrllllxldllll\n11-12 d: lddkdddddddddd\n4-12 z: zwlzzqsvvclw\n15-18 c: tflkxflqccwkmlckck\n8-12 h: hhhhrhhhhfwhhh\n6-9 r: rrrmrjrrvkrrjrdr\n8-19 k: tmdlkcktbkkskvkbkbz\n1-13 s: ssssssssssssss\n4-5 v: xtcfvx\n1-15 s: shtssssswghsfsss\n2-4 k: krkl\n7-12 q: dxqqqhqqqtqj\n1-5 r: rrrvvrrrr\n5-9 t: ttrttftxztnhtvkcmtth\n7-9 g: gngkcggrg\n12-13 z: zzzzzzzzzzzbk\n3-4 l: lslllc\n4-5 t: pqchpk\n5-6 h: hhhhml\n15-16 h: qhhhhhhhhhhhhhhhhhmh\n5-6 d: ddddfs\n1-6 d: dthfdzddfnsdddm\n14-15 d: ddddddddndhtddsdnddh\n5-7 j: qjjjpjplljjc\n2-6 q: qqqqqqq\n3-4 b: bbbb\n5-6 s: ssssjjs\n1-6 b: bbbbfb\n7-13 m: cpvmbbnnrmzwf\n2-4 f: fnhn\n6-7 d: gdktlddrdlvmqdtddpzv\n4-5 w: cwlww\n6-9 z: zmdrgzlmzx\n5-10 m: mjbjmmmlmmwxm\n2-4 s: ssnhht\n8-16 t: tttttttdtttdtttjtt\n3-5 z: wxgzhtswb\n13-14 l: lllllblpllllchll\n4-8 g: vggqzvggnggggggggg\n8-9 h: hhthkshhpnhqhgh\n4-7 l: hglxklpl\n2-3 w: mkmgkwzwmw\n6-7 h: hhhhhrz\n4-6 f: tdffff\n1-7 l: lmfhgqschjqglrvwwnnz\n12-13 g: gggggggggggfhggg\n2-10 m: ntmmmdjmmmsvmm\n1-2 s: sqsssssssssss\n3-8 f: frnthfcfxfft\n7-16 w: wwwwwwmwwnwwwwmk\n11-13 j: jjkjrjjjjjjjjjd\n3-4 s: ssvc\n3-13 t: ddttttnnqpqztzbbdv\n3-5 g: gggggg\n3-4 r: rrrzlklfljqvz\n5-12 m: kdgmmrszqpfsbmz\n11-14 g: gqgvzgxhgbtngxggh\n7-16 k: qqptmhhfkhsgkhbmkx\n5-8 p: ppppptppp\n2-17 x: vxtxpvnxvlctrcpfxxx\n2-3 w: bwmwjwwwwwtnqwxc\n4-5 g: gggggg\n4-7 z: znzwzzzs\n4-5 v: kvvvv\n10-14 r: rrrrrrrrrbrrrx\n4-5 t: xtttt\n4-11 w: wwwwwwdcmkrwx\n6-7 c: ccccccjc\n12-14 w: wkwwwwwwwhwlwwwk\n5-9 s: hmssssssssss\n2-4 r: rrrjsh\n12-13 m: mmmmmmmmmmmmmx\n3-5 s: sshsvssss\n4-5 l: xllwnshchfdfk\n16-18 w: wwpwwpwwwmwwzwwtwwww\n2-9 c: zsfgbrrprlszrr\n10-11 n: thnmfndnnbnnbhnnnn\n13-14 v: vvvcvvvvvvvvvvv\n3-4 n: rwcm\n14-15 t: tttttztttttttpjt\n5-8 j: jjjjwhjdj\n5-8 f: fffffffm\n6-7 z: zzqzzzz\n12-19 j: kjrfnsjnhjdzgqpcjzvh\n3-7 p: pxpdvcqdptxpvk\n1-4 c: wccqj\n1-2 w: pwsqhwjdhcm\n2-3 b: bfbn\n10-19 q: qqqqqqqqqtqqqsqrqqq\n5-8 p: pppppppppppppppp\n10-13 g: rjghslgbggkgjdgvp\n3-9 z: fnnfzzdnf\n9-16 g: tlgtgggggggggggqm\n3-12 r: rrrvccrrvbdbncrrr\n1-4 v: jvhbv\n15-16 v: vvvdvvcvvvpvvvvvvvv\n3-4 g: ggggg\n7-8 w: wwwwwwmpw\n4-9 s: gkxshzkpssc\n2-8 q: pqttnpqqqfwqs\n3-4 n: nlnp\n3-4 h: phhh\n16-18 g: hqsxhmmpfgggmgpqhrjg\n15-17 w: wwgwwwwwwwwwwwmwl\n7-9 x: xxxxxxmxv\n1-7 s: ssssssbs\n1-3 x: dxlxxx\n3-6 p: wpxppq\n5-7 c: clcclcfcbccc\n3-5 s: sbmst\n3-5 d: xqmdn\n5-10 z: zbzwfddkzzc\n18-19 d: dpdddddddddddddddwcv\n3-4 r: rrrr\n8-9 f: ffnffffwd\n12-16 h: hxnlxghzkpcpdhqqvl\n6-7 m: jngclmm\n5-13 m: rmmmpmrmmmmmjmmmm\n1-3 f: dfwfs\n7-8 l: lclllwllll\n16-18 j: jjjjjjjjjjpjnjjgjbj\n12-15 f: fffffffffffnfftfffff\n9-13 w: qbnpmklbwxdbbwkklpwb\n1-3 j: mjkj\n1-3 h: hphwdh\n3-5 x: vrxxb\n1-15 v: mvvvvvvvvvvvvvj\n8-12 v: lvzfvzvvnjtvvvvvv\n4-6 g: gggvggggg\n13-20 l: zdllslnlxslqpnvgwlwm\n2-4 l: lllhll\n3-14 k: kckzddtzbmvkpkhb\n4-5 k: hbkgbxzj\n2-4 p: lppp\n2-11 r: rrrrrbrrbrrrrwrrrwr\n4-7 k: knkqkkk\n7-8 r: njrrrrrrrr\n4-7 h: phhhnhdhc\n4-5 r: nsnrrfktwbbhrrrh\n4-5 q: qqzvzsvkq\n6-7 b: jbbbbbbbkbgb\n5-11 f: ffffcfffffgff\n12-16 v: vfvvqvvdvvvlhvzjvmzv\n9-10 v: vvzvvtvvjgvv\n1-7 v: vvvvvvvv\n3-12 z: zmncxhrdzdmtcbxtlrzq\n3-7 r: rbfbrjrr\n12-13 n: nnnnnnnnpnnldnn\n5-12 j: jjjjjjjjjjjlj\n3-6 s: sssfsgss\n13-17 h: hhhfhhhhhhhhvhhhhhhh\n5-10 g: gggxkggggvg\n1-4 b: hbbtb\n14-15 t: tttttttttttttnb\n7-14 d: ddddddddddddddhdw\n15-16 d: rbdddbdwmjdhmpdd\n7-13 s: kxpdntprmskcs\n3-5 m: gmdmpj\n4-5 j: jjjjjvjsj\n1-14 q: qmqqqntqcqxmqsqkqq\n4-5 q: tqxqq\n8-11 m: mmmxmmmvmmv\n5-8 q: qqqqrqqw\n5-15 d: dqdddddddddddvspd\n1-4 t: tttttt\n12-15 z: dhzzxwfjgnzhzxt\n8-11 v: kdnmfnmqvvdvqdlvk\n16-18 s: sssssssssssssssjstss\n2-5 c: ccwdccc\n3-11 x: nxjxxxxxxxxxxx\n13-15 v: vvvvfvvqvlvvvvvv\n4-11 s: sssszssnqjsbsvs\n9-18 j: gxhjjjjjnjjsjjjrjjjj\n3-4 t: ttfttt\n9-10 s: qzgxjhpsss\n9-10 q: qqqqnrtxrqqsqqq\n6-13 k: kkkrgspkkkpwjshmk\n9-12 h: xhhhmhbhhhdhh\n3-12 g: gfgggggjgggggggggg\n13-15 z: zzfzzszzzzzzvzzzhzw\n5-15 m: mzrmmmmndchfzmmrr\n5-10 z: zzzzzzmzzh\n4-5 s: wscfc\n3-6 m: kmmtmm\n5-7 h: hhhhchh\n4-14 w: wwgwwwqwsfrjvmbwj\n2-9 v: vwvvwvmjcqnxv\n9-10 x: qxxhfxchrx\n17-18 x: xxxxxxxxxxxxxxxxwp\n4-6 v: vvvvvtv\n11-14 l: lcqlclllllrlddxlzll\n1-4 r: trrn\n10-13 h: thnsxphnfgvhvq\n11-12 s: sssssssssshsss\n10-11 z: rqpzszzmzdz\n3-6 q: qrqqqq\n5-6 h: hhkhhhh\n8-16 h: wqhjvhlgwtsgvlpf\n3-6 f: gqfpcfsfhfjgrbqv\n3-5 l: lqllsrjlv\n6-10 v: vvvvvvkvjv\n3-8 t: gtttcbnkxxstttd\n2-6 j: jjjjbjz\n10-15 j: ftwjtfzjmjsvwjj\n8-11 d: ddsksddddzfd\n2-3 b: bhngbb\n14-16 c: mllmcbfxwxrqlcjcw\n7-8 q: qqqqqqdv\n5-6 s: slkzlfvg\n3-5 q: qqqqq\n3-12 p: pddpdpszrppxcpjgv\n11-18 r: rrrrrrrrrrprqrrrrr\n3-4 c: cscckcxcn\n11-13 g: gvgwmlngggqggrtgtkg\n5-7 b: smfpbbb\n7-18 l: sqmlklvkqfjrgtqhzr\n6-7 k: kkpqqfgz\n7-9 n: fnndnnxnndn\n11-12 c: ccccccccccxv\n6-9 d: dddvdddddddvd\n1-4 v: nvjfv\n3-8 g: gggggggg\n6-7 x: xxxxxxx\n4-5 c: scccc\n6-13 q: nchlfqqqlqnqqqtq\n1-8 r: srrrrrrnrrrrrrr\n4-5 p: sdpscpppp\n9-11 r: rrrdrkrrrqg\n3-5 m: xwmscmmm\n7-11 r: rrrfrrrrkfrvrf\n4-6 b: bbfbbb\n4-9 v: rblrvrvpvz\n2-8 w: wwwvwgwww\n8-9 w: wwwwwwwwzw\n16-18 t: ttttttttttttttttttt\n8-9 n: nfrzsdjxr\n1-2 w: wwfwwmfbww\n2-4 s: dfsksft\n1-13 d: dwdrccnddqmndcl\n1-4 f: qfflf\n3-7 d: dnddjpdfgc\n3-5 l: lkglx\n1-13 j: jjjjgbjtpjjhd\n15-16 j: jjjjjjvjjjjjjjjjdjj\n4-13 t: tttxztttttttkt\n4-5 m: mmmmm\n4-6 j: jgjrjwjfjx\n8-11 t: ttbjttvttsttdqtn\n14-15 w: wwwwwwwwwwwwwkjw\n1-6 x: nrscxn\n3-13 w: wwwwwvgnwbwwwwwwv\n12-15 p: pppppppppppjpppppppp\n5-11 v: vvvkvbvwkwnvvvsxvv\n1-11 m: mmmmmmmmmmtm\n10-11 c: cccccccccwc\n5-8 f: xvgstwfxfhxknds\n12-13 t: tttttztttttttt\n4-17 k: vkpkfkkkqnkqnkgkkknk\n9-12 k: kkwpzpdzxhxk\n14-17 v: fgvvvvvvvvvvvjpvv\n7-8 k: kndnqkkk\n7-11 c: gfcbccccjvcskcmrcxc\n4-5 n: ngtnr\n3-12 k: mtcfszkdhkdkd"
-
-    val regex = Pattern.compile("(\\d+)-(\\d+) (.): (.+)")
-
-    println(input.split('\n').to(LazyList)
-      .map(regex.matcher)
-      .count(mtch => {
-        mtch.matches()
-
-        val howMany = mtch.group(4).count(c => mtch.group(3).equals(c.toString))
-
-        mtch.group(1).toInt <= howMany && howMany <= mtch.group(2).toInt
-      }))
-  }
-
-  def _02_2() {
-    val input = "13-17 s: ssssssssssssgsssj\n7-9 p: pnlzhcppvl\n5-6 z: zzbwrv\n7-15 w: wwwwwcqwwwwwwwww\n8-9 h: hhhhhhhhsh\n12-13 l: mtrkqfllrglll\n3-4 t: nntt\n10-14 g: svgggggfhqggghlg\n3-10 x: djxxxpmcxx\n1-10 z: ztzzzgzzzz\n3-4 j: jjjjj\n3-4 c: fbhnsccbc\n2-4 q: thql\n15-18 w: tcwzpwzfwwqftvczbw\n5-7 l: lblwblbmllqnlbl\n3-4 b: bvbfnbb\n4-5 f: ffgzf\n7-9 c: cctcccccc\n5-6 k: szkgsk\n2-5 n: nqvnmgnqnsxfn\n2-10 s: dsfsbsssltss\n12-14 s: ssskrssssssfsxpsqsp\n13-15 j: jjjjjjjjjjjjjjjj\n16-17 l: jllclhllkgcpljzwk\n6-9 f: fsffwffbffbfff\n5-9 s: xtzscxsstcss\n6-9 h: hmnhshhhhqhrthmh\n3-6 x: xxxxxx\n9-11 m: mcmmmktmtmmmmm\n2-6 r: grrprrsmr\n15-16 v: vvvvvvvvvvvvvvwdv\n2-14 j: jmjdbhjjjjjjjxnjj\n5-9 r: rrwrrrrrr\n3-9 b: wbbcbvbtbbbbb\n2-9 r: rrtjlplkrgmrrrrxq\n9-12 s: sssshsmsssss\n2-14 t: qzxxvthtthtthv\n2-11 p: pvlsppwmgcspl\n12-15 s: dsjqsshdssbssstsssss\n3-9 k: tkkkpkkklk\n7-15 s: sssssssssssssgsssss\n6-9 x: xxxsxzxglpwxl\n2-3 x: xxmh\n7-9 g: ggsggggggg\n4-17 t: vtttrrrhtghfjvnqtttj\n1-2 n: jgnpn\n11-12 k: vlkkbkkkkklk\n9-12 v: vvvvvvvvwvvvv\n10-11 j: jjjjljjjjvm\n1-5 x: vvxlx\n3-4 l: flll\n10-11 q: qqqqqqqqqlq\n4-5 d: vdvhmpdtrdhdzfdsgk\n14-16 b: bbqbbbxbhrbpbtmhbbb\n18-19 d: dzpnbxngpgsjtnvtcdd\n2-16 v: jtrhtlfhvkbbflfq\n15-17 q: qqqsqqqqqqjqqqlqm\n5-6 q: tqrhtqqspdqvzqbx\n4-8 b: lbbzbwtv\n10-13 f: hrsnsfflfffsfjff\n10-14 k: kkjkkkkkkkkkkkkkkhk\n4-9 c: ccwcccccc\n5-7 r: grkrrrwjrrgbdtd\n8-12 p: tgspnljpfhtr\n5-6 v: vvvvvvm\n5-7 v: vvvvvvv\n6-8 q: qqqqqdqqq\n2-4 v: xfvqvlp\n3-4 g: gghb\n3-14 z: xflgzzrkzqzxzzxgd\n1-7 g: ldsgswgdgdmjlggb\n5-10 h: rhhmhhnbdhhqhwp\n16-18 m: mmmmmmwmmmmmmmmmmmm\n5-11 j: scssgcjjjjpzjjn\n13-14 s: wznksvkfvfskfs\n3-9 q: qdqqkqtbcnqqqsrklqbf\n7-9 c: cccchcmcvdcczwt\n15-16 g: gggdgggggggggggg\n1-5 k: kkkkkk\n7-10 q: qqqqqqqqqqqqq\n3-12 b: bbbrhbspbrbfbbqbbbb\n3-5 r: gczrgr\n2-6 h: whnlvhsftp\n8-20 h: fhhhsdfhhfzgmhpqhhjh\n15-16 z: sbzzzzzzdzjbwhgb\n4-8 h: jgtbmhwb\n8-9 s: hssksrnhlsdslssvcsbs\n2-6 v: vvvvvvv\n13-14 x: xxxxxxxxxxxxnsbx\n12-13 l: pgllllltpdfpxlll\n2-17 v: vqnncnxkssmhvkzstzgp\n4-5 r: rrrhw\n8-11 p: pppdppphppp\n6-7 t: lrmqtts\n8-12 h: hhlthhhhhhhhhtrzh\n9-10 g: jpgpmpcrwfkz\n12-13 w: wwwwwwwwwwhwwww\n5-13 l: jwtlqglllllxkllls\n2-4 c: ncpcgd\n3-7 l: llslllll\n2-5 n: jdfwnn\n3-5 z: xctxd\n4-6 l: lllrll\n6-7 q: fqqqqqqzqqqqq\n9-12 c: gccwgwqlcxccchchcrcb\n7-12 p: pppgpppppppps\n10-16 g: xrgfggzzghtgxgbggg\n7-11 d: tdddddqdwdgxldj\n11-14 p: pppppbpppppppppp\n1-2 j: jhjj\n8-9 v: pvvvvvvvcgvq\n5-6 n: qcrhnnkktxvdxcp\n7-12 x: lfnxxxdxxxxp\n10-12 s: sssssssssbsgs\n2-3 x: rqhzxmjnjxx\n13-14 w: wwwwvwwwwwwwqhww\n10-11 t: dgttttgttttt\n7-9 w: wdzwwwwjj\n4-10 v: vvsjvvbvvncvvvvxv\n6-8 w: wwwwwpwq\n7-9 j: jtjjjjrjmxg\n5-11 z: zzdgztshstzczw\n1-4 s: ssss\n3-10 p: pqppgrphrpkdmp\n9-12 x: xxxqxxxlxxsx\n3-4 v: nvswzv\n3-9 d: vdjvdcxdkdd\n2-10 v: vvjcvvvzvkb\n11-14 g: gggzggggggjgghg\n4-19 p: pppqppppppppppppppvq\n3-5 n: ngznkgdpn\n13-14 t: tttwtbtttttxxmttt\n3-7 q: qnqqqqqtqqz\n3-4 r: rkgr\n9-12 l: lqkfljzllqtl\n5-8 m: vpmmsnfmtxbcknlmpb\n15-19 s: ssptssnhsksnszsdpqss\n13-14 j: jsjjjjnjjjjjjjj\n6-9 p: ppppppppz\n5-8 r: tqzbnvgsvfrx\n7-9 h: hhhhhhhhp\n8-11 s: ssssssssssss\n3-10 t: ttttttttttttt\n4-5 g: ngtgg\n7-8 n: tnnnnnnqn\n12-16 x: bxxxxxzxxxxxxxbxx\n1-3 z: zzzt\n11-14 x: xxgxxxxxxxwcxfx\n6-11 v: rgkkhpkjvvf\n16-17 n: nfgnnnnnnnnnnnngvn\n3-7 w: wkwwwww\n5-16 g: ggggcggwgggggggggg\n1-3 l: sjvjlkzrpn\n17-19 p: pppppppppppppppptpd\n2-6 b: vjjwbqbdbpwsbw\n4-8 g: gggggjvgd\n2-6 w: pzzpwdm\n3-10 g: ggggcggggggg\n1-3 v: vxvjcvcpslpfn\n9-10 r: qrrwrrjrrr\n5-10 m: qbmdzvqmxnnmmmtk\n2-14 q: vqttpdjhqvkqzqxz\n12-13 r: rrxbrrrrrcrptjrr\n13-15 s: cscssssssssssssss\n7-10 z: zxbrzxbdbhzzzfzz\n2-16 n: skqkxqmqbvbrnspnrgq\n14-18 b: bbxsbbdpbnbbbnbbbpbb\n5-8 b: bmcrdhbbk\n11-13 f: fffffffzfffwffb\n5-6 w: wwbwlww\n12-14 d: gzxclqbnmnxjlzd\n13-19 z: fzzzctzzdsbgzzzzwzpb\n2-12 w: lcwwtwwpwffdjqwms\n9-10 x: hznxnxxxxxxxgkxx\n2-4 g: cwfjtlgf\n11-12 k: bqfkkkklkkkkb\n2-4 l: llllgxk\n5-9 m: cpwmqkzhxs\n8-15 b: bbwlwkblvbwdvrjbgsb\n3-4 f: ffmff\n4-10 c: nqxcqcfcqc\n10-12 d: ddqddddddtdhd\n3-4 n: dghnnqnn\n6-15 s: rsbxbrfmssdknppw\n2-13 m: mmkmmmmqdmkmmm\n8-12 n: nnnnnnnnnnnn\n2-5 d: mdkwd\n14-15 l: hllllllllllllgl\n6-11 g: jgccfnggggpgkt\n4-6 j: jjjtjjjjjj\n9-12 v: hgpjjlvhvhljb\n2-4 g: jfwdgfr\n4-5 f: fhffffchf\n16-17 j: jjjjjjjjjjjjjjjjj\n1-4 t: tcttwtttttrkt\n6-13 x: qmxfxxwxrxdxxxfvgv\n6-7 c: cckpccc\n3-10 m: qwmmpknsgmdmfvrcvw\n2-4 r: rszv\n1-7 q: qqqqqqqq\n1-6 l: lzlmjl\n11-12 q: mqqqqxbqqrqqqdqqbkq\n8-11 h: jtvbhxhhphhhh\n9-11 b: bbbbfbzbzbb\n5-7 l: gljssxddzscvhwrlrb\n3-6 c: fxcbccddxxcn\n3-5 s: ssssz\n17-18 n: jctqdsjchwknkffpnn\n8-10 c: crtwbwcfcqkf\n3-5 r: qrzqc\n7-12 z: zzzzzzzzsfzsz\n13-15 p: pppppppppzrllplp\n11-12 h: nhhhhjhqhhplhhh\n4-5 l: lllll\n4-10 g: gggzgggggg\n15-16 h: qhhhshrhhfhhhhch\n5-7 k: jsrvnjlkq\n16-17 p: pppplppppppppppml\n13-14 m: xqhmjdmrmmmffr\n6-11 c: cccccgccccrccccccc\n9-10 l: llllllllxllt\n8-9 b: bbbpbbbwbdnkbb\n5-9 m: sfmmpmmmzn\n5-15 k: kkkkkkkkkkkkkkkckkkk\n3-4 h: hhhh\n15-19 g: gggwgggrslgxjgfgbfs\n12-13 d: dddddddddddxb\n6-14 k: kkrgdvhscffwwm\n14-16 m: zmmmmqmmmmmmmgmm\n8-9 w: wwwwgwwww\n1-12 x: xxxxxxxxxxxx\n1-5 b: bwrbbfwgh\n1-3 t: ftttttttttttttt\n10-11 m: mwmmmwkmxsf\n8-9 t: tqtttttttdtttt\n7-8 z: hzbkpzfg\n2-3 h: htblhdhh\n3-4 p: pptnlpc\n4-7 l: lqlnqhljt\n12-14 n: nnnnnnnnnnjnnt\n1-4 d: dszd\n12-13 f: fvrffffffffkfff\n9-15 t: bkzstqvtgtwxtzqttd\n6-7 v: vrvvvkgv\n6-9 n: knbpnnjnnsn\n5-8 n: nznkzzqrbn\n9-12 s: ssscssssgsqwsspssp\n4-16 q: slqqqqqhqqqqqqdqc\n1-3 d: gdttdd\n2-6 h: hhqzmh\n4-9 k: lkckkpbzkl\n8-17 b: bbbbkbbpbbbbbbbbn\n1-2 r: bqmrw\n8-9 s: ztssqssbz\n1-13 s: stbqmhspsdgjs\n8-10 z: trnsmlxzmzzzz\n6-8 v: vlxvgzbzvvv\n3-6 c: xddhwtcbpcf\n13-19 g: rjvsgpnfjvzlxxzzgxgm\n5-9 m: qffpmxmfmk\n14-15 c: dcfwcpxcmvpxxtcbmzhc\n6-10 k: jkpbkvghpxksfkkk\n2-3 j: jmgjjjj\n7-12 z: szrznmkjxnzj\n4-9 s: ssssssssss\n2-4 h: hkqvfgpsxlnhtrqr\n6-11 g: gggglrlgggwv\n1-6 v: vrbvlvrvv\n4-10 j: jjjjjjjjjz\n3-12 k: kkjkkkkkfckgkkkkk\n3-15 x: xxrxxxxxxxwxpxg\n5-8 v: vvhvvdzvqfrvgqvt\n5-17 z: zznzzzzzzzzzzzzvf\n4-12 n: fdrncrggjlbgbmwv\n6-17 d: tcddhdbdvtsnrdstd\n1-10 m: mqljmgchlmsx\n4-6 l: fbllbl\n8-9 p: kpmtpphwp\n3-6 j: sjjjjj\n5-6 n: nnnnbs\n7-12 p: ppqppngcphpk\n5-15 x: xxxxgxxxxxxxxxqx\n1-3 p: ppppm\n3-6 w: wwwkwww\n1-4 r: rrrx\n3-4 z: zzznp\n12-14 n: tnzwtmnnnnndnfnnn\n9-10 d: ddfdddtdddd\n4-14 l: hvllsgncmhxhwl\n10-15 j: kxftfsmjvjgslsjpp\n2-10 g: gtgvggqggggxggdg\n1-9 v: cvnxhhffnbcv\n1-14 c: ccccccccccccccc\n3-8 q: fnqmkbhqtbjqvczq\n2-4 v: fvphvgv\n10-11 k: kkkkkkkkkpt\n1-10 d: dddldkddmdd\n4-6 x: xxxxxxxx\n4-11 k: kxkfkkkskkrkmk\n1-6 n: nlmngnkqww\n1-9 j: jjjjjjjjsj\n5-6 b: bbbbbb\n9-11 q: qqnqqmqqqpq\n4-7 r: drrrrsrrz\n10-15 n: nnhncnpsnxtnpnqnnghn\n3-6 d: ddddddd\n4-7 x: bxvxbxjwxgx\n1-2 n: cmpwsbjzghgndj\n2-4 r: przczfjfrbwj\n3-17 k: qbvmqxxffdfpkwxdgxv\n7-12 v: llmjvscvjvpvp\n2-7 z: vzxzxxz\n9-18 r: srdrjcrrrvhrcqrrrr\n7-8 f: rffffvtrfsqff\n5-8 w: pdcwwvmwwwkzwwjwxks\n3-4 c: cccc\n15-19 c: cccccccnccclbnplccqc\n2-13 f: fffffffffffffff\n3-9 t: thrkzdttb\n6-11 p: pppppgppppjp\n5-8 q: qqqrqfqs\n7-11 r: crprgtwwrzp\n2-10 f: qfhgfftfdbfnlffffff\n4-5 h: dhvvlcw\n3-6 s: sffsfws\n2-7 b: jrvxbhbp\n10-11 w: wwwgwwwwwkb\n5-9 d: dddddddddddddkdv\n4-14 b: bbbbbbbbbbbbbbb\n4-8 x: xxxxdxbqxc\n5-9 l: phlkhlllbc\n1-2 s: ssgskqvrsrwdt\n19-20 x: xcxqxvxxxtstnqxhxxxx\n6-7 x: gxxxxgb\n4-5 c: cccccc\n3-10 q: pzzjqjqpfbqws\n4-10 t: ttdttsccttpt\n14-18 c: ccrccccccccccccccrc\n2-5 x: xxzqx\n5-11 c: qbckpccckccccws\n5-6 g: gggggg\n3-5 n: wknbnnlnnq\n9-13 j: jjxjjjnjjjdqjjj\n5-9 s: smsxsrfss\n2-4 v: vxphvvk\n6-7 x: xxxqxxbbx\n1-3 g: zrgtghjwd\n4-5 n: nmntwnn\n6-8 g: ggznggggg\n4-5 k: tkkzs\n9-10 g: hqdhggggqggk\n12-16 w: wwztwmwwspwhmwwz\n3-5 m: mmmvmjmmm\n3-4 j: jjjj\n6-16 z: kzbtzqrkzzcfhvpwjszx\n3-9 f: ffwffffffff\n13-15 b: bbbmsbbbbbbwbbbbwb\n11-14 n: nnnnnnnnnnnnnnv\n8-9 v: bvvflvvbjvbtjqrvsxf\n1-5 g: gjgnggg\n15-16 n: nnnnnnnnnnnnnnwvqnn\n12-13 l: lllhllrlllllllll\n5-6 l: llmlll\n8-9 z: zzqzdzvzzzz\n3-4 d: ntdd\n4-5 b: sbndbkjxldqpzfbzbqx\n2-5 t: bwcfzdlfslsd\n4-5 d: dddddd\n1-3 t: jtpt\n10-12 q: qqqqgqqqqqqmqq\n10-19 m: mmmmmmmmmsmmmmmmmgs\n2-14 x: xxxxxgxvxmxxxbx\n3-13 l: tzljvxspwxdnl\n6-10 p: lhnppkpswrppgppzpppn\n2-6 j: djgwjjhvpwkcgsfbf\n10-11 p: ppppppppprpppppp\n7-12 l: lvgcbnlclvjlk\n6-11 z: zzzzzzzzzzn\n3-5 x: xzvhxnpwbxgxx\n4-6 h: hhlrhnl\n8-10 b: brngbslhqhb\n6-11 s: ssqjxsssdsssss\n2-5 s: dbjdjknspbdldfbjd\n10-16 z: ctzzzzznzcwzztzl\n14-15 j: jjjjjjjjjjjjjjjj\n9-13 d: qvdpwddddddddd\n10-12 c: ccccchcccccq\n3-5 x: xxxxx\n13-14 t: ttstttztttttltt\n3-16 k: qvfjklgzlwfwnbjklqdx\n12-13 g: ggggggtgggggggvg\n11-15 j: vqzjvzvxjhjnjtj\n7-8 t: tttttttg\n6-7 x: xxxxfmn\n10-13 m: mmlmmgvfmmmdmmcmmq\n7-14 w: wwkwwwwwwwmptwww\n17-18 b: bbbbbbbbbbbbbbbbbb\n2-8 c: kbvcnrpx\n16-17 n: nnnnnnnnpnnnnnngl\n10-15 k: kbkjkdkwpkkhkkksklx\n2-12 p: pjhdkvwdltvqrhm\n2-4 x: vxhxxk\n3-5 m: mmfmf\n6-8 n: ngnggfnwnn\n9-16 b: bnbbbtcbbbvbbrbbbbb\n2-8 p: ppphpzppcgptpprhpp\n9-12 x: xxxxxxxxxxxx\n5-7 d: dknbddd\n7-9 r: rrrrrrrrj\n3-6 f: fjmfxk\n3-4 n: jnnnbjckggnpjnddhx\n6-12 d: dznpsfdtdsdndjddvcdd\n7-12 x: xxxxxxhqxxxbxbsx\n1-2 k: kkfkkkk\n4-7 l: lprlqllll\n10-13 x: qxxqkjxxxxxxmz\n3-8 f: rxfffwcff\n4-5 s: cpscszdssc\n5-6 t: tttftt\n2-6 p: kpvtkprdqhxzpxrbfp\n6-12 s: ksnsklwqfssdsbsss\n6-7 l: lzlflfl\n7-9 t: ttktttttt\n8-14 d: hdddzndhsddpfpdxt\n8-10 d: ddnddrddvxbzskddzd\n11-13 v: mvhvrtvvvvzvrvdv\n4-5 g: gggjw\n3-6 m: xhmrpm\n7-13 k: kkkkkkkkkkkkk\n10-13 n: nnnnnnmnngnnnnnn\n7-15 m: kcmtmjtmwjmcsbbms\n13-16 j: jjjjjjjjjjjjpjjlj\n2-4 n: nnmnkn\n12-14 j: jjjjjjjjjjjdjj\n1-2 c: cchc\n9-12 m: mmmqbmlwmmjmzmmml\n3-8 r: rrkkrrrt\n4-5 k: wckknxkpkktmks\n9-10 d: ddddpdddnpx\n1-8 q: qmpmxzfqnqctwpprm\n10-12 q: sgkfxhqjkqqq\n5-11 p: rwpppdlmhpptkbrlp\n1-2 d: lzrbxnlcsqhssgdpvjs\n5-6 c: scccck\n4-5 s: ssshxsssws\n3-11 l: nslwgmxmtblzttxxg\n10-11 z: zzzzzzzzzzz\n7-8 j: bjjjjjqdbjj\n7-8 d: glpfjbpdxd\n2-4 k: wkvwrk\n1-5 c: vtcgcbc\n16-17 x: xxxxxxxxxxxxxxxxm\n2-11 f: ffffffsfffffffffffff\n4-8 m: mrmmmmmmmmmmmm\n2-3 x: xvzxx\n2-4 h: hxpn\n13-17 x: xxxxxxxxxxxrxpxxxx\n4-13 q: wcqqtwwmwqqtqmqtqrqq\n10-12 d: dddtdqdnjxdzdfdd\n13-14 f: jffffsffmffqjff\n6-7 d: sxdvddddwqxd\n5-11 d: ddddfdddddcdddd\n11-14 b: bbbbbbbbbhlbbj\n3-4 p: pfpp\n4-7 j: jjjgwkgwnfffvfzfl\n6-9 l: gknrmjgxtlzcbvkgxb\n4-10 n: nnnnnnnnntnnnnnnxn\n1-4 r: rjrr\n3-4 m: nkmr\n9-10 f: fffffffffff\n2-6 r: rnhtml\n3-12 q: qqkqqqqqqqqkqqqqq\n3-10 r: rrnrrrrrrrrrrrrr\n6-8 m: bmmmdmmmmdmh\n10-12 g: ppdvvxgkhfskfqc\n2-10 v: txpvkgvcbqvqfk\n1-5 p: tqcvbtp\n6-7 z: zzzzzzz\n7-16 z: fdjbbkrzjzdbdblhxb\n2-4 x: xlkz\n4-11 m: glhknzmvqzfmnsbn\n6-8 q: qhmdgqvq\n8-16 t: tttdtttdtttttttbvt\n4-7 v: vptcmgvvjvvvfvvmht\n7-8 t: ttttttngt\n5-15 t: ztgbttrntttdvgtptjdb\n10-12 b: bvndbbbbbbbbbbbbtb\n2-3 t: sbqts\n12-16 w: wwwwwwwwwwwwwnwww\n2-3 m: kmmr\n12-15 k: kkmkrkkkkzkdkrgkk\n5-13 h: hhhhzhhhhhhhzhhhh\n16-17 l: lllllllllllllcllq\n1-3 b: bbjb\n2-5 t: vtjhmf\n4-5 q: vzdqfd\n2-3 n: nnnn\n7-14 j: jpjjjjjjljjjjjjjj\n12-15 j: crsjpjjjqjjfjkv\n10-11 t: ttttttttttttt\n1-2 j: jjbjsjvhbfssznf\n9-13 f: ffmfffffzfffgfff\n15-18 g: sgblrqdftpwzggvgqt\n19-20 r: rrrrrrrrrrrrrrrrrrvj\n1-4 j: mjlt\n3-10 x: jxxxxxxxxgxxx\n6-7 p: ppptppp\n7-8 b: jbmnbmsc\n2-6 f: zshfcbhsh\n2-5 r: rrrqwtrvclf\n3-8 g: chgtwlbg\n11-15 n: knnhnnhnnnpnlknnkhnn\n3-8 z: zkqlzzpmjph\n5-10 b: bbbbhbbbbbbbbbbbbb\n6-15 c: ccccdkrncwrcclc\n6-7 c: ckcccwdhcc\n12-13 t: ntttktttttftttt\n5-8 j: jpjzcjjjjj\n7-8 k: kpkkksbkkkkkr\n9-12 s: sqssssbssxcs\n3-7 p: pbbplhtfpktpnppx\n5-7 n: ncnnnsn\n4-6 s: sssrss\n3-16 v: kvvvvvvczvvvvvvvkww\n3-6 k: kkkpmk\n8-9 h: hhhhhhhmb\n7-9 c: czwcccmtckccc\n5-7 t: tstglltkt\n2-6 c: cccccv\n11-14 z: zzzkzqzzzzzvzzz\n3-4 d: ddch\n4-6 j: jjjjjvj\n5-9 c: cccccccccccccc\n7-9 j: hjkwpjjbjfsq\n4-6 t: tttlcdb\n13-15 z: zzzzzzzzzsqzzgzzzkz\n8-10 t: ptckdjtsptlmzrktwcw\n2-5 n: tnbhnnkvnq\n11-12 m: fqmmmmmqmtmkcmmvnwmm\n7-8 z: zzzzztvl\n6-7 j: jbpjjjjpxmqxcbwsjrjj\n2-8 t: ttkdzfwdtflfswlkntt\n6-11 j: rnjjjjjjjjj\n10-13 f: fwhkbfdfffvfs\n5-8 l: pflplntf\n3-5 r: rrrrr\n16-20 g: gggggggggggggggggggg\n1-12 t: xtnsxbjtttxtt\n7-9 r: zrrrrrkrnrrrrrrrrr\n2-11 n: nrjnnnwnnprnnn\n10-12 z: zzzkxzzhznzdz\n7-13 d: pdcjdddqddlddcdkdxk\n10-15 r: rrrbrrrmvrcphrrqr\n1-3 l: rsll\n7-8 x: xxxqvrxx\n5-14 p: ppcdvppbppppjpbg\n5-11 x: dxdxdhkxxvxqxxxfx\n6-12 r: pvbfrhqrhftqrrxcrr\n4-7 t: zdbtzst\n1-13 k: kkkpkghkqkskk\n5-17 g: gvggqgggggqmgxggfgg\n12-13 l: llllwllllllwf\n2-11 b: bbbbbbbbbbbbb\n7-17 h: hhhhhhhhhhhhhhhhhh\n2-3 w: lwww\n2-3 s: vstcs\n4-6 s: vssssxss\n14-16 r: fmrrrrrrrrrrwwrvrr\n6-7 g: dgwgggp\n12-16 g: wggxgggrggggzgbgggg\n7-9 s: tjhxrscnsscssssnms\n3-12 n: nnnnnnnnjnnlnnnnn\n7-8 b: zncbmbglqbbbbpgm\n12-18 l: lllllllllllvlllllwl\n10-16 q: qqqgqfqqqqqrrqgqq\n1-7 n: nnnsnrnnnp\n10-11 q: qqqqqbqqqqq\n5-8 k: bnsfzkkwfkknkccwqkm\n4-5 c: ccdcccc\n5-9 z: nlsgzzzdz\n2-3 k: xkgxhqkpftx\n10-13 d: ddddddddhdddddddd\n4-6 x: xmjxwxx\n14-16 z: zzzzzzzzzzzzzrzq\n3-14 n: qnznjvlzdnnrgwfr\n7-13 s: sssxsspsssssrz\n4-11 k: pkkkcklbkkkk\n9-11 x: xxxxxxxxxxtx\n4-6 z: zzjzzfz\n3-6 d: zddddddz\n10-20 c: smccccvcpbccbqcxbccq\n13-15 m: mmmmmvmmmjmnnmlpsm\n12-13 j: jjjjjjjjjjjjj\n3-5 m: pmmbm\n7-17 m: mmmmmwcmmdfqrmmhlmm\n3-17 g: gcggnggbgdggggggg\n10-16 b: sqpsqkfbnbqzswbb\n3-4 w: wwwn\n5-6 d: dddddpd\n12-15 m: mmmmmmdmmmqlmnvms\n13-15 j: jjjjjtjjjjjjzjcj\n17-18 t: ttttttttjtttttttttt\n17-18 d: dcddmchdwbqrllsxjdv\n16-17 h: hshghhndnhhfhhhhh\n3-5 g: ggggch\n6-13 c: ccvccpccccccvcccccc\n2-4 h: hbfh\n3-9 l: lllptphml\n2-3 s: xwrqssfzr\n9-13 m: mktmmwmmmmrmmrqmmmcm\n8-11 l: rkvxvrglltlc\n1-8 q: jfqqxngqnqchq\n2-6 x: zxgjxx\n5-14 g: mlbgvdglrgcqjgz\n4-5 t: tqgtv\n2-6 s: ssfnzs\n9-11 m: mxwhmmmmmmzm\n1-5 d: dbhjzd\n2-16 f: fffffffffffkffffff\n6-10 b: dbbbbbbbbbbb\n5-6 x: xxxmxxcx\n7-8 l: lllpllcb\n16-17 d: dzdddddddddddddddd\n3-10 t: tklkqftnwcj\n9-15 g: jccxpmfjgntpptgkkcvt\n1-8 r: sdrwgcvprt\n4-5 v: nhbwx\n17-18 z: zzzzzzzzzzzzzzzzxsz\n5-13 l: cbdklqhnklkmwhpp\n4-5 f: fffff\n1-4 q: qpwq\n9-11 x: vxxgrgxxfvvgx\n3-4 r: rrfd\n3-16 x: sxqxxxxxxxxxxxxwxxx\n5-8 s: ssssssss\n5-13 b: jbbhqqbblbbbbfcb\n17-18 p: ppppppppprpppwppphp\n3-9 p: kdpnpfphppvffpwf\n7-8 r: jvrrrrrnrpqrn\n5-6 m: ncmmdxmm\n10-11 v: vvvvvvvvvdl\n3-8 g: jlsscggg\n1-3 m: gwms\n4-5 f: fpkvnf\n6-15 j: zfdrxjgxtbkbbjjctmsk\n5-6 v: qnnbvvwqvbwqblqd\n2-5 v: kvhvt\n2-6 x: lxrgsxg\n6-8 s: sssxswsjsssxss\n1-6 t: dxwjttnqkt\n13-14 s: sssssssssssvjg\n1-2 t: rtttnf\n9-17 h: hjhhhvhhbhhwhhhhphd\n2-3 m: mmmwcm\n10-12 v: vvvvvvvvvvvv\n6-10 j: flcjtzwhwsnjjjrjrj\n3-8 d: ddmrddddd\n5-8 c: ccccslcrwrc\n13-14 s: hpsqtxvkrsssshh\n11-15 t: hrxtjgtwtlmpqfx\n11-14 l: llbxlsllllqrqlllzlb\n3-4 c: crlcvcxwd\n3-9 s: vjbsxhwwdvshfxstc\n7-8 f: jfhnffbgsfjfwf\n15-16 x: ksxqvfsvnxsgvwgpjzl\n16-19 g: gggggggggggggggdggg\n9-11 h: hhhhhchhzmqhh\n2-6 f: kfdlcfrxftzgq\n2-8 p: ppnppkpwwmgp\n1-2 g: ggfgrg\n1-7 n: nnnnnnqwnnn\n5-6 h: hzjgthqphwhnjh\n10-15 l: xlllllllvlllllll\n3-4 l: lzjb\n4-6 l: ldzlql\n7-12 b: mzlbqzbrqjjbddr\n10-13 t: tttttxtttpttt\n4-6 d: dddddd\n4-5 k: kzvkkrkf\n1-2 h: fqml\n17-18 f: jhtfdfgmnchprbwbfrf\n3-10 s: sjsnsksbdsss\n1-17 j: jgcbjmjjjjjcxsjjjjhj\n15-16 k: tkjkjbkkrkkkklmgkj\n2-13 k: kkkkfkqkkkkkkkkpkk\n4-13 d: ddkkdtddzdnnddjdd\n1-3 v: fxvnxdvxnrbjs\n8-11 s: sssssssssspssssss\n1-8 k: khkkkkkkhk\n8-9 k: kkktmkkzk\n4-9 w: wrqdfwxwnfwwwnmzww\n3-5 l: lgtcvl\n13-18 n: nnnnnnnnnnqnnnvnnnn\n15-17 z: zzzzzzzzzhzzzzqzv\n3-5 q: lxnqsvk\n8-13 g: rzzwdlzgbcmggct\n10-14 d: ddvdbddddnddpd\n6-10 d: dphsszqrvz\n8-16 b: bnbgfbkbbbqlbcbbgqbb\n13-15 d: dddjddmqndddddddvjdd\n8-11 q: tnhgqkqfpjhwqgktq\n1-3 n: nnnbn\n11-12 f: fwfffbfcvflff\n7-8 s: sssglsvssvls\n2-5 z: zzzzzzzzzqjzz\n2-11 m: mhzrwphzxgj\n4-6 k: kkkwkd\n8-15 m: mrmmmmmzmmmmzmgmm\n1-7 w: wwwwwwxwwww\n1-3 l: nllllllll\n4-5 m: zbshj\n3-9 x: xssxqxxxsxkzx\n5-9 d: dnpfhsdmrxbvgxqrs\n7-12 r: vhrgcnxrvksg\n8-10 c: ccccccctzv\n16-17 l: lllllllllllllllhl\n6-15 n: nnnnnnngnnfhwwzsnnn\n6-19 z: zzzzzlzzzzzzzzzzzjv\n7-12 b: bxfbbwsjtbbvbqvbmbpb\n5-14 g: gggcjggggggggmgg\n5-8 d: wdddddkdpd\n4-17 x: xtxxxxxxxxxxxxxxmkx\n4-10 c: cccccccccjcscv\n9-11 v: ttjzfjldvvmswpqt\n5-8 r: tjrwhzgfrkgfq\n5-8 m: cvhsmnzmncsfmbqmm\n7-8 z: krctfpzz\n2-6 w: mlnlswvh\n12-15 t: zrntxzttqlthfdttt\n1-8 r: pkcsjkrrrzxxsfnjw\n5-15 p: xbhpkpnngpqvpcwdppg\n3-4 t: tttt\n8-11 j: wjjjmjjjjjqp\n8-14 h: gjslnklhhkhthxh\n2-6 q: qqlqqq\n3-6 p: pppgppppppppppp\n7-11 l: llhllllrllllxldllll\n11-12 d: lddkdddddddddd\n4-12 z: zwlzzqsvvclw\n15-18 c: tflkxflqccwkmlckck\n8-12 h: hhhhrhhhhfwhhh\n6-9 r: rrrmrjrrvkrrjrdr\n8-19 k: tmdlkcktbkkskvkbkbz\n1-13 s: ssssssssssssss\n4-5 v: xtcfvx\n1-15 s: shtssssswghsfsss\n2-4 k: krkl\n7-12 q: dxqqqhqqqtqj\n1-5 r: rrrvvrrrr\n5-9 t: ttrttftxztnhtvkcmtth\n7-9 g: gngkcggrg\n12-13 z: zzzzzzzzzzzbk\n3-4 l: lslllc\n4-5 t: pqchpk\n5-6 h: hhhhml\n15-16 h: qhhhhhhhhhhhhhhhhhmh\n5-6 d: ddddfs\n1-6 d: dthfdzddfnsdddm\n14-15 d: ddddddddndhtddsdnddh\n5-7 j: qjjjpjplljjc\n2-6 q: qqqqqqq\n3-4 b: bbbb\n5-6 s: ssssjjs\n1-6 b: bbbbfb\n7-13 m: cpvmbbnnrmzwf\n2-4 f: fnhn\n6-7 d: gdktlddrdlvmqdtddpzv\n4-5 w: cwlww\n6-9 z: zmdrgzlmzx\n5-10 m: mjbjmmmlmmwxm\n2-4 s: ssnhht\n8-16 t: tttttttdtttdtttjtt\n3-5 z: wxgzhtswb\n13-14 l: lllllblpllllchll\n4-8 g: vggqzvggnggggggggg\n8-9 h: hhthkshhpnhqhgh\n4-7 l: hglxklpl\n2-3 w: mkmgkwzwmw\n6-7 h: hhhhhrz\n4-6 f: tdffff\n1-7 l: lmfhgqschjqglrvwwnnz\n12-13 g: gggggggggggfhggg\n2-10 m: ntmmmdjmmmsvmm\n1-2 s: sqsssssssssss\n3-8 f: frnthfcfxfft\n7-16 w: wwwwwwmwwnwwwwmk\n11-13 j: jjkjrjjjjjjjjjd\n3-4 s: ssvc\n3-13 t: ddttttnnqpqztzbbdv\n3-5 g: gggggg\n3-4 r: rrrzlklfljqvz\n5-12 m: kdgmmrszqpfsbmz\n11-14 g: gqgvzgxhgbtngxggh\n7-16 k: qqptmhhfkhsgkhbmkx\n5-8 p: ppppptppp\n2-17 x: vxtxpvnxvlctrcpfxxx\n2-3 w: bwmwjwwwwwtnqwxc\n4-5 g: gggggg\n4-7 z: znzwzzzs\n4-5 v: kvvvv\n10-14 r: rrrrrrrrrbrrrx\n4-5 t: xtttt\n4-11 w: wwwwwwdcmkrwx\n6-7 c: ccccccjc\n12-14 w: wkwwwwwwwhwlwwwk\n5-9 s: hmssssssssss\n2-4 r: rrrjsh\n12-13 m: mmmmmmmmmmmmmx\n3-5 s: sshsvssss\n4-5 l: xllwnshchfdfk\n16-18 w: wwpwwpwwwmwwzwwtwwww\n2-9 c: zsfgbrrprlszrr\n10-11 n: thnmfndnnbnnbhnnnn\n13-14 v: vvvcvvvvvvvvvvv\n3-4 n: rwcm\n14-15 t: tttttztttttttpjt\n5-8 j: jjjjwhjdj\n5-8 f: fffffffm\n6-7 z: zzqzzzz\n12-19 j: kjrfnsjnhjdzgqpcjzvh\n3-7 p: pxpdvcqdptxpvk\n1-4 c: wccqj\n1-2 w: pwsqhwjdhcm\n2-3 b: bfbn\n10-19 q: qqqqqqqqqtqqqsqrqqq\n5-8 p: pppppppppppppppp\n10-13 g: rjghslgbggkgjdgvp\n3-9 z: fnnfzzdnf\n9-16 g: tlgtgggggggggggqm\n3-12 r: rrrvccrrvbdbncrrr\n1-4 v: jvhbv\n15-16 v: vvvdvvcvvvpvvvvvvvv\n3-4 g: ggggg\n7-8 w: wwwwwwmpw\n4-9 s: gkxshzkpssc\n2-8 q: pqttnpqqqfwqs\n3-4 n: nlnp\n3-4 h: phhh\n16-18 g: hqsxhmmpfgggmgpqhrjg\n15-17 w: wwgwwwwwwwwwwwmwl\n7-9 x: xxxxxxmxv\n1-7 s: ssssssbs\n1-3 x: dxlxxx\n3-6 p: wpxppq\n5-7 c: clcclcfcbccc\n3-5 s: sbmst\n3-5 d: xqmdn\n5-10 z: zbzwfddkzzc\n18-19 d: dpdddddddddddddddwcv\n3-4 r: rrrr\n8-9 f: ffnffffwd\n12-16 h: hxnlxghzkpcpdhqqvl\n6-7 m: jngclmm\n5-13 m: rmmmpmrmmmmmjmmmm\n1-3 f: dfwfs\n7-8 l: lclllwllll\n16-18 j: jjjjjjjjjjpjnjjgjbj\n12-15 f: fffffffffffnfftfffff\n9-13 w: qbnpmklbwxdbbwkklpwb\n1-3 j: mjkj\n1-3 h: hphwdh\n3-5 x: vrxxb\n1-15 v: mvvvvvvvvvvvvvj\n8-12 v: lvzfvzvvnjtvvvvvv\n4-6 g: gggvggggg\n13-20 l: zdllslnlxslqpnvgwlwm\n2-4 l: lllhll\n3-14 k: kckzddtzbmvkpkhb\n4-5 k: hbkgbxzj\n2-4 p: lppp\n2-11 r: rrrrrbrrbrrrrwrrrwr\n4-7 k: knkqkkk\n7-8 r: njrrrrrrrr\n4-7 h: phhhnhdhc\n4-5 r: nsnrrfktwbbhrrrh\n4-5 q: qqzvzsvkq\n6-7 b: jbbbbbbbkbgb\n5-11 f: ffffcfffffgff\n12-16 v: vfvvqvvdvvvlhvzjvmzv\n9-10 v: vvzvvtvvjgvv\n1-7 v: vvvvvvvv\n3-12 z: zmncxhrdzdmtcbxtlrzq\n3-7 r: rbfbrjrr\n12-13 n: nnnnnnnnpnnldnn\n5-12 j: jjjjjjjjjjjlj\n3-6 s: sssfsgss\n13-17 h: hhhfhhhhhhhhvhhhhhhh\n5-10 g: gggxkggggvg\n1-4 b: hbbtb\n14-15 t: tttttttttttttnb\n7-14 d: ddddddddddddddhdw\n15-16 d: rbdddbdwmjdhmpdd\n7-13 s: kxpdntprmskcs\n3-5 m: gmdmpj\n4-5 j: jjjjjvjsj\n1-14 q: qmqqqntqcqxmqsqkqq\n4-5 q: tqxqq\n8-11 m: mmmxmmmvmmv\n5-8 q: qqqqrqqw\n5-15 d: dqdddddddddddvspd\n1-4 t: tttttt\n12-15 z: dhzzxwfjgnzhzxt\n8-11 v: kdnmfnmqvvdvqdlvk\n16-18 s: sssssssssssssssjstss\n2-5 c: ccwdccc\n3-11 x: nxjxxxxxxxxxxx\n13-15 v: vvvvfvvqvlvvvvvv\n4-11 s: sssszssnqjsbsvs\n9-18 j: gxhjjjjjnjjsjjjrjjjj\n3-4 t: ttfttt\n9-10 s: qzgxjhpsss\n9-10 q: qqqqnrtxrqqsqqq\n6-13 k: kkkrgspkkkpwjshmk\n9-12 h: xhhhmhbhhhdhh\n3-12 g: gfgggggjgggggggggg\n13-15 z: zzfzzszzzzzzvzzzhzw\n5-15 m: mzrmmmmndchfzmmrr\n5-10 z: zzzzzzmzzh\n4-5 s: wscfc\n3-6 m: kmmtmm\n5-7 h: hhhhchh\n4-14 w: wwgwwwqwsfrjvmbwj\n2-9 v: vwvvwvmjcqnxv\n9-10 x: qxxhfxchrx\n17-18 x: xxxxxxxxxxxxxxxxwp\n4-6 v: vvvvvtv\n11-14 l: lcqlclllllrlddxlzll\n1-4 r: trrn\n10-13 h: thnsxphnfgvhvq\n11-12 s: sssssssssshsss\n10-11 z: rqpzszzmzdz\n3-6 q: qrqqqq\n5-6 h: hhkhhhh\n8-16 h: wqhjvhlgwtsgvlpf\n3-6 f: gqfpcfsfhfjgrbqv\n3-5 l: lqllsrjlv\n6-10 v: vvvvvvkvjv\n3-8 t: gtttcbnkxxstttd\n2-6 j: jjjjbjz\n10-15 j: ftwjtfzjmjsvwjj\n8-11 d: ddsksddddzfd\n2-3 b: bhngbb\n14-16 c: mllmcbfxwxrqlcjcw\n7-8 q: qqqqqqdv\n5-6 s: slkzlfvg\n3-5 q: qqqqq\n3-12 p: pddpdpszrppxcpjgv\n11-18 r: rrrrrrrrrrprqrrrrr\n3-4 c: cscckcxcn\n11-13 g: gvgwmlngggqggrtgtkg\n5-7 b: smfpbbb\n7-18 l: sqmlklvkqfjrgtqhzr\n6-7 k: kkpqqfgz\n7-9 n: fnndnnxnndn\n11-12 c: ccccccccccxv\n6-9 d: dddvdddddddvd\n1-4 v: nvjfv\n3-8 g: gggggggg\n6-7 x: xxxxxxx\n4-5 c: scccc\n6-13 q: nchlfqqqlqnqqqtq\n1-8 r: srrrrrrnrrrrrrr\n4-5 p: sdpscpppp\n9-11 r: rrrdrkrrrqg\n3-5 m: xwmscmmm\n7-11 r: rrrfrrrrkfrvrf\n4-6 b: bbfbbb\n4-9 v: rblrvrvpvz\n2-8 w: wwwvwgwww\n8-9 w: wwwwwwwwzw\n16-18 t: ttttttttttttttttttt\n8-9 n: nfrzsdjxr\n1-2 w: wwfwwmfbww\n2-4 s: dfsksft\n1-13 d: dwdrccnddqmndcl\n1-4 f: qfflf\n3-7 d: dnddjpdfgc\n3-5 l: lkglx\n1-13 j: jjjjgbjtpjjhd\n15-16 j: jjjjjjvjjjjjjjjjdjj\n4-13 t: tttxztttttttkt\n4-5 m: mmmmm\n4-6 j: jgjrjwjfjx\n8-11 t: ttbjttvttsttdqtn\n14-15 w: wwwwwwwwwwwwwkjw\n1-6 x: nrscxn\n3-13 w: wwwwwvgnwbwwwwwwv\n12-15 p: pppppppppppjpppppppp\n5-11 v: vvvkvbvwkwnvvvsxvv\n1-11 m: mmmmmmmmmmtm\n10-11 c: cccccccccwc\n5-8 f: xvgstwfxfhxknds\n12-13 t: tttttztttttttt\n4-17 k: vkpkfkkkqnkqnkgkkknk\n9-12 k: kkwpzpdzxhxk\n14-17 v: fgvvvvvvvvvvvjpvv\n7-8 k: kndnqkkk\n7-11 c: gfcbccccjvcskcmrcxc\n4-5 n: ngtnr\n3-12 k: mtcfszkdhkdkd"
-
-    val regex = Pattern.compile("(\\d+)-(\\d+) (.): (.+)")
-
-    println(input.split('\n').to(LazyList)
-      .map(regex.matcher)
-      .count(mtch => {
-        mtch.matches()
-
-        val first = mtch.group(4)(mtch.group(1).toInt - 1).toString
-        val second = mtch.group(4)(mtch.group(2).toInt - 1).toString
-
-        LazyList(first, second).count(mtch.group(3).equals) == 1
-      }))
-  }
-
-  def _03_1() {
-    val input = "...#...#....#....##...###....#.\n#.#...#...#....#.........#..#..\n.#....##..#.#..##..##..........\n.....#.#.............#..#......\n.......#...#.##.#......#..#.#..\n#.#....#......##........##.....\n.....##.#....#..#...#...##...#.\n...#...#..#.......#..#...##...#\n..........#...........##.......\n..#..#..#...................#..\n#..#....#.....##.#.#..........#\n.#.##.......###.....#.#...#....\n.#..##....##....#.......#...###\n#.#..##...#.#..#...............\n.........#....#.......##.#.#...\n...###...##....##...#..##.#..#.\n....#.........#..#...#.......#.\n....................#..#.#.#...\n..#....#..........#...........#\n.#.....#..#.....##........##..#\n#..##..#...##............#..##.\n.#..##....#..........#..#.##.#.\n..#####..#.#............##.....\n...###.#....##..#.#....#.....#.\n.#.......##....#...#.#.....##..\n...#....#...##.#...#..#........\n.####.....#....#.#.#...#.......\n...#....#.....#.......#........\n#..#.#.......#...#............#\n...#.....###.##....#.#.###.#...\n.#.........#.......#.#....##...\n#.#..#...#.#...##......#..#....\n.....#...#..#.#...#..###..#....\n......#.........#...###........\n.....#..##...#..........#.....#\n..#..#.#.##.#...#....#....##..#\n##....#.##...#.##.#..##....#...\n.....#.#.#.#..#....##.#...#.#..\n.....##.......#........#.......\n...#.#.....#...#...##.#......##\n........#..#.#...#.#.....#.#..#\n#..##...#.#...##..##...#.#...##\n.##.#.#..#...#.....#.#.##.#...#\n.#.####.........##.........#..#\n.##..............#....#...#...#\n......#...#..#...#..#..###.#...\n.......##...#.#.#..##..#......#\n.....#....#..##..#.........#...\n.....#..#.#.#........#.#.####..\n#..#.......###....##...........\n#..##..........#.#......#.#....\n.....##........#...#..##.......\n###...#.##.#.#.#.#.##...##.....\n....#...#........##.#.##..##...\n.#..#.#.#......#.......##..#..#\n.#...#.................#....#..\n.##..#..........#..##.......#..\n.#.#.#.....#..#.#.........##..#\n...#......##...#.......#...##..\n##...###....#.###.............#\n#.....#.#..#.#..#........#.#.#.\n.....#.#......##..#.#.....#.##.\n.......#...........#..#.......#\n..#....#.#.#......#.....#...#..\n.....##........#..##..#..##....\n#.#........#...##....#.#..##...\n#......#......#....#..#...#.##.\n....#.#.......#.#.#............\n......####.#.##...#.#.##.....##\n..###.#.#..#.........#.####....\n.#.......#..#.#....#.#..#.#.##.\n#....#....#............##...##.\n....#....#............#....#..#\n..#........#..#....#..#..#...#.\n.#......##....#..........#....#\n#.##.....#..........#.###.#....\n....##...#.....#.#......#.##...\n#.#.....#.......###.###..#..#.#\n..###..##.............#.####.##\n#....#.....#....#..##.......#..\n.....#....#...#.#.#.#..#...#.##\n...#.....#..#....###......#.#.#\n##.........#.#..#..#.#..#.....#\n.#.....#.#....#.........##..#.#\n.#.#..#.###..#..#..........#...\n.##....#.#.#...#......##.....#.\n#.#....#....#...#...##...#..#..\n#...#........#....#....#......#\n#......#...#..#.#.##.....##..#.\n....#...#......##...#..#....#..\n.#......##.##.......#.......#..\n.#...#..####...........#.#.#...\n.........#...#.#.........#.....\n#.##.....#.#..#.#.###...###..#.\n#...##.###......#.###..##.#.##.\n...##.#.....#....#..#......#...\n#....###.#..#...##.....#......#\n........###...#...#............\n........#....#...#...#....#...#\n#....#..#..#....#.#........#.#.\n##...#.....#.#..........#..#..#\n#.#...##.....#........#...#...#\n##.#.#.......#...#..#.###....#.\n.#.......#....##..##...#.....#.\n#....#....#.....#.......#......\n.##.##.##...##...#.#.#..#..#...\n#..#..#.##....#......##....###.\n.......#.#.........#..##.#...##\n.#..##...#....#.....#..........\n..#.#...#......#.#..#..........\n.##....#.#.#.##.......###...#..\n..##.#...#.#.#.#.......#..#....\n#..#.......#...#........#.....#\n.....#.......#......###..#.....\n...##.#.......#.....##.....##..\n##..#.......#.#.....#....#.....\n..#....#.##.##...#...#......#..\n.#..#.###.#....###........#...#\n....##.##...##..#..#.#....#....\n..###...##.....##..............\n#....#...##...#....#..........#\n.##........#......##...##...#.#\n..#.#.##..........#......#.....\n...#...#.........#.##........##\n..#.#..#.#..#...#....#...#.....\n...##...#..#.###.#..#.#...#....\n....###........#..#..##...#....\n#.#....##.......#.#........#...\n.###...#..#.#.#.#..#...#..##.##\n..#.........#####.#......#..#..\n#.....#.....##..#....#...#...#.\n...#..#....##....##.....##.#...\n.........#............#.##.....\n....##.#..#....#.##.......#..##\n.###....#.#..#......#.#.......#\n.###...###.#.........#.#..#...#\n.....#........#..#.#..#.#..##.#\n.###..#....##.........#..##....\n..#.......#..#..##...#.###.#...\n#.......#...........#.#...#.###\n#.##.##...##.#...##..#.....#...\n..#..#........###.#.....##.....\n#.....##....#...##...####..#..#\n....#........#...#...#.........\n......#.#.#.#.......#..#.....##\n..#..#....#.....#.#...##......#\n..#....#...#.###.........#.###.\n...#......##..#.#.....#...#....\n...#.......#...#...#........##.\n............#...#..#....#.....#\n....##......................#..\n#.#.#....#....#..........##....\n#.#.....#.#.##..#...#.##....##.\n...#...#..#...#..#.#.#.......#.\n#.....#..........#.........##.#\n#...##..#..#.#.......###....#..\n.#...#..##....#.....##.......#.\n....#.##.....#.........#.#....#\n........#.#...####..#.......#.#\n.####...#.#......####.....#.##.\n###..#....#..#.......#.#..##..#\n#......#.#....##..#.##.#....#.#\n...###...#...#..##.#..#..#.#...\n...##..##....#..#.....#........\n.....#..............#......#..#\n......#....#......#..#.........\n#..#.....#.##...........##.....\n.#..#.#..................##....\n#.#..#..##...#....#.#......#...\n.##.#.##......#.##...#...#...#.\n..#...#.........#.#..#.#....#..\n.#.####.#..#.#......##.#..#....\n#..#.......#....#..............\n....#............#..#..........\n.....#####.....#.....#..##...##\n#.#....#.#...............#..##.\n.#.#..#...#......#.....#.#.#...\n.#....#.#.#......#.....##....#.\n....#....#.##..#.......###...##\n.....#..#.##...#...#...#..#.#..\n##..#........#.#..#..##......#.\n.#..#..##.......#..#.....#.....\n.#.#.....###..##.#.#...........\n..##..##.####..........#..#....\n..##..#..#...#....#......#.#...\n#...#.#......##.....##.#..###..\n#..#..............#........##.#\n.........#.##..#.#..#..##.##.#.\n#....##....#.#..#.#...##..#....\n.#....#.......#............##..\n.......#.#.......#...#.#......#\n......##...#.......#.#........#\n..###..#.#.##....##...#....##..\n..##.##..........##..###.......\n.#.#.#..#..#.#.......#.#...##..\n..#..##.........#.###..#......#\n....#.#.#...##.#...#...##..###.\n..###..##.........##...#...#..#\n.#..##...#.......#.......#..#.#\n........##....##....#.#.###.#.#\n#.....#.#.................#.#..\n....#.#.#.....##.####.#......#.\n....#.......#.#.##.##..........\n...#...........#...#.##...#.###\n#....#....#..........#.##......\n##..#...........##.....##.##...\n.#.##...##..##....#..#.....####\n#...#...#.##..........##..##...\n....##..#....#.....#.#...#....#\n..#....#..##...###.#.#.........\n#......#.#.#...#...#.........#.\n#............###.#.#.#..##...#.\n.##.....####...##..##..#..##.#.\n#..#........#.....#.#.....#...#\n#............#....#.#.#........\n......##...##.#....#.....#...#.\n..#........##......#.#.....##..\n.#..#..#.....##.......#..#.#..#\n.#....#..#....##.#.#.#..#..#.##\n.####.#..........#...#..##.....\n...###..###...##..#............\n#..#.....##.#...#..##..#.......\n.....##....#...###.##...#......\n...##..#...#..#..##....##....#.\n...###....#.###.#.#.##....#....\n##.#.#.....#....#.#....#..#....\n.......##.....#.#..##...##...#.\n.#....#.#...##.#..#....#.....#.\n..#...#..#...#.##........#...#.\n#....#......##.#....##...#.#..#\n.....#..#..#..#......#...#.#.#.\n..###....#........#...#.......#\n###...#.......#.#.......##.##..\n......##.....#.#........#....#.\n#.##..#.#.#.#..#....#.##.....#.\n..........#.##.#...#...#..#..#.\n..#...##.#..........#..##.###..\n..###..##.##..#.#...##.####..#.\n#.#.#...............##....###.#\n....#.........#.#....#.#....#.#\n..#...#.###...#....###.....#...\n..#..#....#...#............#...\n.#..#....#..##.....##..........\n..#....#.#...#.#.#.#.......##.#\n.........#....##........#.#....\n...#..##.#..#.##...#...#.#....#\n....####...#...####.#....###..#\n......##...#.##.#.......#..#...\n#.#...#.#...#.#...#....#.#.#...\n.#.....##...#.....###.#....#...\n......##.....###...#.#...#.#...\n#..#..##.#.#......#....#..#..#.\n....#.###.....#..#...#.##.....#\n##.##........#......#....#..##.\n##.....##.#.....#.....##.....#.\n.....#.##...#.#..#.#.#.....#...\n.#.##..#...#.#..#.....#.#......\n.....##.......#..#...##..#..#..\n#.....#..#.####......#........#\n.#..#..##.#..##............#..#\n.##..#.#....##.##.....#......#.\n.......##.........#..#.........\n.#...#.......................#.\n#......#.#....##.#.......#..#..\n..##..##......#.......#....#.#.\n##......#......##...##.........\n..#....####....#.#.....##.#.#..\n..........#..#.#.#.....#..#.#..\n##..##...........##.......#....\n##....#.#....#..#......###....#\n...#.#.#..#.......##.......#...\n#....#.......#.......#.........\n...##......##....#...#......#.#\n#......#####.#.........#.....#.\n#..#.............#..#....#...#.\n.......#.##..#..#..#..#....####\n......#.##..##..........###...#\n.#.##....###..#........#....##.\n#......#..#...###.#...#.....#..\n.#.#.......#....##.......#.#...\n..#.##..#..##.....#.........#.#\n#.#...#..#.##....#.......##....\n.#.....###....#.#..#...#.....#.\n#...#..#.......#.#.....##...#.#\n#.#####.........#....##.....#..\n#....#..##...#....#.##.......#.\n.#.#.........##....##....#.....\n...#..##.......#....#.#.#......\n#.###.##...###....#.....#.####.\n.#...#.#.#..##.#..........#....\n#.#.....#.##.#..####.....##.#..\n...###.##..####.......#......##\n.##..#.........#...#.#.....#.##\n..#.....##....###.....#.#...##.\n#....#....#..#....#.##.........\n......###....#.#..#..#....##...\n.#.#................#.......##.\n...#.......#.........#.#.......\n...#..........#...##.....###...\n....#......#...#...............\n.##...#....#.....#.##......#...\n.#.....###...##..##...#.#......\n....##........#.....#...#....#.\n#.........#.#...##...#.#..#....\n...#.#.....#.#........#.#....#.\n.#........#.....#.#.#.#.#..#...\n....#...#.....#.#....#........#\n..###.#....#.#....##...##..#.##\n.#....#.#.####.#.#.....#.......\n.#...#...#.................##.#\n..................##..#..#.#.#.\n.#..#............##....###.....\n.......#....#...........#......\n....#.#.#.....###.........#..##\n...#.#....#.#.##.#.##.....##..#\n.#.##.#...##...#.......#.....##\n.#............#...#..##...#.#.#\n#.##..#.##..#..##.###.#........\n..............##....#...#..#.#.\n.#.#...#.#....#....###........#\n.#....#.#....#......###........\n..#.......##......#.##.....#...\n.....#......#..#...#.#.....#..."
-      .split("\n")
-
-    def navigate(row: Int, col: Int, right: Int, down: Int, encountered: Int) {
-      if (row == input.length) {
-        println(encountered)
-        return
-      }
-      navigate(row + down, (col + right) % input(row).length, right, down, if (input(row)(col) == '#') encountered + 1 else encountered)
-    }
-
-    navigate(0, 0, 3, 1, 0)
-  }
-
-  def _03_2() {
-    val input = "...#...#....#....##...###....#.\n#.#...#...#....#.........#..#..\n.#....##..#.#..##..##..........\n.....#.#.............#..#......\n.......#...#.##.#......#..#.#..\n#.#....#......##........##.....\n.....##.#....#..#...#...##...#.\n...#...#..#.......#..#...##...#\n..........#...........##.......\n..#..#..#...................#..\n#..#....#.....##.#.#..........#\n.#.##.......###.....#.#...#....\n.#..##....##....#.......#...###\n#.#..##...#.#..#...............\n.........#....#.......##.#.#...\n...###...##....##...#..##.#..#.\n....#.........#..#...#.......#.\n....................#..#.#.#...\n..#....#..........#...........#\n.#.....#..#.....##........##..#\n#..##..#...##............#..##.\n.#..##....#..........#..#.##.#.\n..#####..#.#............##.....\n...###.#....##..#.#....#.....#.\n.#.......##....#...#.#.....##..\n...#....#...##.#...#..#........\n.####.....#....#.#.#...#.......\n...#....#.....#.......#........\n#..#.#.......#...#............#\n...#.....###.##....#.#.###.#...\n.#.........#.......#.#....##...\n#.#..#...#.#...##......#..#....\n.....#...#..#.#...#..###..#....\n......#.........#...###........\n.....#..##...#..........#.....#\n..#..#.#.##.#...#....#....##..#\n##....#.##...#.##.#..##....#...\n.....#.#.#.#..#....##.#...#.#..\n.....##.......#........#.......\n...#.#.....#...#...##.#......##\n........#..#.#...#.#.....#.#..#\n#..##...#.#...##..##...#.#...##\n.##.#.#..#...#.....#.#.##.#...#\n.#.####.........##.........#..#\n.##..............#....#...#...#\n......#...#..#...#..#..###.#...\n.......##...#.#.#..##..#......#\n.....#....#..##..#.........#...\n.....#..#.#.#........#.#.####..\n#..#.......###....##...........\n#..##..........#.#......#.#....\n.....##........#...#..##.......\n###...#.##.#.#.#.#.##...##.....\n....#...#........##.#.##..##...\n.#..#.#.#......#.......##..#..#\n.#...#.................#....#..\n.##..#..........#..##.......#..\n.#.#.#.....#..#.#.........##..#\n...#......##...#.......#...##..\n##...###....#.###.............#\n#.....#.#..#.#..#........#.#.#.\n.....#.#......##..#.#.....#.##.\n.......#...........#..#.......#\n..#....#.#.#......#.....#...#..\n.....##........#..##..#..##....\n#.#........#...##....#.#..##...\n#......#......#....#..#...#.##.\n....#.#.......#.#.#............\n......####.#.##...#.#.##.....##\n..###.#.#..#.........#.####....\n.#.......#..#.#....#.#..#.#.##.\n#....#....#............##...##.\n....#....#............#....#..#\n..#........#..#....#..#..#...#.\n.#......##....#..........#....#\n#.##.....#..........#.###.#....\n....##...#.....#.#......#.##...\n#.#.....#.......###.###..#..#.#\n..###..##.............#.####.##\n#....#.....#....#..##.......#..\n.....#....#...#.#.#.#..#...#.##\n...#.....#..#....###......#.#.#\n##.........#.#..#..#.#..#.....#\n.#.....#.#....#.........##..#.#\n.#.#..#.###..#..#..........#...\n.##....#.#.#...#......##.....#.\n#.#....#....#...#...##...#..#..\n#...#........#....#....#......#\n#......#...#..#.#.##.....##..#.\n....#...#......##...#..#....#..\n.#......##.##.......#.......#..\n.#...#..####...........#.#.#...\n.........#...#.#.........#.....\n#.##.....#.#..#.#.###...###..#.\n#...##.###......#.###..##.#.##.\n...##.#.....#....#..#......#...\n#....###.#..#...##.....#......#\n........###...#...#............\n........#....#...#...#....#...#\n#....#..#..#....#.#........#.#.\n##...#.....#.#..........#..#..#\n#.#...##.....#........#...#...#\n##.#.#.......#...#..#.###....#.\n.#.......#....##..##...#.....#.\n#....#....#.....#.......#......\n.##.##.##...##...#.#.#..#..#...\n#..#..#.##....#......##....###.\n.......#.#.........#..##.#...##\n.#..##...#....#.....#..........\n..#.#...#......#.#..#..........\n.##....#.#.#.##.......###...#..\n..##.#...#.#.#.#.......#..#....\n#..#.......#...#........#.....#\n.....#.......#......###..#.....\n...##.#.......#.....##.....##..\n##..#.......#.#.....#....#.....\n..#....#.##.##...#...#......#..\n.#..#.###.#....###........#...#\n....##.##...##..#..#.#....#....\n..###...##.....##..............\n#....#...##...#....#..........#\n.##........#......##...##...#.#\n..#.#.##..........#......#.....\n...#...#.........#.##........##\n..#.#..#.#..#...#....#...#.....\n...##...#..#.###.#..#.#...#....\n....###........#..#..##...#....\n#.#....##.......#.#........#...\n.###...#..#.#.#.#..#...#..##.##\n..#.........#####.#......#..#..\n#.....#.....##..#....#...#...#.\n...#..#....##....##.....##.#...\n.........#............#.##.....\n....##.#..#....#.##.......#..##\n.###....#.#..#......#.#.......#\n.###...###.#.........#.#..#...#\n.....#........#..#.#..#.#..##.#\n.###..#....##.........#..##....\n..#.......#..#..##...#.###.#...\n#.......#...........#.#...#.###\n#.##.##...##.#...##..#.....#...\n..#..#........###.#.....##.....\n#.....##....#...##...####..#..#\n....#........#...#...#.........\n......#.#.#.#.......#..#.....##\n..#..#....#.....#.#...##......#\n..#....#...#.###.........#.###.\n...#......##..#.#.....#...#....\n...#.......#...#...#........##.\n............#...#..#....#.....#\n....##......................#..\n#.#.#....#....#..........##....\n#.#.....#.#.##..#...#.##....##.\n...#...#..#...#..#.#.#.......#.\n#.....#..........#.........##.#\n#...##..#..#.#.......###....#..\n.#...#..##....#.....##.......#.\n....#.##.....#.........#.#....#\n........#.#...####..#.......#.#\n.####...#.#......####.....#.##.\n###..#....#..#.......#.#..##..#\n#......#.#....##..#.##.#....#.#\n...###...#...#..##.#..#..#.#...\n...##..##....#..#.....#........\n.....#..............#......#..#\n......#....#......#..#.........\n#..#.....#.##...........##.....\n.#..#.#..................##....\n#.#..#..##...#....#.#......#...\n.##.#.##......#.##...#...#...#.\n..#...#.........#.#..#.#....#..\n.#.####.#..#.#......##.#..#....\n#..#.......#....#..............\n....#............#..#..........\n.....#####.....#.....#..##...##\n#.#....#.#...............#..##.\n.#.#..#...#......#.....#.#.#...\n.#....#.#.#......#.....##....#.\n....#....#.##..#.......###...##\n.....#..#.##...#...#...#..#.#..\n##..#........#.#..#..##......#.\n.#..#..##.......#..#.....#.....\n.#.#.....###..##.#.#...........\n..##..##.####..........#..#....\n..##..#..#...#....#......#.#...\n#...#.#......##.....##.#..###..\n#..#..............#........##.#\n.........#.##..#.#..#..##.##.#.\n#....##....#.#..#.#...##..#....\n.#....#.......#............##..\n.......#.#.......#...#.#......#\n......##...#.......#.#........#\n..###..#.#.##....##...#....##..\n..##.##..........##..###.......\n.#.#.#..#..#.#.......#.#...##..\n..#..##.........#.###..#......#\n....#.#.#...##.#...#...##..###.\n..###..##.........##...#...#..#\n.#..##...#.......#.......#..#.#\n........##....##....#.#.###.#.#\n#.....#.#.................#.#..\n....#.#.#.....##.####.#......#.\n....#.......#.#.##.##..........\n...#...........#...#.##...#.###\n#....#....#..........#.##......\n##..#...........##.....##.##...\n.#.##...##..##....#..#.....####\n#...#...#.##..........##..##...\n....##..#....#.....#.#...#....#\n..#....#..##...###.#.#.........\n#......#.#.#...#...#.........#.\n#............###.#.#.#..##...#.\n.##.....####...##..##..#..##.#.\n#..#........#.....#.#.....#...#\n#............#....#.#.#........\n......##...##.#....#.....#...#.\n..#........##......#.#.....##..\n.#..#..#.....##.......#..#.#..#\n.#....#..#....##.#.#.#..#..#.##\n.####.#..........#...#..##.....\n...###..###...##..#............\n#..#.....##.#...#..##..#.......\n.....##....#...###.##...#......\n...##..#...#..#..##....##....#.\n...###....#.###.#.#.##....#....\n##.#.#.....#....#.#....#..#....\n.......##.....#.#..##...##...#.\n.#....#.#...##.#..#....#.....#.\n..#...#..#...#.##........#...#.\n#....#......##.#....##...#.#..#\n.....#..#..#..#......#...#.#.#.\n..###....#........#...#.......#\n###...#.......#.#.......##.##..\n......##.....#.#........#....#.\n#.##..#.#.#.#..#....#.##.....#.\n..........#.##.#...#...#..#..#.\n..#...##.#..........#..##.###..\n..###..##.##..#.#...##.####..#.\n#.#.#...............##....###.#\n....#.........#.#....#.#....#.#\n..#...#.###...#....###.....#...\n..#..#....#...#............#...\n.#..#....#..##.....##..........\n..#....#.#...#.#.#.#.......##.#\n.........#....##........#.#....\n...#..##.#..#.##...#...#.#....#\n....####...#...####.#....###..#\n......##...#.##.#.......#..#...\n#.#...#.#...#.#...#....#.#.#...\n.#.....##...#.....###.#....#...\n......##.....###...#.#...#.#...\n#..#..##.#.#......#....#..#..#.\n....#.###.....#..#...#.##.....#\n##.##........#......#....#..##.\n##.....##.#.....#.....##.....#.\n.....#.##...#.#..#.#.#.....#...\n.#.##..#...#.#..#.....#.#......\n.....##.......#..#...##..#..#..\n#.....#..#.####......#........#\n.#..#..##.#..##............#..#\n.##..#.#....##.##.....#......#.\n.......##.........#..#.........\n.#...#.......................#.\n#......#.#....##.#.......#..#..\n..##..##......#.......#....#.#.\n##......#......##...##.........\n..#....####....#.#.....##.#.#..\n..........#..#.#.#.....#..#.#..\n##..##...........##.......#....\n##....#.#....#..#......###....#\n...#.#.#..#.......##.......#...\n#....#.......#.......#.........\n...##......##....#...#......#.#\n#......#####.#.........#.....#.\n#..#.............#..#....#...#.\n.......#.##..#..#..#..#....####\n......#.##..##..........###...#\n.#.##....###..#........#....##.\n#......#..#...###.#...#.....#..\n.#.#.......#....##.......#.#...\n..#.##..#..##.....#.........#.#\n#.#...#..#.##....#.......##....\n.#.....###....#.#..#...#.....#.\n#...#..#.......#.#.....##...#.#\n#.#####.........#....##.....#..\n#....#..##...#....#.##.......#.\n.#.#.........##....##....#.....\n...#..##.......#....#.#.#......\n#.###.##...###....#.....#.####.\n.#...#.#.#..##.#..........#....\n#.#.....#.##.#..####.....##.#..\n...###.##..####.......#......##\n.##..#.........#...#.#.....#.##\n..#.....##....###.....#.#...##.\n#....#....#..#....#.##.........\n......###....#.#..#..#....##...\n.#.#................#.......##.\n...#.......#.........#.#.......\n...#..........#...##.....###...\n....#......#...#...............\n.##...#....#.....#.##......#...\n.#.....###...##..##...#.#......\n....##........#.....#...#....#.\n#.........#.#...##...#.#..#....\n...#.#.....#.#........#.#....#.\n.#........#.....#.#.#.#.#..#...\n....#...#.....#.#....#........#\n..###.#....#.#....##...##..#.##\n.#....#.#.####.#.#.....#.......\n.#...#...#.................##.#\n..................##..#..#.#.#.\n.#..#............##....###.....\n.......#....#...........#......\n....#.#.#.....###.........#..##\n...#.#....#.#.##.#.##.....##..#\n.#.##.#...##...#.......#.....##\n.#............#...#..##...#.#.#\n#.##..#.##..#..##.###.#........\n..............##....#...#..#.#.\n.#.#...#.#....#....###........#\n.#....#.#....#......###........\n..#.......##......#.##.....#...\n.....#......#..#...#.#.....#..."
-      .split("\n")
-
-    def navigate(row: Int, col: Int, right: Int, down: Int, encountered: Int): Long = {
-      if (row >= input.length) {
-        return encountered
-      }
-      navigate(row + down, (col + right) % input(row).length, right, down, if (input(row)(col) == '#') encountered + 1 else encountered)
-    }
-
-    val results = List(
-      navigate(0, 0, 1, 1, 0),
-      navigate(0, 0, 3, 1, 0),
-      navigate(0, 0, 5, 1, 0),
-      navigate(0, 0, 7, 1, 0),
-      navigate(0, 0, 1, 2, 0))
-
-    println(results)
-    println(results.product)
-  }
-
-  def _04_1() {
-    val req = "byr\niyr\neyr\nhgt\nhcl\necl\npid"
-      .split('\n')
-      .map(_ + ":")
-    val input = "eyr:2029 iyr:2013\nhcl:#ceb3a1 byr:1939 ecl:blu\nhgt:163cm\npid:660456119\n\nhcl:#0f8b2e ecl:grn\nbyr:1975 iyr:2011\neyr:2028 cid:207 hgt:158cm\npid:755567813\n\nbyr:2002 pid:867936514 eyr:2021 iyr:2012\nhcl:#18171d ecl:brn cid:293 hgt:177cm\n\nhgt:193cm\niyr:2010 pid:214278931 ecl:grn byr:1953\neyr:2021\nhcl:#733820\n\niyr:2010 eyr:2020 hcl:#866857\nbyr:1934 pid:022785900\nhgt:161cm ecl:oth\n\nhgt:166cm\nhcl:#602927\ncid:262 ecl:brn pid:393738288\neyr:2021 byr:1928 iyr:2010\n\necl:grn hcl:#6b5442\ncid:317 byr:2001\neyr:2023 iyr:2016 pid:407685013\nhgt:177cm\n\nhcl:#86127d ecl:grn pid:113577635 iyr:2018\nhgt:180cm eyr:2022 cid:59 byr:1921\n\nbyr:1984 eyr:2023\niyr:2015 hgt:152cm cid:177 ecl:amb\nhcl:#fffffd pid:379600323\n\nhgt:154cm byr:1930 ecl:amb\ncid:101 hcl:#fffffd pid:919013176\neyr:2024\n\nhgt:76in hcl:#a97842 byr:1920 pid:612193949\ncid:337 eyr:2026\necl:gry\n\ncid:262 iyr:2016 ecl:hzl hcl:#efcc98\nhgt:159cm eyr:2020 byr:1974\npid:520627197\n\ncid:302 eyr:1956 hgt:158cm hcl:#3355a2 pid:282247859\niyr:2015\nbyr:1979 ecl:gry\n\npid:156126542\neyr:2025\necl:amb hcl:#733820 hgt:187cm\nbyr:1938 iyr:2020\n\nhcl:#341e13 byr:1952 ecl:blu iyr:2015 cid:230 eyr:2023 hgt:59in pid:955240866\n\npid:783635907 eyr:2020 hgt:186cm ecl:grn byr:1925 iyr:2013 hcl:#341e13\n\nbyr:1921 iyr:2019 eyr:2027 hgt:175cm\npid:355931973 hcl:#18171d\necl:oth\n\nhcl:z\nbyr:1947 pid:164cm hgt:146 ecl:hzl\neyr:1976 iyr:2014 cid:160\n\npid:686203227\niyr:2014 hgt:170cm\nbyr:1985 ecl:gry eyr:2023\nhcl:#c0946f\n\nhgt:192cm\niyr:2016 byr:1954\nhcl:#c0946f eyr:2022\necl:oth pid:371575068\n\niyr:2013 pid:877594317 hgt:72in\ncid:236 byr:1922\neyr:2023 ecl:brn hcl:#ceb3a1\n\nhcl:#602927 pid:277638047 byr:1944 hgt:178cm iyr:2019 eyr:2028 ecl:blu\n\necl:hzl byr:1979 hcl:#a97842 iyr:2019 eyr:2022 cid:91 hgt:164cm pid:371132831\n\ncid:346 iyr:2019 pid:335209743 hgt:178cm byr:1993\necl:hzl hcl:#341e13 eyr:2030\n\nhcl:#ceb3a1 pid:863130968 byr:1964 ecl:brn\ncid:94 iyr:2012\nhgt:153cm\n\nbyr:1980 hcl:z eyr:2039 hgt:65cm\necl:lzr pid:96101128 iyr:2024\n\niyr:2017 eyr:2027\necl:grn\nbyr:1956 hcl:#888785 hgt:174cm\npid:565437685\n\npid:769106108 cid:77\neyr:2024\nhcl:#602927 hgt:164cm ecl:gry\niyr:2017 byr:1947\n\niyr:2019\necl:gry\nhcl:#fffffd hgt:187cm eyr:2022 cid:212 pid:475618502 byr:1993\n\nhcl:#623a2f ecl:grn\nhgt:191cm pid:166515049 iyr:2018 byr:1971 eyr:2022\n\niyr:2011 eyr:2029 byr:2029\nhcl:2912ec\npid:15469000 hgt:150cm\necl:#4ddfb1\n\necl:grn byr:1941 pid:273390626 hgt:166cm eyr:2024 iyr:2010\nhcl:#888785\n\niyr:2015 ecl:amb eyr:2030\nhgt:156cm\ncid:301 hcl:#18171d pid:677032916\nbyr:1957\n\necl:amb iyr:2010\nbyr:1967\neyr:2021 cid:128 pid:104999760 hgt:162cm\nhcl:#c83b73\n\neyr:2028 cid:173 hcl:#7d3b0c\nhgt:161cm pid:810244270 ecl:gry byr:1930 iyr:2013\n\nbyr:1975 pid:7243957480 eyr:2028 hcl:#cfa07d iyr:2019 hgt:184cm ecl:blu\n\nbyr:1953 pid:807621409 eyr:2020 iyr:2017 hcl:#ceb3a1 ecl:blu hgt:157cm\n\npid:976089116 byr:1920 eyr:2020 iyr:2011 hcl:#733820 hgt:166cm cid:330\n\npid:896621814 hgt:179cm ecl:blu\neyr:2027\niyr:2014\nbyr:1965\nhcl:#a97842\n\niyr:2015\npid:827722366 eyr:2024 hgt:188cm hcl:#ceb3a1 byr:1963\necl:oth\n\niyr:2014\nbyr:1959 hgt:150cm hcl:#602927 pid:948589059 eyr:2027 ecl:oth cid:215\n\neyr:2029\nhgt:171cm iyr:2019 hcl:#e9317f byr:1926\necl:gry\n\neyr:2020\ncid:260 hgt:155cm iyr:2011\nbyr:1948 hcl:#602927 pid:651156700\n\neyr:2027 hgt:188cm\nhcl:#873cdf iyr:2019 pid:738493109 byr:1979\necl:hzl\n\neyr:2025\necl:oth iyr:2012 pid:563787480\nbyr:1978 hgt:164cm\nhcl:#18171d\n\niyr:2012 eyr:2028 hgt:190cm ecl:blu byr:1971\nhcl:#6b5442 pid:758307028\n\ncid:186\necl:amb eyr:2027\niyr:2019 hcl:#c0946f hgt:165cm pid:904275084 byr:1997\n\npid:382971064 byr:1932 cid:77 iyr:2010 hgt:158cm ecl:brn\neyr:2027 hcl:#602927\n\nhgt:156cm eyr:2028 pid:249987568 iyr:2020\nbyr:1956\necl:grn hcl:#fffffd\n\nbyr:1947\npid:190618020\necl:blu\neyr:2023\nhgt:192cm iyr:2012 hcl:#b62955\n\neyr:2022 hgt:191cm pid:699194379\nhcl:#f0a94b iyr:2020 byr:1996\necl:gry\n\necl:brn hgt:69in\npid:368841807 byr:1971 eyr:2020 iyr:2020\nhcl:498a57\n\nhcl:#cfa07d\niyr:2012 ecl:gry\npid:034993671\neyr:2028 byr:1934\nhgt:156cm\n\niyr:2013 hgt:65in byr:1937 hcl:#fffffd\necl:blu eyr:2028 pid:777520867\n\nhcl:#c0946f\nbyr:1965\neyr:2023 pid:289861622 ecl:gry\nhgt:185cm\niyr:2012\n\neyr:2023 hcl:#cfa07d\niyr:2012 hgt:62in pid:447717379\ncid:309 byr:1977 ecl:hzl\n\nbyr:1934\necl:amb pid:385150170\neyr:2025 hgt:178cm\nhcl:#efcc98 cid:294\niyr:2012\n\neyr:2029\necl:gry iyr:2015 pid:396708902 cid:107 byr:1983 hgt:160cm hcl:#341e13\n\nhgt:158cm ecl:oth hcl:#cfa07d\npid:032310453\niyr:2012\nbyr:1945\ncid:60\n\npid:304883665 byr:1932 ecl:grn eyr:2022 hcl:#a97842 hgt:174cm iyr:2014\n\ncid:321 hgt:186cm eyr:2022 hcl:#623a2f byr:1952\niyr:2019 pid:857917879\n\nhcl:#ceb3a1 hgt:157cm byr:2002\niyr:2017 cid:245 ecl:amb\npid:142383109 eyr:2022\n\nhgt:176cm\nbyr:1999 ecl:oth hcl:#5e0906 pid:101108193 eyr:2029\n\nhgt:168in pid:8803058749 byr:2013 hcl:z iyr:2027\n\nbyr:1986\neyr:2023 pid:108222056\nhcl:#c0946f iyr:2012 hgt:71in ecl:grn\n\nbyr:1955 hcl:#a97842 ecl:blu\neyr:2020 pid:164459538\niyr:2018 hgt:175cm\n\nbyr:1950\nhcl:#341e13\necl:grn hgt:177cm cid:273 pid:473932418 eyr:2025 iyr:2010\n\npid:366720897 hgt:187cm hcl:#866857\nbyr:1967 cid:178 iyr:2014 ecl:amb\neyr:2027\n\niyr:2018\nhgt:174cm\neyr:2024\npid:478505944 byr:1938 hcl:#341e13 ecl:amb\ncid:181\n\nhgt:180cm pid:178969784 eyr:2025 ecl:brn\nbyr:1938 hcl:z\niyr:2013\n\niyr:2016 pid:447434977 hgt:71in hcl:#733820\neyr:2023 ecl:amb byr:1950\n\ncid:173\nbyr:1945 ecl:oth iyr:2010\nhgt:163cm hcl:#a97842 eyr:2025\npid:418871498\n\nhcl:#623a2f\necl:hzl byr:1937 iyr:2018 pid:017817627 eyr:2026 hgt:163cm\n\nhgt:162cm\nbyr:1982 iyr:2012 cid:123 pid:099838535\necl:hzl\nhcl:#efcc98 eyr:2021\n\npid:041801541 ecl:brn\nhgt:184cm iyr:2013 hcl:#6b5442\nbyr:1969\neyr:2020\n\necl:oth pid:314689468\nhcl:#fffffd byr:1934\neyr:2023\niyr:2019 hgt:151cm\n\neyr:2021\npid:206164494 ecl:amb hgt:158cm iyr:2014\n\neyr:2024 hgt:162cm hcl:#cfa07d\niyr:2015 pid:269887041 byr:2002\n\nhgt:179cm\nbyr:1953 hcl:#a87847 pid:966876004 iyr:2016 cid:193 eyr:2022\necl:oth\n\necl:blu cid:111\niyr:2019\nbyr:1955 pid:659491942\neyr:2028\nhcl:#602927\nhgt:179cm\n\neyr:2030 iyr:2011 ecl:oth cid:83 pid:160688997 byr:1994 hcl:#602927\nhgt:180cm\n\necl:brn eyr:2025\nhcl:#130738 iyr:2012 byr:1950 hgt:167cm pid:686738997\n\nhcl:#c0946f\nbyr:1937 pid:208013639 eyr:2026 cid:309 hgt:155cm ecl:hzl\niyr:2015\n\necl:amb\neyr:2022 pid:181247866\nbyr:1949 hgt:191cm hcl:#18171d iyr:2015\n\nhcl:#3fd612 eyr:2025 byr:1981 ecl:gry hgt:184cm iyr:2010\n\nbyr:1969 ecl:amb pid:770110016 iyr:2012 hgt:193cm hcl:#7d3b0c eyr:2020 cid:171\n\nbyr:1937\nhgt:162cm iyr:2020 ecl:blu hcl:#866857 eyr:2020\npid:057170032\n\npid:516772675 iyr:2018\neyr:2027 byr:1962 cid:233 hgt:176cm hcl:#623a2f\necl:hzl\n\npid:909808739 hgt:165cm iyr:2018 hcl:#18171d eyr:2028 ecl:hzl\ncid:254\n\neyr:2025\niyr:2011 hcl:#a97842 hgt:65in\nbyr:1981 ecl:blu pid:722094416\n\neyr:2027 hcl:#866857 byr:1953 pid:532811620 ecl:gry hgt:192cm iyr:2010\n\nhgt:161cm hcl:#7d3b0c byr:1983 ecl:grn\npid:242124004 eyr:2021\niyr:2017\n\neyr:2027 cid:240 byr:1955 pid:969478946 ecl:brn\niyr:2011 hcl:#341e13\n\niyr:2013\neyr:2030\npid:032502598 byr:1945 hcl:#efcc98 hgt:162cm\necl:gry\n\nhgt:165cm ecl:hzl\npid:851962288 eyr:2023\niyr:2020 hcl:#888785 byr:1981\n\nhgt:170cm\nhcl:#7d3b0c pid:897882367\ncid:207 ecl:hzl eyr:2030\niyr:2020\n\niyr:2014 hcl:#623a2f\nhgt:176cm eyr:2030\nbyr:1932\necl:brn pid:676358652\n\niyr:1963 eyr:2034\necl:#698113 hgt:76cm\nhcl:z pid:174cm byr:2029\n\nhgt:181cm byr:1966 eyr:2022 hcl:#888785 ecl:amb pid:692426080\n\niyr:2019 hcl:#c0946f byr:1948\neyr:2028 hgt:156cm cid:204 ecl:oth\npid:229984095\n\npid:859107750 hgt:188cm hcl:#efcc98 byr:1992 eyr:2029 ecl:amb\niyr:2010\n\npid:754939860 hcl:#602927\nhgt:164cm cid:261 ecl:gry byr:1966 eyr:2027 iyr:2011\n\neyr:2027 hcl:#602927 iyr:2019 hgt:71in pid:368543014 ecl:brn\nbyr:1991\n\nbyr:1934 hcl:#341e13 pid:410490656 cid:66 ecl:grt eyr:2028 hgt:188cm\niyr:2014\n\nbyr:1932\necl:amb\nhgt:188cm eyr:2020 pid:29447176 iyr:1947 hcl:#a6304b\n\nbyr:1980\nhcl:#733820 pid:800459957 hgt:191cm eyr:2020 cid:337 iyr:2010\n\neyr:2029 hgt:155cm byr:1994 hcl:#6b5442 ecl:grn iyr:2010 pid:978117883\n\niyr:2019\nbyr:1958\npid:466450124 ecl:grn hcl:#7d3b0c\nhgt:182cm\neyr:2029\n\nbyr:1970 pid:810234340 iyr:2014 cid:247 eyr:2021 hgt:172cm\nhcl:#cfa07d\necl:blu\n\neyr:2024 hcl:#7d3b0c byr:1930 cid:194 pid:216814907 hgt:63in ecl:gry\n\nbyr:1946 ecl:gry pid:579575823\nhgt:170cm\niyr:2015\neyr:2029\nhcl:#7d3b0c cid:88\n\nbyr:1936\nhcl:#6b5442 eyr:2020 cid:272 pid:752699834 iyr:2015 hgt:190cm\necl:grn\n\ncid:85 iyr:2018 eyr:2025 ecl:hzl hgt:192cm byr:1996 pid:478812793 hcl:#602927\n\nhgt:160cm byr:1968 ecl:hzl\nhcl:#9b96f3 iyr:2016 eyr:2027 pid:054732103\n\nhcl:#efcc98\nhgt:152cm pid:140347821\ncid:346\neyr:2022 ecl:oth iyr:2010 byr:1931\n\nhgt:187cm ecl:blu hcl:#fffffd\npid:061318772 byr:1969 iyr:2012\n\neyr:2021\nhgt:190cm\necl:amb byr:1963 pid:158cm\niyr:2015\nhcl:#b6652a\n\ncid:144 hcl:#c0946f iyr:2013 pid:384713923\necl:brn eyr:2028 hgt:157cm\nbyr:1981\n\nbyr:1981\necl:gry\neyr:2027 hgt:150cm\npid:769266043 hcl:#006f93 iyr:2015\n\nhgt:154cm\neyr:2022 pid:516617153\necl:amb byr:1993\niyr:2020\nhcl:#341e13\n\nhcl:#b6652a\nbyr:1943\npid:323629477 cid:311 ecl:grn eyr:2030 hgt:182cm\n\nhcl:#7d3b0c hgt:70in byr:1996\npid:205918254 eyr:2022 cid:178 iyr:2010\necl:brn\n\nbyr:2001\nhgt:186cm iyr:2019 ecl:brn\neyr:2024\nhcl:#888785\npid:218031016\n\necl:oth\nbyr:1971 hcl:#a97842 pid:673909751 hgt:152cm eyr:2026\niyr:2017\n\niyr:2017 byr:1966 cid:334 pid:#8a11cd\nhgt:188cm eyr:2023 hcl:#ceb3a1 ecl:brn\n\necl:amb iyr:2019 pid:835997489 hcl:#b6652a eyr:2026\nhgt:154cm\nbyr:1989\n\neyr:2033 cid:288 byr:1951\necl:zzz iyr:1962 hcl:#cfa07d\nhgt:72cm\n\ncid:238 hgt:162cm eyr:2020 byr:1995\necl:amb iyr:2010 pid:700982289\n\niyr:2010\npid:741394760 ecl:blu eyr:2030\nbyr:1934 hgt:68in hcl:#fffffd\n\npid:434939593 iyr:2020 hcl:#cfa07d cid:282 hgt:168cm\necl:blu byr:1939\n\nhgt:154cm byr:1981\necl:brn eyr:2029\niyr:2019 hcl:#602927 pid:329288264\ncid:307\n\niyr:2016\nhcl:#866857\nbyr:1968 eyr:2029 hgt:152cm pid:347204193 ecl:brn\n\necl:hzl\npid:975616547\nhgt:166cm\niyr:2015 hcl:#efcc98 eyr:2020\nbyr:1927\n\nbyr:1960 cid:309 eyr:2022\nhcl:#a97842 pid:186837033\niyr:2019 hgt:156cm\n\nhgt:165cm ecl:hzl pid:776872855 byr:1923\ncid:313 eyr:2029\n\neyr:2021 hcl:#18171d\nbyr:1938 hgt:160cm iyr:2020\npid:938987284 ecl:gry\n\niyr:2011\neyr:2026\necl:amb hgt:166cm\npid:727980371 cid:154 byr:1967 hcl:#888785\n\nbyr:1928\necl:grn hcl:#888785 pid:852102448 hgt:150cm eyr:2024\niyr:2018\n\nbyr:1997 cid:201 ecl:gry\neyr:2020\nhgt:163cm hcl:#7d3b0c\niyr:2011 pid:052314445\n\necl:hzl eyr:2024 pid:460808964 iyr:2015 byr:1965\nhcl:#c0946f hgt:189cm\n\niyr:2010 cid:163 byr:1944 ecl:grn pid:731085710\nhcl:#efcc98 hgt:159cm eyr:2027\n\nbyr:1937 hcl:#ceb3a1 eyr:2026 ecl:hzl iyr:2019\nhgt:185cm\n\niyr:2016 ecl:gry byr:1964 hcl:526fbd eyr:2023 pid:981371510\nhgt:71cm\n\neyr:2028 hcl:#18171d\npid:264437557\niyr:2014 byr:1987 hgt:168cm\necl:gry\n\npid:574867413\nhcl:#2b965f eyr:2025 hgt:154cm\nbyr:2001 ecl:hzl\niyr:2011\n\niyr:2020\ncid:212\necl:gry hgt:174cm\nbyr:1939\neyr:2029 hcl:#cfa07d\n\nhcl:#b6652a cid:249 pid:447524365 ecl:brn hgt:177cm byr:1959\neyr:2040\niyr:2011\n\nbyr:1935\necl:amb hcl:#fffffd pid:270076583 cid:128 hgt:60in iyr:2016\neyr:2027\n\nhcl:#c0946f pid:149533201 cid:332\necl:blu byr:1935 hgt:185cm\niyr:2016 eyr:2025\n\nhcl:#f27d9b cid:275 hgt:59in\nbyr:1928 iyr:2017\npid:311342224\neyr:2022 ecl:grn\n\necl:gry\nbyr:1985 cid:131\nhgt:191cm hcl:#6b5442 pid:957166785 iyr:2018 eyr:2029\n\npid:741921163 hgt:192cm byr:1982\niyr:2012 ecl:blu hcl:#623a2f eyr:2020\n\nbyr:1995 hgt:164cm eyr:2027 ecl:gry pid:086846266 hcl:#ceb3a1\niyr:2017\n\necl:brn\nhcl:#c0946f hgt:158cm\neyr:2020\npid:548013549 cid:107\niyr:2014 byr:1966\n\neyr:2021 hgt:192cm\niyr:2015 hcl:#888785 byr:1942 cid:104 pid:582902279 ecl:grn\n\nbyr:1923 hcl:#fffffd\nhgt:185cm pid:216803187\necl:amb\neyr:2030\niyr:2020 cid:153\n\npid:129687562 hgt:156cm eyr:2020 cid:336 byr:1964 hcl:#733820\niyr:2011\necl:gry\n\niyr:2012 hcl:#866857 pid:814749853 hgt:156cm cid:155 byr:1924 eyr:2024 ecl:oth\n\nhgt:151cm pid:832407555\ncid:188 byr:1949\niyr:2010 ecl:oth eyr:2022\nhcl:#adeffb\n\ncid:102 ecl:grn byr:1999\nhcl:#a97842 iyr:2013 hgt:173cm\npid:199221595 eyr:2029\n\nbyr:1963\nhcl:#fffffd pid:980136208 cid:230 ecl:amb iyr:2010 hgt:171cm eyr:2029\n\nbyr:1969 pid:524832668 hcl:#efcc98\necl:oth iyr:2019 eyr:2029\n\nhgt:184cm hcl:#13682c eyr:2022\niyr:2016\nbyr:1944 ecl:amb\npid:764280754\n\nhgt:162cm\nbyr:1922 eyr:2023 ecl:hzl pid:870409472 iyr:2012\n\nhcl:#888785 byr:2008\necl:utc iyr:1921 pid:993871206 hgt:152cm eyr:2026\n\nhcl:#18171d\nbyr:1924\nhgt:191cm\npid:130883621 iyr:2010 eyr:2028\n\nbyr:2001 hgt:185cm ecl:blu eyr:2020 iyr:2013 hcl:#888785\n\npid:043166927 cid:287 hgt:179cm iyr:2016 eyr:2021 byr:1960 ecl:amb hcl:#888785\n\necl:oth iyr:2020 hgt:183cm\nhcl:#fffffd byr:2013 eyr:2026 pid:042844334\n\npid:36857936 byr:1995 ecl:hzl hcl:#c0946f eyr:2025 hgt:162cm iyr:2011\n\niyr:2013 ecl:blu cid:92 byr:1946 pid:150720364 hcl:#7d3b0c hgt:180cm eyr:2027\n\ncid:276 eyr:2020 iyr:2011 hgt:59in hcl:#6b5442 ecl:amb byr:1992\npid:137604720\n\nbyr:1998 pid:239200796 hcl:#957b4b hgt:171cm iyr:2011\n\ncid:176\npid:872650041 byr:1987 hcl:#623a2f\nhgt:69in\necl:oth\neyr:2024 iyr:2015\n\npid:388875093 eyr:2024\necl:amb hgt:175cm\niyr:2019 byr:1970 hcl:#ceb3a1\n\npid:040506316 byr:1999\ncid:186\nhcl:#7d3b0c ecl:blu eyr:2022\nhgt:160cm iyr:2019\n\ncid:54\npid:314873473 iyr:2013 eyr:2020 hcl:#c0946f hgt:158cm ecl:amb\n\niyr:2016 ecl:oth hcl:#fffffd\nhgt:189cm eyr:2020\n\nbyr:1981\nhgt:163cm\necl:brn\nhcl:#7d3b0c iyr:2016\npid:500276094\ncid:98\neyr:2029\n\necl:oth hgt:158cm\neyr:2023 hcl:#c0946f iyr:1966 pid:544377825\nbyr:1938\n\ncid:170 hcl:#733820\niyr:2020 byr:1970 hgt:150cm pid:469096877\neyr:2020\n\npid:144977701 hgt:180cm byr:1975 hcl:#cfa07d\necl:oth cid:57 eyr:2022 iyr:2016\n\npid:591688826 hgt:156cm hcl:#a97842\necl:amb iyr:2011 byr:1955\neyr:2020\n\nhcl:#a97842 hgt:174cm\niyr:2012 eyr:2024 pid:928395919 ecl:blu byr:1927\n\necl:amb hgt:156 hcl:#888785 iyr:2019 eyr:2021\nbyr:1997 pid:980577052\n\niyr:2014 eyr:2028\nhgt:177cm ecl:oth byr:1930 pid:846909255 hcl:#602927\n\nhcl:#efcc98 iyr:2015\nbyr:1941\npid:387565513\necl:gry cid:54 eyr:2026 hgt:166cm\n\nhcl:#fffffd iyr:2014\necl:brn pid:241522887 byr:1963\nhgt:178cm\neyr:2024\n\necl:grn hgt:70in pid:202158837 eyr:2026 byr:1964\niyr:2018 hcl:#c0946f cid:77\n\nhgt:176cm iyr:2016\npid:927969731 byr:1939\neyr:2025 ecl:grn\nhcl:#fffffd\n\nbyr:1943 hgt:191cm ecl:hzl pid:422579553 iyr:2017 eyr:2020 hcl:#c2a152\n\neyr:2028 byr:1952\npid:506339509 hcl:#602927 hgt:179cm\necl:grn\n\nhcl:#341e13 iyr:2014 ecl:brn pid:823049712 eyr:2030 hgt:157cm\nbyr:1924\n\necl:hzl\nbyr:1954 pid:146052548 cid:194 eyr:2030 hcl:#ceb3a1 hgt:183cm\niyr:2013\n\nhcl:#6b5442\neyr:2026 iyr:2014 hgt:65in ecl:hzl pid:694135829 byr:1920\n\niyr:2012 byr:1984 hcl:#623a2f pid:270343526 ecl:gry cid:246 hgt:185cm\n\neyr:2026 ecl:grn\npid:853148268 iyr:2014 hgt:183cm hcl:#a97842 byr:1939\n\neyr:2021 hgt:182cm byr:1922 pid:199346431 cid:165 iyr:2016 ecl:hzl\nhcl:#b6652a\n\necl:#47904b pid:66740994\niyr:2025\nbyr:2029 eyr:2023 hcl:a6b541\n\necl:hzl\nbyr:1947\nhcl:#341e13 pid:466649892 eyr:2030 hgt:164cm iyr:2015\n\nhcl:#a97842 iyr:2016 eyr:2029\nbyr:1921 hgt:191cm ecl:brn pid:776471818\n\niyr:2014 pid:605101404 ecl:amb hcl:#7d3b0c byr:1991 eyr:2026 hgt:158cm\n\nbyr:1995 cid:271 ecl:hzl iyr:2012 hcl:#18171d pid:723865532\nhgt:165cm eyr:2029\n\necl:hzl\nhcl:#7d3b0c byr:2023\npid:83552498 eyr:2025 hgt:65cm iyr:2010\n\neyr:2022 byr:1975 hcl:#888785 hgt:165cm\ncid:166 ecl:brn iyr:2013\npid:261135534\n\nbyr:1999 cid:222 iyr:2020 ecl:hzl hgt:182cm pid:411279683 hcl:#1ca912\n\nbyr:1954 cid:273 ecl:#915d48\npid:004261180 hcl:#cfa07d iyr:2018 hgt:167cm eyr:2021\n\nhcl:#733820 byr:1931 iyr:2015\ncid:130 eyr:2029 hgt:183cm pid:962443962 ecl:gry\n\nhgt:192cm eyr:2029\npid:643859724\niyr:2016 ecl:blu byr:1951\n\nhgt:155cm\npid:293590926\nhcl:#602927 eyr:2027 byr:1959\niyr:2020\necl:amb\n\necl:oth hcl:#fffffd\npid:469517546 eyr:2026 iyr:2020 hgt:181cm\nbyr:1929\n\neyr:2026 iyr:2015 ecl:grn pid:644263519\ncid:254 hgt:64cm\nbyr:1944 hcl:#341e13\n\nhgt:150cm hcl:#341e13 ecl:oth\neyr:2028\nbyr:1922\npid:618759433 iyr:2017\n\necl:hzl hgt:169cm\nhcl:z eyr:2035 pid:620999812 byr:1998\n\npid:897252903 hgt:158cm iyr:2011 byr:1987 cid:221 eyr:2030\nhcl:#efcc98 ecl:gry\n\necl:gry eyr:2027\nbyr:1951\niyr:2012\npid:542287454 hcl:#6b5442\n\npid:#185456\nbyr:2028 cid:148 iyr:1957 hcl:df9a7d ecl:grt hgt:101\neyr:2036\n\nhcl:#b6652a eyr:2023\nhgt:150cm cid:316 ecl:gry byr:1922\npid:858150885 iyr:2013\n\nhgt:181in\nbyr:1986 ecl:#b9ca4d\niyr:2014 cid:77 eyr:2026\npid:277491106\n\nbyr:1996 ecl:oth pid:539623368 hcl:#602927 eyr:2024 iyr:2019\nhgt:179cm\n\npid:499804970 hcl:#18171d ecl:grn\nhgt:189cm eyr:2029\niyr:2012 byr:1986\n\npid:093626459 eyr:2024 iyr:2018\necl:grn byr:1922 hgt:153cm\n\nhgt:179cm byr:1990 eyr:2026 hcl:#623a2f ecl:blu\npid:306676089\niyr:2011\n\neyr:2022 ecl:oth\npid:894642127\nhcl:#5a9d3e iyr:2013 hgt:68in byr:1987\n\npid:059006259 byr:1953\nhcl:#7d3b0c ecl:oth iyr:2011 hgt:157cm eyr:2023\n\necl:oth\npid:255405447\neyr:2021 cid:130 hcl:#b6652a hgt:75in iyr:2020 byr:1959\n\ncid:124\npid:353288807 hcl:#7d3b0c\nbyr:1963 eyr:2020 ecl:oth\niyr:2020\nhgt:66in\n\nhgt:150cm\nhcl:#3a4c6e pid:027677435 ecl:hzl cid:256 byr:1960\niyr:2015\neyr:2026\n\necl:brn hcl:#c0946f\neyr:2028 pid:801870217\niyr:2010\n\ncid:170 hgt:175cm\neyr:2028\npid:620159472\niyr:2020 hcl:#7d3b0c ecl:brn\nbyr:1939\n\niyr:2019 hgt:161cm cid:152 ecl:brn hcl:#fffffd\nbyr:1938\neyr:2030 pid:311625365\n\nhgt:174cm hcl:#efcc98\nbyr:1934 pid:389966928 ecl:amb eyr:2023\ncid:235 iyr:2018\n\niyr:2010 byr:1960 pid:556639427 ecl:grn\nhgt:152cm\neyr:2022\nhcl:#623a2f\n\niyr:2014 hcl:#623a2f\nhgt:169cm\npid:243755754 ecl:oth eyr:2025\n\niyr:2016 hcl:#623a2f cid:152\necl:oth hgt:62in byr:1946\neyr:2028 pid:007367220\n\niyr:2013\necl:oth hgt:163cm\nbyr:1987\nhcl:#2cc1fd pid:346162920 eyr:2028\n\npid:636103282 byr:1952 eyr:2029 ecl:oth\nhcl:#fffffd cid:290 iyr:2013 hgt:63in\n\neyr:2021 hcl:#866857 iyr:2013 byr:1951\nhgt:157cm\npid:258369047 ecl:amb\n\nbyr:1983 iyr:2026 hgt:182cm eyr:2029\nhcl:#623a2f ecl:hzl pid:373940630\n\npid:283714222\nhgt:173cm ecl:gry byr:1939\niyr:2015 cid:81\n\nhgt:164cm cid:215\necl:amb eyr:2025 pid:459309210\nhcl:#ceb3a1 byr:1954\niyr:2015\n\nbyr:1931\necl:oth eyr:2028 hgt:177cm\nhcl:#6b5442\npid:494763488 iyr:2015 cid:293\n\necl:oth hcl:#fffffd hgt:188cm cid:252 pid:987529683 iyr:2020 byr:1945\n\neyr:2029\npid:745721513\niyr:2017 ecl:brn\nhgt:184cm hcl:#a97842\n\neyr:2024 ecl:oth\nhgt:153cm\niyr:2014 byr:1953 hcl:#888785\ncid:330 pid:532786321\n\nhgt:180cm\npid:441199844 eyr:2023 ecl:grn hcl:#b6652a byr:1958 iyr:2012\n\necl:brn\npid:835022632 hgt:174cm\nbyr:1928 hcl:#341e13 eyr:2028\ncid:176 iyr:2011\n\nhgt:72in byr:1964 hcl:#ceb3a1 ecl:gry\niyr:2020 eyr:2028\ncid:180\n\niyr:2025 byr:1999\neyr:2024 hcl:#c0946f\nhgt:86\n\neyr:2023 byr:1974 pid:067761346 cid:178 ecl:hzl iyr:2020 hgt:168cm hcl:#6b5442\n\nhgt:157cm ecl:gry eyr:2025\npid:803912278 byr:1963 iyr:2011 hcl:#ceb3a1\n\nhgt:190cm eyr:2026 ecl:grn\niyr:2012\nhcl:#c0946f byr:1929\n\neyr:2023 ecl:hzl pid:#fb639f\niyr:2030\ncid:132\nbyr:2016 hcl:#ce01c8\n\neyr:2021\npid:228596339\nbyr:1974\necl:blu\nhgt:174in iyr:2020\nhcl:#733820\n\necl:amb hgt:165cm\nbyr:1953 hcl:#118493 pid:458049702 cid:99\niyr:2019 eyr:2023\n\necl:hzl pid:277988952\neyr:2025 hgt:179cm byr:1992 cid:322\n\nhcl:#a97842 eyr:2021 cid:261\nhgt:61in pid:162402242\necl:amb byr:1938 iyr:2016\n\nhcl:#d125e3 iyr:2016 byr:1982 eyr:2027\nhgt:154cm\npid:365548961\n\niyr:2016 hgt:187cm byr:1980 pid:977322718\neyr:2027 ecl:brn hcl:#ceb3a1\n\niyr:2010 ecl:oth\npid:455361219 hgt:153cm eyr:2027 hcl:#6b5442\nbyr:1965"
-      .split("\n\n")
-
-    println(input.count(p => req.count(p.contains) == req.length))
-  }
-
-  def _04_2() {
-    val input = "eyr:2029 iyr:2013\nhcl:#ceb3a1 byr:1939 ecl:blu\nhgt:163cm\npid:660456119\n\nhcl:#0f8b2e ecl:grn\nbyr:1975 iyr:2011\neyr:2028 cid:207 hgt:158cm\npid:755567813\n\nbyr:2002 pid:867936514 eyr:2021 iyr:2012\nhcl:#18171d ecl:brn cid:293 hgt:177cm\n\nhgt:193cm\niyr:2010 pid:214278931 ecl:grn byr:1953\neyr:2021\nhcl:#733820\n\niyr:2010 eyr:2020 hcl:#866857\nbyr:1934 pid:022785900\nhgt:161cm ecl:oth\n\nhgt:166cm\nhcl:#602927\ncid:262 ecl:brn pid:393738288\neyr:2021 byr:1928 iyr:2010\n\necl:grn hcl:#6b5442\ncid:317 byr:2001\neyr:2023 iyr:2016 pid:407685013\nhgt:177cm\n\nhcl:#86127d ecl:grn pid:113577635 iyr:2018\nhgt:180cm eyr:2022 cid:59 byr:1921\n\nbyr:1984 eyr:2023\niyr:2015 hgt:152cm cid:177 ecl:amb\nhcl:#fffffd pid:379600323\n\nhgt:154cm byr:1930 ecl:amb\ncid:101 hcl:#fffffd pid:919013176\neyr:2024\n\nhgt:76in hcl:#a97842 byr:1920 pid:612193949\ncid:337 eyr:2026\necl:gry\n\ncid:262 iyr:2016 ecl:hzl hcl:#efcc98\nhgt:159cm eyr:2020 byr:1974\npid:520627197\n\ncid:302 eyr:1956 hgt:158cm hcl:#3355a2 pid:282247859\niyr:2015\nbyr:1979 ecl:gry\n\npid:156126542\neyr:2025\necl:amb hcl:#733820 hgt:187cm\nbyr:1938 iyr:2020\n\nhcl:#341e13 byr:1952 ecl:blu iyr:2015 cid:230 eyr:2023 hgt:59in pid:955240866\n\npid:783635907 eyr:2020 hgt:186cm ecl:grn byr:1925 iyr:2013 hcl:#341e13\n\nbyr:1921 iyr:2019 eyr:2027 hgt:175cm\npid:355931973 hcl:#18171d\necl:oth\n\nhcl:z\nbyr:1947 pid:164cm hgt:146 ecl:hzl\neyr:1976 iyr:2014 cid:160\n\npid:686203227\niyr:2014 hgt:170cm\nbyr:1985 ecl:gry eyr:2023\nhcl:#c0946f\n\nhgt:192cm\niyr:2016 byr:1954\nhcl:#c0946f eyr:2022\necl:oth pid:371575068\n\niyr:2013 pid:877594317 hgt:72in\ncid:236 byr:1922\neyr:2023 ecl:brn hcl:#ceb3a1\n\nhcl:#602927 pid:277638047 byr:1944 hgt:178cm iyr:2019 eyr:2028 ecl:blu\n\necl:hzl byr:1979 hcl:#a97842 iyr:2019 eyr:2022 cid:91 hgt:164cm pid:371132831\n\ncid:346 iyr:2019 pid:335209743 hgt:178cm byr:1993\necl:hzl hcl:#341e13 eyr:2030\n\nhcl:#ceb3a1 pid:863130968 byr:1964 ecl:brn\ncid:94 iyr:2012\nhgt:153cm\n\nbyr:1980 hcl:z eyr:2039 hgt:65cm\necl:lzr pid:96101128 iyr:2024\n\niyr:2017 eyr:2027\necl:grn\nbyr:1956 hcl:#888785 hgt:174cm\npid:565437685\n\npid:769106108 cid:77\neyr:2024\nhcl:#602927 hgt:164cm ecl:gry\niyr:2017 byr:1947\n\niyr:2019\necl:gry\nhcl:#fffffd hgt:187cm eyr:2022 cid:212 pid:475618502 byr:1993\n\nhcl:#623a2f ecl:grn\nhgt:191cm pid:166515049 iyr:2018 byr:1971 eyr:2022\n\niyr:2011 eyr:2029 byr:2029\nhcl:2912ec\npid:15469000 hgt:150cm\necl:#4ddfb1\n\necl:grn byr:1941 pid:273390626 hgt:166cm eyr:2024 iyr:2010\nhcl:#888785\n\niyr:2015 ecl:amb eyr:2030\nhgt:156cm\ncid:301 hcl:#18171d pid:677032916\nbyr:1957\n\necl:amb iyr:2010\nbyr:1967\neyr:2021 cid:128 pid:104999760 hgt:162cm\nhcl:#c83b73\n\neyr:2028 cid:173 hcl:#7d3b0c\nhgt:161cm pid:810244270 ecl:gry byr:1930 iyr:2013\n\nbyr:1975 pid:7243957480 eyr:2028 hcl:#cfa07d iyr:2019 hgt:184cm ecl:blu\n\nbyr:1953 pid:807621409 eyr:2020 iyr:2017 hcl:#ceb3a1 ecl:blu hgt:157cm\n\npid:976089116 byr:1920 eyr:2020 iyr:2011 hcl:#733820 hgt:166cm cid:330\n\npid:896621814 hgt:179cm ecl:blu\neyr:2027\niyr:2014\nbyr:1965\nhcl:#a97842\n\niyr:2015\npid:827722366 eyr:2024 hgt:188cm hcl:#ceb3a1 byr:1963\necl:oth\n\niyr:2014\nbyr:1959 hgt:150cm hcl:#602927 pid:948589059 eyr:2027 ecl:oth cid:215\n\neyr:2029\nhgt:171cm iyr:2019 hcl:#e9317f byr:1926\necl:gry\n\neyr:2020\ncid:260 hgt:155cm iyr:2011\nbyr:1948 hcl:#602927 pid:651156700\n\neyr:2027 hgt:188cm\nhcl:#873cdf iyr:2019 pid:738493109 byr:1979\necl:hzl\n\neyr:2025\necl:oth iyr:2012 pid:563787480\nbyr:1978 hgt:164cm\nhcl:#18171d\n\niyr:2012 eyr:2028 hgt:190cm ecl:blu byr:1971\nhcl:#6b5442 pid:758307028\n\ncid:186\necl:amb eyr:2027\niyr:2019 hcl:#c0946f hgt:165cm pid:904275084 byr:1997\n\npid:382971064 byr:1932 cid:77 iyr:2010 hgt:158cm ecl:brn\neyr:2027 hcl:#602927\n\nhgt:156cm eyr:2028 pid:249987568 iyr:2020\nbyr:1956\necl:grn hcl:#fffffd\n\nbyr:1947\npid:190618020\necl:blu\neyr:2023\nhgt:192cm iyr:2012 hcl:#b62955\n\neyr:2022 hgt:191cm pid:699194379\nhcl:#f0a94b iyr:2020 byr:1996\necl:gry\n\necl:brn hgt:69in\npid:368841807 byr:1971 eyr:2020 iyr:2020\nhcl:498a57\n\nhcl:#cfa07d\niyr:2012 ecl:gry\npid:034993671\neyr:2028 byr:1934\nhgt:156cm\n\niyr:2013 hgt:65in byr:1937 hcl:#fffffd\necl:blu eyr:2028 pid:777520867\n\nhcl:#c0946f\nbyr:1965\neyr:2023 pid:289861622 ecl:gry\nhgt:185cm\niyr:2012\n\neyr:2023 hcl:#cfa07d\niyr:2012 hgt:62in pid:447717379\ncid:309 byr:1977 ecl:hzl\n\nbyr:1934\necl:amb pid:385150170\neyr:2025 hgt:178cm\nhcl:#efcc98 cid:294\niyr:2012\n\neyr:2029\necl:gry iyr:2015 pid:396708902 cid:107 byr:1983 hgt:160cm hcl:#341e13\n\nhgt:158cm ecl:oth hcl:#cfa07d\npid:032310453\niyr:2012\nbyr:1945\ncid:60\n\npid:304883665 byr:1932 ecl:grn eyr:2022 hcl:#a97842 hgt:174cm iyr:2014\n\ncid:321 hgt:186cm eyr:2022 hcl:#623a2f byr:1952\niyr:2019 pid:857917879\n\nhcl:#ceb3a1 hgt:157cm byr:2002\niyr:2017 cid:245 ecl:amb\npid:142383109 eyr:2022\n\nhgt:176cm\nbyr:1999 ecl:oth hcl:#5e0906 pid:101108193 eyr:2029\n\nhgt:168in pid:8803058749 byr:2013 hcl:z iyr:2027\n\nbyr:1986\neyr:2023 pid:108222056\nhcl:#c0946f iyr:2012 hgt:71in ecl:grn\n\nbyr:1955 hcl:#a97842 ecl:blu\neyr:2020 pid:164459538\niyr:2018 hgt:175cm\n\nbyr:1950\nhcl:#341e13\necl:grn hgt:177cm cid:273 pid:473932418 eyr:2025 iyr:2010\n\npid:366720897 hgt:187cm hcl:#866857\nbyr:1967 cid:178 iyr:2014 ecl:amb\neyr:2027\n\niyr:2018\nhgt:174cm\neyr:2024\npid:478505944 byr:1938 hcl:#341e13 ecl:amb\ncid:181\n\nhgt:180cm pid:178969784 eyr:2025 ecl:brn\nbyr:1938 hcl:z\niyr:2013\n\niyr:2016 pid:447434977 hgt:71in hcl:#733820\neyr:2023 ecl:amb byr:1950\n\ncid:173\nbyr:1945 ecl:oth iyr:2010\nhgt:163cm hcl:#a97842 eyr:2025\npid:418871498\n\nhcl:#623a2f\necl:hzl byr:1937 iyr:2018 pid:017817627 eyr:2026 hgt:163cm\n\nhgt:162cm\nbyr:1982 iyr:2012 cid:123 pid:099838535\necl:hzl\nhcl:#efcc98 eyr:2021\n\npid:041801541 ecl:brn\nhgt:184cm iyr:2013 hcl:#6b5442\nbyr:1969\neyr:2020\n\necl:oth pid:314689468\nhcl:#fffffd byr:1934\neyr:2023\niyr:2019 hgt:151cm\n\neyr:2021\npid:206164494 ecl:amb hgt:158cm iyr:2014\n\neyr:2024 hgt:162cm hcl:#cfa07d\niyr:2015 pid:269887041 byr:2002\n\nhgt:179cm\nbyr:1953 hcl:#a87847 pid:966876004 iyr:2016 cid:193 eyr:2022\necl:oth\n\necl:blu cid:111\niyr:2019\nbyr:1955 pid:659491942\neyr:2028\nhcl:#602927\nhgt:179cm\n\neyr:2030 iyr:2011 ecl:oth cid:83 pid:160688997 byr:1994 hcl:#602927\nhgt:180cm\n\necl:brn eyr:2025\nhcl:#130738 iyr:2012 byr:1950 hgt:167cm pid:686738997\n\nhcl:#c0946f\nbyr:1937 pid:208013639 eyr:2026 cid:309 hgt:155cm ecl:hzl\niyr:2015\n\necl:amb\neyr:2022 pid:181247866\nbyr:1949 hgt:191cm hcl:#18171d iyr:2015\n\nhcl:#3fd612 eyr:2025 byr:1981 ecl:gry hgt:184cm iyr:2010\n\nbyr:1969 ecl:amb pid:770110016 iyr:2012 hgt:193cm hcl:#7d3b0c eyr:2020 cid:171\n\nbyr:1937\nhgt:162cm iyr:2020 ecl:blu hcl:#866857 eyr:2020\npid:057170032\n\npid:516772675 iyr:2018\neyr:2027 byr:1962 cid:233 hgt:176cm hcl:#623a2f\necl:hzl\n\npid:909808739 hgt:165cm iyr:2018 hcl:#18171d eyr:2028 ecl:hzl\ncid:254\n\neyr:2025\niyr:2011 hcl:#a97842 hgt:65in\nbyr:1981 ecl:blu pid:722094416\n\neyr:2027 hcl:#866857 byr:1953 pid:532811620 ecl:gry hgt:192cm iyr:2010\n\nhgt:161cm hcl:#7d3b0c byr:1983 ecl:grn\npid:242124004 eyr:2021\niyr:2017\n\neyr:2027 cid:240 byr:1955 pid:969478946 ecl:brn\niyr:2011 hcl:#341e13\n\niyr:2013\neyr:2030\npid:032502598 byr:1945 hcl:#efcc98 hgt:162cm\necl:gry\n\nhgt:165cm ecl:hzl\npid:851962288 eyr:2023\niyr:2020 hcl:#888785 byr:1981\n\nhgt:170cm\nhcl:#7d3b0c pid:897882367\ncid:207 ecl:hzl eyr:2030\niyr:2020\n\niyr:2014 hcl:#623a2f\nhgt:176cm eyr:2030\nbyr:1932\necl:brn pid:676358652\n\niyr:1963 eyr:2034\necl:#698113 hgt:76cm\nhcl:z pid:174cm byr:2029\n\nhgt:181cm byr:1966 eyr:2022 hcl:#888785 ecl:amb pid:692426080\n\niyr:2019 hcl:#c0946f byr:1948\neyr:2028 hgt:156cm cid:204 ecl:oth\npid:229984095\n\npid:859107750 hgt:188cm hcl:#efcc98 byr:1992 eyr:2029 ecl:amb\niyr:2010\n\npid:754939860 hcl:#602927\nhgt:164cm cid:261 ecl:gry byr:1966 eyr:2027 iyr:2011\n\neyr:2027 hcl:#602927 iyr:2019 hgt:71in pid:368543014 ecl:brn\nbyr:1991\n\nbyr:1934 hcl:#341e13 pid:410490656 cid:66 ecl:grt eyr:2028 hgt:188cm\niyr:2014\n\nbyr:1932\necl:amb\nhgt:188cm eyr:2020 pid:29447176 iyr:1947 hcl:#a6304b\n\nbyr:1980\nhcl:#733820 pid:800459957 hgt:191cm eyr:2020 cid:337 iyr:2010\n\neyr:2029 hgt:155cm byr:1994 hcl:#6b5442 ecl:grn iyr:2010 pid:978117883\n\niyr:2019\nbyr:1958\npid:466450124 ecl:grn hcl:#7d3b0c\nhgt:182cm\neyr:2029\n\nbyr:1970 pid:810234340 iyr:2014 cid:247 eyr:2021 hgt:172cm\nhcl:#cfa07d\necl:blu\n\neyr:2024 hcl:#7d3b0c byr:1930 cid:194 pid:216814907 hgt:63in ecl:gry\n\nbyr:1946 ecl:gry pid:579575823\nhgt:170cm\niyr:2015\neyr:2029\nhcl:#7d3b0c cid:88\n\nbyr:1936\nhcl:#6b5442 eyr:2020 cid:272 pid:752699834 iyr:2015 hgt:190cm\necl:grn\n\ncid:85 iyr:2018 eyr:2025 ecl:hzl hgt:192cm byr:1996 pid:478812793 hcl:#602927\n\nhgt:160cm byr:1968 ecl:hzl\nhcl:#9b96f3 iyr:2016 eyr:2027 pid:054732103\n\nhcl:#efcc98\nhgt:152cm pid:140347821\ncid:346\neyr:2022 ecl:oth iyr:2010 byr:1931\n\nhgt:187cm ecl:blu hcl:#fffffd\npid:061318772 byr:1969 iyr:2012\n\neyr:2021\nhgt:190cm\necl:amb byr:1963 pid:158cm\niyr:2015\nhcl:#b6652a\n\ncid:144 hcl:#c0946f iyr:2013 pid:384713923\necl:brn eyr:2028 hgt:157cm\nbyr:1981\n\nbyr:1981\necl:gry\neyr:2027 hgt:150cm\npid:769266043 hcl:#006f93 iyr:2015\n\nhgt:154cm\neyr:2022 pid:516617153\necl:amb byr:1993\niyr:2020\nhcl:#341e13\n\nhcl:#b6652a\nbyr:1943\npid:323629477 cid:311 ecl:grn eyr:2030 hgt:182cm\n\nhcl:#7d3b0c hgt:70in byr:1996\npid:205918254 eyr:2022 cid:178 iyr:2010\necl:brn\n\nbyr:2001\nhgt:186cm iyr:2019 ecl:brn\neyr:2024\nhcl:#888785\npid:218031016\n\necl:oth\nbyr:1971 hcl:#a97842 pid:673909751 hgt:152cm eyr:2026\niyr:2017\n\niyr:2017 byr:1966 cid:334 pid:#8a11cd\nhgt:188cm eyr:2023 hcl:#ceb3a1 ecl:brn\n\necl:amb iyr:2019 pid:835997489 hcl:#b6652a eyr:2026\nhgt:154cm\nbyr:1989\n\neyr:2033 cid:288 byr:1951\necl:zzz iyr:1962 hcl:#cfa07d\nhgt:72cm\n\ncid:238 hgt:162cm eyr:2020 byr:1995\necl:amb iyr:2010 pid:700982289\n\niyr:2010\npid:741394760 ecl:blu eyr:2030\nbyr:1934 hgt:68in hcl:#fffffd\n\npid:434939593 iyr:2020 hcl:#cfa07d cid:282 hgt:168cm\necl:blu byr:1939\n\nhgt:154cm byr:1981\necl:brn eyr:2029\niyr:2019 hcl:#602927 pid:329288264\ncid:307\n\niyr:2016\nhcl:#866857\nbyr:1968 eyr:2029 hgt:152cm pid:347204193 ecl:brn\n\necl:hzl\npid:975616547\nhgt:166cm\niyr:2015 hcl:#efcc98 eyr:2020\nbyr:1927\n\nbyr:1960 cid:309 eyr:2022\nhcl:#a97842 pid:186837033\niyr:2019 hgt:156cm\n\nhgt:165cm ecl:hzl pid:776872855 byr:1923\ncid:313 eyr:2029\n\neyr:2021 hcl:#18171d\nbyr:1938 hgt:160cm iyr:2020\npid:938987284 ecl:gry\n\niyr:2011\neyr:2026\necl:amb hgt:166cm\npid:727980371 cid:154 byr:1967 hcl:#888785\n\nbyr:1928\necl:grn hcl:#888785 pid:852102448 hgt:150cm eyr:2024\niyr:2018\n\nbyr:1997 cid:201 ecl:gry\neyr:2020\nhgt:163cm hcl:#7d3b0c\niyr:2011 pid:052314445\n\necl:hzl eyr:2024 pid:460808964 iyr:2015 byr:1965\nhcl:#c0946f hgt:189cm\n\niyr:2010 cid:163 byr:1944 ecl:grn pid:731085710\nhcl:#efcc98 hgt:159cm eyr:2027\n\nbyr:1937 hcl:#ceb3a1 eyr:2026 ecl:hzl iyr:2019\nhgt:185cm\n\niyr:2016 ecl:gry byr:1964 hcl:526fbd eyr:2023 pid:981371510\nhgt:71cm\n\neyr:2028 hcl:#18171d\npid:264437557\niyr:2014 byr:1987 hgt:168cm\necl:gry\n\npid:574867413\nhcl:#2b965f eyr:2025 hgt:154cm\nbyr:2001 ecl:hzl\niyr:2011\n\niyr:2020\ncid:212\necl:gry hgt:174cm\nbyr:1939\neyr:2029 hcl:#cfa07d\n\nhcl:#b6652a cid:249 pid:447524365 ecl:brn hgt:177cm byr:1959\neyr:2040\niyr:2011\n\nbyr:1935\necl:amb hcl:#fffffd pid:270076583 cid:128 hgt:60in iyr:2016\neyr:2027\n\nhcl:#c0946f pid:149533201 cid:332\necl:blu byr:1935 hgt:185cm\niyr:2016 eyr:2025\n\nhcl:#f27d9b cid:275 hgt:59in\nbyr:1928 iyr:2017\npid:311342224\neyr:2022 ecl:grn\n\necl:gry\nbyr:1985 cid:131\nhgt:191cm hcl:#6b5442 pid:957166785 iyr:2018 eyr:2029\n\npid:741921163 hgt:192cm byr:1982\niyr:2012 ecl:blu hcl:#623a2f eyr:2020\n\nbyr:1995 hgt:164cm eyr:2027 ecl:gry pid:086846266 hcl:#ceb3a1\niyr:2017\n\necl:brn\nhcl:#c0946f hgt:158cm\neyr:2020\npid:548013549 cid:107\niyr:2014 byr:1966\n\neyr:2021 hgt:192cm\niyr:2015 hcl:#888785 byr:1942 cid:104 pid:582902279 ecl:grn\n\nbyr:1923 hcl:#fffffd\nhgt:185cm pid:216803187\necl:amb\neyr:2030\niyr:2020 cid:153\n\npid:129687562 hgt:156cm eyr:2020 cid:336 byr:1964 hcl:#733820\niyr:2011\necl:gry\n\niyr:2012 hcl:#866857 pid:814749853 hgt:156cm cid:155 byr:1924 eyr:2024 ecl:oth\n\nhgt:151cm pid:832407555\ncid:188 byr:1949\niyr:2010 ecl:oth eyr:2022\nhcl:#adeffb\n\ncid:102 ecl:grn byr:1999\nhcl:#a97842 iyr:2013 hgt:173cm\npid:199221595 eyr:2029\n\nbyr:1963\nhcl:#fffffd pid:980136208 cid:230 ecl:amb iyr:2010 hgt:171cm eyr:2029\n\nbyr:1969 pid:524832668 hcl:#efcc98\necl:oth iyr:2019 eyr:2029\n\nhgt:184cm hcl:#13682c eyr:2022\niyr:2016\nbyr:1944 ecl:amb\npid:764280754\n\nhgt:162cm\nbyr:1922 eyr:2023 ecl:hzl pid:870409472 iyr:2012\n\nhcl:#888785 byr:2008\necl:utc iyr:1921 pid:993871206 hgt:152cm eyr:2026\n\nhcl:#18171d\nbyr:1924\nhgt:191cm\npid:130883621 iyr:2010 eyr:2028\n\nbyr:2001 hgt:185cm ecl:blu eyr:2020 iyr:2013 hcl:#888785\n\npid:043166927 cid:287 hgt:179cm iyr:2016 eyr:2021 byr:1960 ecl:amb hcl:#888785\n\necl:oth iyr:2020 hgt:183cm\nhcl:#fffffd byr:2013 eyr:2026 pid:042844334\n\npid:36857936 byr:1995 ecl:hzl hcl:#c0946f eyr:2025 hgt:162cm iyr:2011\n\niyr:2013 ecl:blu cid:92 byr:1946 pid:150720364 hcl:#7d3b0c hgt:180cm eyr:2027\n\ncid:276 eyr:2020 iyr:2011 hgt:59in hcl:#6b5442 ecl:amb byr:1992\npid:137604720\n\nbyr:1998 pid:239200796 hcl:#957b4b hgt:171cm iyr:2011\n\ncid:176\npid:872650041 byr:1987 hcl:#623a2f\nhgt:69in\necl:oth\neyr:2024 iyr:2015\n\npid:388875093 eyr:2024\necl:amb hgt:175cm\niyr:2019 byr:1970 hcl:#ceb3a1\n\npid:040506316 byr:1999\ncid:186\nhcl:#7d3b0c ecl:blu eyr:2022\nhgt:160cm iyr:2019\n\ncid:54\npid:314873473 iyr:2013 eyr:2020 hcl:#c0946f hgt:158cm ecl:amb\n\niyr:2016 ecl:oth hcl:#fffffd\nhgt:189cm eyr:2020\n\nbyr:1981\nhgt:163cm\necl:brn\nhcl:#7d3b0c iyr:2016\npid:500276094\ncid:98\neyr:2029\n\necl:oth hgt:158cm\neyr:2023 hcl:#c0946f iyr:1966 pid:544377825\nbyr:1938\n\ncid:170 hcl:#733820\niyr:2020 byr:1970 hgt:150cm pid:469096877\neyr:2020\n\npid:144977701 hgt:180cm byr:1975 hcl:#cfa07d\necl:oth cid:57 eyr:2022 iyr:2016\n\npid:591688826 hgt:156cm hcl:#a97842\necl:amb iyr:2011 byr:1955\neyr:2020\n\nhcl:#a97842 hgt:174cm\niyr:2012 eyr:2024 pid:928395919 ecl:blu byr:1927\n\necl:amb hgt:156 hcl:#888785 iyr:2019 eyr:2021\nbyr:1997 pid:980577052\n\niyr:2014 eyr:2028\nhgt:177cm ecl:oth byr:1930 pid:846909255 hcl:#602927\n\nhcl:#efcc98 iyr:2015\nbyr:1941\npid:387565513\necl:gry cid:54 eyr:2026 hgt:166cm\n\nhcl:#fffffd iyr:2014\necl:brn pid:241522887 byr:1963\nhgt:178cm\neyr:2024\n\necl:grn hgt:70in pid:202158837 eyr:2026 byr:1964\niyr:2018 hcl:#c0946f cid:77\n\nhgt:176cm iyr:2016\npid:927969731 byr:1939\neyr:2025 ecl:grn\nhcl:#fffffd\n\nbyr:1943 hgt:191cm ecl:hzl pid:422579553 iyr:2017 eyr:2020 hcl:#c2a152\n\neyr:2028 byr:1952\npid:506339509 hcl:#602927 hgt:179cm\necl:grn\n\nhcl:#341e13 iyr:2014 ecl:brn pid:823049712 eyr:2030 hgt:157cm\nbyr:1924\n\necl:hzl\nbyr:1954 pid:146052548 cid:194 eyr:2030 hcl:#ceb3a1 hgt:183cm\niyr:2013\n\nhcl:#6b5442\neyr:2026 iyr:2014 hgt:65in ecl:hzl pid:694135829 byr:1920\n\niyr:2012 byr:1984 hcl:#623a2f pid:270343526 ecl:gry cid:246 hgt:185cm\n\neyr:2026 ecl:grn\npid:853148268 iyr:2014 hgt:183cm hcl:#a97842 byr:1939\n\neyr:2021 hgt:182cm byr:1922 pid:199346431 cid:165 iyr:2016 ecl:hzl\nhcl:#b6652a\n\necl:#47904b pid:66740994\niyr:2025\nbyr:2029 eyr:2023 hcl:a6b541\n\necl:hzl\nbyr:1947\nhcl:#341e13 pid:466649892 eyr:2030 hgt:164cm iyr:2015\n\nhcl:#a97842 iyr:2016 eyr:2029\nbyr:1921 hgt:191cm ecl:brn pid:776471818\n\niyr:2014 pid:605101404 ecl:amb hcl:#7d3b0c byr:1991 eyr:2026 hgt:158cm\n\nbyr:1995 cid:271 ecl:hzl iyr:2012 hcl:#18171d pid:723865532\nhgt:165cm eyr:2029\n\necl:hzl\nhcl:#7d3b0c byr:2023\npid:83552498 eyr:2025 hgt:65cm iyr:2010\n\neyr:2022 byr:1975 hcl:#888785 hgt:165cm\ncid:166 ecl:brn iyr:2013\npid:261135534\n\nbyr:1999 cid:222 iyr:2020 ecl:hzl hgt:182cm pid:411279683 hcl:#1ca912\n\nbyr:1954 cid:273 ecl:#915d48\npid:004261180 hcl:#cfa07d iyr:2018 hgt:167cm eyr:2021\n\nhcl:#733820 byr:1931 iyr:2015\ncid:130 eyr:2029 hgt:183cm pid:962443962 ecl:gry\n\nhgt:192cm eyr:2029\npid:643859724\niyr:2016 ecl:blu byr:1951\n\nhgt:155cm\npid:293590926\nhcl:#602927 eyr:2027 byr:1959\niyr:2020\necl:amb\n\necl:oth hcl:#fffffd\npid:469517546 eyr:2026 iyr:2020 hgt:181cm\nbyr:1929\n\neyr:2026 iyr:2015 ecl:grn pid:644263519\ncid:254 hgt:64cm\nbyr:1944 hcl:#341e13\n\nhgt:150cm hcl:#341e13 ecl:oth\neyr:2028\nbyr:1922\npid:618759433 iyr:2017\n\necl:hzl hgt:169cm\nhcl:z eyr:2035 pid:620999812 byr:1998\n\npid:897252903 hgt:158cm iyr:2011 byr:1987 cid:221 eyr:2030\nhcl:#efcc98 ecl:gry\n\necl:gry eyr:2027\nbyr:1951\niyr:2012\npid:542287454 hcl:#6b5442\n\npid:#185456\nbyr:2028 cid:148 iyr:1957 hcl:df9a7d ecl:grt hgt:101\neyr:2036\n\nhcl:#b6652a eyr:2023\nhgt:150cm cid:316 ecl:gry byr:1922\npid:858150885 iyr:2013\n\nhgt:181in\nbyr:1986 ecl:#b9ca4d\niyr:2014 cid:77 eyr:2026\npid:277491106\n\nbyr:1996 ecl:oth pid:539623368 hcl:#602927 eyr:2024 iyr:2019\nhgt:179cm\n\npid:499804970 hcl:#18171d ecl:grn\nhgt:189cm eyr:2029\niyr:2012 byr:1986\n\npid:093626459 eyr:2024 iyr:2018\necl:grn byr:1922 hgt:153cm\n\nhgt:179cm byr:1990 eyr:2026 hcl:#623a2f ecl:blu\npid:306676089\niyr:2011\n\neyr:2022 ecl:oth\npid:894642127\nhcl:#5a9d3e iyr:2013 hgt:68in byr:1987\n\npid:059006259 byr:1953\nhcl:#7d3b0c ecl:oth iyr:2011 hgt:157cm eyr:2023\n\necl:oth\npid:255405447\neyr:2021 cid:130 hcl:#b6652a hgt:75in iyr:2020 byr:1959\n\ncid:124\npid:353288807 hcl:#7d3b0c\nbyr:1963 eyr:2020 ecl:oth\niyr:2020\nhgt:66in\n\nhgt:150cm\nhcl:#3a4c6e pid:027677435 ecl:hzl cid:256 byr:1960\niyr:2015\neyr:2026\n\necl:brn hcl:#c0946f\neyr:2028 pid:801870217\niyr:2010\n\ncid:170 hgt:175cm\neyr:2028\npid:620159472\niyr:2020 hcl:#7d3b0c ecl:brn\nbyr:1939\n\niyr:2019 hgt:161cm cid:152 ecl:brn hcl:#fffffd\nbyr:1938\neyr:2030 pid:311625365\n\nhgt:174cm hcl:#efcc98\nbyr:1934 pid:389966928 ecl:amb eyr:2023\ncid:235 iyr:2018\n\niyr:2010 byr:1960 pid:556639427 ecl:grn\nhgt:152cm\neyr:2022\nhcl:#623a2f\n\niyr:2014 hcl:#623a2f\nhgt:169cm\npid:243755754 ecl:oth eyr:2025\n\niyr:2016 hcl:#623a2f cid:152\necl:oth hgt:62in byr:1946\neyr:2028 pid:007367220\n\niyr:2013\necl:oth hgt:163cm\nbyr:1987\nhcl:#2cc1fd pid:346162920 eyr:2028\n\npid:636103282 byr:1952 eyr:2029 ecl:oth\nhcl:#fffffd cid:290 iyr:2013 hgt:63in\n\neyr:2021 hcl:#866857 iyr:2013 byr:1951\nhgt:157cm\npid:258369047 ecl:amb\n\nbyr:1983 iyr:2026 hgt:182cm eyr:2029\nhcl:#623a2f ecl:hzl pid:373940630\n\npid:283714222\nhgt:173cm ecl:gry byr:1939\niyr:2015 cid:81\n\nhgt:164cm cid:215\necl:amb eyr:2025 pid:459309210\nhcl:#ceb3a1 byr:1954\niyr:2015\n\nbyr:1931\necl:oth eyr:2028 hgt:177cm\nhcl:#6b5442\npid:494763488 iyr:2015 cid:293\n\necl:oth hcl:#fffffd hgt:188cm cid:252 pid:987529683 iyr:2020 byr:1945\n\neyr:2029\npid:745721513\niyr:2017 ecl:brn\nhgt:184cm hcl:#a97842\n\neyr:2024 ecl:oth\nhgt:153cm\niyr:2014 byr:1953 hcl:#888785\ncid:330 pid:532786321\n\nhgt:180cm\npid:441199844 eyr:2023 ecl:grn hcl:#b6652a byr:1958 iyr:2012\n\necl:brn\npid:835022632 hgt:174cm\nbyr:1928 hcl:#341e13 eyr:2028\ncid:176 iyr:2011\n\nhgt:72in byr:1964 hcl:#ceb3a1 ecl:gry\niyr:2020 eyr:2028\ncid:180\n\niyr:2025 byr:1999\neyr:2024 hcl:#c0946f\nhgt:86\n\neyr:2023 byr:1974 pid:067761346 cid:178 ecl:hzl iyr:2020 hgt:168cm hcl:#6b5442\n\nhgt:157cm ecl:gry eyr:2025\npid:803912278 byr:1963 iyr:2011 hcl:#ceb3a1\n\nhgt:190cm eyr:2026 ecl:grn\niyr:2012\nhcl:#c0946f byr:1929\n\neyr:2023 ecl:hzl pid:#fb639f\niyr:2030\ncid:132\nbyr:2016 hcl:#ce01c8\n\neyr:2021\npid:228596339\nbyr:1974\necl:blu\nhgt:174in iyr:2020\nhcl:#733820\n\necl:amb hgt:165cm\nbyr:1953 hcl:#118493 pid:458049702 cid:99\niyr:2019 eyr:2023\n\necl:hzl pid:277988952\neyr:2025 hgt:179cm byr:1992 cid:322\n\nhcl:#a97842 eyr:2021 cid:261\nhgt:61in pid:162402242\necl:amb byr:1938 iyr:2016\n\nhcl:#d125e3 iyr:2016 byr:1982 eyr:2027\nhgt:154cm\npid:365548961\n\niyr:2016 hgt:187cm byr:1980 pid:977322718\neyr:2027 ecl:brn hcl:#ceb3a1\n\niyr:2010 ecl:oth\npid:455361219 hgt:153cm eyr:2027 hcl:#6b5442\nbyr:1965"
-      .split("\n\n")
-
-    def forNum(min: Int, max: Int): String => Boolean = {
-      s => Try(s.toInt).map(i => min <= i && i <= max).getOrElse(false)
-    }
-
-    val hgtCm = "(\\d{3})cm".r
-    val hgtIn = "(\\d{2})in".r
-
-    def hgt(s: String): Boolean = s match {
-      case hgtCm(h) => 150 <= h.toInt && h.toInt <= 193
-      case hgtIn(h) => 59 <= h.toInt && h.toInt <= 76
-      case _ => false
-    }
-
-    val validations = HashMap(("byr", forNum(1920, 2002)),
-      ("iyr", forNum(2010, 2020)),
-      ("eyr", forNum(2020, 2030)),
-      ("hgt", hgt(_)),
-      ("hcl", (s: String) => s.matches("#[0-9a-f]{6}")),
-      ("ecl", (s: String) => s.matches("amb|blu|brn|gry|grn|hzl|oth")),
-      ("pid", (s: String) => s.matches("[0-9]{9}")))
-
-    val reqs = "byr\niyr\neyr\nhgt\nhcl\necl\npid"
-      .split('\n')
-      .map(f => Pattern.compile(s"($f):([\\S]+)"))
-
-    def valid(passport: String): Boolean = {
-      reqs.map(req => {
-        val mtch = req.matcher(passport)
-        mtch.find() && validations.get(mtch.group(1)).exists(_(mtch.group(2)))
-      })
-        .reduce(_ && _)
-    }
-
-    println(input.count(valid))
-  }
-
-  def _05_1() {
-    val input = "FFBFFBFLLR\nBFBFBFBRLR\nFBFFBFBRRL\nBBFFFBFLLR\nBFFFFFBLRL\nBFBBFBBRLR\nFBFBFBBRRL\nFBBBFBFLRL\nBFFFFBBRRL\nFBBBBBFLLR\nFBFFFBFLLR\nFFBBBFBLRR\nBFFFBBFLRL\nFFBBBFBLLL\nFBFBBFFLRR\nFBFBBBBLLR\nFFBBBFFRLR\nFBFBBFFRRL\nFBBBFFFLRR\nFFBBFFBLLR\nBFBFBFBLLR\nBFBBFFFLLL\nFBFBBFBRRL\nFFBBFFBLRR\nFFFBFBBLRR\nFFFFBBBRRL\nBFBBBBFRLL\nFBFBBFFLLR\nFFFBFBBLLR\nFFBFBBBRRL\nBFFBBFFLLR\nBFFFBBFLLR\nFBBFBBFRRR\nFFFBBFFRLL\nBFBFBBFRRR\nFBBFFBBRLR\nFBFFBBFRRL\nBFFFFBFLLR\nBBFFBFBLRL\nFBFBBBBLRL\nFBFBFBFRLL\nFFBFBFFLLR\nBFBFBBBRLR\nBFBBFFBLRR\nFFBFBFFLLL\nFFFFBBBLLR\nFFBBBBBLRR\nFFBFBBFLLR\nBFFFBBBRLL\nBFBFBFFLRR\nFBBBBFFRLR\nBFBFFBBRLL\nFBFBBFBRRR\nBFBFFBBLLL\nBFBFBBFLRL\nBBFFFFBLLR\nFFFBFFFRLL\nFBFFFBBRLL\nFFFBBBBLRR\nFBBBFBBLRL\nBBFFBFFRRR\nBFFBFFFLRR\nFBBBBBBLLR\nBFBBFFBLLL\nBFFBBBFRLR\nFBBBFFFLRL\nBBFFFBFRRL\nBBFFBFFRRL\nFFBBFFFLRR\nBBFFBFBRLR\nBBFFFFBRLR\nFFFFBBFLRL\nFFBBBFBRRR\nFFBFBBFLRR\nFFFBBFBRRR\nFBFFBFBRRR\nBFFFFFBRRR\nBFBFFBFLRL\nFFBBBBBLLL\nBFFFFBBRRR\nFBFBFFBRLR\nFBFFFFBLLR\nFFFBFFFLLL\nBFFFFFFLLR\nBBFFFFBRRL\nFBBFBBBRRL\nBFFFFFFLLL\nFBFBFFBLRR\nFFFBBFFRLR\nFBBFFBBLLL\nBFBBBFFLRL\nBFBFBBBRLL\nFBFBBFFLLL\nFFBBFFBRLL\nBFFFFBBLLR\nFFBBFBFLLL\nFFFFBBBRRR\nBFFFBBBLLL\nBFFBBFBRLR\nFFBBBFFLLR\nFBBBFFBRLR\nFFBFBFBRLL\nFFFBFFFLRR\nFFBBFBBLRL\nBFFBFBFLRR\nFFBFFFFRLR\nBFBFFFFLLL\nBFBBFBFRRR\nBFBFBFFRLL\nBFFFBFBRLL\nFFBBFBFLRR\nFFBFFFFRRL\nBFBBBBFLLL\nBFBFBFBRRR\nFFFFBBFLRR\nFBBBBFFLRL\nBFBFFFBLRL\nBFFBBFBRLL\nFBFFBBFLLL\nBFBBBBBLRL\nFFBFBBBLLR\nFBBFFBBLRR\nFFBFFFBRRR\nFFFBBBBRRR\nFBBBBBBLRL\nFFFBBFFLLL\nBBFFBFBLRR\nBFFBFFBRLR\nFFBBFBBRLL\nBFFBFBBRLR\nFBFBFBFRRL\nBFBBBFBRLL\nFFFFBBBLRR\nBBFFFFFLLR\nFFBFFFFLLL\nBFBBFBFRLR\nFFBBBBFRLR\nBFBFBBBLRL\nFBBFFFFRLL\nBFFFBBBLRL\nFBBBBFBRLL\nFFBFFFFRRR\nBFFFBFBRLR\nFBBBFBBLLL\nFFBFFBBLLL\nFBFFFBFRLL\nFFFFBBBRLR\nBBFFFBBLLL\nBBFFFFBLRL\nFBFBBBFRRL\nBFBFFBBRLR\nFBBFFBBLLR\nBFFBBFFLLL\nFFBFBBBLLL\nBFFBFFFRRL\nBFBBBFBRLR\nFBFBBBFLLL\nBFFBBBBLRR\nBFFBFBFLLL\nFFFBBBBLLR\nBFFFBFFLLR\nFFFBFBBRLL\nFFBBBBFRRL\nFBFBFBBLLR\nBBFFBFFLRL\nFBFBFBFLLR\nFBFBFFFLLR\nBFFFBFFLRR\nBFFBFFBRRR\nFFFFBBBLLL\nFBBFBFFLLR\nFFBBFBFRRL\nBFBFBBBLRR\nBFFFFFBRLR\nBFBBBBFRRL\nFFFBBBFLRR\nFBBFBFBRLL\nFFFBFFBRLR\nBFFFFFFRRL\nBFFBFBFRRL\nBFBBBBBRRL\nFBBFBFFRLL\nFFBFBBBRRR\nBBFFFBFLRL\nFBBBFBFRRR\nFBFFFFFLRL\nFFFBBBBRLR\nFBFFFFBLRR\nFBBFFBBLRL\nBFFFBBFRRL\nFFFBFBFRRR\nBFFBFFFRLL\nBFBFBFFLLR\nBFFBFBFRLL\nFFBFFBBLRR\nBFBBBBBLLR\nFBBBFFBRRL\nFFBBBFFRLL\nBFFBFBFLLR\nFFBBBBFRLL\nBFBFFFBRLR\nFBFFFFBLRL\nFFFBBFBRLR\nFFFBBFFLLR\nBFBBBBFLRR\nFFBFBFFRRR\nBFFBBBBRLR\nFBFFFFFRRL\nBFFFFFBRLL\nFFBBBBFLRL\nBFFFBFBRRL\nFFBFBFFLRL\nFFFBFFBLLR\nFBBBBFBRRL\nFFFFBBFLLR\nBFBFFBFRLR\nFBFFBFBRLL\nBFBBFBFLRL\nFBBFBBFLLL\nBFFFFFFRLL\nFBBBBBBRLR\nBFFFFBBRLL\nFBFBBFFLRL\nFBFFBBFLLR\nFBFBFBBLLL\nFFBBBFBLRL\nFFBBBBFRRR\nFFFFBBBLRL\nBFBBBBFRLR\nBFBFBBFLRR\nFFFBBFFRRL\nFFFBBFFRRR\nFBBBBFBRRR\nFFBBBBBRLL\nFBBBFFBRLL\nFBFBBFBLRR\nBFFFBBBRRR\nBFFBBBBLLR\nBFFFFBFRLR\nFFBFFFFLRR\nBFBBBFBLRR\nFBFFFBBRLR\nFFBFFBFLRR\nFFFBFBBLLL\nBFFBFFBRLL\nFBFFFBFLRR\nFBBBFBFLLL\nFBBFFBFLRR\nFFBFBFBRLR\nFFBBBBBRRR\nBFFBFBBRRL\nBFBFFBBRRR\nBFFBBFBLRR\nFBBFFFBRLL\nFBBBBFFLLL\nFBBBBBFLRR\nFBFFBBBLLL\nBFFBFFBLRR\nBBFFFFBRRR\nFBFFFBBLLL\nFFFFBFBRRL\nFBBFBBBRLR\nFBFBFBFRLR\nFBFFBFFRRR\nBBFFFBFRLL\nBFBBBFFRLR\nFBBFFFFRRL\nBBFFFFFLLL\nFBBFFFBRLR\nFFBFBFFRLR\nBFBFBBFRRL\nBFBBBBFLLR\nBFFFBBFLRR\nFBFBFBBLRR\nFBBFFFBLRR\nBFBFFBFLLR\nBFBFBFBRRL\nFBFBFFBLLR\nFBBFBFBLRL\nBFBFFFFRLR\nBFFBBFBLLR\nBFBBFBBRLL\nBFBFBFFRRR\nFFBBFBBLLR\nFFBFBFFLRR\nFFFBFFBLRR\nBFBFFFFLLR\nFBFBFFBRRL\nFBFBFFFRRL\nFBFFFFBRRR\nFFBBFFFRRL\nFFBFFBBRLR\nBFBBBFFLRR\nBBFFBFBRLL\nFBBBBBFRRL\nBBFFFFBRLL\nFFBBBBFLLR\nFBBBBFBRLR\nFFBFFBFRLR\nFFFFBFBRLR\nFBBFBFBLRR\nBFBBFBFLLR\nFFBFBBFRRR\nFFBBFBFRRR\nFFBBBBBLLR\nBFBBFFBLRL\nBFFBBFFRLR\nFFFBBFBRRL\nBFFFFFBRRL\nBFBBFFBRRL\nBFFBFBBLLR\nFBBFFFBRRR\nFBBBFFFLLR\nBFFFFFFRLR\nFBBFFFFLRL\nBFFFBBBLLR\nBFFFBFBLLR\nBFFFBFBLLL\nFFFBFBFLRL\nFBFBBBBRLR\nFBFFFFFRLL\nBFFBFFBLLL\nFFBBFBBLLL\nFBFFBFFRLL\nFBFFBFBLLR\nFFBBFBBLRR\nFFBFFFFLLR\nFBBFBFBRLR\nFFBBBFBRRL\nFBFBBFBRLL\nFBFFBFFLLR\nBFFBFBFRRR\nFBBFBBBLLR\nBFBBFFBRLR\nBFFBBFBRRR\nFBBBBFFLRR\nBBFFFFFLRR\nFBFFFFFRRR\nFFBBFFFLLR\nFFBFFBBRRL\nFBBFBBFRRL\nBFFFFFBLLL\nFFFBBBFLLL\nFBBBBFBLRR\nFFFBBFFLRL\nBFBBBFBLRL\nFBBFBBFLRL\nFBBFFBBRRR\nFFBFBBFRLL\nFBFFFFFLLL\nFFFBBBFRLL\nBFBBFBBRRR\nFFFBBFBRLL\nFFBBBFBRLL\nFBBBBBBLRR\nFBFBBBBRLL\nFFBBFFFRRR\nFFFBFFBLRL\nFBBFBFFRLR\nBFFFBFFRRR\nBFFFBFFRLL\nFBBBFFFRLR\nFFBBFBBRRL\nFFBBFFFRLL\nFFBBFBFRLL\nFBBFFBFLLL\nFBFFBBBRRL\nFBFBFFBRRR\nFFBFBFFRLL\nFFFBFFFRRR\nFFBFFFBLRR\nFFBFFFFRLL\nFBFFFFBLLL\nBFBBFFFLLR\nBFBFBFBLRR\nFBBFFFFLLR\nFBBBBFBLLR\nFBBBFBFRRL\nBFBBBBBRLR\nFBFFBFBLLL\nFFBBFFFLLL\nFFFFBBFRRR\nFBFBBBFLRR\nBBFFBBFLLL\nFFBBBFBLLR\nFBBFFFFLRR\nFBBFBFBLLR\nFFBFFFBLRL\nFBFBBBFLLR\nBFBBBFFLLR\nBFFFFBBRLR\nFBBFBFBLLL\nFBBFFBFRLR\nBFBBFFBRLL\nFBBBBFFLLR\nBFFBBBBLRL\nFBBFBFBRRL\nFBBFFBFLRL\nFBBBFFFRRL\nBFBFBFFLRL\nBFBFFFFLRR\nBBFFBFBLLL\nBBFFBFFLLR\nFBFBBFFRLL\nBFFBBBFLLR\nFFBFBBFRRL\nFBFFBBBLRR\nFBBBFBBLLR\nFBFFFBBRRL\nFFBBBBFLRR\nBFBFFBFLRR\nFFFBBBFLRL\nFBFFBFFRRL\nFBBFBFFLRL\nFBBFFFBLRL\nFFBBBFFRRR\nBBFFFBFLLL\nFBFFBFBLRR\nBBFFBFBLLR\nFBFFFFBRLR\nBFFFFFFLRL\nFBBFBBBLLL\nBFBFFBBLRR\nFFBBFFBRRL\nBFFFBFBLRL\nBFBBBBBRLL\nFBFFFBBLRR\nBFBBFBBLLR\nFBBBFFBLLL\nFBFFFBFRLR\nFFBBBFFRRL\nFFBFBBFLRL\nFFBFBFBLLR\nFBBFFBFRLL\nFFBBBFFLLL\nBFBBBBBLLL\nFBBFFFBLLL\nBFBFFFBLRR\nFFBFFBFLRL\nBFFFBBBRRL\nBFBFBBBLLR\nBFBFFFFLRL\nBFBFBFFRLR\nFBBFBBBRLL\nFFBFBBBLRL\nFBBBFFFRRR\nFBBFBFFRRR\nFFFFBBBRLL\nFBFFBBBLRL\nFBBBFBBRRL\nFBFFBBBLLR\nFFFBBBFRLR\nBFBFBFBLLL\nBBFFFFFRRR\nBFBBFBFRLL\nFBBBFBBLRR\nBFFFFBFLRL\nFBBBFFBLRL\nBBFFBFFLRR\nFFBFBFFRRL\nFBFFFBBLRL\nFFFFBBFRLL\nFFBFFFBLLR\nFBBBFBFLRR\nFFFBFFBRRR\nBFFBBBFLRR\nFBFFBFBLRL\nFFBFFFBRLR\nFFBFFFBRRL\nFFBBFFBLRL\nFBFBFBBRLR\nFFFBBBBLLL\nBFBBBFFLLL\nFFBBFBFLRL\nBFBFFFBRLL\nBFFBFFBRRL\nFBBFFFFLLL\nFBFFFFBRRL\nFFFBFFFLLR\nFBFFFBFLRL\nFFBFBFBLRL\nFBBBFFBLRR\nBFBFFBBRRL\nBBFFFFFLRL\nBFFBBBFRRR\nBFFFBBFRLR\nBFBBBBBLRR\nBFFFBFBLRR\nFBBBBFFRLL\nBFBBFFFLRR\nBBFFFBFLRR\nFBFFBBFRLL\nFBFFFFFLRR\nFBBBBBFLLL\nFBFBFBBRRR\nFBFBBFBLRL\nBFBBFFBRRR\nBBFFBFFLLL\nFBBFBBFLLR\nBFFBBFFRLL\nFBFBBBBRRL\nFBBFBBFRLL\nFBBFBBFLRR\nFBBFFBBRRL\nFFFBBFFLRR\nBFBBFBFLRR\nBFBFBFFRRL\nFFBBFFFRLR\nFBFBBBBLRR\nBBFFFFBLLL\nBBFFFFFRLR\nFBFBFFBLRL\nFBBFBBBLRR\nBFBFFBFRRL\nBFFFFFFLRR\nFFFBBBFLLR\nFBFFBFFLLL\nBFFBFFFRLR\nFBBFFBFRRR\nFFFBFFFRRL\nFFFFBBFLLL\nBFBFBBFRLL\nBFFFFBBLRL\nFFFBFBFLLR\nFBBBFBFRLR\nBFBBBFBLLR\nBFFBFBFLRL\nBFBFFFBLLL\nFFBFFBBLRL\nBFBBFFFRRR\nFFFBFFBLLL\nBBFFFFBLRR\nFFBBFBFLLR\nFBFBBBBRRR\nFBFBBBFLRL\nBFFBFFFLLR\nFBFBFFFRLL\nFBBBBFBLLL\nFBBBBBBRRR\nFBBBBBFLRL\nBFBBFFFRLL\nFFFBFFFRLR\nFFBFFBBLLR\nFFBFBBBRLL\nBBFFFFFRLL\nFFFBFBBRLR\nFFBFBBBRLR\nBFFFFBFLRR\nBBFFFBBLRL\nFBBBBFFRRL\nFBBFBBBRRR\nBFBFFFFRLL\nFBBBBBBRRL\nBFFBFFBLRL\nBFFBBFBRRL\nFBFFFBBRRR\nBFFBBFBLLL\nBFBFBFBRLL\nBBFFFBBRLL\nFBFBBBBLLL\nBBFFFBBLRR\nBFFFFBFRRL\nBFBBFBFLLL\nFBBBFBBRLR\nBFBBFBBLRL\nBFBBBFBRRL\nFBFFBBBRLL\nFBBBFBBRLL\nFFBBBBBRRL\nFBFBBBFRLL\nBFFFFFBLLR\nBFFBBBBRRR\nBFBBFBBLLL\nFFBFBBBLRR\nFFBFBFBRRR\nBFBFFFBRRL\nFBBBFFBRRR\nBFBFBBBLLL\nBBFFFBBRLR\nBBFFFBBRRL\nFBBBBFBLRL\nFBBBBFFRRR\nFBBBFFBLLR\nBFFFFBFLLL\nBFBFBFFLLL\nBFFFFBFRLL\nBFFBFBBLRL\nFFBFFBBRLL\nFBFBBFFRLR\nBFBFFFBRRR\nFBFFBFBRLR\nFBBFBFFLLL\nFBFFBBFLRR\nFFBFBFBLRR\nFFBBFFBRRR\nFBBFFFBLLR\nFBFFFBFRRR\nBFBFFFFRRL\nFBBBBBFRRR\nFFFFBFBRRR\nFFBBBBBLRL\nFBFBFFFLRR\nFBBBFFFLLL\nBFFFBFFRLR\nFFBFFBFRRL\nFBFBFFFRRR\nFFBFBFBLLL\nBFBBBFFRRL\nBFBFBFBLRL\nBFFFBFFLRL\nFBBBBBFRLR\nFBFBBFBLLL\nFFFBBBBLRL\nFBBFBFFLRR\nBFBBBFBRRR\nFBBBFBBRRR\nFFFBFBFRLL\nFFBBBBFLLL\nFBFFBFFRLR\nFBFBFBFLLL\nFBFFFFBRLL\nFBBBBBBRLL\nFBBBBBFRLL\nFBBFFBFLLR\nBFFBBBFRRL\nFBFFFBFLLL\nFBBBFFFRLL\nBFFFBBBRLR\nFBFBBBFRRR\nBFFBFBBRLL\nFBBFFFBRRL\nFBFFBBBRLR\nBFBFFBBLRL\nBFFBBFFLRL\nFBFBBBFRLR\nBBFFFBBLLR\nFFFBFFBRLL\nBFBFBBFLLL\nBFFBBFBLRL\nFFFBBFBLRL\nBFBBBFFRRR\nFBFBFFFRLR\nBFFBBFFLRR\nBFFBBBFLLL\nFBFFBBBRRR\nFBFBFFFLLL\nFFBFBBFLLL\nFBFFFFFLLR\nBFBBFBFRRL\nBFBBBBBRRR\nBFFBFBBLLL\nBFFFBFBRRR\nFFFBFBFLLL\nFFBFFFBRLL\nFBBFBFBRRR\nFBBBBBBLLL\nFBFBFFFLRL\nFFFBFFBRRL\nFBFBBFFRRR\nFFBBFBFRLR\nBFFFFBBLLL\nFBBFBFFRRL\nBFFBFFFRRR\nBFFBFFFLRL\nBFBBBBFRRR\nBFBFBBBRRL\nFFFBBFBLLL\nFBBFFBBRLL\nBFBBBFFRLL\nBFFFBBFRRR\nFBFBFBBRLL\nBFFFBFFRRL\nFFFBFBBLRL\nFBFFFBBLLR\nBFFFBBFLLL\nBFFBFFFLLL\nBFFBBBBLLL\nBFFFFFFRRR\nFFBFFBFRLL\nFBBFFBFRRL\nBBFFFFFRRL\nFFFBBFBLRR\nBFBFFBFLLL\nFBFBBFBRLR\nBFBBFBBLRR\nFBFFBFFLRL\nBFBFFFBLLR\nFFFBBBFRRR\nFFBBBFFLRR\nFBFBFFBLLL\nBFBFBBFRLR\nFBFBFBFRRR\nFFBBFBBRRR\nFBBFFFFRLR\nFBFBFBBLRL\nFBBFFFFRRR\nFFBBFFBRLR\nBFBFBBFLLR\nBFFBFBFRLR\nBBFFBFFRLL\nBFFBBFFRRR\nBFFFFBFRRR\nFBFBFBFLRL\nFFBFFBFRRR\nFBBFBBBLRL\nFFFBBBFRRL\nFBFFBFFLRR\nFFBFBBFRLR\nBBFFBFBRRL\nFBFFBBFLRL\nFFFFBBFRRL\nBFBFFBFRLL\nBFBBFBBRRL\nFFFBBFBLLR\nFFBFFBFLLL\nFFBBBBBRLR\nBFFFFFBLRR\nBFFBBBFLRL\nBFBBFFBLLR\nFBFFFFFRLR\nFFFFBBFRLR\nFBFFFBFRRL\nBFFBBBBRRL\nFFBFFFFLRL\nBFBBFFFRRL\nFFFBBBBRLL\nBFBFBBBRRR\nBBFFFBFRLR\nBFFBFFBLLR\nFBFBBFBLLR\nBFFBFBBRRR\nBBFFFBFRRR\nBFFBBFFRRL\nFBBBFBFRLL\nFFBFFFBLLL\nBFFBBBFRLL\nFBFFBBFRRR\nFFFBFBFLRR\nBFFBFBBLRR\nFBFFBBFRLR\nFFFBFBFRRL\nFFFBBBBRRL\nBFBFFBBLLR\nBFBFFFFRRR\nFFBBBFFLRL\nFBBFBBFRLR\nFFFBFFFLRL\nBFFBBBBRLL\nBFBBFFFRLR\nFFBFFBBRRR\nFBFBFFBRLL\nBFFFBFFLLL\nFFBFBFBRRL\nFFFBFBBRRL\nBFBBBBFLRL\nBBFFBFFRLR\nBFFFBBFRLL\nBFFFBBBLRR\nBFBFFBFRRR\nFBBBFBFLLR\nBFBBFFFLRL\nFFFBFBFRLR\nFBFBFBFLRR\nBBFFBFBRRR\nBBFFFBBRRR\nFFBBFFBLLL\nBFBBBFBLLL\nFFBBFFFLRL\nFFFBFBBRRR\nFFBBBFBRLR\nFFBBFBBRLR"
-      .split("\n")
-
-    println(input.map(_
-      .replaceAll("[BR]", "1")
-      .replaceAll("[FL]", "0"))
-      .map(parseInt(_, 2))
-      .max)
-  }
-
-  def _05_2() {
-    val input = "FFBFFBFLLR\nBFBFBFBRLR\nFBFFBFBRRL\nBBFFFBFLLR\nBFFFFFBLRL\nBFBBFBBRLR\nFBFBFBBRRL\nFBBBFBFLRL\nBFFFFBBRRL\nFBBBBBFLLR\nFBFFFBFLLR\nFFBBBFBLRR\nBFFFBBFLRL\nFFBBBFBLLL\nFBFBBFFLRR\nFBFBBBBLLR\nFFBBBFFRLR\nFBFBBFFRRL\nFBBBFFFLRR\nFFBBFFBLLR\nBFBFBFBLLR\nBFBBFFFLLL\nFBFBBFBRRL\nFFBBFFBLRR\nFFFBFBBLRR\nFFFFBBBRRL\nBFBBBBFRLL\nFBFBBFFLLR\nFFFBFBBLLR\nFFBFBBBRRL\nBFFBBFFLLR\nBFFFBBFLLR\nFBBFBBFRRR\nFFFBBFFRLL\nBFBFBBFRRR\nFBBFFBBRLR\nFBFFBBFRRL\nBFFFFBFLLR\nBBFFBFBLRL\nFBFBBBBLRL\nFBFBFBFRLL\nFFBFBFFLLR\nBFBFBBBRLR\nBFBBFFBLRR\nFFBFBFFLLL\nFFFFBBBLLR\nFFBBBBBLRR\nFFBFBBFLLR\nBFFFBBBRLL\nBFBFBFFLRR\nFBBBBFFRLR\nBFBFFBBRLL\nFBFBBFBRRR\nBFBFFBBLLL\nBFBFBBFLRL\nBBFFFFBLLR\nFFFBFFFRLL\nFBFFFBBRLL\nFFFBBBBLRR\nFBBBFBBLRL\nBBFFBFFRRR\nBFFBFFFLRR\nFBBBBBBLLR\nBFBBFFBLLL\nBFFBBBFRLR\nFBBBFFFLRL\nBBFFFBFRRL\nBBFFBFFRRL\nFFBBFFFLRR\nBBFFBFBRLR\nBBFFFFBRLR\nFFFFBBFLRL\nFFBBBFBRRR\nFFBFBBFLRR\nFFFBBFBRRR\nFBFFBFBRRR\nBFFFFFBRRR\nBFBFFBFLRL\nFFBBBBBLLL\nBFFFFBBRRR\nFBFBFFBRLR\nFBFFFFBLLR\nFFFBFFFLLL\nBFFFFFFLLR\nBBFFFFBRRL\nFBBFBBBRRL\nBFFFFFFLLL\nFBFBFFBLRR\nFFFBBFFRLR\nFBBFFBBLLL\nBFBBBFFLRL\nBFBFBBBRLL\nFBFBBFFLLL\nFFBBFFBRLL\nBFFFFBBLLR\nFFBBFBFLLL\nFFFFBBBRRR\nBFFFBBBLLL\nBFFBBFBRLR\nFFBBBFFLLR\nFBBBFFBRLR\nFFBFBFBRLL\nFFFBFFFLRR\nFFBBFBBLRL\nBFFBFBFLRR\nFFBFFFFRLR\nBFBFFFFLLL\nBFBBFBFRRR\nBFBFBFFRLL\nBFFFBFBRLL\nFFBBFBFLRR\nFFBFFFFRRL\nBFBBBBFLLL\nBFBFBFBRRR\nFFFFBBFLRR\nFBBBBFFLRL\nBFBFFFBLRL\nBFFBBFBRLL\nFBFFBBFLLL\nBFBBBBBLRL\nFFBFBBBLLR\nFBBFFBBLRR\nFFBFFFBRRR\nFFFBBBBRRR\nFBBBBBBLRL\nFFFBBFFLLL\nBBFFBFBLRR\nBFFBFFBRLR\nFFBBFBBRLL\nBFFBFBBRLR\nFBFBFBFRRL\nBFBBBFBRLL\nFFFFBBBLRR\nBBFFFFFLLR\nFFBFFFFLLL\nBFBBFBFRLR\nFFBBBBFRLR\nBFBFBBBLRL\nFBBFFFFRLL\nBFFFBBBLRL\nFBBBBFBRLL\nFFBFFFFRRR\nBFFFBFBRLR\nFBBBFBBLLL\nFFBFFBBLLL\nFBFFFBFRLL\nFFFFBBBRLR\nBBFFFBBLLL\nBBFFFFBLRL\nFBFBBBFRRL\nBFBFFBBRLR\nFBBFFBBLLR\nBFFBBFFLLL\nFFBFBBBLLL\nBFFBFFFRRL\nBFBBBFBRLR\nFBFBBBFLLL\nBFFBBBBLRR\nBFFBFBFLLL\nFFFBBBBLLR\nBFFFBFFLLR\nFFFBFBBRLL\nFFBBBBFRRL\nFBFBFBBLLR\nBBFFBFFLRL\nFBFBFBFLLR\nFBFBFFFLLR\nBFFFBFFLRR\nBFFBFFBRRR\nFFFFBBBLLL\nFBBFBFFLLR\nFFBBFBFRRL\nBFBFBBBLRR\nBFFFFFBRLR\nBFBBBBFRRL\nFFFBBBFLRR\nFBBFBFBRLL\nFFFBFFBRLR\nBFFFFFFRRL\nBFFBFBFRRL\nBFBBBBBRRL\nFBBFBFFRLL\nFFBFBBBRRR\nBBFFFBFLRL\nFBBBFBFRRR\nFBFFFFFLRL\nFFFBBBBRLR\nFBFFFFBLRR\nFBBFFBBLRL\nBFFFBBFRRL\nFFFBFBFRRR\nBFFBFFFRLL\nBFBFBFFLLR\nBFFBFBFRLL\nFFBFFBBLRR\nBFBBBBBLLR\nFBBBFFBRRL\nFFBBBFFRLL\nBFFBFBFLLR\nFFBBBBFRLL\nBFBFFFBRLR\nFBFFFFBLRL\nFFFBBFBRLR\nFFFBBFFLLR\nBFBBBBFLRR\nFFBFBFFRRR\nBFFBBBBRLR\nFBFFFFFRRL\nBFFFFFBRLL\nFFBBBBFLRL\nBFFFBFBRRL\nFFBFBFFLRL\nFFFBFFBLLR\nFBBBBFBRRL\nFFFFBBFLLR\nBFBFFBFRLR\nFBFFBFBRLL\nBFBBFBFLRL\nFBBFBBFLLL\nBFFFFFFRLL\nFBBBBBBRLR\nBFFFFBBRLL\nFBFBBFFLRL\nFBFFBBFLLR\nFBFBFBBLLL\nFFBBBFBLRL\nFFBBBBFRRR\nFFFFBBBLRL\nBFBBBBFRLR\nBFBFBBFLRR\nFFFBBFFRRL\nFFFBBFFRRR\nFBBBBFBRRR\nFFBBBBBRLL\nFBBBFFBRLL\nFBFBBFBLRR\nBFFFBBBRRR\nBFFBBBBLLR\nBFFFFBFRLR\nFFBFFFFLRR\nBFBBBFBLRR\nFBFFFBBRLR\nFFBFFBFLRR\nFFFBFBBLLL\nBFFBFFBRLL\nFBFFFBFLRR\nFBBBFBFLLL\nFBBFFBFLRR\nFFBFBFBRLR\nFFBBBBBRRR\nBFFBFBBRRL\nBFBFFBBRRR\nBFFBBFBLRR\nFBBFFFBRLL\nFBBBBFFLLL\nFBBBBBFLRR\nFBFFBBBLLL\nBFFBFFBLRR\nBBFFFFBRRR\nFBFFFBBLLL\nFFFFBFBRRL\nFBBFBBBRLR\nFBFBFBFRLR\nFBFFBFFRRR\nBBFFFBFRLL\nBFBBBFFRLR\nFBBFFFFRRL\nBBFFFFFLLL\nFBBFFFBRLR\nFFBFBFFRLR\nBFBFBBFRRL\nBFBBBBFLLR\nBFFFBBFLRR\nFBFBFBBLRR\nFBBFFFBLRR\nBFBFFBFLLR\nBFBFBFBRRL\nFBFBFFBLLR\nFBBFBFBLRL\nBFBFFFFRLR\nBFFBBFBLLR\nBFBBFBBRLL\nBFBFBFFRRR\nFFBBFBBLLR\nFFBFBFFLRR\nFFFBFFBLRR\nBFBFFFFLLR\nFBFBFFBRRL\nFBFBFFFRRL\nFBFFFFBRRR\nFFBBFFFRRL\nFFBFFBBRLR\nBFBBBFFLRR\nBBFFBFBRLL\nFBBBBBFRRL\nBBFFFFBRLL\nFFBBBBFLLR\nFBBBBFBRLR\nFFBFFBFRLR\nFFFFBFBRLR\nFBBFBFBLRR\nBFBBFBFLLR\nFFBFBBFRRR\nFFBBFBFRRR\nFFBBBBBLLR\nBFBBFFBLRL\nBFFBBFFRLR\nFFFBBFBRRL\nBFFFFFBRRL\nBFBBFFBRRL\nBFFBFBBLLR\nFBBFFFBRRR\nFBBBFFFLLR\nBFFFFFFRLR\nFBBFFFFLRL\nBFFFBBBLLR\nBFFFBFBLLR\nBFFFBFBLLL\nFFFBFBFLRL\nFBFBBBBRLR\nFBFFFFFRLL\nBFFBFFBLLL\nFFBBFBBLLL\nFBFFBFFRLL\nFBFFBFBLLR\nFFBBFBBLRR\nFFBFFFFLLR\nFBBFBFBRLR\nFFBBBFBRRL\nFBFBBFBRLL\nFBFFBFFLLR\nBFFBFBFRRR\nFBBFBBBLLR\nBFBBFFBRLR\nBFFBBFBRRR\nFBBBBFFLRR\nBBFFFFFLRR\nFBFFFFFRRR\nFFBBFFFLLR\nFFBFFBBRRL\nFBBFBBFRRL\nBFFFFFBLLL\nFFFBBBFLLL\nFBBBBFBLRR\nFFFBBFFLRL\nBFBBBFBLRL\nFBBFBBFLRL\nFBBFFBBRRR\nFFBFBBFRLL\nFBFFFFFLLL\nFFFBBBFRLL\nBFBBFBBRRR\nFFFBBFBRLL\nFFBBBFBRLL\nFBBBBBBLRR\nFBFBBBBRLL\nFFBBFFFRRR\nFFFBFFBLRL\nFBBFBFFRLR\nBFFFBFFRRR\nBFFFBFFRLL\nFBBBFFFRLR\nFFBBFBBRRL\nFFBBFFFRLL\nFFBBFBFRLL\nFBBFFBFLLL\nFBFFBBBRRL\nFBFBFFBRRR\nFFBFBFFRLL\nFFFBFFFRRR\nFFBFFFBLRR\nFFBFFFFRLL\nFBFFFFBLLL\nBFBBFFFLLR\nBFBFBFBLRR\nFBBFFFFLLR\nFBBBBFBLLR\nFBBBFBFRRL\nBFBBBBBRLR\nFBFFBFBLLL\nFFBBFFFLLL\nFFFFBBFRRR\nFBFBBBFLRR\nBBFFBBFLLL\nFFBBBFBLLR\nFBBFFFFLRR\nFBBFBFBLLR\nFFBFFFBLRL\nFBFBBBFLLR\nBFBBBFFLLR\nBFFFFBBRLR\nFBBFBFBLLL\nFBBFFBFRLR\nBFBBFFBRLL\nFBBBBFFLLR\nBFFBBBBLRL\nFBBFBFBRRL\nFBBFFBFLRL\nFBBBFFFRRL\nBFBFBFFLRL\nBFBFFFFLRR\nBBFFBFBLLL\nBBFFBFFLLR\nFBFBBFFRLL\nBFFBBBFLLR\nFFBFBBFRRL\nFBFFBBBLRR\nFBBBFBBLLR\nFBFFFBBRRL\nFFBBBBFLRR\nBFBFFBFLRR\nFFFBBBFLRL\nFBFFBFFRRL\nFBBFBFFLRL\nFBBFFFBLRL\nFFBBBFFRRR\nBBFFFBFLLL\nFBFFBFBLRR\nBBFFBFBLLR\nFBFFFFBRLR\nBFFFFFFLRL\nFBBFBBBLLL\nBFBFFBBLRR\nFFBBFFBRRL\nBFFFBFBLRL\nBFBBBBBRLL\nFBFFFBBLRR\nBFBBFBBLLR\nFBBBFFBLLL\nFBFFFBFRLR\nFFBBBFFRRL\nFFBFBBFLRL\nFFBFBFBLLR\nFBBFFBFRLL\nFFBBBFFLLL\nBFBBBBBLLL\nFBBFFFBLLL\nBFBFFFBLRR\nFFBFFBFLRL\nBFFFBBBRRL\nBFBFBBBLLR\nBFBFFFFLRL\nBFBFBFFRLR\nFBBFBBBRLL\nFFBFBBBLRL\nFBBBFFFRRR\nFBBFBFFRRR\nFFFFBBBRLL\nFBFFBBBLRL\nFBBBFBBRRL\nFBFFBBBLLR\nFFFBBBFRLR\nBFBFBFBLLL\nBBFFFFFRRR\nBFBBFBFRLL\nFBBBFBBLRR\nBFFFFBFLRL\nFBBBFFBLRL\nBBFFBFFLRR\nFFBFBFFRRL\nFBFFFBBLRL\nFFFFBBFRLL\nFFBFFFBLLR\nFBBBFBFLRR\nFFFBFFBRRR\nBFFBBBFLRR\nFBFFBFBLRL\nFFBFFFBRLR\nFFBFFFBRRL\nFFBBFFBLRL\nFBFBFBBRLR\nFFFBBBBLLL\nBFBBBFFLLL\nFFBBFBFLRL\nBFBFFFBRLL\nBFFBFFBRRL\nFBBFFFFLLL\nFBFFFFBRRL\nFFFBFFFLLR\nFBFFFBFLRL\nFFBFBFBLRL\nFBBBFFBLRR\nBFBFFBBRRL\nBBFFFFFLRL\nBFFBBBFRRR\nBFFFBBFRLR\nBFBBBBBLRR\nBFFFBFBLRR\nFBBBBFFRLL\nBFBBFFFLRR\nBBFFFBFLRR\nFBFFBBFRLL\nFBFFFFFLRR\nFBBBBBFLLL\nFBFBFBBRRR\nFBFBBFBLRL\nBFBBFFBRRR\nBBFFBFFLLL\nFBBFBBFLLR\nBFFBBFFRLL\nFBFBBBBRRL\nFBBFBBFRLL\nFBBFBBFLRR\nFBBFFBBRRL\nFFFBBFFLRR\nBFBBFBFLRR\nBFBFBFFRRL\nFFBBFFFRLR\nFBFBBBBLRR\nBBFFFFBLLL\nBBFFFFFRLR\nFBFBFFBLRL\nFBBFBBBLRR\nBFBFFBFRRL\nBFFFFFFLRR\nFFFBBBFLLR\nFBFFBFFLLL\nBFFBFFFRLR\nFBBFFBFRRR\nFFFBFFFRRL\nFFFFBBFLLL\nBFBFBBFRLL\nBFFFFBBLRL\nFFFBFBFLLR\nFBBBFBFRLR\nBFBBBFBLLR\nBFFBFBFLRL\nBFBFFFBLLL\nFFBFFBBLRL\nBFBBFFFRRR\nFFFBFFBLLL\nBBFFFFBLRR\nFFBBFBFLLR\nFBFBBBBRRR\nFBFBBBFLRL\nBFFBFFFLLR\nFBFBFFFRLL\nFBBBBFBLLL\nFBBBBBBRRR\nFBBBBBFLRL\nBFBBFFFRLL\nFFFBFFFRLR\nFFBFFBBLLR\nFFBFBBBRLL\nBBFFFFFRLL\nFFFBFBBRLR\nFFBFBBBRLR\nBFFFFBFLRR\nBBFFFBBLRL\nFBBBBFFRRL\nFBBFBBBRRR\nBFBFFFFRLL\nFBBBBBBRRL\nBFFBFFBLRL\nBFFBBFBRRL\nFBFFFBBRRR\nBFFBBFBLLL\nBFBFBFBRLL\nBBFFFBBRLL\nFBFBBBBLLL\nBBFFFBBLRR\nBFFFFBFRRL\nBFBBFBFLLL\nFBBBFBBRLR\nBFBBFBBLRL\nBFBBBFBRRL\nFBFFBBBRLL\nFBBBFBBRLL\nFFBBBBBRRL\nFBFBBBFRLL\nBFFFFFBLLR\nBFFBBBBRRR\nBFBBFBBLLL\nFFBFBBBLRR\nFFBFBFBRRR\nBFBFFFBRRL\nFBBBFFBRRR\nBFBFBBBLLL\nBBFFFBBRLR\nBBFFFBBRRL\nFBBBBFBLRL\nFBBBBFFRRR\nFBBBFFBLLR\nBFFFFBFLLL\nBFBFBFFLLL\nBFFFFBFRLL\nBFFBFBBLRL\nFFBFFBBRLL\nFBFBBFFRLR\nBFBFFFBRRR\nFBFFBFBRLR\nFBBFBFFLLL\nFBFFBBFLRR\nFFBFBFBLRR\nFFBBFFBRRR\nFBBFFFBLLR\nFBFFFBFRRR\nBFBFFFFRRL\nFBBBBBFRRR\nFFFFBFBRRR\nFFBBBBBLRL\nFBFBFFFLRR\nFBBBFFFLLL\nBFFFBFFRLR\nFFBFFBFRRL\nFBFBFFFRRR\nFFBFBFBLLL\nBFBBBFFRRL\nBFBFBFBLRL\nBFFFBFFLRL\nFBBBBBFRLR\nFBFBBFBLLL\nFFFBBBBLRL\nFBBFBFFLRR\nBFBBBFBRRR\nFBBBFBBRRR\nFFFBFBFRLL\nFFBBBBFLLL\nFBFFBFFRLR\nFBFBFBFLLL\nFBFFFFBRLL\nFBBBBBBRLL\nFBBBBBFRLL\nFBBFFBFLLR\nBFFBBBFRRL\nFBFFFBFLLL\nFBBBFFFRLL\nBFFFBBBRLR\nFBFBBBFRRR\nBFFBFBBRLL\nFBBFFFBRRL\nFBFFBBBRLR\nBFBFFBBLRL\nBFFBBFFLRL\nFBFBBBFRLR\nBBFFFBBLLR\nFFFBFFBRLL\nBFBFBBFLLL\nBFFBBFBLRL\nFFFBBFBLRL\nBFBBBFFRRR\nFBFBFFFRLR\nBFFBBFFLRR\nBFFBBBFLLL\nFBFFBBBRRR\nFBFBFFFLLL\nFFBFBBFLLL\nFBFFFFFLLR\nBFBBFBFRRL\nBFBBBBBRRR\nBFFBFBBLLL\nBFFFBFBRRR\nFFFBFBFLLL\nFFBFFFBRLL\nFBBFBFBRRR\nFBBBBBBLLL\nFBFBFFFLRL\nFFFBFFBRRL\nFBFBBFFRRR\nFFBBFBFRLR\nBFFFFBBLLL\nFBBFBFFRRL\nBFFBFFFRRR\nBFFBFFFLRL\nBFBBBBFRRR\nBFBFBBBRRL\nFFFBBFBLLL\nFBBFFBBRLL\nBFBBBFFRLL\nBFFFBBFRRR\nFBFBFBBRLL\nBFFFBFFRRL\nFFFBFBBLRL\nFBFFFBBLLR\nBFFFBBFLLL\nBFFBFFFLLL\nBFFBBBBLLL\nBFFFFFFRRR\nFFBFFBFRLL\nFBBFFBFRRL\nBBFFFFFRRL\nFFFBBFBLRR\nBFBFFBFLLL\nFBFBBFBRLR\nBFBBFBBLRR\nFBFFBFFLRL\nBFBFFFBLLR\nFFFBBBFRRR\nFFBBBFFLRR\nFBFBFFBLLL\nBFBFBBFRLR\nFBFBFBFRRR\nFFBBFBBRRR\nFBBFFFFRLR\nFBFBFBBLRL\nFBBFFFFRRR\nFFBBFFBRLR\nBFBFBBFLLR\nBFFBFBFRLR\nBBFFBFFRLL\nBFFBBFFRRR\nBFFFFBFRRR\nFBFBFBFLRL\nFFBFFBFRRR\nFBBFBBBLRL\nFFFBBBFRRL\nFBFFBFFLRR\nFFBFBBFRLR\nBBFFBFBRRL\nFBFFBBFLRL\nFFFFBBFRRL\nBFBFFBFRLL\nBFBBFBBRRL\nFFFBBFBLLR\nFFBFFBFLLL\nFFBBBBBRLR\nBFFFFFBLRR\nBFFBBBFLRL\nBFBBFFBLLR\nFBFFFFFRLR\nFFFFBBFRLR\nFBFFFBFRRL\nBFFBBBBRRL\nFFBFFFFLRL\nBFBBFFFRRL\nFFFBBBBRLL\nBFBFBBBRRR\nBBFFFBFRLR\nBFFBFFBLLR\nFBFBBFBLLR\nBFFBFBBRRR\nBBFFFBFRRR\nBFFBBFFRRL\nFBBBFBFRLL\nFFBFFFBLLL\nBFFBBBFRLL\nFBFFBBFRRR\nFFFBFBFLRR\nBFFBFBBLRR\nFBFFBBFRLR\nFFFBFBFRRL\nFFFBBBBRRL\nBFBFFBBLLR\nBFBFFFFRRR\nFFBBBFFLRL\nFBBFBBFRLR\nFFFBFFFLRL\nBFFBBBBRLL\nBFBBFFFRLR\nFFBFFBBRRR\nFBFBFFBRLL\nBFFFBFFLLL\nFFBFBFBRRL\nFFFBFBBRRL\nBFBBBBFLRL\nBBFFBFFRLR\nBFFFBBFRLL\nBFFFBBBLRR\nBFBFFBFRRR\nFBBBFBFLLR\nBFBBFFFLRL\nFFFBFBFRLR\nFBFBFBFLRR\nBBFFBFBRRR\nBBFFFBBRRR\nFFBBFFBLLL\nBFBBBFBLLL\nFFBBFFFLRL\nFFFBFBBRRR\nFFBBBFBRLR\nFFBBFBBRLR"
-      .split("\n").toSet
-
-    val existingTix = input.map(_
-      .replaceAll("[BR]", "1")
-      .replaceAll("[FL]", "0"))
-      .map(parseInt(_, 2))
-
-    val unassigned = (0 to 2047).toSet &~ existingTix
-
-    println(unassigned.filter(u => (existingTix contains u + 1) && (existingTix contains u - 1)))
-  }
-
-  def _06_1() {
-    val input = "gsvdkufnoawjmhp\nwvhusojpnikgfadb\nvshnpfedgwajkou\naujodhskfvnpgw\nokpdnwhsfvjguqa\n\njuedvq\nvqeduj\nveqdju\neqduvj\n\ncdajbuernxm\nmnucjearxbd\naxrmdejuncb\njrebucmdxna\n\nkutplibdoqzfvhw\nqcewmrkdvhl\nqobdwvlkgxhpasyjn\n\npsbjhnatxe\npnaxesjhbt\ntpxhbnseaj\n\nyskvjqacinfdrphwltxgzomueb\njwprnxzvkeofthuqbyimadgslc\n\nhzwepdg\nlwf\nwy\nlw\nrw\n\nsdgateubqmznhjk\nxyphltd\n\ndbhl\nhdlb\ndbhl\nhbld\n\ninkxywsbgjuvpztrecadqmhf\nqkaxszcujwvmbhdietpgr\nkrupsdgibzecxwvhqatmj\nqtpxerobhisgcumkwjvdaz\n\nbgakcmh\ngxasvzu\n\ncjtrvhinaqsbpk\njvxtspaqbhkfnrc\n\nnab\nbex\nqmkudf\n\nuot\nout\ntuo\nuto\nout\n\nsbtkzf\nhbkfzs\nszbkf\nfuzksbx\njskrzbf\n\ndkphfoybsagcrnq\noybunchapqrdks\nbokpysdhqnarc\n\nbchumknwairvlf\ncvbuhwzirkalfy\n\nbwnseiaoqyjf\njosqanfbew\nnealmqbfwojs\njqnbfaoesw\n\nsowzuji\nwojizu\nwzojiu\njwozui\nujiwoz\n\nkcopqbt\ntpkudorbc\npmtwogjb\n\nuxkfthszqdbelomwri\ntlusezbomwfqrdxhki\nrsdfkbteoqilzmxhuw\nbwkqmdsxhrfulietoz\nizkwomsdlbfxetqhru\n\nkzhwyodlvpt\nkvfhldwpo\nlvhypokdw\nopwjhvdkcl\nkpahtdolvw\n\ncldyjpuitah\nupcthlijyda\nicdauhpytj\nhaypidtucj\njidaptcyhuo\n\nmi\nti\nqjwiz\noudehsxyiv\nbpirjl\n\nstlyornuk\ncxapqwie\n\nkharunjiebzvoy\ngvaihtklnqmcebf\n\nt\nlt\nt\n\nohca\nlznkumqfgyjsbeot\nowxdivpr\n\ndy\nyo\nxsy\nyo\npyrvh\n\niaxp\nafpikux\n\nhpsfqrnevzxl\nfnrsvexzhqup\n\ncp\nfpc\ncp\n\nnbuojxgqckrwzspehf\nmhguoxsfkejwcq\nkedqujsomfxhwcvg\nouqefsxchgjkw\nqcjohwxfseugk\n\nrybnlomq\nfmxsngauo\nembkon\nnlmvob\nbomhnv\n\ngptixlduzjeqkorsamwnh\npgzwohfjkuvbtylmc\n\nthcgo\nvhrlxmo\n\nkbfgmhpzoqysac\nockazbsmhpgqfy\nkmcqfyzhapgsxb\nqcfdlwyarzuskhmbg\n\nsurovlcihp\nirvlpuhosc\n\nrd\nd\ntd\nd\n\nkoasgulh\nkroumtzl\nyubkwolnpj\n\ndqykvasneb\neasknpbovyjd\n\nwqvop\nnwqvpo\nwovfrqizh\nonqpwv\n\nuskotfndjeplixhycbqrmw\nmbxoewypcfdhistn\nydtvpbfemcnsihoxw\nynwbtosmcipxhfde\n\nurgvbqechawznsjltixody\nswdbnevcatrflguhziyoqxj\nsgiwelhanovrbjuzyqtdcx\nvyenojuhxwzbtigsdqarkcl\n\njxcwblszo\nkhrwvnbdolgjufyz\n\nn\nr\nd\nr\nbd\n\ntolrypvdwmb\nlwrmopdvybt\nrdybwmvplto\nvlwpyrbmotd\nybolwtvprmd\n\nradvoplnq\nlrqpondva\nlpanovqrd\nlpanvoqdr\ndqorvalpn\n\nfvxca\nvacfx\nfvcxa\n\njbeixdakgolpfhqyntw\npurbsxvjfwhetlmcnz\n\njztfgbmnkadiehv\ndjhpntcizmkgvbfe\nbtegdfknizvmhjl\nigbdeztjfkhmnv\niqvtgejanzfdbhmk\n\njpi\nijp\nijp\n\nkydp\nviykd\ndgyqmftwx\nvdye\n\nsewpkrvh\nekorswnp\n\nbplerysifmzvwoncjxqk\nnsbeqowitcpjymxvl\n\nnexbskamprd\nxsmdbkaupner\ndsbnekarctvmpx\nnkapxcbsmrlde\nezfdhnakwrsmyxpb\n\nolnuzvxcgpdtfb\ntgzcfdbnxp\n\nvjliksorxpmwecaznu\ngiuoxsvnzewljpmc\n\nfuljpkt\ncplftukj\nlfptjku\nfutpjlk\n\ntqcgihveyswu\nywvqecthgs\ncqhwetvsyg\nqvshwcegty\ngyctwqsfvhe\n\nuhezlyxwvpcidr\nhvucplszbxywer\n\njchyntk\nrevwzsm\nkdn\n\ncotikdlpxauzb\nuxzptoksqbdiw\n\nnqvruwhtombs\ntebfnykzmlgcj\n\nciwkhzbmgfqo\nuwmzgbcoihqkfs\nfkwmziohcgbq\n\nshezgrvq\ngprveokq\n\njvqcdiuyrnpgh\ndyjqvhrpic\nxjiklpfzmthrwdsyqv\npdyvochrunaqjgi\n\ndenpjgywacrk\njbypwarecukdgn\nzacqyxdgejrmpwkfnh\nuwldpkrgaenjcy\n\nq\nmq\nq\n\nd\nc\nd\nd\n\ncdazeitlrpxnfqs\nnrqxelfocsihpz\ncnkveyiujlbzfgrpxq\ncqznxrfliedp\n\nqicgkpjzlxonv\ncijozlhkqxnp\npgkcnloyxzj\njapmxlknszotc\n\nbytg\nmwetjayhgiup\ngyt\nygt\ntyg\n\nzx\nelxz\nz\nvncto\n\nnryw\nrywn\n\ntakqgijzforxpshm\nztmrxfqgsiphjaok\n\nfsaoqe\nwcavsldeytn\nauoekhzsi\nhgseja\n\nwhqyvldpintefaxgurjc\nxluqrgsvnacykptefbdhmj\n\noylknhsgmpawziqte\nznhmatkpolsw\npkvmtznhlwsoa\nonzamhltkpsw\n\nzs\nzs\nzs\nsz\n\ne\na\n\nxmi\ngahfrd\nyt\njmy\nnqy\n\nzkbtfdjom\not\nuoqtr\neot\n\ng\ner\ng\ng\n\nncvwupezd\ncnvdwzpeu\ncuwnzvedp\ncdnzevwpu\n\nbkuwcsv\nbqcsgwve\nmbhcovjsw\nqwrnbeycsv\nlisvbwc\n\nvlys\nomvzl\nptlejv\nxylv\nvxol\n\nmqgohwtz\nrqomlcw\n\nrgopxhkslqcjmzbatnd\nphgzmojlstadqrxkbn\n\npcbziv\npivzbc\njilczbpv\n\ngmrecszl\nzgrecslm\nmelzcrgs\ngezcrlsm\nmcsezrlg\n\nceotvxaydplfiu\nvbzcaxuokwd\n\ngy\ng\n\nlj\nneocbiqjt\n\nfoyawivmhs\nfayowsmzrv\nmksfowvzayg\n\nsivthk\nyh\nlh\nybhq\nhy\n\nfk\nka\nvka\nqhjxrzmckpb\nkf\n\ns\ns\ns\n\niyth\ntyhri\nyrtaih\nhtyi\nutifydhq\n\nhwxd\nxhwd\nhxwdb\n\nkfhsvbjdwxcnt\ndbfjhsctkvw\nsogzdhjqiwmlyuvf\n\nv\nnyx\nv\ni\n\nncimwelvbudtp\npiwvudaer\n\naktpdrhm\ndtapkhr\ndhapktri\ntaprnhdko\nkrpdhta\n\nfvpaxk\nzgmrwja\nar\n\nvxoisla\nolisvxa\nkioqvrlxjhsya\noixgalsv\nsavloix\n\nac\ngr\nc\ntl\nc\n\nfndwqsruzagei\ndwunqfaesgzri\nbzarmfnqwuisgde\n\njr\nrj\njr\njr\n\nzrevhj\nxigdmqatk\nwjlpzvsy\n\nqijznvobp\nnzbojivdpc\nvzipojnb\n\nsrbtecvpixam\nhdvctplmfxuow\n\nbf\nbf\nbf\nbfex\n\ng\ng\ng\ng\nmag\n\nb\ny\nb\n\nvmbwhdfiqteglnpkxucjoa\nwmqdconxbuaetjlgphivkf\nupadfnhkwtjioqcbvxelmg\ngofjubptnmkelihwxacqvd\n\nalev\nevla\nalve\navel\naelv\n\nuceyogrhl\noqcafluvhewgr\nugcxlhorpez\nnduzshrceglox\nodmgbeucrplh\n\nugesfliovyzja\nnlizsujfadgo\nlfguaszotij\nosznifulgadkj\nxadguljioszf\n\nkpgve\nvqsfezungalxidb\ngwteov\ngpvcyjrmeh\n\napuonkxtc\nxithys\nhfvtsx\n\npgtzamyqeu\ngtiuwpcqzvya\ntpqauezyg\n\nsouvmglcqbekzi\nmiuzvksocgqbel\ngzqkmsubevolci\n\nndvahciz\njaoyhepsbirw\n\nbghijfkotcxwu\nuaokxsfbleyhgqr\nhucfokpwbxg\nnbfuztgovhxk\n\nzghianxtlq\nqthdvingxalkz\nwfqnixgteza\n\nlhzcrodupwqsv\nphsvrzcldoqwu\nhogspqwcvukdz\nhwzpucdsqvrol\n\ntxkqmsfbglhnoypaduiv\nnavlozmpthwsbkiqxdfu\n\ngidnojlbtfpskuwavyxczr\nirvsyznbtedxougfwc\nsovxhcqdbwytzmugrinf\n\nmcjnraketsgvqyxpbuwlhzdof\nkqudanprgiejcftzlywhsoxmb\nrcxmkdqfohjzgtsalbpunywe\n\no\ncm\nrc\nm\n\ntmkbedxgzsw\ntdgmsxbzkernw\nktdbxzsemgw\ngwkmestbdxz\nxzdewgtmskb\n\nevpdgzcu\ngepcvzu\nvegzpuc\nvpuzgec\ncvpgeuz\n\nwfsadgtuizejmylxq\naspmuqigbxeljdtwfy\n\nygbtdp\ntye\nfryaitjcukhl\nytdm\n\njfm\nm\nm\n\ncl\ncl\nlc\n\nqbodnftrsygi\niztorbfskqdvng\nonpsgqrdtif\nmnrofcuqwdtils\n\nmhavcgzlrtbo\njrldgavzqbcxoh\n\npshtdbifrvzqmklyae\ntfajgrpkidzyhslebmvq\n\ncuzbjepnvis\nisjbenvyzlcup\nnagieshzcpqmfuvbj\nezvpsjrnciub\n\nywgdsafzuxbrmhvqkljnoipc\nhjcntvrzwdfbxksyiqolpumag\nkwvpybmjeoualxcsqgidhfnrz\n\nzubvkn\ncuwfsnv\n\nisknfdalujeg\nhikagvpzx\n\nszgfhjtxid\nstdgjzxfih\nxghsjfdzit\ngfjixszhtd\nfzjshdtixg\n\nuibgxka\nabigukx\nxubakgi\ngxabiuky\nbkaxgui\n\ntjzor\nslaore\ngxwnoifudmcqvpk\nobhr\ntyo\n\nq\ny\ne\nr\nifdcxzm\n\nfvtqrjsblkhuoiey\ngyfnizlqupvscobt\nbouiftlcvqsy\nqdvlyfiusgotb\nqvbycwsifluto\n\nifmnxbcvpkylruh\nnckrmfphluixvby\nvbcmxyhurnpilfk\nsuiqbvyrxcphmfnlk\nynuxkrfvibclpmh\n\npmvdyexsoilatqfgk\nlpidgatmqevkxyso\ngkpsaoleymqxvtdi\ngmvxeyplqdoasitk\nvkpdylstquegaiomx\n\nymogxuswkce\nvbxznpscg\nxlcgs\nxclbdsng\n\nzovwtiy\nzrjlciov\nksxefuvqdam\n\ncwvsx\nxeuwyhsfb\n\nhxwgolksbrdmpifye\nlwpyefxmiohdnbksgr\ngzoehypbldxikmwtrfs\nsdlwpeymxifrbghok\nglmbhykiofdxpwervs\n\nndcphgqfuvsekaxozri\nxofkqnbschdaveruizgp\n\nptkvdhjaw\nskhtapidwjn\nthunakpjwd\neaglphkdtjcfw\n\npnezvfgiouahmslkdxcrqb\nfpiqsecxaukbvzgnorjhwdm\n\nemcwblgxuvnkzrdsoyjf\nrvsxeugkmwzfncojlb\nnhbiclvskjrupgxzfeomw\ngcxlbjozeiufwnskvrm\neonjkwbzfxgpruvcmsl\n\nntgia\nia\nai\n\nscoernukhwtigpd\nbogiserhpucnkt\n\neaipzg\ngne\nlughe\n\ntbipylmnrague\ntbargijepmhuynl\ntriyzgupnbmale\nprtnumgaibley\npeblaytmgnrui\n\nrqlsepfhiv\nfvelqphsri\nilfhvqerps\nfheqplsvir\nsfpvehirlq\n\njciwxfbhqkadgzloprnuvsty\nqwouetsmhykavxfbjzrglc\n\nfjmz\nmjczf\nunkzjm\n\ngebyncxsphfokwm\ngmchsnbwkpfyxeo\nfybkhcesogxpwmn\n\nud\nwhjn\n\nihm\nhmi\n\np\np\nyp\nsp\n\nytrwx\nyrtxw\n\np\nd\n\nieltygdukac\nyiudeakg\nkpegdaiyu\naudykegi\naeqydwiguk\n\nlb\nei\n\nzjvctfndgiqbeu\nabfwndutvecqigz\nfznbdutevigcq\nqtvzfnueibdcg\n\ng\ni\nx\ng\ng\n\nqhm\nlq\nnq\npqsb\nlvq\n\nqnxpslvajf\nxuavcnst\nnsyabxv\n\nocu\nuco\nacoux\ncluo\ncoun\n\njtgzodebq\nojztqdgbe\noejrqzgtbda\ntjzbqedgo\ngzqtejobd\n\nubglorcsyvmxd\noentwfgulpqikdcbzhj\n\njwkmxdnyvbiz\nqebnoyfz\n\nbepdfi\nikepbof\n\ncofgekndbatpruhqmwvsiy\nbsqtrakeudopghvyfmcnw\nqgusdwvabytmfkcnehorp\nqfhruygwkvmbpaoestcdn\nprmkbuvynseatdoqwfhcg\n\nenczdvbkiyjgruqp\nbdiqvwjurpycngzk\nbnrcejupvqiwgzkyd\npqgybrcjzkudonvi\n\nfygasxjwrb\nijsgafwyrbx\nsyjawbxgfr\nwfbrgvsjyax\n\nxtp\nbfnzuadh\nxlce\n\nrxmeqcbv\n\nfjtdihkrbyn\nhtfjyinkgdrb\nfibgtkdryhnj\nhkbnjrftidy\nintbxjrkfdyh\n\nnhkiqjocfgxmrvwetd\nghnfdjvqtxwermokci\nxnrtfkhdwcmjvoeigq\nmcvfjhwiqtnrekogdx\nimhxgjwrodtcknqvef\n\nsabwqu\nyqamws\n\nusei\nxkvnyfi\n\nrcewysztbiamhpgldj\nkibrfndmgeazcsh\nadrzgchmwjsbie\n\npnrwvx\nlwjxnu\nvwxinb\npfxybwng\n\nwnvxgfdkubzrmip\nkmjhoigltwbz\n\nznlgbx\nfla\ndl\nl\nl\n\njgeqcikplduzxrwby\nkwplxgcbndjqyizuer\newrqujilgyxpdzcbk\n\nguteykxsqrfwhz\nmhiurkswtqbxz\nsdlqxvowujrhkz\n\nzrejn\nnzrje\njnerz\nrzjen\n\nfvkouigderaqbjth\neqrvwjinfltzpcuag\n\nstqcrpy\nctsqpyr\nptyqrsc\nymsqrcaptb\n\negftpqosiv\necjdqlnmwouf\nspxaqofre\nepogyfqi\n\noxwgtshjqef\npbtgjfeqhnoxw\njfhewygoxtq\n\njwavgo\nogevbwdjai\nogvahjw\nwrajvog\nogsjwarv\n\nrpucvg\ncgpr\nhtfelpc\n\nhqrwapscubelovxntd\njisztpnbcahlfdoqmvexgk\n\npmadwutiejznofcgvsbqyxh\nqvmzbaisgfcunwjexhpdktyo\nxtvsdnqaucyehgzbjiopwmf\n\nvjqih\nhimjwdv\nhsojivb\njtrklivehyg\nmvnaizoqpfxhj\n\nj\nj\nj\nj\nj\n\ndpnwmtvklc\nndjpkm\nndmpk\nndmpk\nndmpk\n\nucojsptiqymhewbxvrfdkngl\nebpjwuxrdqkmsochvfgliznt\n\nesroiaqdbxnjhut\nxibuqhmensdajo\nfeukjbahdxwqnsiocv\n\nvkxzclpmtujngerawod\nazdrgcenvjxptkuolmw\nonrluajvcgewpdzkmtx\nxamlobkrvzwpjgnucdte\ntvenjdmlzpwugkoracx\n\nywpfhztklgjmoien\nnyelduixgvwmojpzhtk\nlehpigtmoknjyzw\ngmhwzokylpnteji\ntmypejhwgzlniko\n\nfmd\njf\nfj\n\nl\nv\nntobg\n\ndrmzvt\nrxtsfuk\n\nvxdlpjkrsz\ndgxkprvzsl\ndsklxpvzr\npdvxzlrsk\n\ngwip\ngakpw\n\nkbcadn\nbakcd\n\nmhpstqcgvoalkwxinebryzujd\nqwhirveoczyatgjdsmkpxnl\nmejazlincwodtshxkrpgqyv\nfsaizhnkcvwolgjpdqxtmrye\n\nrxubznovy\nyvudnrbz\nzvunwcyrkb\nntzbryuav\nndvytuzrb\n\nutvklrxypgbnimsdofwa\nfyzliqswjmtpvdka\n\nbyctfzx\npxzf\n\ng\nik\nykei\n\nbcajzwogvetd\ngedbcwtzvoaj\nhcwogaxjbzvted\neoczjagtdwbv\natvdjbcozgew\n\nwsclketugyaz\nqvfbjcxtmilgow\n\namgirw\nyf\n\niletugq\nhnoxbfkaspt\n\nqjdfcpvksmxzywahgn\nwnzceyskgpoxdhlmravjqf\nxjgvhcsqyfankpzmdw\njyncdxwsqkvphfamgz\ndvznygwjkcmqsxahfp\n\nfkuwiezpljbdayhsr\nrphwbeujdzlfiasyk\nerdhpafkzbwjlusiy\njwfhkrzdeabpliusy\nwfipeujhylazbsrkd\n\nfqmbwapvouelydgicj\nuelvqwomtyfjgcpadi\nqkpcsdeyjiwarovnglfum\nijvgqdauoefpcylwbm\nydmqpfewaovlicugj\n\nsgbq\nltqambf\ndjbuvn\nzmtbgy\n\negvifnsaztxqcrywumj\naqcxujnzvfwetrhgm\nfxdmtkgawplobqjenv\n\nbxuwpnl\nztdkau\n\necylsbdv\ncvzlsbeyd\nyqcveblsd\nbeyvlqcsd\nebsvdycl\n\neb\nxpnyter\n\ncbhqyzresnpxfwm\ncwksfvhxqperbimz\nwhqzpbcsjxafre\n\nbeodiasvj\njbisdoave\niedaovbsj\njxovacsiebd\n\nprxsnyehgwifbzo\nozpbsxynfiwrge\nribozypewsxngf\nnyoxspzbfihgewr\nrxzewpbnimfoysg\n\nshviadjcue\nweduziolmcbngp\nvqdieycu\ncudeif\ndicyuqke\n\nnvcwhuo\nsltchuwnq\nunwhci\n\nekrisbvx\nrsxiev\nexnrvsi\nsrxevi\nxhsvnrei\n\nrs\nrs\nsr\n\nfuzjmek\ntihf\n\nptlzxreqfdchibogvysmku\npceyizvdkqxbhrtulsfmog\ngvkrxfsbipqedlyzomhtncu\nsuyzgrobilxhmcfdtvkeqp\n\nokfvjneqyicuthwsmpxl\nwdojlsqknyhmptvicfe\nnyfjthmpivqgsulwkboec\n\nufrpkliqjgwxeybmcztasovh\nqvkpyghxblewsaitrmzuonjc\ngiptzywrejquxkoavslcmhb\npeaywubjozmcgrihlqsvxktf\nlehjpzguiksacobxywtmvrq\n\negryz\nergyz\nyzegr\nyfcgzre\n\nj\nj\nj\np\nj\n\nbjt\ntbj\ntjb\ntbj\njbt\n\ncklp\nlpck\nlckp\n\nfea\neaf\n\nqgm\ndz\n\nyri\nyi\nzypdi\n\nekwjxhtmgiudpz\nxepuhgwjizkdrmt\nezghmliwtxujdpk\nwdhxpjizugkemtl\n\ngyokbspj\nkbgsopyj\npykhdsjbgio\nopjykusbg\njygbsopakc\n\nixorlekwgnbcjm\nlrjkwxgie\nwgeikxjrlh\nrlkjxipeqgw\n\necgmdrykq\nmyrclqdgek\nguqmzyceprkd\ndygckwerqm\nyrcgkiqempd\n\nu\nu\n\nbcwnvl\nfcvsib\nhcnvb\ncpgveoaby\nivbcn\n\nn\nn\nn\n\nmacuqyp\nfs\n\npbout\nsuota\ncuotm\natmuo\n\nnlravibecfj\neilcjafn\naeibjfl\njxfeilauo\njafesdnli\n\nxfwroij\nfxwijor\nirxjofw\nrfjwboxi\n\ndj\ndvj\ndj\ndj\n\nmikhxcqensrodu\nimcvkhewgljupsxonq\nkceuqxltinhsom\nkzqanoemcxhius\nvmuoxineqpzhacbks\n\nhklcsoaywpfrdntmu\nxirbzckwjvpfygeno\n\no\nto\no\n\nzvirefunh\nezhirnvfu\nrunifzhev\nezivhnrfu\n\nzfsoducbpai\ncupdisobazf\naozbsucfdiypv\nbcfazodiups\n\nnjfprukxamlg\nkpfrjgynulxma\npijflgnuxkram\npkuxalrmjfgn\nnkvrlmxustpfjga\n\ndszgm\nrsmu\n\niaw\naiw\nwai\naiw\niwa\n\ns\ns\ns\nsl\ns\n\neyavhnk\nkeov\nqeiwfkcgvrpsld\nkuvze\n\na\nmh\n\nilqchdfsog\nigscolmdzq\n\nfnyhxmpct\nlchdpamfu\n\nvijz\nvj\n\ndhpxjrmyuefclvqgni\nuicskweohgvfryxmdq\ntqgdifymexbwckruvh\n\nc\nc\nc\n\nlthidumvycszw\ncvpqjeuhlwrfyzg\n\nkeocuw\nuckowe\noekwcu\ncwukoe\n\nc\nc\n\nlkxjv\njkvxl\nvxklj\njkvxlwg\nzkxvjly\n\nsqlamxeuhpgwjbdcioy\nbumyjxsiqweholpgcda\ngecswjlqhadiuboxmpy\nsyahibgopexwlucdqjm\n\nigebmv\ntawqvezdsmpunk\nrimocvgje\nevmbhc\n\nykjisvfatux\njfaisyvktux\nitksajfxuvy\n\nfclkush\nslcykdbgu\nfskuchl\n\nxjhulywagzeioq\noejlfsquhbriayxwg\njiexgylwzhuqao\nwyxhqljagieou\nhwladixguqejyo\n\nsofcbhuwyqtid\nicofdwsb\ndfwovimbcs\ncwibodnsxf\n\njqnestpyaczfmxhwrbuvldi\nmxanrhbfcujvyzditwlo\n\nntfgivkyszmarc\nqrajdhowuvpclbex\n\nkliwy\nukwlryi\nilkyw\n\npmgltzsk\nljmipztvgow\nzltmgup\numztglp\nftzmlpgk\n\ntwepi\nteq\n\nohyd\nmytjrwk\nhfy\nexyh\nye\n\nrzsqhtlowavy\ntwvhzyrjsqaol\n\nyzagcqrjeimwxokh\nfmxtiagboleqcwdrhzk\nopvemhasrguxkcqzwi\n\natcpeyosnrjbg\nkcesobgua\nfagsev\n\nca\nac\n\nxahmtil\nhximtla\n\nhyutnjdrbsmofl\ndswjfyothmblrnq\n\net\nte\neqt\ntuje\n\nnbhpwzgtrcx\ngqcxobhriznw\nhwnbugrejxtczs\ncujrbxfhnwzg\ngcwzhrxndjb\n\nqwrfxhlyzcptvmkdeauigob\nzbchavlkmryxiptegwfouqd\ncfhuiptgqkoalrewzmvbyxds\nhudmwcgbfepyitovkaqzxrl\n\neysuozkm\nndgloqyv\nignhwrovy\n\niqufgnvxspdcjtylkbrozh\ninuzqktrvcxyhbgfosljdp\n\nqhfknsyeuirzagvcwd\nyfcgqnxewlkrsuazhi\n\nlj\njl\nlj\nlyj\njl\n\nkpholfxewrdt\nopdfhtlmkwex\nhpwftoxderlk\nylxhwdofnztkpe\nrwfpxlhtdkoe\n\nubml\nyvohg\nzeasf\nplum\nqhnjxg\n\ncjgzwmdyuakvixlprhfe\nkmplfuxywergchzjvani\nxcvlzmewyujgirhpfka\nfwhgrxkjcvpeuilzyam\n\neclsiqp\nivgspoe\nscepi\npmiadtskyze\n\nakmdfivwgct\nyvgcwakhdimts\nlcvgirakwudtp\naftnckwmidgv\n\nzuqblv\nblvzqu\nuvbqlz\nqbuzvl\nbzqxvlmu\n\nby\nbetlcdvry\nyb\nbxy\ngybjx\n\nwaoqdxpkmfnyrhjebcuvzs\nqprvfczedowxknhausmyjb\ncufdqvknwrajzybospmxeh\nqkjzybeadvncwhxumpfsro\nnwfkjxumpabzodvqychsre\n\nshdfctjrkwbvemoxluyzipnqag\nsgbwpeuqcilxojfynmrdhvktaz\nlhgkxfwuzbevojmrpnqydsacti\ngflnzjasbuyimxqcvwrdpohetk\nynwedhbskjtvoucgapiqlzfxrm\n\nwlmjxp\nylnxoajk\nluxjidte\nxjrvlcdeu\n\npv\nvbpw\nwvuap\nkceavp\nrdphvoitqf\n\ndlxzwkhtmc\ntwzhlmkdc\ntcylhkdwzm\nzwmlkdcth\ngdhikmctlwsz\n\nhpekwtlufyragqvb\nkbrwatlupqyfvh\nucbvwlhfqatpkry\nhlakvyuqrfpwbt\nrubvwpfqlatkhy\n\nnpxobl\nckuhnplvx\n\nedkpjycofasulrthw\nrpuhcdyxlftosewa\n\nstuvhwaexykzdorbfpi\nshjuwgpfqmtacvlbirkzd\n\nfrph\nwo\nh\nzumtyngcj\npfse\n\nzivsgurp\nszirpvg\ngikvsprz\nzigpvsrk\nnvpsgzir\n\nrfimu\nhmqrlfxpbgu\nnrmoyftu\nnmujfr\nrfum\n\nbidzwmearovjkgcl\nogdbirkzqmclejvaw\ngealcqozmijkbrvdw\ngzakwmdbejclrivo\neolwifvgmrbjdazkc\n\niurfyqgwk\nymkvqfpwrig\naqrfywikg\nbfdryxikgnwoq\ncjfgkryiuwq\n\njvb\nx\n\nqkxnvfuopdytrzjgbm\nbmqkuyarjgtdpfzsnox\nobwjxyeqtingzmdrfkplu\ngamupstryfokbzqdxnj\n\nymanp\nnpfshjya\npundyklatc\nypfanh\n\ngarp\ngrpa\narpg\nargp\n\ninedmtpwrvub\nvwibertnudmp\nidurbepwtmnv\n\nyczoaplwmxibqgsfn\nwbiuxyqzpncvgmsl\ngpoltrbisqfznwxcymj\n\na\na\na\na\n\ncgjviyufped\npcuvygzkjid\nptvigyjdcku\n\nu\ni\ntqa\nj\nul\n\nzitbfuer\nuzfithernb\nuftbzire\nuitefbzr\n\njcwkroamnegxvqsh\nkejhanqmvcrxwsog\ngaqmkexjvcsnrowh\n\ntaegslmuyqcipz\nasytxrgdpfbcqliwz\ntpgoschavqjlkyzi\n\noc\noc\noc\nco\n\nmrfvjlykquz\nfzmvuleykqt\nvykmzlgquf\nkqfzuvygetml\n\na\ns\n\nionve\neailo\njiomexwh\nbeziqpdco\nihtoe\n\ncfjvrqosnwyedaubmhgtlxzk\nolxzjfwrcytkbgudasvqhnme\ntuwfnckrzsvjgbxdmyhqlaeo\ntxnydgufbmjaqkrzlwhoevcs\nbgakxqmfsdwyjvelntchzour\n\nnjvkobzs\nbsjvkwn\n\nqmnluxw\novgtikpdbrya\nnchfsjqez\n\nkhtdryej\nhdrtv\nqnaodtm\n\nas\naus\n\nomjaixvgklq\nvgkaxjihdrlq\nycxkzatpusiljqgv\n\njtepylmrqbiwk\nxerlpmynqtkfbjwi\ntimwpbkegljqry\nlqijkebmgpwyrt\nqrkljpbytemiw\n\nq\nxqint\ndkqb\n\ntrjmdxpoeqcwf\ntfedxqcprwom\nkmftwxqporncde\n\nfpcmndhr\ndzepjngqikcfrh\nnhfpcrd\n\nc\nc\nc\ny\nc\n\nt\nt\nti\n\ntgfspxdy\nsytgpfxd\n\nzov\nvoz\novz\n\nic\nieuc\n\nvhesndkizjgwcpobulam\nszjymxwopckfqtgdib\n\njdfzgqsbrmnticl\nfbcdgqtznmjri\ndoitrczjqfbgmn\notzbjfidmnrqcg\n\ncb\nbc\ncb\ncb\n\naspjvghyurx\nvcarwsbh\nlaeshvzrb\nhsrvia\n\nkpxzdhasfmltouvrc\nduvpkascromz\nvmcskupzorad\n\npguayqbwm\ngpyaqbwum\nywaqgpmbu\n\nypdlnjse\nelskbyz\njesylo\nlyse\n\nzlokqmjiurpwgbn\nxzvurhgmlkwqnibjpo\n\nwpzenbojcykqxvt\nwnyzlbjkeoqtvcpx\nitkjzydbpvoxcewnq\n\ngfuvenodsizwmh\nhwfuvsntdg\n\ngkibnpalc\ninkcgpal\ngnpzliakfc\n\nibektdpuhg\niwbgjke\niycnergfqkbo\negxadvlmbijk\n\nl\nyjg\nl\n\nymuoc\noymcu\noymcu\nyoucm\nycomu\n\nwifoqy\nzksfrict\nbdpvhuilgejn\nacoi\nmqi\n\nqrcukhwvzdjilpto\nhqrmwkupclvjtizo\n\ngdizlfawobuk\ndihqfswcgukzrbla\ngmukawbdfizlo\n\nzwdrpv\ndzvpwr\n\nycidkpfrqjelha\npjahykfecdqrli\nkfqlajhdepyirc\npkhaceijlqdfry\n\npwuxbqotey\nokbphqutfx\nxvsgodubtlpqijr\npuoewaxtbmq\n\nqtbaoex\nxpoqa\nagiweqtop\nnozuadmjrq\n\nhuvqdaij\nfkuvaijdhq\n\nbiewrdlzc\nrjlcikbzmdwve\nleucwfpoizndbxr\n\navgtnubqcpxrhwdoml\nvrkqgyndcmohtupswa\n\ndmjqfesnxyciz\nhfzudcjxsmno\nmfjhrcsxznd\nlcmbjnxzfadrs\n\ntvayrhpqskd\nqrtkvhyus\nepjsytzhvrqk\nyzqohmvtgsrk\nhvlysqtwirbk\n\nieyd\naipeyg\n\nr\nrg\nrm\nr\n\nqcweajhfmlgyzpv\nouqdxmsphkbatenr\n\nfba\nkfba\nbaf\n\nvxpblsyjnrmkai\nvpyhagcwqbzxmik\nrpbytexdijnfavkom\n\nfpmytnk\ncyaw\n\njhyimdg\nmhyij\nhmvyij\nmihezcovyj\nyhmciej\n\nsqfaopkgze\nkxsecfphozq\nqsotezpfk\n\newcvoaurmjhq\ncuqvjwreaohm\newvjmqacuorh\nwauhojecqmvrs\n\nlyieuqdjcb\nvkpctdmywzsreubn\ndelchbujy\njbcyegdou\nxduayceb\n\nnidxswarcqtk\nkciqdarnbewsxt\nacqrdkntipxlsw\nikpawtcdsqnrx\n\ngikrfhwltveap\nyeplvatowncgxirf\ngriadtewmvblpf\n\nhtjkmzxlvbygcwqioe\njzwtymxvigbheolckq\ncxmvjyklobhtwgqzie\nteczbxvlmgojqiwkyh\noxvejlbtiqywhzgmkc\n\njkfapozdwcie\nbmushwjgltrn\n\nrpnbxkoij\nexjdri\nuirjegxsa\n\nzawoncbymgsrhv\ndvctxgeusr\ncvizgyhlmrwsfp\n\nshnyujedctwfq\nyscahxndftreq\n\nxrnysi\nitcluvmd\nniqwozjber\n\ngjobshlayxvctnfq\nljcxaonvbstqfgyh\nybhxgntqcsvfaojl\nrjoxblqavncusgheytf\nsncqlvybxfjtghao\n\nuozb\nldewrck\nxbg\nxhg\nz\n\nyubqaznf\nimwhrgnkjls\n\nmsvgjrnkcloaup\ncsxjnkgrpiuoveamy\npknvrajmcudswgzo\ntnupacsxrgkvmoji\nbpnkfarqumsvhcojg\n\nbzvyuenaihfdcgwk\npuchyenmzgksrijoa\nugebciktnlzhay\n\ncedfutpnsivjlobrkaxy\ndenscvpbmlkxaojrfh\n\nvugjrq\njqgtuwcnrv\ngrvkuqj\ngbmvqsyxjueplzirhf\navudnjqgr\n\nvpidajtmgozxwuksh\nughiwmakpjszdtvox\nwugipxzdvastkhmj\nsxughjakmnwfpdivzt\nwimvaupjhgxstkzd\n\nmqsuwhdxpcozfr\nrsdwhfpakzxjmo\neysprldfqzowmhx\nfodhrtmswxyzp\n\nvsmzfbjhwqxiugtlc\ngcqdsfmpletnhbxr\n\nuxbgkfjwesmlchr\nfeakldxqnpwcjgvrzbhs\nhjosrixfmlcgbkew\n\ndszv\nzsvd\nvsdz\nzvds\nsdzv\n\npdw\nlxudjpiwz\nadvnkpqw\nhswpdog\n\ngslpira\nneuzf\n\nifjt\ntjif\njfit\njitf\nifktj\n\njrnzb\nunkjshpe\nnjza\n\nydigxaqm\niaeokqgmyhr\n\nbkxmngtdlyev\nnmbuzoveixlprkys\n\nzjloghmtu\nermvzn\nqzdncfm\nafczxmp\n\naptscynieqmfgjhbr\nbmsirnafyepthgcqj\n\nq\nqlr\nfq\nqk\n\nv\nv\n\nucrmbsozjkfy\nbuscrmofkzy\nuosmfkryzbc\nsqirbfzmocuynk\nmbucykofrzs\n\nlrp\nplz\nlrp\nlp\nlp\n\nyfzmpgcsw\nprycgfm\nrfcmgpy\nscgmyqpf\nflcyvpohmgu\n\ng\ng\ng\ne\n\nlxoaibjch\nwudyrfvkzn\ngtpmkeyqd\n\nzjhlwbtfuvick\nwujtvblzchkif\n\nksdfpvqbtxaluneo\ndqofprhnxamskujeb\n\nhacwigyftqblx\nxwlqahitbgcyf\nahyxlwgtqifbc\n\nmb\nm\nmg\nm\n\necrasdtyljuqimhnwfb\nqleiaydjcugrwn\njdyewaqnrcliu\n\ng\nlqfio\nadhstybz\n\ngryzuxenhlvopbtmfwqcas\nyeqzalmogbncxpsfujwh\nmflqzhoaupcseygnbwx\nmyswzuonfqlaxecbhpgj\n\nuwvzati\nuzcjvtai\nsavuizt\natiuzv\n\ncozuin\nabuizo\ntiozua\nwilfzeoqu\nroictxuz\n\ngndpkojqu\ngunkpsjoa\nrtmzjpgkovwxlnuc\n\nlrj\nrjl\nmrjl\n\nudlpevjgzrfawitqs\nqgrtiujslfeovza\nsurivgtfqjleaz\nezavgjtlfrusiq\n\ng\nfx\nmf\ntx\niyqd\n\nsljgmdztynvrqexhwfi\ngvwdyerlqnxthmjifsz\nqgmjlfwrxztynivdhse\n\nzkvuir\nkvruiz\nvriukz\nkriumzv\nzruikv\n\nzyqgnvda\nkdygvqanz\nzyagvqdnu\nqnwdmlaxgzyfvo\ndnzgyavq\n\no\ns\nq\n\ncbroy\nqiyocmjdt\nuyoc\n\nlkqj\nkjql\njqlk\nljqk"
-      .split("\n\n")
-
-    println(input.map(_.toSet.count(_ != '\n')).sum)
-  }
-
-  def _06_2() {
-    val input = "gsvdkufnoawjmhp\nwvhusojpnikgfadb\nvshnpfedgwajkou\naujodhskfvnpgw\nokpdnwhsfvjguqa\n\njuedvq\nvqeduj\nveqdju\neqduvj\n\ncdajbuernxm\nmnucjearxbd\naxrmdejuncb\njrebucmdxna\n\nkutplibdoqzfvhw\nqcewmrkdvhl\nqobdwvlkgxhpasyjn\n\npsbjhnatxe\npnaxesjhbt\ntpxhbnseaj\n\nyskvjqacinfdrphwltxgzomueb\njwprnxzvkeofthuqbyimadgslc\n\nhzwepdg\nlwf\nwy\nlw\nrw\n\nsdgateubqmznhjk\nxyphltd\n\ndbhl\nhdlb\ndbhl\nhbld\n\ninkxywsbgjuvpztrecadqmhf\nqkaxszcujwvmbhdietpgr\nkrupsdgibzecxwvhqatmj\nqtpxerobhisgcumkwjvdaz\n\nbgakcmh\ngxasvzu\n\ncjtrvhinaqsbpk\njvxtspaqbhkfnrc\n\nnab\nbex\nqmkudf\n\nuot\nout\ntuo\nuto\nout\n\nsbtkzf\nhbkfzs\nszbkf\nfuzksbx\njskrzbf\n\ndkphfoybsagcrnq\noybunchapqrdks\nbokpysdhqnarc\n\nbchumknwairvlf\ncvbuhwzirkalfy\n\nbwnseiaoqyjf\njosqanfbew\nnealmqbfwojs\njqnbfaoesw\n\nsowzuji\nwojizu\nwzojiu\njwozui\nujiwoz\n\nkcopqbt\ntpkudorbc\npmtwogjb\n\nuxkfthszqdbelomwri\ntlusezbomwfqrdxhki\nrsdfkbteoqilzmxhuw\nbwkqmdsxhrfulietoz\nizkwomsdlbfxetqhru\n\nkzhwyodlvpt\nkvfhldwpo\nlvhypokdw\nopwjhvdkcl\nkpahtdolvw\n\ncldyjpuitah\nupcthlijyda\nicdauhpytj\nhaypidtucj\njidaptcyhuo\n\nmi\nti\nqjwiz\noudehsxyiv\nbpirjl\n\nstlyornuk\ncxapqwie\n\nkharunjiebzvoy\ngvaihtklnqmcebf\n\nt\nlt\nt\n\nohca\nlznkumqfgyjsbeot\nowxdivpr\n\ndy\nyo\nxsy\nyo\npyrvh\n\niaxp\nafpikux\n\nhpsfqrnevzxl\nfnrsvexzhqup\n\ncp\nfpc\ncp\n\nnbuojxgqckrwzspehf\nmhguoxsfkejwcq\nkedqujsomfxhwcvg\nouqefsxchgjkw\nqcjohwxfseugk\n\nrybnlomq\nfmxsngauo\nembkon\nnlmvob\nbomhnv\n\ngptixlduzjeqkorsamwnh\npgzwohfjkuvbtylmc\n\nthcgo\nvhrlxmo\n\nkbfgmhpzoqysac\nockazbsmhpgqfy\nkmcqfyzhapgsxb\nqcfdlwyarzuskhmbg\n\nsurovlcihp\nirvlpuhosc\n\nrd\nd\ntd\nd\n\nkoasgulh\nkroumtzl\nyubkwolnpj\n\ndqykvasneb\neasknpbovyjd\n\nwqvop\nnwqvpo\nwovfrqizh\nonqpwv\n\nuskotfndjeplixhycbqrmw\nmbxoewypcfdhistn\nydtvpbfemcnsihoxw\nynwbtosmcipxhfde\n\nurgvbqechawznsjltixody\nswdbnevcatrflguhziyoqxj\nsgiwelhanovrbjuzyqtdcx\nvyenojuhxwzbtigsdqarkcl\n\njxcwblszo\nkhrwvnbdolgjufyz\n\nn\nr\nd\nr\nbd\n\ntolrypvdwmb\nlwrmopdvybt\nrdybwmvplto\nvlwpyrbmotd\nybolwtvprmd\n\nradvoplnq\nlrqpondva\nlpanovqrd\nlpanvoqdr\ndqorvalpn\n\nfvxca\nvacfx\nfvcxa\n\njbeixdakgolpfhqyntw\npurbsxvjfwhetlmcnz\n\njztfgbmnkadiehv\ndjhpntcizmkgvbfe\nbtegdfknizvmhjl\nigbdeztjfkhmnv\niqvtgejanzfdbhmk\n\njpi\nijp\nijp\n\nkydp\nviykd\ndgyqmftwx\nvdye\n\nsewpkrvh\nekorswnp\n\nbplerysifmzvwoncjxqk\nnsbeqowitcpjymxvl\n\nnexbskamprd\nxsmdbkaupner\ndsbnekarctvmpx\nnkapxcbsmrlde\nezfdhnakwrsmyxpb\n\nolnuzvxcgpdtfb\ntgzcfdbnxp\n\nvjliksorxpmwecaznu\ngiuoxsvnzewljpmc\n\nfuljpkt\ncplftukj\nlfptjku\nfutpjlk\n\ntqcgihveyswu\nywvqecthgs\ncqhwetvsyg\nqvshwcegty\ngyctwqsfvhe\n\nuhezlyxwvpcidr\nhvucplszbxywer\n\njchyntk\nrevwzsm\nkdn\n\ncotikdlpxauzb\nuxzptoksqbdiw\n\nnqvruwhtombs\ntebfnykzmlgcj\n\nciwkhzbmgfqo\nuwmzgbcoihqkfs\nfkwmziohcgbq\n\nshezgrvq\ngprveokq\n\njvqcdiuyrnpgh\ndyjqvhrpic\nxjiklpfzmthrwdsyqv\npdyvochrunaqjgi\n\ndenpjgywacrk\njbypwarecukdgn\nzacqyxdgejrmpwkfnh\nuwldpkrgaenjcy\n\nq\nmq\nq\n\nd\nc\nd\nd\n\ncdazeitlrpxnfqs\nnrqxelfocsihpz\ncnkveyiujlbzfgrpxq\ncqznxrfliedp\n\nqicgkpjzlxonv\ncijozlhkqxnp\npgkcnloyxzj\njapmxlknszotc\n\nbytg\nmwetjayhgiup\ngyt\nygt\ntyg\n\nzx\nelxz\nz\nvncto\n\nnryw\nrywn\n\ntakqgijzforxpshm\nztmrxfqgsiphjaok\n\nfsaoqe\nwcavsldeytn\nauoekhzsi\nhgseja\n\nwhqyvldpintefaxgurjc\nxluqrgsvnacykptefbdhmj\n\noylknhsgmpawziqte\nznhmatkpolsw\npkvmtznhlwsoa\nonzamhltkpsw\n\nzs\nzs\nzs\nsz\n\ne\na\n\nxmi\ngahfrd\nyt\njmy\nnqy\n\nzkbtfdjom\not\nuoqtr\neot\n\ng\ner\ng\ng\n\nncvwupezd\ncnvdwzpeu\ncuwnzvedp\ncdnzevwpu\n\nbkuwcsv\nbqcsgwve\nmbhcovjsw\nqwrnbeycsv\nlisvbwc\n\nvlys\nomvzl\nptlejv\nxylv\nvxol\n\nmqgohwtz\nrqomlcw\n\nrgopxhkslqcjmzbatnd\nphgzmojlstadqrxkbn\n\npcbziv\npivzbc\njilczbpv\n\ngmrecszl\nzgrecslm\nmelzcrgs\ngezcrlsm\nmcsezrlg\n\nceotvxaydplfiu\nvbzcaxuokwd\n\ngy\ng\n\nlj\nneocbiqjt\n\nfoyawivmhs\nfayowsmzrv\nmksfowvzayg\n\nsivthk\nyh\nlh\nybhq\nhy\n\nfk\nka\nvka\nqhjxrzmckpb\nkf\n\ns\ns\ns\n\niyth\ntyhri\nyrtaih\nhtyi\nutifydhq\n\nhwxd\nxhwd\nhxwdb\n\nkfhsvbjdwxcnt\ndbfjhsctkvw\nsogzdhjqiwmlyuvf\n\nv\nnyx\nv\ni\n\nncimwelvbudtp\npiwvudaer\n\naktpdrhm\ndtapkhr\ndhapktri\ntaprnhdko\nkrpdhta\n\nfvpaxk\nzgmrwja\nar\n\nvxoisla\nolisvxa\nkioqvrlxjhsya\noixgalsv\nsavloix\n\nac\ngr\nc\ntl\nc\n\nfndwqsruzagei\ndwunqfaesgzri\nbzarmfnqwuisgde\n\njr\nrj\njr\njr\n\nzrevhj\nxigdmqatk\nwjlpzvsy\n\nqijznvobp\nnzbojivdpc\nvzipojnb\n\nsrbtecvpixam\nhdvctplmfxuow\n\nbf\nbf\nbf\nbfex\n\ng\ng\ng\ng\nmag\n\nb\ny\nb\n\nvmbwhdfiqteglnpkxucjoa\nwmqdconxbuaetjlgphivkf\nupadfnhkwtjioqcbvxelmg\ngofjubptnmkelihwxacqvd\n\nalev\nevla\nalve\navel\naelv\n\nuceyogrhl\noqcafluvhewgr\nugcxlhorpez\nnduzshrceglox\nodmgbeucrplh\n\nugesfliovyzja\nnlizsujfadgo\nlfguaszotij\nosznifulgadkj\nxadguljioszf\n\nkpgve\nvqsfezungalxidb\ngwteov\ngpvcyjrmeh\n\napuonkxtc\nxithys\nhfvtsx\n\npgtzamyqeu\ngtiuwpcqzvya\ntpqauezyg\n\nsouvmglcqbekzi\nmiuzvksocgqbel\ngzqkmsubevolci\n\nndvahciz\njaoyhepsbirw\n\nbghijfkotcxwu\nuaokxsfbleyhgqr\nhucfokpwbxg\nnbfuztgovhxk\n\nzghianxtlq\nqthdvingxalkz\nwfqnixgteza\n\nlhzcrodupwqsv\nphsvrzcldoqwu\nhogspqwcvukdz\nhwzpucdsqvrol\n\ntxkqmsfbglhnoypaduiv\nnavlozmpthwsbkiqxdfu\n\ngidnojlbtfpskuwavyxczr\nirvsyznbtedxougfwc\nsovxhcqdbwytzmugrinf\n\nmcjnraketsgvqyxpbuwlhzdof\nkqudanprgiejcftzlywhsoxmb\nrcxmkdqfohjzgtsalbpunywe\n\no\ncm\nrc\nm\n\ntmkbedxgzsw\ntdgmsxbzkernw\nktdbxzsemgw\ngwkmestbdxz\nxzdewgtmskb\n\nevpdgzcu\ngepcvzu\nvegzpuc\nvpuzgec\ncvpgeuz\n\nwfsadgtuizejmylxq\naspmuqigbxeljdtwfy\n\nygbtdp\ntye\nfryaitjcukhl\nytdm\n\njfm\nm\nm\n\ncl\ncl\nlc\n\nqbodnftrsygi\niztorbfskqdvng\nonpsgqrdtif\nmnrofcuqwdtils\n\nmhavcgzlrtbo\njrldgavzqbcxoh\n\npshtdbifrvzqmklyae\ntfajgrpkidzyhslebmvq\n\ncuzbjepnvis\nisjbenvyzlcup\nnagieshzcpqmfuvbj\nezvpsjrnciub\n\nywgdsafzuxbrmhvqkljnoipc\nhjcntvrzwdfbxksyiqolpumag\nkwvpybmjeoualxcsqgidhfnrz\n\nzubvkn\ncuwfsnv\n\nisknfdalujeg\nhikagvpzx\n\nszgfhjtxid\nstdgjzxfih\nxghsjfdzit\ngfjixszhtd\nfzjshdtixg\n\nuibgxka\nabigukx\nxubakgi\ngxabiuky\nbkaxgui\n\ntjzor\nslaore\ngxwnoifudmcqvpk\nobhr\ntyo\n\nq\ny\ne\nr\nifdcxzm\n\nfvtqrjsblkhuoiey\ngyfnizlqupvscobt\nbouiftlcvqsy\nqdvlyfiusgotb\nqvbycwsifluto\n\nifmnxbcvpkylruh\nnckrmfphluixvby\nvbcmxyhurnpilfk\nsuiqbvyrxcphmfnlk\nynuxkrfvibclpmh\n\npmvdyexsoilatqfgk\nlpidgatmqevkxyso\ngkpsaoleymqxvtdi\ngmvxeyplqdoasitk\nvkpdylstquegaiomx\n\nymogxuswkce\nvbxznpscg\nxlcgs\nxclbdsng\n\nzovwtiy\nzrjlciov\nksxefuvqdam\n\ncwvsx\nxeuwyhsfb\n\nhxwgolksbrdmpifye\nlwpyefxmiohdnbksgr\ngzoehypbldxikmwtrfs\nsdlwpeymxifrbghok\nglmbhykiofdxpwervs\n\nndcphgqfuvsekaxozri\nxofkqnbschdaveruizgp\n\nptkvdhjaw\nskhtapidwjn\nthunakpjwd\neaglphkdtjcfw\n\npnezvfgiouahmslkdxcrqb\nfpiqsecxaukbvzgnorjhwdm\n\nemcwblgxuvnkzrdsoyjf\nrvsxeugkmwzfncojlb\nnhbiclvskjrupgxzfeomw\ngcxlbjozeiufwnskvrm\neonjkwbzfxgpruvcmsl\n\nntgia\nia\nai\n\nscoernukhwtigpd\nbogiserhpucnkt\n\neaipzg\ngne\nlughe\n\ntbipylmnrague\ntbargijepmhuynl\ntriyzgupnbmale\nprtnumgaibley\npeblaytmgnrui\n\nrqlsepfhiv\nfvelqphsri\nilfhvqerps\nfheqplsvir\nsfpvehirlq\n\njciwxfbhqkadgzloprnuvsty\nqwouetsmhykavxfbjzrglc\n\nfjmz\nmjczf\nunkzjm\n\ngebyncxsphfokwm\ngmchsnbwkpfyxeo\nfybkhcesogxpwmn\n\nud\nwhjn\n\nihm\nhmi\n\np\np\nyp\nsp\n\nytrwx\nyrtxw\n\np\nd\n\nieltygdukac\nyiudeakg\nkpegdaiyu\naudykegi\naeqydwiguk\n\nlb\nei\n\nzjvctfndgiqbeu\nabfwndutvecqigz\nfznbdutevigcq\nqtvzfnueibdcg\n\ng\ni\nx\ng\ng\n\nqhm\nlq\nnq\npqsb\nlvq\n\nqnxpslvajf\nxuavcnst\nnsyabxv\n\nocu\nuco\nacoux\ncluo\ncoun\n\njtgzodebq\nojztqdgbe\noejrqzgtbda\ntjzbqedgo\ngzqtejobd\n\nubglorcsyvmxd\noentwfgulpqikdcbzhj\n\njwkmxdnyvbiz\nqebnoyfz\n\nbepdfi\nikepbof\n\ncofgekndbatpruhqmwvsiy\nbsqtrakeudopghvyfmcnw\nqgusdwvabytmfkcnehorp\nqfhruygwkvmbpaoestcdn\nprmkbuvynseatdoqwfhcg\n\nenczdvbkiyjgruqp\nbdiqvwjurpycngzk\nbnrcejupvqiwgzkyd\npqgybrcjzkudonvi\n\nfygasxjwrb\nijsgafwyrbx\nsyjawbxgfr\nwfbrgvsjyax\n\nxtp\nbfnzuadh\nxlce\n\nrxmeqcbv\n\nfjtdihkrbyn\nhtfjyinkgdrb\nfibgtkdryhnj\nhkbnjrftidy\nintbxjrkfdyh\n\nnhkiqjocfgxmrvwetd\nghnfdjvqtxwermokci\nxnrtfkhdwcmjvoeigq\nmcvfjhwiqtnrekogdx\nimhxgjwrodtcknqvef\n\nsabwqu\nyqamws\n\nusei\nxkvnyfi\n\nrcewysztbiamhpgldj\nkibrfndmgeazcsh\nadrzgchmwjsbie\n\npnrwvx\nlwjxnu\nvwxinb\npfxybwng\n\nwnvxgfdkubzrmip\nkmjhoigltwbz\n\nznlgbx\nfla\ndl\nl\nl\n\njgeqcikplduzxrwby\nkwplxgcbndjqyizuer\newrqujilgyxpdzcbk\n\nguteykxsqrfwhz\nmhiurkswtqbxz\nsdlqxvowujrhkz\n\nzrejn\nnzrje\njnerz\nrzjen\n\nfvkouigderaqbjth\neqrvwjinfltzpcuag\n\nstqcrpy\nctsqpyr\nptyqrsc\nymsqrcaptb\n\negftpqosiv\necjdqlnmwouf\nspxaqofre\nepogyfqi\n\noxwgtshjqef\npbtgjfeqhnoxw\njfhewygoxtq\n\njwavgo\nogevbwdjai\nogvahjw\nwrajvog\nogsjwarv\n\nrpucvg\ncgpr\nhtfelpc\n\nhqrwapscubelovxntd\njisztpnbcahlfdoqmvexgk\n\npmadwutiejznofcgvsbqyxh\nqvmzbaisgfcunwjexhpdktyo\nxtvsdnqaucyehgzbjiopwmf\n\nvjqih\nhimjwdv\nhsojivb\njtrklivehyg\nmvnaizoqpfxhj\n\nj\nj\nj\nj\nj\n\ndpnwmtvklc\nndjpkm\nndmpk\nndmpk\nndmpk\n\nucojsptiqymhewbxvrfdkngl\nebpjwuxrdqkmsochvfgliznt\n\nesroiaqdbxnjhut\nxibuqhmensdajo\nfeukjbahdxwqnsiocv\n\nvkxzclpmtujngerawod\nazdrgcenvjxptkuolmw\nonrluajvcgewpdzkmtx\nxamlobkrvzwpjgnucdte\ntvenjdmlzpwugkoracx\n\nywpfhztklgjmoien\nnyelduixgvwmojpzhtk\nlehpigtmoknjyzw\ngmhwzokylpnteji\ntmypejhwgzlniko\n\nfmd\njf\nfj\n\nl\nv\nntobg\n\ndrmzvt\nrxtsfuk\n\nvxdlpjkrsz\ndgxkprvzsl\ndsklxpvzr\npdvxzlrsk\n\ngwip\ngakpw\n\nkbcadn\nbakcd\n\nmhpstqcgvoalkwxinebryzujd\nqwhirveoczyatgjdsmkpxnl\nmejazlincwodtshxkrpgqyv\nfsaizhnkcvwolgjpdqxtmrye\n\nrxubznovy\nyvudnrbz\nzvunwcyrkb\nntzbryuav\nndvytuzrb\n\nutvklrxypgbnimsdofwa\nfyzliqswjmtpvdka\n\nbyctfzx\npxzf\n\ng\nik\nykei\n\nbcajzwogvetd\ngedbcwtzvoaj\nhcwogaxjbzvted\neoczjagtdwbv\natvdjbcozgew\n\nwsclketugyaz\nqvfbjcxtmilgow\n\namgirw\nyf\n\niletugq\nhnoxbfkaspt\n\nqjdfcpvksmxzywahgn\nwnzceyskgpoxdhlmravjqf\nxjgvhcsqyfankpzmdw\njyncdxwsqkvphfamgz\ndvznygwjkcmqsxahfp\n\nfkuwiezpljbdayhsr\nrphwbeujdzlfiasyk\nerdhpafkzbwjlusiy\njwfhkrzdeabpliusy\nwfipeujhylazbsrkd\n\nfqmbwapvouelydgicj\nuelvqwomtyfjgcpadi\nqkpcsdeyjiwarovnglfum\nijvgqdauoefpcylwbm\nydmqpfewaovlicugj\n\nsgbq\nltqambf\ndjbuvn\nzmtbgy\n\negvifnsaztxqcrywumj\naqcxujnzvfwetrhgm\nfxdmtkgawplobqjenv\n\nbxuwpnl\nztdkau\n\necylsbdv\ncvzlsbeyd\nyqcveblsd\nbeyvlqcsd\nebsvdycl\n\neb\nxpnyter\n\ncbhqyzresnpxfwm\ncwksfvhxqperbimz\nwhqzpbcsjxafre\n\nbeodiasvj\njbisdoave\niedaovbsj\njxovacsiebd\n\nprxsnyehgwifbzo\nozpbsxynfiwrge\nribozypewsxngf\nnyoxspzbfihgewr\nrxzewpbnimfoysg\n\nshviadjcue\nweduziolmcbngp\nvqdieycu\ncudeif\ndicyuqke\n\nnvcwhuo\nsltchuwnq\nunwhci\n\nekrisbvx\nrsxiev\nexnrvsi\nsrxevi\nxhsvnrei\n\nrs\nrs\nsr\n\nfuzjmek\ntihf\n\nptlzxreqfdchibogvysmku\npceyizvdkqxbhrtulsfmog\ngvkrxfsbipqedlyzomhtncu\nsuyzgrobilxhmcfdtvkeqp\n\nokfvjneqyicuthwsmpxl\nwdojlsqknyhmptvicfe\nnyfjthmpivqgsulwkboec\n\nufrpkliqjgwxeybmcztasovh\nqvkpyghxblewsaitrmzuonjc\ngiptzywrejquxkoavslcmhb\npeaywubjozmcgrihlqsvxktf\nlehjpzguiksacobxywtmvrq\n\negryz\nergyz\nyzegr\nyfcgzre\n\nj\nj\nj\np\nj\n\nbjt\ntbj\ntjb\ntbj\njbt\n\ncklp\nlpck\nlckp\n\nfea\neaf\n\nqgm\ndz\n\nyri\nyi\nzypdi\n\nekwjxhtmgiudpz\nxepuhgwjizkdrmt\nezghmliwtxujdpk\nwdhxpjizugkemtl\n\ngyokbspj\nkbgsopyj\npykhdsjbgio\nopjykusbg\njygbsopakc\n\nixorlekwgnbcjm\nlrjkwxgie\nwgeikxjrlh\nrlkjxipeqgw\n\necgmdrykq\nmyrclqdgek\nguqmzyceprkd\ndygckwerqm\nyrcgkiqempd\n\nu\nu\n\nbcwnvl\nfcvsib\nhcnvb\ncpgveoaby\nivbcn\n\nn\nn\nn\n\nmacuqyp\nfs\n\npbout\nsuota\ncuotm\natmuo\n\nnlravibecfj\neilcjafn\naeibjfl\njxfeilauo\njafesdnli\n\nxfwroij\nfxwijor\nirxjofw\nrfjwboxi\n\ndj\ndvj\ndj\ndj\n\nmikhxcqensrodu\nimcvkhewgljupsxonq\nkceuqxltinhsom\nkzqanoemcxhius\nvmuoxineqpzhacbks\n\nhklcsoaywpfrdntmu\nxirbzckwjvpfygeno\n\no\nto\no\n\nzvirefunh\nezhirnvfu\nrunifzhev\nezivhnrfu\n\nzfsoducbpai\ncupdisobazf\naozbsucfdiypv\nbcfazodiups\n\nnjfprukxamlg\nkpfrjgynulxma\npijflgnuxkram\npkuxalrmjfgn\nnkvrlmxustpfjga\n\ndszgm\nrsmu\n\niaw\naiw\nwai\naiw\niwa\n\ns\ns\ns\nsl\ns\n\neyavhnk\nkeov\nqeiwfkcgvrpsld\nkuvze\n\na\nmh\n\nilqchdfsog\nigscolmdzq\n\nfnyhxmpct\nlchdpamfu\n\nvijz\nvj\n\ndhpxjrmyuefclvqgni\nuicskweohgvfryxmdq\ntqgdifymexbwckruvh\n\nc\nc\nc\n\nlthidumvycszw\ncvpqjeuhlwrfyzg\n\nkeocuw\nuckowe\noekwcu\ncwukoe\n\nc\nc\n\nlkxjv\njkvxl\nvxklj\njkvxlwg\nzkxvjly\n\nsqlamxeuhpgwjbdcioy\nbumyjxsiqweholpgcda\ngecswjlqhadiuboxmpy\nsyahibgopexwlucdqjm\n\nigebmv\ntawqvezdsmpunk\nrimocvgje\nevmbhc\n\nykjisvfatux\njfaisyvktux\nitksajfxuvy\n\nfclkush\nslcykdbgu\nfskuchl\n\nxjhulywagzeioq\noejlfsquhbriayxwg\njiexgylwzhuqao\nwyxhqljagieou\nhwladixguqejyo\n\nsofcbhuwyqtid\nicofdwsb\ndfwovimbcs\ncwibodnsxf\n\njqnestpyaczfmxhwrbuvldi\nmxanrhbfcujvyzditwlo\n\nntfgivkyszmarc\nqrajdhowuvpclbex\n\nkliwy\nukwlryi\nilkyw\n\npmgltzsk\nljmipztvgow\nzltmgup\numztglp\nftzmlpgk\n\ntwepi\nteq\n\nohyd\nmytjrwk\nhfy\nexyh\nye\n\nrzsqhtlowavy\ntwvhzyrjsqaol\n\nyzagcqrjeimwxokh\nfmxtiagboleqcwdrhzk\nopvemhasrguxkcqzwi\n\natcpeyosnrjbg\nkcesobgua\nfagsev\n\nca\nac\n\nxahmtil\nhximtla\n\nhyutnjdrbsmofl\ndswjfyothmblrnq\n\net\nte\neqt\ntuje\n\nnbhpwzgtrcx\ngqcxobhriznw\nhwnbugrejxtczs\ncujrbxfhnwzg\ngcwzhrxndjb\n\nqwrfxhlyzcptvmkdeauigob\nzbchavlkmryxiptegwfouqd\ncfhuiptgqkoalrewzmvbyxds\nhudmwcgbfepyitovkaqzxrl\n\neysuozkm\nndgloqyv\nignhwrovy\n\niqufgnvxspdcjtylkbrozh\ninuzqktrvcxyhbgfosljdp\n\nqhfknsyeuirzagvcwd\nyfcgqnxewlkrsuazhi\n\nlj\njl\nlj\nlyj\njl\n\nkpholfxewrdt\nopdfhtlmkwex\nhpwftoxderlk\nylxhwdofnztkpe\nrwfpxlhtdkoe\n\nubml\nyvohg\nzeasf\nplum\nqhnjxg\n\ncjgzwmdyuakvixlprhfe\nkmplfuxywergchzjvani\nxcvlzmewyujgirhpfka\nfwhgrxkjcvpeuilzyam\n\neclsiqp\nivgspoe\nscepi\npmiadtskyze\n\nakmdfivwgct\nyvgcwakhdimts\nlcvgirakwudtp\naftnckwmidgv\n\nzuqblv\nblvzqu\nuvbqlz\nqbuzvl\nbzqxvlmu\n\nby\nbetlcdvry\nyb\nbxy\ngybjx\n\nwaoqdxpkmfnyrhjebcuvzs\nqprvfczedowxknhausmyjb\ncufdqvknwrajzybospmxeh\nqkjzybeadvncwhxumpfsro\nnwfkjxumpabzodvqychsre\n\nshdfctjrkwbvemoxluyzipnqag\nsgbwpeuqcilxojfynmrdhvktaz\nlhgkxfwuzbevojmrpnqydsacti\ngflnzjasbuyimxqcvwrdpohetk\nynwedhbskjtvoucgapiqlzfxrm\n\nwlmjxp\nylnxoajk\nluxjidte\nxjrvlcdeu\n\npv\nvbpw\nwvuap\nkceavp\nrdphvoitqf\n\ndlxzwkhtmc\ntwzhlmkdc\ntcylhkdwzm\nzwmlkdcth\ngdhikmctlwsz\n\nhpekwtlufyragqvb\nkbrwatlupqyfvh\nucbvwlhfqatpkry\nhlakvyuqrfpwbt\nrubvwpfqlatkhy\n\nnpxobl\nckuhnplvx\n\nedkpjycofasulrthw\nrpuhcdyxlftosewa\n\nstuvhwaexykzdorbfpi\nshjuwgpfqmtacvlbirkzd\n\nfrph\nwo\nh\nzumtyngcj\npfse\n\nzivsgurp\nszirpvg\ngikvsprz\nzigpvsrk\nnvpsgzir\n\nrfimu\nhmqrlfxpbgu\nnrmoyftu\nnmujfr\nrfum\n\nbidzwmearovjkgcl\nogdbirkzqmclejvaw\ngealcqozmijkbrvdw\ngzakwmdbejclrivo\neolwifvgmrbjdazkc\n\niurfyqgwk\nymkvqfpwrig\naqrfywikg\nbfdryxikgnwoq\ncjfgkryiuwq\n\njvb\nx\n\nqkxnvfuopdytrzjgbm\nbmqkuyarjgtdpfzsnox\nobwjxyeqtingzmdrfkplu\ngamupstryfokbzqdxnj\n\nymanp\nnpfshjya\npundyklatc\nypfanh\n\ngarp\ngrpa\narpg\nargp\n\ninedmtpwrvub\nvwibertnudmp\nidurbepwtmnv\n\nyczoaplwmxibqgsfn\nwbiuxyqzpncvgmsl\ngpoltrbisqfznwxcymj\n\na\na\na\na\n\ncgjviyufped\npcuvygzkjid\nptvigyjdcku\n\nu\ni\ntqa\nj\nul\n\nzitbfuer\nuzfithernb\nuftbzire\nuitefbzr\n\njcwkroamnegxvqsh\nkejhanqmvcrxwsog\ngaqmkexjvcsnrowh\n\ntaegslmuyqcipz\nasytxrgdpfbcqliwz\ntpgoschavqjlkyzi\n\noc\noc\noc\nco\n\nmrfvjlykquz\nfzmvuleykqt\nvykmzlgquf\nkqfzuvygetml\n\na\ns\n\nionve\neailo\njiomexwh\nbeziqpdco\nihtoe\n\ncfjvrqosnwyedaubmhgtlxzk\nolxzjfwrcytkbgudasvqhnme\ntuwfnckrzsvjgbxdmyhqlaeo\ntxnydgufbmjaqkrzlwhoevcs\nbgakxqmfsdwyjvelntchzour\n\nnjvkobzs\nbsjvkwn\n\nqmnluxw\novgtikpdbrya\nnchfsjqez\n\nkhtdryej\nhdrtv\nqnaodtm\n\nas\naus\n\nomjaixvgklq\nvgkaxjihdrlq\nycxkzatpusiljqgv\n\njtepylmrqbiwk\nxerlpmynqtkfbjwi\ntimwpbkegljqry\nlqijkebmgpwyrt\nqrkljpbytemiw\n\nq\nxqint\ndkqb\n\ntrjmdxpoeqcwf\ntfedxqcprwom\nkmftwxqporncde\n\nfpcmndhr\ndzepjngqikcfrh\nnhfpcrd\n\nc\nc\nc\ny\nc\n\nt\nt\nti\n\ntgfspxdy\nsytgpfxd\n\nzov\nvoz\novz\n\nic\nieuc\n\nvhesndkizjgwcpobulam\nszjymxwopckfqtgdib\n\njdfzgqsbrmnticl\nfbcdgqtznmjri\ndoitrczjqfbgmn\notzbjfidmnrqcg\n\ncb\nbc\ncb\ncb\n\naspjvghyurx\nvcarwsbh\nlaeshvzrb\nhsrvia\n\nkpxzdhasfmltouvrc\nduvpkascromz\nvmcskupzorad\n\npguayqbwm\ngpyaqbwum\nywaqgpmbu\n\nypdlnjse\nelskbyz\njesylo\nlyse\n\nzlokqmjiurpwgbn\nxzvurhgmlkwqnibjpo\n\nwpzenbojcykqxvt\nwnyzlbjkeoqtvcpx\nitkjzydbpvoxcewnq\n\ngfuvenodsizwmh\nhwfuvsntdg\n\ngkibnpalc\ninkcgpal\ngnpzliakfc\n\nibektdpuhg\niwbgjke\niycnergfqkbo\negxadvlmbijk\n\nl\nyjg\nl\n\nymuoc\noymcu\noymcu\nyoucm\nycomu\n\nwifoqy\nzksfrict\nbdpvhuilgejn\nacoi\nmqi\n\nqrcukhwvzdjilpto\nhqrmwkupclvjtizo\n\ngdizlfawobuk\ndihqfswcgukzrbla\ngmukawbdfizlo\n\nzwdrpv\ndzvpwr\n\nycidkpfrqjelha\npjahykfecdqrli\nkfqlajhdepyirc\npkhaceijlqdfry\n\npwuxbqotey\nokbphqutfx\nxvsgodubtlpqijr\npuoewaxtbmq\n\nqtbaoex\nxpoqa\nagiweqtop\nnozuadmjrq\n\nhuvqdaij\nfkuvaijdhq\n\nbiewrdlzc\nrjlcikbzmdwve\nleucwfpoizndbxr\n\navgtnubqcpxrhwdoml\nvrkqgyndcmohtupswa\n\ndmjqfesnxyciz\nhfzudcjxsmno\nmfjhrcsxznd\nlcmbjnxzfadrs\n\ntvayrhpqskd\nqrtkvhyus\nepjsytzhvrqk\nyzqohmvtgsrk\nhvlysqtwirbk\n\nieyd\naipeyg\n\nr\nrg\nrm\nr\n\nqcweajhfmlgyzpv\nouqdxmsphkbatenr\n\nfba\nkfba\nbaf\n\nvxpblsyjnrmkai\nvpyhagcwqbzxmik\nrpbytexdijnfavkom\n\nfpmytnk\ncyaw\n\njhyimdg\nmhyij\nhmvyij\nmihezcovyj\nyhmciej\n\nsqfaopkgze\nkxsecfphozq\nqsotezpfk\n\newcvoaurmjhq\ncuqvjwreaohm\newvjmqacuorh\nwauhojecqmvrs\n\nlyieuqdjcb\nvkpctdmywzsreubn\ndelchbujy\njbcyegdou\nxduayceb\n\nnidxswarcqtk\nkciqdarnbewsxt\nacqrdkntipxlsw\nikpawtcdsqnrx\n\ngikrfhwltveap\nyeplvatowncgxirf\ngriadtewmvblpf\n\nhtjkmzxlvbygcwqioe\njzwtymxvigbheolckq\ncxmvjyklobhtwgqzie\nteczbxvlmgojqiwkyh\noxvejlbtiqywhzgmkc\n\njkfapozdwcie\nbmushwjgltrn\n\nrpnbxkoij\nexjdri\nuirjegxsa\n\nzawoncbymgsrhv\ndvctxgeusr\ncvizgyhlmrwsfp\n\nshnyujedctwfq\nyscahxndftreq\n\nxrnysi\nitcluvmd\nniqwozjber\n\ngjobshlayxvctnfq\nljcxaonvbstqfgyh\nybhxgntqcsvfaojl\nrjoxblqavncusgheytf\nsncqlvybxfjtghao\n\nuozb\nldewrck\nxbg\nxhg\nz\n\nyubqaznf\nimwhrgnkjls\n\nmsvgjrnkcloaup\ncsxjnkgrpiuoveamy\npknvrajmcudswgzo\ntnupacsxrgkvmoji\nbpnkfarqumsvhcojg\n\nbzvyuenaihfdcgwk\npuchyenmzgksrijoa\nugebciktnlzhay\n\ncedfutpnsivjlobrkaxy\ndenscvpbmlkxaojrfh\n\nvugjrq\njqgtuwcnrv\ngrvkuqj\ngbmvqsyxjueplzirhf\navudnjqgr\n\nvpidajtmgozxwuksh\nughiwmakpjszdtvox\nwugipxzdvastkhmj\nsxughjakmnwfpdivzt\nwimvaupjhgxstkzd\n\nmqsuwhdxpcozfr\nrsdwhfpakzxjmo\neysprldfqzowmhx\nfodhrtmswxyzp\n\nvsmzfbjhwqxiugtlc\ngcqdsfmpletnhbxr\n\nuxbgkfjwesmlchr\nfeakldxqnpwcjgvrzbhs\nhjosrixfmlcgbkew\n\ndszv\nzsvd\nvsdz\nzvds\nsdzv\n\npdw\nlxudjpiwz\nadvnkpqw\nhswpdog\n\ngslpira\nneuzf\n\nifjt\ntjif\njfit\njitf\nifktj\n\njrnzb\nunkjshpe\nnjza\n\nydigxaqm\niaeokqgmyhr\n\nbkxmngtdlyev\nnmbuzoveixlprkys\n\nzjloghmtu\nermvzn\nqzdncfm\nafczxmp\n\naptscynieqmfgjhbr\nbmsirnafyepthgcqj\n\nq\nqlr\nfq\nqk\n\nv\nv\n\nucrmbsozjkfy\nbuscrmofkzy\nuosmfkryzbc\nsqirbfzmocuynk\nmbucykofrzs\n\nlrp\nplz\nlrp\nlp\nlp\n\nyfzmpgcsw\nprycgfm\nrfcmgpy\nscgmyqpf\nflcyvpohmgu\n\ng\ng\ng\ne\n\nlxoaibjch\nwudyrfvkzn\ngtpmkeyqd\n\nzjhlwbtfuvick\nwujtvblzchkif\n\nksdfpvqbtxaluneo\ndqofprhnxamskujeb\n\nhacwigyftqblx\nxwlqahitbgcyf\nahyxlwgtqifbc\n\nmb\nm\nmg\nm\n\necrasdtyljuqimhnwfb\nqleiaydjcugrwn\njdyewaqnrcliu\n\ng\nlqfio\nadhstybz\n\ngryzuxenhlvopbtmfwqcas\nyeqzalmogbncxpsfujwh\nmflqzhoaupcseygnbwx\nmyswzuonfqlaxecbhpgj\n\nuwvzati\nuzcjvtai\nsavuizt\natiuzv\n\ncozuin\nabuizo\ntiozua\nwilfzeoqu\nroictxuz\n\ngndpkojqu\ngunkpsjoa\nrtmzjpgkovwxlnuc\n\nlrj\nrjl\nmrjl\n\nudlpevjgzrfawitqs\nqgrtiujslfeovza\nsurivgtfqjleaz\nezavgjtlfrusiq\n\ng\nfx\nmf\ntx\niyqd\n\nsljgmdztynvrqexhwfi\ngvwdyerlqnxthmjifsz\nqgmjlfwrxztynivdhse\n\nzkvuir\nkvruiz\nvriukz\nkriumzv\nzruikv\n\nzyqgnvda\nkdygvqanz\nzyagvqdnu\nqnwdmlaxgzyfvo\ndnzgyavq\n\no\ns\nq\n\ncbroy\nqiyocmjdt\nuyoc\n\nlkqj\nkjql\njqlk\nljqk"
-      .split("\n\n")
-
-    println(input.map(_.split("\n")
-      .map(_.toSet))
-      .map(_.reduce(_ & _))
-      .map(_.size)
-      .sum)
-  }
-
-  def _08_1() {
-    val input = "jmp +1\nacc -18\nacc +19\nacc +19\njmp +202\nacc +15\nacc +42\nacc +30\nacc -7\njmp +535\nacc +31\nacc +9\njmp +581\nnop +501\nacc +44\nacc +18\nacc -4\njmp +545\nacc +9\nacc +5\nnop -2\nacc +3\njmp +475\nacc -10\nacc +14\nacc +29\nnop +471\njmp +470\nacc +2\nnop +375\nacc +31\nacc +6\njmp +420\nacc -1\nacc +2\nnop +185\njmp +490\nacc +2\njmp +317\nnop +282\njmp +457\nacc +37\njmp +254\nacc +19\njmp +436\njmp +458\nacc -7\nacc -2\nacc -17\njmp +454\nacc +37\njmp +212\nacc +6\nacc +5\nacc -7\njmp +104\nacc +5\nnop +134\nacc +46\njmp +541\nacc +4\nacc +46\nacc +18\njmp -53\nacc +10\njmp +285\nacc +39\nacc +34\nnop +109\nacc +47\njmp +32\njmp +1\njmp +143\nacc +36\njmp +429\nacc +45\nacc +22\njmp -59\nacc +0\nacc +23\nacc +30\nacc -7\njmp -45\nacc +46\nacc +31\njmp +164\nacc +37\nacc +34\nacc +40\nacc -1\njmp +246\nnop -46\nacc +2\nacc +31\njmp +221\nnop +413\njmp -51\nacc -14\njmp +145\nacc +1\nnop +77\njmp +131\njmp +370\nnop +513\nacc +7\njmp +476\nacc +22\nacc +37\nacc +44\njmp +334\nacc +9\nacc -1\nacc +5\nacc +27\njmp +351\nacc +31\njmp +220\nnop -61\nacc +34\njmp +504\nnop +471\nacc +6\nacc +48\njmp -17\njmp +217\nacc +13\nacc +0\nacc +25\njmp +144\nacc -5\nacc +17\nnop +341\njmp -26\nacc -10\nacc +34\njmp +168\nnop -16\nacc -6\nacc -10\nacc +38\njmp +30\nacc -2\nacc -14\njmp +419\nacc +40\nacc -17\nacc +27\nacc +8\njmp +101\nnop +370\nacc +2\nacc -10\nacc -7\njmp +224\nnop +437\nacc +42\nacc +50\nacc +39\njmp +81\njmp +11\njmp +143\nacc +6\nacc +46\njmp -107\nacc +13\njmp -109\nacc +5\njmp +1\njmp +467\njmp -159\nnop +421\njmp +243\nacc +44\nnop +412\nacc -6\nacc +13\njmp +56\nacc -12\nacc +18\njmp +313\nnop +151\nacc +5\nacc +49\njmp +120\nacc +46\nacc +23\nnop -122\nacc +21\njmp -55\nacc -15\njmp -115\nacc +19\nnop +116\nacc +32\nacc +21\njmp +16\nacc +27\nacc +32\njmp +359\nacc +16\nacc +18\nacc +15\njmp +54\nnop +182\nacc +4\njmp +361\nacc +4\nacc +38\nacc +49\njmp -94\njmp +428\nacc +0\nacc +9\njmp +224\njmp +82\nnop +57\nacc +6\njmp +1\njmp +144\njmp +20\njmp +265\njmp +402\nnop +114\nacc -12\nacc -11\nacc +1\njmp +412\nnop -163\nacc +50\nacc +1\nacc -9\njmp -20\nacc +10\nacc +6\njmp +323\nacc +10\njmp +1\nacc +42\njmp +46\nacc +35\nacc +31\njmp -139\njmp +129\njmp -193\nacc -4\nnop +247\nnop -163\nacc +25\njmp -26\nacc +34\nacc +26\nacc +40\njmp +220\nacc -6\nacc +6\njmp +311\nacc +0\nacc +14\nacc +41\nacc +6\njmp +284\nacc +32\njmp -13\nnop +157\nacc -4\nacc +47\njmp -146\nacc +34\nacc +6\nnop +196\nacc +5\njmp +268\nacc -8\njmp -176\nacc +34\nacc +17\njmp +1\njmp +114\nacc +32\nacc +39\nnop +313\nacc +38\njmp -237\njmp -273\nacc +21\nacc +26\nacc +31\njmp -231\nacc +17\njmp -105\nnop +333\nnop +17\njmp +11\nacc -9\nacc +2\njmp -162\nacc +3\nacc +0\nnop +318\njmp +215\nacc +14\nacc +32\njmp -196\njmp +33\nacc -6\nacc +45\nacc +27\njmp -166\nacc -1\njmp -25\nacc +0\nacc +4\nacc -14\nacc +0\njmp -115\nnop +118\nacc +28\nnop +175\nacc +45\njmp -97\njmp +78\njmp -284\nacc +35\nnop -248\nacc -18\nacc -6\njmp -308\njmp -95\nacc -2\nacc +0\njmp +255\nacc +7\njmp -24\nacc +17\nacc +20\njmp -220\njmp +172\nacc +40\nacc +39\nacc +19\njmp -238\nnop +44\nnop -99\nnop +238\njmp +195\nacc -14\nacc +36\nacc +40\nacc -11\njmp -65\nnop -54\nnop +47\nacc -11\njmp +18\njmp +98\njmp +252\nnop -1\nacc +1\nacc +10\njmp -4\njmp -319\njmp -46\nacc -8\nacc +50\nacc +43\njmp -174\nacc +22\nacc +4\nacc +32\nacc -6\njmp +73\nacc +8\njmp -31\nacc +16\nnop +11\nacc +26\njmp -98\nacc -11\nacc +40\njmp +101\nacc +28\nacc +30\nacc -16\nacc +7\njmp +239\njmp -179\nacc +47\nacc +15\nacc +42\nacc +26\njmp +30\nacc +17\nacc +3\nacc -5\nnop -98\njmp -236\nacc +2\njmp +196\nacc +39\nacc -14\nacc +36\nacc +49\njmp +189\njmp +235\nacc +27\nacc -2\nacc +16\nacc +40\njmp -81\nacc -5\nacc +17\nacc -1\njmp +1\njmp -129\nnop -171\nacc +47\njmp +169\nacc -16\nacc -5\njmp +172\nnop -84\nacc +8\nacc +40\nacc +43\njmp -33\nacc +39\nnop -12\njmp +53\nacc +36\njmp -270\nacc +17\nacc -1\njmp -255\nacc +0\nacc -12\njmp -371\njmp -216\nacc +45\nacc -18\nacc +23\nacc -17\njmp -37\njmp -386\nnop -302\nacc -19\nacc -16\njmp -297\nacc -18\nacc -7\nacc +17\nacc +17\njmp -173\njmp +114\nacc +4\nacc +19\nnop -296\njmp -36\nacc -18\nacc -14\nacc +6\nnop -27\njmp -101\nacc +15\nnop -407\njmp +44\nacc -18\nacc -6\nacc +33\nacc +36\njmp +80\nacc +43\njmp +73\nacc -2\nacc +7\nacc +4\njmp -10\nacc +46\nnop -49\nacc +7\njmp +65\nacc +24\njmp +144\nacc +13\nacc +26\njmp -351\njmp +1\nacc +34\nnop +62\njmp -363\nacc -4\nacc -5\njmp +23\njmp +82\nacc -7\nacc -7\nacc +15\njmp -468\nacc +7\nnop -423\njmp -178\nnop -425\nacc +23\njmp -181\nacc +6\nacc -11\njmp -321\nacc +3\njmp -122\nacc +12\njmp +44\nacc -5\nacc -16\nacc +16\njmp -281\nacc +33\nacc -4\nacc +15\njmp +9\njmp +63\nacc +35\nnop -12\nacc +25\nacc -10\njmp -452\nacc +1\njmp -148\nacc +8\nacc +40\nacc +48\njmp +2\njmp -315\nnop -325\nacc -4\nacc +16\nacc -4\njmp -369\nacc +21\nacc +3\njmp -153\nnop -25\nacc +0\njmp -84\nacc +32\njmp +19\nacc -18\nacc -1\njmp -385\njmp +1\njmp -357\nacc -13\nacc -13\njmp -360\njmp -393\nacc +4\nnop -102\njmp -316\njmp -248\nacc +4\nnop -487\njmp -339\nacc +3\njmp -190\nacc +24\nacc +31\njmp -166\njmp -482\nacc +22\nacc +32\njmp -290\nacc +22\nacc +48\nacc +5\nacc -6\njmp -330\nnop -203\nacc +7\nacc +1\njmp -287\nacc +15\nacc +3\njmp -230\nacc +37\nnop -162\njmp +33\njmp -147\nacc +16\nacc +39\nacc +49\njmp -560\nacc +26\njmp +26\njmp -283\njmp -486\nacc -9\njmp +1\nacc +25\nacc +1\njmp -514\nacc +46\njmp -166\nacc -5\nacc +35\nnop -204\njmp -175\nnop -30\nnop +11\njmp -400\nacc +15\nacc -7\nacc -1\njmp +18\nacc +31\nacc +16\nacc +43\nacc +33\njmp -260\nacc +1\nacc +23\nacc +25\nacc -1\njmp -200\nacc -15\njmp -314\njmp -238\njmp -75\njmp -192\nacc +30\njmp -525\nacc -14\njmp +17\nacc +7\nacc +9\nacc -6\nnop -312\njmp -559\nacc +28\njmp -305\njmp -239\nacc +0\nacc -5\nacc +9\njmp -471\nnop -327\nacc -5\nacc -19\njmp -496\nacc +17\njmp +1\njmp +1\nacc +29\njmp +1"
-      .split("\n")
-
-    val accRx = "acc (.+)".r
-    val jmpRx = "jmp (.+)".r
-    val nopRx = "nop .+".r
-
-    def loop(acc: Int, index: Int, visited: Set[Int]) {
-      if (visited contains index) {
-        println(acc)
-        return
-      }
-
-      input(index) match {
-        case accRx(delta) => loop(acc + delta.toInt, index + 1, visited + index)
-        case jmpRx(delta) => loop(acc, index + delta.toInt, visited + index)
-        case nopRx() => loop(acc, index + 1, visited + index)
-      }
-    }
-    loop(0, 0, Set.empty)
-  }
-
-  def _08_2() {
-    val input = "jmp +1\nacc -18\nacc +19\nacc +19\njmp +202\nacc +15\nacc +42\nacc +30\nacc -7\njmp +535\nacc +31\nacc +9\njmp +581\nnop +501\nacc +44\nacc +18\nacc -4\njmp +545\nacc +9\nacc +5\nnop -2\nacc +3\njmp +475\nacc -10\nacc +14\nacc +29\nnop +471\njmp +470\nacc +2\nnop +375\nacc +31\nacc +6\njmp +420\nacc -1\nacc +2\nnop +185\njmp +490\nacc +2\njmp +317\nnop +282\njmp +457\nacc +37\njmp +254\nacc +19\njmp +436\njmp +458\nacc -7\nacc -2\nacc -17\njmp +454\nacc +37\njmp +212\nacc +6\nacc +5\nacc -7\njmp +104\nacc +5\nnop +134\nacc +46\njmp +541\nacc +4\nacc +46\nacc +18\njmp -53\nacc +10\njmp +285\nacc +39\nacc +34\nnop +109\nacc +47\njmp +32\njmp +1\njmp +143\nacc +36\njmp +429\nacc +45\nacc +22\njmp -59\nacc +0\nacc +23\nacc +30\nacc -7\njmp -45\nacc +46\nacc +31\njmp +164\nacc +37\nacc +34\nacc +40\nacc -1\njmp +246\nnop -46\nacc +2\nacc +31\njmp +221\nnop +413\njmp -51\nacc -14\njmp +145\nacc +1\nnop +77\njmp +131\njmp +370\nnop +513\nacc +7\njmp +476\nacc +22\nacc +37\nacc +44\njmp +334\nacc +9\nacc -1\nacc +5\nacc +27\njmp +351\nacc +31\njmp +220\nnop -61\nacc +34\njmp +504\nnop +471\nacc +6\nacc +48\njmp -17\njmp +217\nacc +13\nacc +0\nacc +25\njmp +144\nacc -5\nacc +17\nnop +341\njmp -26\nacc -10\nacc +34\njmp +168\nnop -16\nacc -6\nacc -10\nacc +38\njmp +30\nacc -2\nacc -14\njmp +419\nacc +40\nacc -17\nacc +27\nacc +8\njmp +101\nnop +370\nacc +2\nacc -10\nacc -7\njmp +224\nnop +437\nacc +42\nacc +50\nacc +39\njmp +81\njmp +11\njmp +143\nacc +6\nacc +46\njmp -107\nacc +13\njmp -109\nacc +5\njmp +1\njmp +467\njmp -159\nnop +421\njmp +243\nacc +44\nnop +412\nacc -6\nacc +13\njmp +56\nacc -12\nacc +18\njmp +313\nnop +151\nacc +5\nacc +49\njmp +120\nacc +46\nacc +23\nnop -122\nacc +21\njmp -55\nacc -15\njmp -115\nacc +19\nnop +116\nacc +32\nacc +21\njmp +16\nacc +27\nacc +32\njmp +359\nacc +16\nacc +18\nacc +15\njmp +54\nnop +182\nacc +4\njmp +361\nacc +4\nacc +38\nacc +49\njmp -94\njmp +428\nacc +0\nacc +9\njmp +224\njmp +82\nnop +57\nacc +6\njmp +1\njmp +144\njmp +20\njmp +265\njmp +402\nnop +114\nacc -12\nacc -11\nacc +1\njmp +412\nnop -163\nacc +50\nacc +1\nacc -9\njmp -20\nacc +10\nacc +6\njmp +323\nacc +10\njmp +1\nacc +42\njmp +46\nacc +35\nacc +31\njmp -139\njmp +129\njmp -193\nacc -4\nnop +247\nnop -163\nacc +25\njmp -26\nacc +34\nacc +26\nacc +40\njmp +220\nacc -6\nacc +6\njmp +311\nacc +0\nacc +14\nacc +41\nacc +6\njmp +284\nacc +32\njmp -13\nnop +157\nacc -4\nacc +47\njmp -146\nacc +34\nacc +6\nnop +196\nacc +5\njmp +268\nacc -8\njmp -176\nacc +34\nacc +17\njmp +1\njmp +114\nacc +32\nacc +39\nnop +313\nacc +38\njmp -237\njmp -273\nacc +21\nacc +26\nacc +31\njmp -231\nacc +17\njmp -105\nnop +333\nnop +17\njmp +11\nacc -9\nacc +2\njmp -162\nacc +3\nacc +0\nnop +318\njmp +215\nacc +14\nacc +32\njmp -196\njmp +33\nacc -6\nacc +45\nacc +27\njmp -166\nacc -1\njmp -25\nacc +0\nacc +4\nacc -14\nacc +0\njmp -115\nnop +118\nacc +28\nnop +175\nacc +45\njmp -97\njmp +78\njmp -284\nacc +35\nnop -248\nacc -18\nacc -6\njmp -308\njmp -95\nacc -2\nacc +0\njmp +255\nacc +7\njmp -24\nacc +17\nacc +20\njmp -220\njmp +172\nacc +40\nacc +39\nacc +19\njmp -238\nnop +44\nnop -99\nnop +238\njmp +195\nacc -14\nacc +36\nacc +40\nacc -11\njmp -65\nnop -54\nnop +47\nacc -11\njmp +18\njmp +98\njmp +252\nnop -1\nacc +1\nacc +10\njmp -4\njmp -319\njmp -46\nacc -8\nacc +50\nacc +43\njmp -174\nacc +22\nacc +4\nacc +32\nacc -6\njmp +73\nacc +8\njmp -31\nacc +16\nnop +11\nacc +26\njmp -98\nacc -11\nacc +40\njmp +101\nacc +28\nacc +30\nacc -16\nacc +7\njmp +239\njmp -179\nacc +47\nacc +15\nacc +42\nacc +26\njmp +30\nacc +17\nacc +3\nacc -5\nnop -98\njmp -236\nacc +2\njmp +196\nacc +39\nacc -14\nacc +36\nacc +49\njmp +189\njmp +235\nacc +27\nacc -2\nacc +16\nacc +40\njmp -81\nacc -5\nacc +17\nacc -1\njmp +1\njmp -129\nnop -171\nacc +47\njmp +169\nacc -16\nacc -5\njmp +172\nnop -84\nacc +8\nacc +40\nacc +43\njmp -33\nacc +39\nnop -12\njmp +53\nacc +36\njmp -270\nacc +17\nacc -1\njmp -255\nacc +0\nacc -12\njmp -371\njmp -216\nacc +45\nacc -18\nacc +23\nacc -17\njmp -37\njmp -386\nnop -302\nacc -19\nacc -16\njmp -297\nacc -18\nacc -7\nacc +17\nacc +17\njmp -173\njmp +114\nacc +4\nacc +19\nnop -296\njmp -36\nacc -18\nacc -14\nacc +6\nnop -27\njmp -101\nacc +15\nnop -407\njmp +44\nacc -18\nacc -6\nacc +33\nacc +36\njmp +80\nacc +43\njmp +73\nacc -2\nacc +7\nacc +4\njmp -10\nacc +46\nnop -49\nacc +7\njmp +65\nacc +24\njmp +144\nacc +13\nacc +26\njmp -351\njmp +1\nacc +34\nnop +62\njmp -363\nacc -4\nacc -5\njmp +23\njmp +82\nacc -7\nacc -7\nacc +15\njmp -468\nacc +7\nnop -423\njmp -178\nnop -425\nacc +23\njmp -181\nacc +6\nacc -11\njmp -321\nacc +3\njmp -122\nacc +12\njmp +44\nacc -5\nacc -16\nacc +16\njmp -281\nacc +33\nacc -4\nacc +15\njmp +9\njmp +63\nacc +35\nnop -12\nacc +25\nacc -10\njmp -452\nacc +1\njmp -148\nacc +8\nacc +40\nacc +48\njmp +2\njmp -315\nnop -325\nacc -4\nacc +16\nacc -4\njmp -369\nacc +21\nacc +3\njmp -153\nnop -25\nacc +0\njmp -84\nacc +32\njmp +19\nacc -18\nacc -1\njmp -385\njmp +1\njmp -357\nacc -13\nacc -13\njmp -360\njmp -393\nacc +4\nnop -102\njmp -316\njmp -248\nacc +4\nnop -487\njmp -339\nacc +3\njmp -190\nacc +24\nacc +31\njmp -166\njmp -482\nacc +22\nacc +32\njmp -290\nacc +22\nacc +48\nacc +5\nacc -6\njmp -330\nnop -203\nacc +7\nacc +1\njmp -287\nacc +15\nacc +3\njmp -230\nacc +37\nnop -162\njmp +33\njmp -147\nacc +16\nacc +39\nacc +49\njmp -560\nacc +26\njmp +26\njmp -283\njmp -486\nacc -9\njmp +1\nacc +25\nacc +1\njmp -514\nacc +46\njmp -166\nacc -5\nacc +35\nnop -204\njmp -175\nnop -30\nnop +11\njmp -400\nacc +15\nacc -7\nacc -1\njmp +18\nacc +31\nacc +16\nacc +43\nacc +33\njmp -260\nacc +1\nacc +23\nacc +25\nacc -1\njmp -200\nacc -15\njmp -314\njmp -238\njmp -75\njmp -192\nacc +30\njmp -525\nacc -14\njmp +17\nacc +7\nacc +9\nacc -6\nnop -312\njmp -559\nacc +28\njmp -305\njmp -239\nacc +0\nacc -5\nacc +9\njmp -471\nnop -327\nacc -5\nacc -19\njmp -496\nacc +17\njmp +1\njmp +1\nacc +29\njmp +1"
-      .split("\n")
-
-    val accRx = "acc (.+)".r
-    val jmpRx = "jmp (.+)".r
-    val nopRx = "nop (.+)".r
-
-    def loopAcc(acc: Int, index: Int, visited: Set[Int], delta: String, flipped: Int) = {
-      loop(acc + delta.toInt, index + 1, visited + index, flipped)
-    }
-
-    def loopJmp(acc: Int, index: Int, visited: Set[Int], delta: String, flipped: Int) = {
-      loop(acc, index + delta.toInt, visited + index, flipped)
-    }
-
-    def loopNop(acc: Int, index: Int, visited: Set[Int], flipped: Int) = {
-      loop(acc, index + 1, visited + index, flipped)
-    }
-
-    def loop(acc: Int, index: Int, visited: Set[Int], flipped: Int): Boolean = {
-      if (index == input.length) {
-        println(acc)
-        return true
-      }
-      if (visited contains index) {
-        return false
-      }
-
-      input(index) match {
-        case accRx(delta) => loopAcc(acc, index, visited, delta, flipped)
-        case nopRx(delta) => -1 == flipped && loopJmp(acc, index, visited, delta, index) ||
-          (if (flipped == index) loopJmp(acc, index, visited, delta, index) else loopNop(acc, index, visited, flipped))
-        case jmpRx(delta) => -1 == flipped && loopNop(acc, index, visited, index) ||
-          (if (flipped == index) loopNop(acc, index, visited, index) else loopJmp(acc, index, visited, delta, flipped))
-      }
-    }
-    loop(0, 0, Set.empty, -1)
-  }
-
-  def _11_1() {
-    val input = "LLLLL.LLLLLLLLLLLLLLLLL.LLLLLL.LLLLL.L.LLLLLL.LLLLLL.LLL.LLLLLLLL.LLLLL.L.LLLLLLLLLLLLLLLL.LLLLLLL\nLLLLL.LLLLLLLLLLL.LLLLLLLLLLLLLLLLLL.LLLLLLLLL.LLLLL.LLL.LLLLLLLLLLLLL.L.LLLLLLLLLL.LLLLLL.LLLLLLL\nLLLLLLLLLLLLLLL.LLLLLLLLLLLLLL.LLLLL.LLLLL.LLL..LLLLLLLL.LLLLLLLL.LLLLLLL.LLLLLLLLL.LLLLLL.LLLLLLL\nLLLLLLLL.LLLLLL.LLLLLLLLLLLLLL.LLLLL.LLLLLLLLL.LLLLLLLLLLLLLLLLLL.LLLLL.L.LLLLLLLLLLLLLLLL.LLLLL.L\nLLLLL.LLLLLLLLL.LLLLLLL.LLLL.L.LLLLL.LLLLLLLLL.LLLLLLLLL.LLLLLLLLLLLLL.LLLLLLLLLLLLL.LLLLLLLLLLLLL\nL.LLLLLLLLLLLLL.LLLLLLL.LLLLLL.LLLLLLLLLLLLLLLLLLLLLLLLLLLLLLLLLL.LL.LLLL.LLLLLLLLL.LLLLLL.LLLLLLL\nLLLLL.LLLLLLLLL.LLLLLLL.LL.LLLLLLLLL.LLLLLLLLL.LLLLLLLLL.LLLLLLLL.LLLLLLLLLLLLLL.LLLLLLLLL.LLLLLLL\nLLLLLLLLLLLLLLLLLLL.LLL.LLLLLL.L.LLL.LLLLLLLLL.LLLLLLLL.LLLLLLLLL.LLLLLLLLLLLLLLLLLLLLLLLL.LL.LLLL\nLLLLL.LLLLLLLLL.LLLLLLLLLLL.LL.LLLLL.LLLLLL.LL.LLLLLLLL..LLLLLLLLLLLLLLLL.LLLLLLLLL.LLLLLL.LLLLLLL\nLL.LL.LLL.LLLLLLLLLLLLL.LLLL.L.LLLLL.LLLLLLLLLLLLLLLLLLLLLLLLLLLL.LLLLLLLLLLLLLLLLL.LLLL.L.LLLLLLL\n......L....L...L...L.....L........LLL..LL..LL..L.L..LL.L..L.....LL......L.LLL.L..L....LLL.......LL\nLLLLL.LLLLLLLL.LLLLLLLL.LLLLL..LLLLL.LLL.LLLL..LLLLLLLLL.LLLLLLLLLLLLLLLLLLLLLLLLLLLLLLLLLLLLLLLLL\nLLLLL.LLLLLLLLL.LL.LLLLLLLLLLLLLLLLLLLLLLLLLLL.LLLLLLLLLLLLLLLLLL.LLLLLLLLLLLLLLLLL.LLLLLL.LL..L.L\nLLLLL.LLLLLLL...LLLLLLLLLLLLLL.LLLLL.LLLLLLLLL.LLLLLLLLL.LLLLLLL.LLLL.LLLLLLLLLLLLL.LLLLLL.LLLLLLL\nLLLLL.LLLLLLLL..LLLLLLL.LLLLL.LL.LLL.LLLLLLLLL.LLL.LLLLL.LLLLLLLLLLLLLLLL.LLLLLLLLL.LLLLLL.LLLLLLL\nLLLLL.LLLLLLLLL.LLLL.LL.LLLLLLLLLLLL..LLLLLLLL.LLLLLLLLL.LLLLLLLLLLL.LLLL.LL.LLLLLL.LLLLLL.LLLLLLL\nLLLLL.LLLLLLLLLLLLLLLLL.LLLLLLLLL.LL.LLLLLLLLL.LLLLLLLLL.LLLLLLLL.LLLLLLLLLLLLLLLLL.LLLLLL.LL.LLLL\nL.LL..LLLLLLLLL.LLLLLLL.L.LLLLLLLLLLLL.LLLLLLL.LLLLLLLLLLLLLLLLLLLLLLLLL..LLLLLLLLL.LLLLLLLLLLLLLL\n.L.......L..LL.....L.L.....L..............L.L.L.LL..LL.L.L..LLL.L..LL.............LLL.L......L.LL.\nLLLLL.LLLLLLLLL.LLLLLLL.LLLLLLLLLLL.LLLLLLLLLLLLLL.LLLLL.LLLLLLLL.LLLLLLL.LLLLLLLLLLLLLLLL.LLLLLLL\nLLLLL.LLLLLLLLL.LLLLLLL.LLLLLL.LLLLL.LL.LLLLLL.LLLLLLLLL.LLLLLLLL.LLLLLLLLLLLLLLLLLLLLLLLL.LLLLLLL\nLLLLL.LLLLLLLLL.LLLLLLL.LLLLLLLLLLLLLLLLLLLLLLLLLLLLLLLL.LLLLLLLLLLLLLLLL.LLLLLLLLL.L.LLLL.LLLLLLL\nLLLLL.LLL.LLLLL.LLLLLLL.LLLLLL..LLLLLLLLLLLLLLLLLLLL.LLL.L.LLLLLLLLLLLLLL.LLLLLLLLLL.LLLLL..L.LLLL\nLLLLL.LLLLLLLLL.LLLLLLL.LLLLLLLLLLLL.LLLLLLLLL.LLLLLLL.LLLLLLLLLL..LLLLLL.LLLLLLLLL.LLLLLLL.LL.LL.\nLLLLL.LLLLLLLLL.LLLLLLL.LLLLLL.LLLLL.LLLLLLL.L.LL.LLLLL..L..LLL.L.LLLLLLL..LLLLLLLL..LLLLL.LLLLLLL\nLLLLL.LLL.LLLLLLLLLLLLL.LLLLLL.LLLLL.LLLLLLLLL.LLLLL.LLL.LLLLL.LL.LL.LLLL.LLLLLL.LL.LLLLLL.LLLLLLL\nLLLLL.LLLLLLLLL.LLLLLLL.LLLLL..LLLLLLLLLLLLL.L.LLLLLLLLL.LLLLLLLLLLLLLLLL.LLLLLLLLL.LLLLLL.LL.LLLL\nLLLLL.LLLLLLLLLLLLLLLLLLLLLLLL.L.LLL.LLLLLL.LL.LLLLLLLLL.LLLLLLLL.LLLLLLL.LLLLLLLLL.LLL.LLLLLLLLLL\n......L.....L...LL.L.L..L...L....L...LLL...LLLLL..L.....L.....L..L......L...L.L.......L..L.LL...L.\nLL.LLL.LLLLLLLLLLLLLLLL.LLLLLL.LLLLL.LLLLLLLLLLLLLLLLLLL.LLLLLLLL.LL.LLLL.LLLLLLLLL.LL.LLLLLLLLLLL\nLLLLL.LLLLLLLLL.LLLLLLL.LLLLLLLLLLL..LLLLLLLLL..LLLLLLLL.LLLL.LLL..LLLLLLLLLLLLLLLL.LLLL.L.LLLLLLL\nLLLLL.LLLLLLLLL.LLLLLLL.LLLLLLLLLLLL.LLLLLLLLL.LLLLLLLLL.LLLLLLLL..LLLLLL.LLLLLL.LLLLLL.LL.LLLLLLL\nLLLLL..LL.LLLLLLLLLLLLL.LLLLLLLLLLLLLLLLLLLL.LLLLLLLLLLL.LLLLLLLL.LLLLLLL.LLLLLLLLLL.LLLL..LLLLLLL\n.LLL...LL.....L.L..L.LL..LL......LL.L.L........L.L.....LLLL.L.L.......L......LL.....LLLL....L.L.LL\nLLLLL.LLLLLLLLL.LLLLLLL.LLLLLL.LLLLL.LLLLLLLLL.LLLLLLLLL.LLLLLLLL.LL.LLLL.LLLLLLLLLLLLLLLL.LLLLLLL\nLLLLL.LLLLLLLLL.LLLLLLL.LLLLL.LLLLLLLLLLLLLLLL.LLLLLLLLLLLLLLLLLL.LLLLLLLLLLLL.LLLL.LLLLLLLLLLLLLL\nLLLL..LLLLLLLLL.LLLLLLLLLLLLLL.LLLLLLLLLLLLLLL.LLLLLLL.L..LLLLLLL..LLLL.L.LLLLLLLLL.LLLLLL.LLLLLLL\nLLLLLLLLLLLL.LL.LLLLLLL.LLLL.L.LLLLLLLLLLLL.LL.LLLLLLLLL.LLLLLLLL.LLLLLLL.LLLLLLLLL..LLLLL.LLLLLLL\nLLLLLLLLLLL.LLL.LLLLLLLL.LLLLL.LLLLL.LLLLLLLLLLLLLLLLLLL.LLLL.LLL.LLLLLLLLLLLLLLL.L.LLLLLLLLLLLLLL\n.L.L...L...L.....L...LLLLLLL.LL....LL..LLL...L...L.LLL..L.L...L.L.L....L.LL.LLL..L....L..LL..L...L\nLLLLLLLLLLL.LL..LLLL..L.LLLLLL.LLLLLLLLLLLL.LL.LLLLLLLLLLLLLLLLLL.LLLLLLLLLLLLLLLLL....LLLLLLLLLLL\nLLLLLLLLLLLLLLLLLLLLL.LLLLLLLL.LLLLL.LL.LLLLLL.LLLLLLLLL.LLLLLLLL.LLLLLLLLLLLLLLLLLLLLLLLL.LLLLLLL\nLLLLL.LLLLLLLLL.LLLLL.LLLLLLLL.LLLL..LLL.LLLLLLLLLLLLLLL.LLLLLL.L.LLLL.LL.LLLL.LLLLLLLLLL..LLLLLLL\nLLLLL..LLLLLLLLLLLLLLLLLLLLLLL.LLLLLLLLLLLLLLL.LLLLLLLLLLLLLLLL.LLLLLLLLLLLLL.LLLLL.LLLLLLLLLLLLLL\nLLLLL.LLLLLLLLL.LLLL.LLLLLLLLL.LLLLL.LLLLLLLLL.LLLLLLLLL.LLLL.LLLLLLLLLLLLLLLLLLLLLLLLLLLL.LLLLLL.\nLLLLL.LLLLLLLLLLLLL.LLL.LLLLLL..LLLL.LLLLLLLLL.LLLLLLLLL.LLLLLLLL.LLLLLL..LLLL.LLLL.LLLLLL.LLLLLLL\nLLLLL.LLLLLLLLL.LLLLLLL.LLLLLL.LLLLL.LLLLLLLLL.LLLLLLLLL.LLLLLLLL.LLLLLLLLLLLLLLLLLLLLLLLL.LLLLLLL\nLLLLLLLLLLLLLLL.LLLLLLL.LLLLLL.LLL.L.LLLLLLLLL..LLLLL.LL..LLLLLLL.LLLLLLL.LLLLLLLLL.LLLLLL.LLLLLLL\nL..L.L.LLL.LL.......LL.LL..LL.L...L..LLL........L.LL....LL...LL.L.....LL......L.L.L...L.L.L..L..L.\nLLLLL.LLLLLLLLL.LLLLLLL.LLLLL.LLLLLL.LLLL.LLLL.LLLLLLLLLLLLLLLLL..LLLL.LL.LLLLLLLLL.LLLLLL.LLLLLLL\nLL.LLLLLLLLLLLLLLLLLLLLLLL.LLLLLLLLL.LLLLLLLLL.LLLLLLLLL.LLLLLLLL.LLLLLLL.LLLL.LLLL.LLLLLL.L.LLLLL\nLLLLL.LLLLLLL.L.L.L.LLL..LLLLL.LLLLLLLLLLLLLLLLLLLLLLLLLLLLLLLLLL.LLLLLLL.LLL.LLLLLLLLLLLLLL.LLLLL\nLL....LLLLLLLLLLLLLLLLLLLLLLLLLLLLLL.LLLLLLLLL.LLLLLLLLLLLLLLLLL.LLLLLLLL.LLLLLLLLLLLLLLLL.LLLLLLL\nL.LLL.LLLLLLLLL.LL.LL.L.LLLLLL.LLLLLLLLLLLLLLL.LLLLLLLLLLLLLLLLLL..LLLLLL.LLLLLLLLL.LLLLLL.LLLLLLL\nLLLLLLLLLLLLLLL.LLLLLLLLLLLLLL..LLLL.LLLLLLLLL.LLLLLLLLL.LLLLLLLLLLLLLLLL.LLLLLLLLL.LLLLLL.LLLLLLL\nLLLLLLLLL.LLLLLLLLLLLLLLLLLLLLLLLLL..LLLLLLLLLLLLLL.LLLL.LLLLLLLL..LLLLLLLLLLLLLLL..LLLLLL..L.LLLL\nLLLLL.LLL.LL.LL.L.LLLLLLLLLLLLL.LLLLLLLLLLLLLL.LLLL.LLLLLLLLLLLL..LLLLLLLLLLLLLLLLLLLLLLLL.LLLLLLL\n..LL....L.LL...L.......L.L...L.L..LLL....L...L.L..L...L......L...LLL.L.L...L..L.LL...L..L..L.L...L\nLLLLL.LLLLLL....LLLLLLL.LLLLLL.LLLLLLLLLLL.LLLLLLLLLLLLL.LLLLLLLLLLLLLLLLLLLLL.LLL..LLLLLL.LLLLLLL\nLLLLL..LLLLLLL.LLLLLLLLLLLLLLL.LLLLLL.LL.LLLLLLLLLL.LLLL..LLLLLLLLLLLLL.L.LLLLLLLLLLLLLLLL.LLLL.LL\nLLLLLLLLLLLLLLL.LLLLLLL.LLLLLLLLLLLL.LLLLLLLLL.LLLLLLLLLLLLLLLLLL.LLLLLLL.LLLLLLLLL.LLLL.L.LLLLLLL\nLLLLL.LLLLLLLLL.L.LLLLL.LLLLLL.LLLLL.LLLLLLLLL.LLLLLLLLL.LLLLLL.LLLLLLLLLLLLL.LLLLL..LLLLL.LLLLLLL\nLLLL..LLLLLLLLLLLLLLLLL.LLLLLL.LLLLL.LLL.LLLLL.LLLLLLLLL.LLLLLLLLLLLLLLLL.LLLLLLLLL.LLLLLL.LLLLL.L\nLLLLL.LLLLLLLLLLLLLLLLL.LLLLLL.LLLLLLLLLLLLLLLL.LLLLLLLL.LLLLLLLL.LLL.LLLLLLLLLLLLL.LLLLLLLLLLLL.L\nLLLLLLLLLLLLLLLLLLLLLLL.LLLLLLLLLLLLLLLLL.LLLL.LLLLLLLLL.LLLLLLLLL.LLLLL..LLLLLLLLL.LLLLLL.LLLLLLL\nLLLLLLLLLLLLLLLLLLLLLLL.LL.LLL.LLLLL.LLLLLLLLL.LLLL.LLLL.LLLLLLLL.LLLLLLL.LLLLLLLLL.LLLLLL.LLLL.LL\nLLLLL.LLLLLLLLL.LLLLLLL..LLLLL.LLLLLLLL.LLLLLL.LLLLLLLLL.LLLLLLLL.LLLLLLL.LLLLLLLLL.L.LLLL.LLLL.LL\n.L.....L.LL.......L.L..L.L...LLL.L.LL..LLL..L.L....L........L..L.L...L....L...L.LL..L.....L..LL..L\nLLLLL.LLLLLLLLL.LLLLLLL.LLLL.LLL.LLLLLLLLLLLLL.LLLLLLLLL.LL.LLLLL.LLLLLLL.LLLLLL.LLLLLLLLL.LLLLLLL\nLLLLLLLLLLLLLLL.LL.LLLL.LLLLLL.LLLLL.LL.LLLLLL.LLLLLLLLL.LLLLLLLL.LLLLLLLLLLL.LLLLL.LLLLLL.LLLLLLL\nLLLLL.LLLL..LLL.LLLLLLL.LL.LLLLLLLLLLLLLLLLLLLLL.LLLLLLL.LLLLLLLLL.LLLLLL.LLLLLLLLL.LLLLLL..LL.LLL\nLLLLL.LLLLLLLLLLLLLLLLLLLLLL.L.LLLLLL.L.LLLLLLLLLLL.LLLL.LL.LLLLLLLLLLLLL.LLLLLLLLLLLLLLL..LLLLLLL\nLLLLL.L.LLLLLL..LLLLL.L.LLLLLL.LLLLLLLLLLLLLLLLLLLLLLLL..LLLLLLLL.LLLLLL..LLLLLLLLLLLLLLLLLLLLLLLL\nLL.LL.LLLLLLLL.LLLLLLL.LLL.LLL...LLLLLLLLLLLLL.LLLLLLLLLLLLLLLLLL.LLLLLLLLLLLL.LL.L.LLLLLL.LLLLLLL\nLLLLL.LLLLLLLLL.LLLLLLL.LLLLLLLLL.LLLLLLLLLLLLLLLLLL.LLLLLLLLLLLL.LLLLLLL.LLLLLL.LL.LLLLLL.LL.LLLL\nLLLLL.LLLLLLLLL.LLLLLLL.LLLLL..LLLLL.LLL.LLLLLLLLL.LLLLL.LLLLLL.L.LLLLLLL.LLLLLL.LL.LLLLL.LLLLLLLL\nLLLLL.LLL.LLLLL.LLLLLLLLLL.LLL.LLLLL.LLLLLLLLLLLLL.LLLLLLLLLLL..LLLLLLLLLLLLLLLLLLLLLLLLLLLLLLLLLL\n........LL..LLL.....L.L...LLL...L.........L.L...L..L...LL.L.....L....L....L.....L....L....L.LLL...\nLLLLL.LLLLL..LLLLLLLLLLLLL.LLL.LLLLL.LLLLLLLLL.LLLLLL.LLLLLLLLLLL.LLLLL.LLLLLLLLLLL.LLL.LL.LLLLLLL\nLLLLL.LLLLLLLLLLLLLLLLL.LLLLLL.LLLLL.LLLLLLLLL.LLLLLLLLL.LLLLLLLL.LLLLLLL.LLLLLLLLLLLLL.LLLL.LL.LL\nLL.LL.LLLLLLLL.LLLLL.LLLLLLLL.LLLLLL.LLLLLLLLL.LLLLLLLLLLLLLLLLLLLLLLLLLL..LLLLLLLLLLLLLLLLLLLLLLL\nLLLLL.LLLLLLLLLLLLLLLLL.LLLLLL.LLLLL.LLLL.LLLLLL.LLLLLLL.LLLLLLLL.LLLLLLL.LLLLLLLLLLLLLLLL.LLLLL.L\nLLLLL..LLLLLLLL.LLLLLLL.LLLLLLLLL.LL.LLLLLLLLL.LLLLLLLLL.LLLLLLLL.LLLLLLL.LLLLLLLLL.LLLLLL.LLLLLLL\nLLLLLLLLLLLLLLL.LLLLLL..LLLLLL.LLLLL.LLLLLLLL.LL.LLLLLLL.LLLLL.LL.LLLLLLL.LLLLLLLLLLLLLLLL.LLLLLLL\nL.LL.....LL......LL.L.LL...LL..........LL.LLL...L..L..LL....L......L.....LL.....LL.L..L.LL.L.L....\nLLLLL.LLLLLLLLLL.LLLLLL.LLLLLL.LLLLLLLLLLLLLLL.LLLL.LL.L.LLLLL.LLLLLLLLLL.LLLLLLLLLLLLLL.L.LLLLLLL\nLLLLLLL.LLLLLLL.LLLLLLL.LLLLLL.LL.LLLLLLLLLLLL.LLLLLLLLL.L.LLLLLL.LLLLLLLLLLLLLLLLL.LLLLLL.LLLLLLL\nLLLLLLLLLLLLLLL.LL.LLLL.LLLLLL.LL.LL.LLL.LLLLL.LLLLLLLLLLLLLLLLLLLLLLLLLL.LLLLLLLLLLLLLLLL.LLLLLLL\nLLLLL.LLLLLLLLL.LLLLLL..LLL.LL.LLLLL.LLLLLLLLLLLLL.LLLLL.LLLLLLLL.LLLLLLL.LLLLLLLLLLLLLLLL.LLLL.L.\nLLLLL.LLLLLLLLL.LLLLLLL.LLLLL.LLLLLLLLLLLLLLLL.LLLLLLLLLLLLLLLLLL.LLLLLLL.LLLLLLLLL.LLLLLL.LLLL.LL\nLLLLL.LLLLLLLLL..LLLLLL.LLLLLL.LLLL.LLLLLL.LLL.LLLLL.LLL.LLLLLLL..LLLLLLLLLLLLL.LLLLLLLLLL.LLLLLLL"
-      .split("\n")
-      .map(_.to(List))
-      .to(List)
-
-    def isOccupied(seats: Seq[Seq[Char]], row: Int, col: Int) =
-      row >= 0 && col >= 0 && row < seats.length && col < seats(row).length && '#' == seats(row)(col)
-
-    def countOccupiedAround(seats: Seq[Seq[Char]], row: Int, col: Int) =
-      LazyList((row - 1, col - 1), (row - 1, col), (row - 1, col + 1), (row, col - 1), (row, col + 1), (row + 1, col - 1), (row + 1, col), (row + 1, col + 1))
-        .count(rowCol => isOccupied(seats, rowCol._1, rowCol._2))
-
-    def loop(seats: Seq[Seq[Char]],prevSeats: Seq[Seq[Char]]): Seq[Seq[Char]] ={
-      if (seats.equals(prevSeats)) {
-        return seats
-      }
-
-      val newSeats =
-        seats.indices.map(row =>
-          seats(row).indices.map(col =>
-            seats(row)(col) match {
-              case '.' => '.'
-              case 'L' => if (countOccupiedAround(seats, row, col) == 0) '#' else 'L'
-              case '#' => if (countOccupiedAround(seats, row, col) > 3) 'L' else '#'
-            }
-          )
-        )
-
-      return loop(newSeats, seats)
-    }
-
-    println(loop(input, null)
-      .map(_.count(_ == '#'))
-      .sum)
-  }
-
-  def _11_2() {
-//    val input = "L.LL.LL.LL\nLLLLLLL.LL\nL.L.L..L..\nLLLL.LL.LL\nL.LL.LL.LL\nL.LLLLL.LL\n..L.L.....\nLLLLLLLLLL\nL.LLLLLL.L\nL.LLLLL.LL"
-    val input = "LLLLL.LLLLLLLLLLLLLLLLL.LLLLLL.LLLLL.L.LLLLLL.LLLLLL.LLL.LLLLLLLL.LLLLL.L.LLLLLLLLLLLLLLLL.LLLLLLL\nLLLLL.LLLLLLLLLLL.LLLLLLLLLLLLLLLLLL.LLLLLLLLL.LLLLL.LLL.LLLLLLLLLLLLL.L.LLLLLLLLLL.LLLLLL.LLLLLLL\nLLLLLLLLLLLLLLL.LLLLLLLLLLLLLL.LLLLL.LLLLL.LLL..LLLLLLLL.LLLLLLLL.LLLLLLL.LLLLLLLLL.LLLLLL.LLLLLLL\nLLLLLLLL.LLLLLL.LLLLLLLLLLLLLL.LLLLL.LLLLLLLLL.LLLLLLLLLLLLLLLLLL.LLLLL.L.LLLLLLLLLLLLLLLL.LLLLL.L\nLLLLL.LLLLLLLLL.LLLLLLL.LLLL.L.LLLLL.LLLLLLLLL.LLLLLLLLL.LLLLLLLLLLLLL.LLLLLLLLLLLLL.LLLLLLLLLLLLL\nL.LLLLLLLLLLLLL.LLLLLLL.LLLLLL.LLLLLLLLLLLLLLLLLLLLLLLLLLLLLLLLLL.LL.LLLL.LLLLLLLLL.LLLLLL.LLLLLLL\nLLLLL.LLLLLLLLL.LLLLLLL.LL.LLLLLLLLL.LLLLLLLLL.LLLLLLLLL.LLLLLLLL.LLLLLLLLLLLLLL.LLLLLLLLL.LLLLLLL\nLLLLLLLLLLLLLLLLLLL.LLL.LLLLLL.L.LLL.LLLLLLLLL.LLLLLLLL.LLLLLLLLL.LLLLLLLLLLLLLLLLLLLLLLLL.LL.LLLL\nLLLLL.LLLLLLLLL.LLLLLLLLLLL.LL.LLLLL.LLLLLL.LL.LLLLLLLL..LLLLLLLLLLLLLLLL.LLLLLLLLL.LLLLLL.LLLLLLL\nLL.LL.LLL.LLLLLLLLLLLLL.LLLL.L.LLLLL.LLLLLLLLLLLLLLLLLLLLLLLLLLLL.LLLLLLLLLLLLLLLLL.LLLL.L.LLLLLLL\n......L....L...L...L.....L........LLL..LL..LL..L.L..LL.L..L.....LL......L.LLL.L..L....LLL.......LL\nLLLLL.LLLLLLLL.LLLLLLLL.LLLLL..LLLLL.LLL.LLLL..LLLLLLLLL.LLLLLLLLLLLLLLLLLLLLLLLLLLLLLLLLLLLLLLLLL\nLLLLL.LLLLLLLLL.LL.LLLLLLLLLLLLLLLLLLLLLLLLLLL.LLLLLLLLLLLLLLLLLL.LLLLLLLLLLLLLLLLL.LLLLLL.LL..L.L\nLLLLL.LLLLLLL...LLLLLLLLLLLLLL.LLLLL.LLLLLLLLL.LLLLLLLLL.LLLLLLL.LLLL.LLLLLLLLLLLLL.LLLLLL.LLLLLLL\nLLLLL.LLLLLLLL..LLLLLLL.LLLLL.LL.LLL.LLLLLLLLL.LLL.LLLLL.LLLLLLLLLLLLLLLL.LLLLLLLLL.LLLLLL.LLLLLLL\nLLLLL.LLLLLLLLL.LLLL.LL.LLLLLLLLLLLL..LLLLLLLL.LLLLLLLLL.LLLLLLLLLLL.LLLL.LL.LLLLLL.LLLLLL.LLLLLLL\nLLLLL.LLLLLLLLLLLLLLLLL.LLLLLLLLL.LL.LLLLLLLLL.LLLLLLLLL.LLLLLLLL.LLLLLLLLLLLLLLLLL.LLLLLL.LL.LLLL\nL.LL..LLLLLLLLL.LLLLLLL.L.LLLLLLLLLLLL.LLLLLLL.LLLLLLLLLLLLLLLLLLLLLLLLL..LLLLLLLLL.LLLLLLLLLLLLLL\n.L.......L..LL.....L.L.....L..............L.L.L.LL..LL.L.L..LLL.L..LL.............LLL.L......L.LL.\nLLLLL.LLLLLLLLL.LLLLLLL.LLLLLLLLLLL.LLLLLLLLLLLLLL.LLLLL.LLLLLLLL.LLLLLLL.LLLLLLLLLLLLLLLL.LLLLLLL\nLLLLL.LLLLLLLLL.LLLLLLL.LLLLLL.LLLLL.LL.LLLLLL.LLLLLLLLL.LLLLLLLL.LLLLLLLLLLLLLLLLLLLLLLLL.LLLLLLL\nLLLLL.LLLLLLLLL.LLLLLLL.LLLLLLLLLLLLLLLLLLLLLLLLLLLLLLLL.LLLLLLLLLLLLLLLL.LLLLLLLLL.L.LLLL.LLLLLLL\nLLLLL.LLL.LLLLL.LLLLLLL.LLLLLL..LLLLLLLLLLLLLLLLLLLL.LLL.L.LLLLLLLLLLLLLL.LLLLLLLLLL.LLLLL..L.LLLL\nLLLLL.LLLLLLLLL.LLLLLLL.LLLLLLLLLLLL.LLLLLLLLL.LLLLLLL.LLLLLLLLLL..LLLLLL.LLLLLLLLL.LLLLLLL.LL.LL.\nLLLLL.LLLLLLLLL.LLLLLLL.LLLLLL.LLLLL.LLLLLLL.L.LL.LLLLL..L..LLL.L.LLLLLLL..LLLLLLLL..LLLLL.LLLLLLL\nLLLLL.LLL.LLLLLLLLLLLLL.LLLLLL.LLLLL.LLLLLLLLL.LLLLL.LLL.LLLLL.LL.LL.LLLL.LLLLLL.LL.LLLLLL.LLLLLLL\nLLLLL.LLLLLLLLL.LLLLLLL.LLLLL..LLLLLLLLLLLLL.L.LLLLLLLLL.LLLLLLLLLLLLLLLL.LLLLLLLLL.LLLLLL.LL.LLLL\nLLLLL.LLLLLLLLLLLLLLLLLLLLLLLL.L.LLL.LLLLLL.LL.LLLLLLLLL.LLLLLLLL.LLLLLLL.LLLLLLLLL.LLL.LLLLLLLLLL\n......L.....L...LL.L.L..L...L....L...LLL...LLLLL..L.....L.....L..L......L...L.L.......L..L.LL...L.\nLL.LLL.LLLLLLLLLLLLLLLL.LLLLLL.LLLLL.LLLLLLLLLLLLLLLLLLL.LLLLLLLL.LL.LLLL.LLLLLLLLL.LL.LLLLLLLLLLL\nLLLLL.LLLLLLLLL.LLLLLLL.LLLLLLLLLLL..LLLLLLLLL..LLLLLLLL.LLLL.LLL..LLLLLLLLLLLLLLLL.LLLL.L.LLLLLLL\nLLLLL.LLLLLLLLL.LLLLLLL.LLLLLLLLLLLL.LLLLLLLLL.LLLLLLLLL.LLLLLLLL..LLLLLL.LLLLLL.LLLLLL.LL.LLLLLLL\nLLLLL..LL.LLLLLLLLLLLLL.LLLLLLLLLLLLLLLLLLLL.LLLLLLLLLLL.LLLLLLLL.LLLLLLL.LLLLLLLLLL.LLLL..LLLLLLL\n.LLL...LL.....L.L..L.LL..LL......LL.L.L........L.L.....LLLL.L.L.......L......LL.....LLLL....L.L.LL\nLLLLL.LLLLLLLLL.LLLLLLL.LLLLLL.LLLLL.LLLLLLLLL.LLLLLLLLL.LLLLLLLL.LL.LLLL.LLLLLLLLLLLLLLLL.LLLLLLL\nLLLLL.LLLLLLLLL.LLLLLLL.LLLLL.LLLLLLLLLLLLLLLL.LLLLLLLLLLLLLLLLLL.LLLLLLLLLLLL.LLLL.LLLLLLLLLLLLLL\nLLLL..LLLLLLLLL.LLLLLLLLLLLLLL.LLLLLLLLLLLLLLL.LLLLLLL.L..LLLLLLL..LLLL.L.LLLLLLLLL.LLLLLL.LLLLLLL\nLLLLLLLLLLLL.LL.LLLLLLL.LLLL.L.LLLLLLLLLLLL.LL.LLLLLLLLL.LLLLLLLL.LLLLLLL.LLLLLLLLL..LLLLL.LLLLLLL\nLLLLLLLLLLL.LLL.LLLLLLLL.LLLLL.LLLLL.LLLLLLLLLLLLLLLLLLL.LLLL.LLL.LLLLLLLLLLLLLLL.L.LLLLLLLLLLLLLL\n.L.L...L...L.....L...LLLLLLL.LL....LL..LLL...L...L.LLL..L.L...L.L.L....L.LL.LLL..L....L..LL..L...L\nLLLLLLLLLLL.LL..LLLL..L.LLLLLL.LLLLLLLLLLLL.LL.LLLLLLLLLLLLLLLLLL.LLLLLLLLLLLLLLLLL....LLLLLLLLLLL\nLLLLLLLLLLLLLLLLLLLLL.LLLLLLLL.LLLLL.LL.LLLLLL.LLLLLLLLL.LLLLLLLL.LLLLLLLLLLLLLLLLLLLLLLLL.LLLLLLL\nLLLLL.LLLLLLLLL.LLLLL.LLLLLLLL.LLLL..LLL.LLLLLLLLLLLLLLL.LLLLLL.L.LLLL.LL.LLLL.LLLLLLLLLL..LLLLLLL\nLLLLL..LLLLLLLLLLLLLLLLLLLLLLL.LLLLLLLLLLLLLLL.LLLLLLLLLLLLLLLL.LLLLLLLLLLLLL.LLLLL.LLLLLLLLLLLLLL\nLLLLL.LLLLLLLLL.LLLL.LLLLLLLLL.LLLLL.LLLLLLLLL.LLLLLLLLL.LLLL.LLLLLLLLLLLLLLLLLLLLLLLLLLLL.LLLLLL.\nLLLLL.LLLLLLLLLLLLL.LLL.LLLLLL..LLLL.LLLLLLLLL.LLLLLLLLL.LLLLLLLL.LLLLLL..LLLL.LLLL.LLLLLL.LLLLLLL\nLLLLL.LLLLLLLLL.LLLLLLL.LLLLLL.LLLLL.LLLLLLLLL.LLLLLLLLL.LLLLLLLL.LLLLLLLLLLLLLLLLLLLLLLLL.LLLLLLL\nLLLLLLLLLLLLLLL.LLLLLLL.LLLLLL.LLL.L.LLLLLLLLL..LLLLL.LL..LLLLLLL.LLLLLLL.LLLLLLLLL.LLLLLL.LLLLLLL\nL..L.L.LLL.LL.......LL.LL..LL.L...L..LLL........L.LL....LL...LL.L.....LL......L.L.L...L.L.L..L..L.\nLLLLL.LLLLLLLLL.LLLLLLL.LLLLL.LLLLLL.LLLL.LLLL.LLLLLLLLLLLLLLLLL..LLLL.LL.LLLLLLLLL.LLLLLL.LLLLLLL\nLL.LLLLLLLLLLLLLLLLLLLLLLL.LLLLLLLLL.LLLLLLLLL.LLLLLLLLL.LLLLLLLL.LLLLLLL.LLLL.LLLL.LLLLLL.L.LLLLL\nLLLLL.LLLLLLL.L.L.L.LLL..LLLLL.LLLLLLLLLLLLLLLLLLLLLLLLLLLLLLLLLL.LLLLLLL.LLL.LLLLLLLLLLLLLL.LLLLL\nLL....LLLLLLLLLLLLLLLLLLLLLLLLLLLLLL.LLLLLLLLL.LLLLLLLLLLLLLLLLL.LLLLLLLL.LLLLLLLLLLLLLLLL.LLLLLLL\nL.LLL.LLLLLLLLL.LL.LL.L.LLLLLL.LLLLLLLLLLLLLLL.LLLLLLLLLLLLLLLLLL..LLLLLL.LLLLLLLLL.LLLLLL.LLLLLLL\nLLLLLLLLLLLLLLL.LLLLLLLLLLLLLL..LLLL.LLLLLLLLL.LLLLLLLLL.LLLLLLLLLLLLLLLL.LLLLLLLLL.LLLLLL.LLLLLLL\nLLLLLLLLL.LLLLLLLLLLLLLLLLLLLLLLLLL..LLLLLLLLLLLLLL.LLLL.LLLLLLLL..LLLLLLLLLLLLLLL..LLLLLL..L.LLLL\nLLLLL.LLL.LL.LL.L.LLLLLLLLLLLLL.LLLLLLLLLLLLLL.LLLL.LLLLLLLLLLLL..LLLLLLLLLLLLLLLLLLLLLLLL.LLLLLLL\n..LL....L.LL...L.......L.L...L.L..LLL....L...L.L..L...L......L...LLL.L.L...L..L.LL...L..L..L.L...L\nLLLLL.LLLLLL....LLLLLLL.LLLLLL.LLLLLLLLLLL.LLLLLLLLLLLLL.LLLLLLLLLLLLLLLLLLLLL.LLL..LLLLLL.LLLLLLL\nLLLLL..LLLLLLL.LLLLLLLLLLLLLLL.LLLLLL.LL.LLLLLLLLLL.LLLL..LLLLLLLLLLLLL.L.LLLLLLLLLLLLLLLL.LLLL.LL\nLLLLLLLLLLLLLLL.LLLLLLL.LLLLLLLLLLLL.LLLLLLLLL.LLLLLLLLLLLLLLLLLL.LLLLLLL.LLLLLLLLL.LLLL.L.LLLLLLL\nLLLLL.LLLLLLLLL.L.LLLLL.LLLLLL.LLLLL.LLLLLLLLL.LLLLLLLLL.LLLLLL.LLLLLLLLLLLLL.LLLLL..LLLLL.LLLLLLL\nLLLL..LLLLLLLLLLLLLLLLL.LLLLLL.LLLLL.LLL.LLLLL.LLLLLLLLL.LLLLLLLLLLLLLLLL.LLLLLLLLL.LLLLLL.LLLLL.L\nLLLLL.LLLLLLLLLLLLLLLLL.LLLLLL.LLLLLLLLLLLLLLLL.LLLLLLLL.LLLLLLLL.LLL.LLLLLLLLLLLLL.LLLLLLLLLLLL.L\nLLLLLLLLLLLLLLLLLLLLLLL.LLLLLLLLLLLLLLLLL.LLLL.LLLLLLLLL.LLLLLLLLL.LLLLL..LLLLLLLLL.LLLLLL.LLLLLLL\nLLLLLLLLLLLLLLLLLLLLLLL.LL.LLL.LLLLL.LLLLLLLLL.LLLL.LLLL.LLLLLLLL.LLLLLLL.LLLLLLLLL.LLLLLL.LLLL.LL\nLLLLL.LLLLLLLLL.LLLLLLL..LLLLL.LLLLLLLL.LLLLLL.LLLLLLLLL.LLLLLLLL.LLLLLLL.LLLLLLLLL.L.LLLL.LLLL.LL\n.L.....L.LL.......L.L..L.L...LLL.L.LL..LLL..L.L....L........L..L.L...L....L...L.LL..L.....L..LL..L\nLLLLL.LLLLLLLLL.LLLLLLL.LLLL.LLL.LLLLLLLLLLLLL.LLLLLLLLL.LL.LLLLL.LLLLLLL.LLLLLL.LLLLLLLLL.LLLLLLL\nLLLLLLLLLLLLLLL.LL.LLLL.LLLLLL.LLLLL.LL.LLLLLL.LLLLLLLLL.LLLLLLLL.LLLLLLLLLLL.LLLLL.LLLLLL.LLLLLLL\nLLLLL.LLLL..LLL.LLLLLLL.LL.LLLLLLLLLLLLLLLLLLLLL.LLLLLLL.LLLLLLLLL.LLLLLL.LLLLLLLLL.LLLLLL..LL.LLL\nLLLLL.LLLLLLLLLLLLLLLLLLLLLL.L.LLLLLL.L.LLLLLLLLLLL.LLLL.LL.LLLLLLLLLLLLL.LLLLLLLLLLLLLLL..LLLLLLL\nLLLLL.L.LLLLLL..LLLLL.L.LLLLLL.LLLLLLLLLLLLLLLLLLLLLLLL..LLLLLLLL.LLLLLL..LLLLLLLLLLLLLLLLLLLLLLLL\nLL.LL.LLLLLLLL.LLLLLLL.LLL.LLL...LLLLLLLLLLLLL.LLLLLLLLLLLLLLLLLL.LLLLLLLLLLLL.LL.L.LLLLLL.LLLLLLL\nLLLLL.LLLLLLLLL.LLLLLLL.LLLLLLLLL.LLLLLLLLLLLLLLLLLL.LLLLLLLLLLLL.LLLLLLL.LLLLLL.LL.LLLLLL.LL.LLLL\nLLLLL.LLLLLLLLL.LLLLLLL.LLLLL..LLLLL.LLL.LLLLLLLLL.LLLLL.LLLLLL.L.LLLLLLL.LLLLLL.LL.LLLLL.LLLLLLLL\nLLLLL.LLL.LLLLL.LLLLLLLLLL.LLL.LLLLL.LLLLLLLLLLLLL.LLLLLLLLLLL..LLLLLLLLLLLLLLLLLLLLLLLLLLLLLLLLLL\n........LL..LLL.....L.L...LLL...L.........L.L...L..L...LL.L.....L....L....L.....L....L....L.LLL...\nLLLLL.LLLLL..LLLLLLLLLLLLL.LLL.LLLLL.LLLLLLLLL.LLLLLL.LLLLLLLLLLL.LLLLL.LLLLLLLLLLL.LLL.LL.LLLLLLL\nLLLLL.LLLLLLLLLLLLLLLLL.LLLLLL.LLLLL.LLLLLLLLL.LLLLLLLLL.LLLLLLLL.LLLLLLL.LLLLLLLLLLLLL.LLLL.LL.LL\nLL.LL.LLLLLLLL.LLLLL.LLLLLLLL.LLLLLL.LLLLLLLLL.LLLLLLLLLLLLLLLLLLLLLLLLLL..LLLLLLLLLLLLLLLLLLLLLLL\nLLLLL.LLLLLLLLLLLLLLLLL.LLLLLL.LLLLL.LLLL.LLLLLL.LLLLLLL.LLLLLLLL.LLLLLLL.LLLLLLLLLLLLLLLL.LLLLL.L\nLLLLL..LLLLLLLL.LLLLLLL.LLLLLLLLL.LL.LLLLLLLLL.LLLLLLLLL.LLLLLLLL.LLLLLLL.LLLLLLLLL.LLLLLL.LLLLLLL\nLLLLLLLLLLLLLLL.LLLLLL..LLLLLL.LLLLL.LLLLLLLL.LL.LLLLLLL.LLLLL.LL.LLLLLLL.LLLLLLLLLLLLLLLL.LLLLLLL\nL.LL.....LL......LL.L.LL...LL..........LL.LLL...L..L..LL....L......L.....LL.....LL.L..L.LL.L.L....\nLLLLL.LLLLLLLLLL.LLLLLL.LLLLLL.LLLLLLLLLLLLLLL.LLLL.LL.L.LLLLL.LLLLLLLLLL.LLLLLLLLLLLLLL.L.LLLLLLL\nLLLLLLL.LLLLLLL.LLLLLLL.LLLLLL.LL.LLLLLLLLLLLL.LLLLLLLLL.L.LLLLLL.LLLLLLLLLLLLLLLLL.LLLLLL.LLLLLLL\nLLLLLLLLLLLLLLL.LL.LLLL.LLLLLL.LL.LL.LLL.LLLLL.LLLLLLLLLLLLLLLLLLLLLLLLLL.LLLLLLLLLLLLLLLL.LLLLLLL\nLLLLL.LLLLLLLLL.LLLLLL..LLL.LL.LLLLL.LLLLLLLLLLLLL.LLLLL.LLLLLLLL.LLLLLLL.LLLLLLLLLLLLLLLL.LLLL.L.\nLLLLL.LLLLLLLLL.LLLLLLL.LLLLL.LLLLLLLLLLLLLLLL.LLLLLLLLLLLLLLLLLL.LLLLLLL.LLLLLLLLL.LLLLLL.LLLL.LL\nLLLLL.LLLLLLLLL..LLLLLL.LLLLLL.LLLL.LLLLLL.LLL.LLLLL.LLL.LLLLLLL..LLLLLLLLLLLLL.LLLLLLLLLL.LLLLLLL"
-      .split("\n")
-      .map(_.to(List))
-      .to(List)
-
-    def isOccupied(seats: Seq[Seq[Char]], rowOrigin: Int, colOrigin: Int, rowDelta: Int, colDelta: Int): Boolean = {
-      val row = rowOrigin + rowDelta
-      val col = colOrigin + colDelta
-      if (!(row >= 0 && col >= 0 && row < seats.length) || !(col < seats(row).length)) {
-        false
-      } else {
-        seats(row)(col) match {
-          case '.' => isOccupied(seats, row, col, rowDelta, colDelta)
-          case 'L' => false
-          case '#' => true
-        }
-      }
-    }
-
-    def countOccupiedAround(seats: Seq[Seq[Char]], row: Int, col: Int) =
-      LazyList((-1, -1), (-1, 0), (-1, 1), (0, -1), (0, 1), (1, -1), (1, 0), (1, 1))
-        .count(deltas => isOccupied(seats, row, col, deltas._1, deltas._2))
-
-    def loop(seats: Seq[Seq[Char]],prevSeats: Seq[Seq[Char]]): Seq[Seq[Char]] ={
-      if (seats.equals(prevSeats)) {
-        return seats
-      }
-
-      val newSeats =
-        seats.indices.map(row =>
-          seats(row).indices.map(col =>
-            seats(row)(col) match {
-              case '.' => '.'
-              case 'L' => if (countOccupiedAround(seats, row, col) == 0) '#' else 'L'
-              case '#' => if (countOccupiedAround(seats, row, col) > 4) 'L' else '#'
-            }
-          )
-        )
-
-      return loop(newSeats, seats)
-    }
-
-    println(loop(input, null)
-      .map(_.count(_ == '#'))
-      .sum)
-  }
-}
