+++ src/test/scala/io/github/teonistor/adventofcode/_02Test.scala
@@ -1,0 +1,10 @@
+package io.github.teonistor.adventofcode
+
+class _02Test extends AdventOfCodeTestBase {
+  testAndRun(2, _02._1, _02._2,
+    "1-3 a: abcde\n1-3 b: cdefg\n2-9 c: ccccccccc",
+    2,
+    1,
+    "13-17 s: ssssssssssssgsssj\n7-9 p: pnlzhcppvl\n5-6 z: zzbwrv\n7-15 w: wwwwwcqwwwwwwwww\n8-9 h: hhhhhhhhsh\n12-13 l: mtrkqfllrglll\n3-4 t: nntt\n10-14 g: svgggggfhqggghlg\n3-10 x: djxxxpmcxx\n1-10 z: ztzzzgzzzz\n3-4 j: jjjjj\n3-4 c: fbhnsccbc\n2-4 q: thql\n15-18 w: tcwzpwzfwwqftvczbw\n5-7 l: lblwblbmllqnlbl\n3-4 b: bvbfnbb\n4-5 f: ffgzf\n7-9 c: cctcccccc\n5-6 k: szkgsk\n2-5 n: nqvnmgnqnsxfn\n2-10 s: dsfsbsssltss\n12-14 s: ssskrssssssfsxpsqsp\n13-15 j: jjjjjjjjjjjjjjjj\n16-17 l: jllclhllkgcpljzwk\n6-9 f: fsffwffbffbfff\n5-9 s: xtzscxsstcss\n6-9 h: hmnhshhhhqhrthmh\n3-6 x: xxxxxx\n9-11 m: mcmmmktmtmmmmm\n2-6 r: grrprrsmr\n15-16 v: vvvvvvvvvvvvvvwdv\n2-14 j: jmjdbhjjjjjjjxnjj\n5-9 r: rrwrrrrrr\n3-9 b: wbbcbvbtbbbbb\n2-9 r: rrtjlplkrgmrrrrxq\n9-12 s: sssshsmsssss\n2-14 t: qzxxvthtthtthv\n2-11 p: pvlsppwmgcspl\n12-15 s: dsjqsshdssbssstsssss\n3-9 k: tkkkpkkklk\n7-15 s: sssssssssssssgsssss\n6-9 x: xxxsxzxglpwxl\n2-3 x: xxmh\n7-9 g: ggsggggggg\n4-17 t: vtttrrrhtghfjvnqtttj\n1-2 n: jgnpn\n11-12 k: vlkkbkkkkklk\n9-12 v: vvvvvvvvwvvvv\n10-11 j: jjjjljjjjvm\n1-5 x: vvxlx\n3-4 l: flll\n10-11 q: qqqqqqqqqlq\n4-5 d: vdvhmpdtrdhdzfdsgk\n14-16 b: bbqbbbxbhrbpbtmhbbb\n18-19 d: dzpnbxngpgsjtnvtcdd\n2-16 v: jtrhtlfhvkbbflfq\n15-17 q: qqqsqqqqqqjqqqlqm\n5-6 q: tqrhtqqspdqvzqbx\n4-8 b: lbbzbwtv\n10-13 f: hrsnsfflfffsfjff\n10-14 k: kkjkkkkkkkkkkkkkkhk\n4-9 c: ccwcccccc\n5-7 r: grkrrrwjrrgbdtd\n8-12 p: tgspnljpfhtr\n5-6 v: vvvvvvm\n5-7 v: vvvvvvv\n6-8 q: qqqqqdqqq\n2-4 v: xfvqvlp\n3-4 g: gghb\n3-14 z: xflgzzrkzqzxzzxgd\n1-7 g: ldsgswgdgdmjlggb\n5-10 h: rhhmhhnbdhhqhwp\n16-18 m: mmmmmmwmmmmmmmmmmmm\n5-11 j: scssgcjjjjpzjjn\n13-14 s: wznksvkfvfskfs\n3-9 q: qdqqkqtbcnqqqsrklqbf\n7-9 c: cccchcmcvdcczwt\n15-16 g: gggdgggggggggggg\n1-5 k: kkkkkk\n7-10 q: qqqqqqqqqqqqq\n3-12 b: bbbrhbspbrbfbbqbbbb\n3-5 r: gczrgr\n2-6 h: whnlvhsftp\n8-20 h: fhhhsdfhhfzgmhpqhhjh\n15-16 z: sbzzzzzzdzjbwhgb\n4-8 h: jgtbmhwb\n8-9 s: hssksrnhlsdslssvcsbs\n2-6 v: vvvvvvv\n13-14 x: xxxxxxxxxxxxnsbx\n12-13 l: pgllllltpdfpxlll\n2-17 v: vqnncnxkssmhvkzstzgp\n4-5 r: rrrhw\n8-11 p: pppdppphppp\n6-7 t: lrmqtts\n8-12 h: hhlthhhhhhhhhtrzh\n9-10 g: jpgpmpcrwfkz\n12-13 w: wwwwwwwwwwhwwww\n5-13 l: jwtlqglllllxkllls\n2-4 c: ncpcgd\n3-7 l: llslllll\n2-5 n: jdfwnn\n3-5 z: xctxd\n4-6 l: lllrll\n6-7 q: fqqqqqqzqqqqq\n9-12 c: gccwgwqlcxccchchcrcb\n7-12 p: pppgpppppppps\n10-16 g: xrgfggzzghtgxgbggg\n7-11 d: tdddddqdwdgxldj\n11-14 p: pppppbpppppppppp\n1-2 j: jhjj\n8-9 v: pvvvvvvvcgvq\n5-6 n: qcrhnnkktxvdxcp\n7-12 x: lfnxxxdxxxxp\n10-12 s: sssssssssbsgs\n2-3 x: rqhzxmjnjxx\n13-14 w: wwwwvwwwwwwwqhww\n10-11 t: dgttttgttttt\n7-9 w: wdzwwwwjj\n4-10 v: vvsjvvbvvncvvvvxv\n6-8 w: wwwwwpwq\n7-9 j: jtjjjjrjmxg\n5-11 z: zzdgztshstzczw\n1-4 s: ssss\n3-10 p: pqppgrphrpkdmp\n9-12 x: xxxqxxxlxxsx\n3-4 v: nvswzv\n3-9 d: vdjvdcxdkdd\n2-10 v: vvjcvvvzvkb\n11-14 g: gggzggggggjgghg\n4-19 p: pppqppppppppppppppvq\n3-5 n: ngznkgdpn\n13-14 t: tttwtbtttttxxmttt\n3-7 q: qnqqqqqtqqz\n3-4 r: rkgr\n9-12 l: lqkfljzllqtl\n5-8 m: vpmmsnfmtxbcknlmpb\n15-19 s: ssptssnhsksnszsdpqss\n13-14 j: jsjjjjnjjjjjjjj\n6-9 p: ppppppppz\n5-8 r: tqzbnvgsvfrx\n7-9 h: hhhhhhhhp\n8-11 s: ssssssssssss\n3-10 t: ttttttttttttt\n4-5 g: ngtgg\n7-8 n: tnnnnnnqn\n12-16 x: bxxxxxzxxxxxxxbxx\n1-3 z: zzzt\n11-14 x: xxgxxxxxxxwcxfx\n6-11 v: rgkkhpkjvvf\n16-17 n: nfgnnnnnnnnnnnngvn\n3-7 w: wkwwwww\n5-16 g: ggggcggwgggggggggg\n1-3 l: sjvjlkzrpn\n17-19 p: pppppppppppppppptpd\n2-6 b: vjjwbqbdbpwsbw\n4-8 g: gggggjvgd\n2-6 w: pzzpwdm\n3-10 g: ggggcggggggg\n1-3 v: vxvjcvcpslpfn\n9-10 r: qrrwrrjrrr\n5-10 m: qbmdzvqmxnnmmmtk\n2-14 q: vqttpdjhqvkqzqxz\n12-13 r: rrxbrrrrrcrptjrr\n13-15 s: cscssssssssssssss\n7-10 z: zxbrzxbdbhzzzfzz\n2-16 n: skqkxqmqbvbrnspnrgq\n14-18 b: bbxsbbdpbnbbbnbbbpbb\n5-8 b: bmcrdhbbk\n11-13 f: fffffffzfffwffb\n5-6 w: wwbwlww\n12-14 d: gzxclqbnmnxjlzd\n13-19 z: fzzzctzzdsbgzzzzwzpb\n2-12 w: lcwwtwwpwffdjqwms\n9-10 x: hznxnxxxxxxxgkxx\n2-4 g: cwfjtlgf\n11-12 k: bqfkkkklkkkkb\n2-4 l: llllgxk\n5-9 m: cpwmqkzhxs\n8-15 b: bbwlwkblvbwdvrjbgsb\n3-4 f: ffmff\n4-10 c: nqxcqcfcqc\n10-12 d: ddqddddddtdhd\n3-4 n: dghnnqnn\n6-15 s: rsbxbrfmssdknppw\n2-13 m: mmkmmmmqdmkmmm\n8-12 n: nnnnnnnnnnnn\n2-5 d: mdkwd\n14-15 l: hllllllllllllgl\n6-11 g: jgccfnggggpgkt\n4-6 j: jjjtjjjjjj\n9-12 v: hgpjjlvhvhljb\n2-4 g: jfwdgfr\n4-5 f: fhffffchf\n16-17 j: jjjjjjjjjjjjjjjjj\n1-4 t: tcttwtttttrkt\n6-13 x: qmxfxxwxrxdxxxfvgv\n6-7 c: cckpccc\n3-10 m: qwmmpknsgmdmfvrcvw\n2-4 r: rszv\n1-7 q: qqqqqqqq\n1-6 l: lzlmjl\n11-12 q: mqqqqxbqqrqqqdqqbkq\n8-11 h: jtvbhxhhphhhh\n9-11 b: bbbbfbzbzbb\n5-7 l: gljssxddzscvhwrlrb\n3-6 c: fxcbccddxxcn\n3-5 s: ssssz\n17-18 n: jctqdsjchwknkffpnn\n8-10 c: crtwbwcfcqkf\n3-5 r: qrzqc\n7-12 z: zzzzzzzzsfzsz\n13-15 p: pppppppppzrllplp\n11-12 h: nhhhhjhqhhplhhh\n4-5 l: lllll\n4-10 g: gggzgggggg\n15-16 h: qhhhshrhhfhhhhch\n5-7 k: jsrvnjlkq\n16-17 p: pppplppppppppppml\n13-14 m: xqhmjdmrmmmffr\n6-11 c: cccccgccccrccccccc\n9-10 l: llllllllxllt\n8-9 b: bbbpbbbwbdnkbb\n5-9 m: sfmmpmmmzn\n5-15 k: kkkkkkkkkkkkkkkckkkk\n3-4 h: hhhh\n15-19 g: gggwgggrslgxjgfgbfs\n12-13 d: dddddddddddxb\n6-14 k: kkrgdvhscffwwm\n14-16 m: zmmmmqmmmmmmmgmm\n8-9 w: wwwwgwwww\n1-12 x: xxxxxxxxxxxx\n1-5 b: bwrbbfwgh\n1-3 t: ftttttttttttttt\n10-11 m: mwmmmwkmxsf\n8-9 t: tqtttttttdtttt\n7-8 z: hzbkpzfg\n2-3 h: htblhdhh\n3-4 p: pptnlpc\n4-7 l: lqlnqhljt\n12-14 n: nnnnnnnnnnjnnt\n1-4 d: dszd\n12-13 f: fvrffffffffkfff\n9-15 t: bkzstqvtgtwxtzqttd\n6-7 v: vrvvvkgv\n6-9 n: knbpnnjnnsn\n5-8 n: nznkzzqrbn\n9-12 s: ssscssssgsqwsspssp\n4-16 q: slqqqqqhqqqqqqdqc\n1-3 d: gdttdd\n2-6 h: hhqzmh\n4-9 k: lkckkpbzkl\n8-17 b: bbbbkbbpbbbbbbbbn\n1-2 r: bqmrw\n8-9 s: ztssqssbz\n1-13 s: stbqmhspsdgjs\n8-10 z: trnsmlxzmzzzz\n6-8 v: vlxvgzbzvvv\n3-6 c: xddhwtcbpcf\n13-19 g: rjvsgpnfjvzlxxzzgxgm\n5-9 m: qffpmxmfmk\n14-15 c: dcfwcpxcmvpxxtcbmzhc\n6-10 k: jkpbkvghpxksfkkk\n2-3 j: jmgjjjj\n7-12 z: szrznmkjxnzj\n4-9 s: ssssssssss\n2-4 h: hkqvfgpsxlnhtrqr\n6-11 g: gggglrlgggwv\n1-6 v: vrbvlvrvv\n4-10 j: jjjjjjjjjz\n3-12 k: kkjkkkkkfckgkkkkk\n3-15 x: xxrxxxxxxxwxpxg\n5-8 v: vvhvvdzvqfrvgqvt\n5-17 z: zznzzzzzzzzzzzzvf\n4-12 n: fdrncrggjlbgbmwv\n6-17 d: tcddhdbdvtsnrdstd\n1-10 m: mqljmgchlmsx\n4-6 l: fbllbl\n8-9 p: kpmtpphwp\n3-6 j: sjjjjj\n5-6 n: nnnnbs\n7-12 p: ppqppngcphpk\n5-15 x: xxxxgxxxxxxxxxqx\n1-3 p: ppppm\n3-6 w: wwwkwww\n1-4 r: rrrx\n3-4 z: zzznp\n12-14 n: tnzwtmnnnnndnfnnn\n9-10 d: ddfdddtdddd\n4-14 l: hvllsgncmhxhwl\n10-15 j: kxftfsmjvjgslsjpp\n2-10 g: gtgvggqggggxggdg\n1-9 v: cvnxhhffnbcv\n1-14 c: ccccccccccccccc\n3-8 q: fnqmkbhqtbjqvczq\n2-4 v: fvphvgv\n10-11 k: kkkkkkkkkpt\n1-10 d: dddldkddmdd\n4-6 x: xxxxxxxx\n4-11 k: kxkfkkkskkrkmk\n1-6 n: nlmngnkqww\n1-9 j: jjjjjjjjsj\n5-6 b: bbbbbb\n9-11 q: qqnqqmqqqpq\n4-7 r: drrrrsrrz\n10-15 n: nnhncnpsnxtnpnqnnghn\n3-6 d: ddddddd\n4-7 x: bxvxbxjwxgx\n1-2 n: cmpwsbjzghgndj\n2-4 r: przczfjfrbwj\n3-17 k: qbvmqxxffdfpkwxdgxv\n7-12 v: llmjvscvjvpvp\n2-7 z: vzxzxxz\n9-18 r: srdrjcrrrvhrcqrrrr\n7-8 f: rffffvtrfsqff\n5-8 w: pdcwwvmwwwkzwwjwxks\n3-4 c: cccc\n15-19 c: cccccccnccclbnplccqc\n2-13 f: fffffffffffffff\n3-9 t: thrkzdttb\n6-11 p: pppppgppppjp\n5-8 q: qqqrqfqs\n7-11 r: crprgtwwrzp\n2-10 f: qfhgfftfdbfnlffffff\n4-5 h: dhvvlcw\n3-6 s: sffsfws\n2-7 b: jrvxbhbp\n10-11 w: wwwgwwwwwkb\n5-9 d: dddddddddddddkdv\n4-14 b: bbbbbbbbbbbbbbb\n4-8 x: xxxxdxbqxc\n5-9 l: phlkhlllbc\n1-2 s: ssgskqvrsrwdt\n19-20 x: xcxqxvxxxtstnqxhxxxx\n6-7 x: gxxxxgb\n4-5 c: cccccc\n3-10 q: pzzjqjqpfbqws\n4-10 t: ttdttsccttpt\n14-18 c: ccrccccccccccccccrc\n2-5 x: xxzqx\n5-11 c: qbckpccckccccws\n5-6 g: gggggg\n3-5 n: wknbnnlnnq\n9-13 j: jjxjjjnjjjdqjjj\n5-9 s: smsxsrfss\n2-4 v: vxphvvk\n6-7 x: xxxqxxbbx\n1-3 g: zrgtghjwd\n4-5 n: nmntwnn\n6-8 g: ggznggggg\n4-5 k: tkkzs\n9-10 g: hqdhggggqggk\n12-16 w: wwztwmwwspwhmwwz\n3-5 m: mmmvmjmmm\n3-4 j: jjjj\n6-16 z: kzbtzqrkzzcfhvpwjszx\n3-9 f: ffwffffffff\n13-15 b: bbbmsbbbbbbwbbbbwb\n11-14 n: nnnnnnnnnnnnnnv\n8-9 v: bvvflvvbjvbtjqrvsxf\n1-5 g: gjgnggg\n15-16 n: nnnnnnnnnnnnnnwvqnn\n12-13 l: lllhllrlllllllll\n5-6 l: llmlll\n8-9 z: zzqzdzvzzzz\n3-4 d: ntdd\n4-5 b: sbndbkjxldqpzfbzbqx\n2-5 t: bwcfzdlfslsd\n4-5 d: dddddd\n1-3 t: jtpt\n10-12 q: qqqqgqqqqqqmqq\n10-19 m: mmmmmmmmmsmmmmmmmgs\n2-14 x: xxxxxgxvxmxxxbx\n3-13 l: tzljvxspwxdnl\n6-10 p: lhnppkpswrppgppzpppn\n2-6 j: djgwjjhvpwkcgsfbf\n10-11 p: ppppppppprpppppp\n7-12 l: lvgcbnlclvjlk\n6-11 z: zzzzzzzzzzn\n3-5 x: xzvhxnpwbxgxx\n4-6 h: hhlrhnl\n8-10 b: brngbslhqhb\n6-11 s: ssqjxsssdsssss\n2-5 s: dbjdjknspbdldfbjd\n10-16 z: ctzzzzznzcwzztzl\n14-15 j: jjjjjjjjjjjjjjjj\n9-13 d: qvdpwddddddddd\n10-12 c: ccccchcccccq\n3-5 x: xxxxx\n13-14 t: ttstttztttttltt\n3-16 k: qvfjklgzlwfwnbjklqdx\n12-13 g: ggggggtgggggggvg\n11-15 j: vqzjvzvxjhjnjtj\n7-8 t: tttttttg\n6-7 x: xxxxfmn\n10-13 m: mmlmmgvfmmmdmmcmmq\n7-14 w: wwkwwwwwwwmptwww\n17-18 b: bbbbbbbbbbbbbbbbbb\n2-8 c: kbvcnrpx\n16-17 n: nnnnnnnnpnnnnnngl\n10-15 k: kbkjkdkwpkkhkkksklx\n2-12 p: pjhdkvwdltvqrhm\n2-4 x: vxhxxk\n3-5 m: mmfmf\n6-8 n: ngnggfnwnn\n9-16 b: bnbbbtcbbbvbbrbbbbb\n2-8 p: ppphpzppcgptpprhpp\n9-12 x: xxxxxxxxxxxx\n5-7 d: dknbddd\n7-9 r: rrrrrrrrj\n3-6 f: fjmfxk\n3-4 n: jnnnbjckggnpjnddhx\n6-12 d: dznpsfdtdsdndjddvcdd\n7-12 x: xxxxxxhqxxxbxbsx\n1-2 k: kkfkkkk\n4-7 l: lprlqllll\n10-13 x: qxxqkjxxxxxxmz\n3-8 f: rxfffwcff\n4-5 s: cpscszdssc\n5-6 t: tttftt\n2-6 p: kpvtkprdqhxzpxrbfp\n6-12 s: ksnsklwqfssdsbsss\n6-7 l: lzlflfl\n7-9 t: ttktttttt\n8-14 d: hdddzndhsddpfpdxt\n8-10 d: ddnddrddvxbzskddzd\n11-13 v: mvhvrtvvvvzvrvdv\n4-5 g: gggjw\n3-6 m: xhmrpm\n7-13 k: kkkkkkkkkkkkk\n10-13 n: nnnnnnmnngnnnnnn\n7-15 m: kcmtmjtmwjmcsbbms\n13-16 j: jjjjjjjjjjjjpjjlj\n2-4 n: nnmnkn\n12-14 j: jjjjjjjjjjjdjj\n1-2 c: cchc\n9-12 m: mmmqbmlwmmjmzmmml\n3-8 r: rrkkrrrt\n4-5 k: wckknxkpkktmks\n9-10 d: ddddpdddnpx\n1-8 q: qmpmxzfqnqctwpprm\n10-12 q: sgkfxhqjkqqq\n5-11 p: rwpppdlmhpptkbrlp\n1-2 d: lzrbxnlcsqhssgdpvjs\n5-6 c: scccck\n4-5 s: ssshxsssws\n3-11 l: nslwgmxmtblzttxxg\n10-11 z: zzzzzzzzzzz\n7-8 j: bjjjjjqdbjj\n7-8 d: glpfjbpdxd\n2-4 k: wkvwrk\n1-5 c: vtcgcbc\n16-17 x: xxxxxxxxxxxxxxxxm\n2-11 f: ffffffsfffffffffffff\n4-8 m: mrmmmmmmmmmmmm\n2-3 x: xvzxx\n2-4 h: hxpn\n13-17 x: xxxxxxxxxxxrxpxxxx\n4-13 q: wcqqtwwmwqqtqmqtqrqq\n10-12 d: dddtdqdnjxdzdfdd\n13-14 f: jffffsffmffqjff\n6-7 d: sxdvddddwqxd\n5-11 d: ddddfdddddcdddd\n11-14 b: bbbbbbbbbhlbbj\n3-4 p: pfpp\n4-7 j: jjjgwkgwnfffvfzfl\n6-9 l: gknrmjgxtlzcbvkgxb\n4-10 n: nnnnnnnnntnnnnnnxn\n1-4 r: rjrr\n3-4 m: nkmr\n9-10 f: fffffffffff\n2-6 r: rnhtml\n3-12 q: qqkqqqqqqqqkqqqqq\n3-10 r: rrnrrrrrrrrrrrrr\n6-8 m: bmmmdmmmmdmh\n10-12 g: ppdvvxgkhfskfqc\n2-10 v: txpvkgvcbqvqfk\n1-5 p: tqcvbtp\n6-7 z: zzzzzzz\n7-16 z: fdjbbkrzjzdbdblhxb\n2-4 x: xlkz\n4-11 m: glhknzmvqzfmnsbn\n6-8 q: qhmdgqvq\n8-16 t: tttdtttdtttttttbvt\n4-7 v: vptcmgvvjvvvfvvmht\n7-8 t: ttttttngt\n5-15 t: ztgbttrntttdvgtptjdb\n10-12 b: bvndbbbbbbbbbbbbtb\n2-3 t: sbqts\n12-16 w: wwwwwwwwwwwwwnwww\n2-3 m: kmmr\n12-15 k: kkmkrkkkkzkdkrgkk\n5-13 h: hhhhzhhhhhhhzhhhh\n16-17 l: lllllllllllllcllq\n1-3 b: bbjb\n2-5 t: vtjhmf\n4-5 q: vzdqfd\n2-3 n: nnnn\n7-14 j: jpjjjjjjljjjjjjjj\n12-15 j: crsjpjjjqjjfjkv\n10-11 t: ttttttttttttt\n1-2 j: jjbjsjvhbfssznf\n9-13 f: ffmfffffzfffgfff\n15-18 g: sgblrqdftpwzggvgqt\n19-20 r: rrrrrrrrrrrrrrrrrrvj\n1-4 j: mjlt\n3-10 x: jxxxxxxxxgxxx\n6-7 p: ppptppp\n7-8 b: jbmnbmsc\n2-6 f: zshfcbhsh\n2-5 r: rrrqwtrvclf\n3-8 g: chgtwlbg\n11-15 n: knnhnnhnnnpnlknnkhnn\n3-8 z: zkqlzzpmjph\n5-10 b: bbbbhbbbbbbbbbbbbb\n6-15 c: ccccdkrncwrcclc\n6-7 c: ckcccwdhcc\n12-13 t: ntttktttttftttt\n5-8 j: jpjzcjjjjj\n7-8 k: kpkkksbkkkkkr\n9-12 s: sqssssbssxcs\n3-7 p: pbbplhtfpktpnppx\n5-7 n: ncnnnsn\n4-6 s: sssrss\n3-16 v: kvvvvvvczvvvvvvvkww\n3-6 k: kkkpmk\n8-9 h: hhhhhhhmb\n7-9 c: czwcccmtckccc\n5-7 t: tstglltkt\n2-6 c: cccccv\n11-14 z: zzzkzqzzzzzvzzz\n3-4 d: ddch\n4-6 j: jjjjjvj\n5-9 c: cccccccccccccc\n7-9 j: hjkwpjjbjfsq\n4-6 t: tttlcdb\n13-15 z: zzzzzzzzzsqzzgzzzkz\n8-10 t: ptckdjtsptlmzrktwcw\n2-5 n: tnbhnnkvnq\n11-12 m: fqmmmmmqmtmkcmmvnwmm\n7-8 z: zzzzztvl\n6-7 j: jbpjjjjpxmqxcbwsjrjj\n2-8 t: ttkdzfwdtflfswlkntt\n6-11 j: rnjjjjjjjjj\n10-13 f: fwhkbfdfffvfs\n5-8 l: pflplntf\n3-5 r: rrrrr\n16-20 g: gggggggggggggggggggg\n1-12 t: xtnsxbjtttxtt\n7-9 r: zrrrrrkrnrrrrrrrrr\n2-11 n: nrjnnnwnnprnnn\n10-12 z: zzzkxzzhznzdz\n7-13 d: pdcjdddqddlddcdkdxk\n10-15 r: rrrbrrrmvrcphrrqr\n1-3 l: rsll\n7-8 x: xxxqvrxx\n5-14 p: ppcdvppbppppjpbg\n5-11 x: dxdxdhkxxvxqxxxfx\n6-12 r: pvbfrhqrhftqrrxcrr\n4-7 t: zdbtzst\n1-13 k: kkkpkghkqkskk\n5-17 g: gvggqgggggqmgxggfgg\n12-13 l: llllwllllllwf\n2-11 b: bbbbbbbbbbbbb\n7-17 h: hhhhhhhhhhhhhhhhhh\n2-3 w: lwww\n2-3 s: vstcs\n4-6 s: vssssxss\n14-16 r: fmrrrrrrrrrrwwrvrr\n6-7 g: dgwgggp\n12-16 g: wggxgggrggggzgbgggg\n7-9 s: tjhxrscnsscssssnms\n3-12 n: nnnnnnnnjnnlnnnnn\n7-8 b: zncbmbglqbbbbpgm\n12-18 l: lllllllllllvlllllwl\n10-16 q: qqqgqfqqqqqrrqgqq\n1-7 n: nnnsnrnnnp\n10-11 q: qqqqqbqqqqq\n5-8 k: bnsfzkkwfkknkccwqkm\n4-5 c: ccdcccc\n5-9 z: nlsgzzzdz\n2-3 k: xkgxhqkpftx\n10-13 d: ddddddddhdddddddd\n4-6 x: xmjxwxx\n14-16 z: zzzzzzzzzzzzzrzq\n3-14 n: qnznjvlzdnnrgwfr\n7-13 s: sssxsspsssssrz\n4-11 k: pkkkcklbkkkk\n9-11 x: xxxxxxxxxxtx\n4-6 z: zzjzzfz\n3-6 d: zddddddz\n10-20 c: smccccvcpbccbqcxbccq\n13-15 m: mmmmmvmmmjmnnmlpsm\n12-13 j: jjjjjjjjjjjjj\n3-5 m: pmmbm\n7-17 m: mmmmmwcmmdfqrmmhlmm\n3-17 g: gcggnggbgdggggggg\n10-16 b: sqpsqkfbnbqzswbb\n3-4 w: wwwn\n5-6 d: dddddpd\n12-15 m: mmmmmmdmmmqlmnvms\n13-15 j: jjjjjtjjjjjjzjcj\n17-18 t: ttttttttjtttttttttt\n17-18 d: dcddmchdwbqrllsxjdv\n16-17 h: hshghhndnhhfhhhhh\n3-5 g: ggggch\n6-13 c: ccvccpccccccvcccccc\n2-4 h: hbfh\n3-9 l: lllptphml\n2-3 s: xwrqssfzr\n9-13 m: mktmmwmmmmrmmrqmmmcm\n8-11 l: rkvxvrglltlc\n1-8 q: jfqqxngqnqchq\n2-6 x: zxgjxx\n5-14 g: mlbgvdglrgcqjgz\n4-5 t: tqgtv\n2-6 s: ssfnzs\n9-11 m: mxwhmmmmmmzm\n1-5 d: dbhjzd\n2-16 f: fffffffffffkffffff\n6-10 b: dbbbbbbbbbbb\n5-6 x: xxxmxxcx\n7-8 l: lllpllcb\n16-17 d: dzdddddddddddddddd\n3-10 t: tklkqftnwcj\n9-15 g: jccxpmfjgntpptgkkcvt\n1-8 r: sdrwgcvprt\n4-5 v: nhbwx\n17-18 z: zzzzzzzzzzzzzzzzxsz\n5-13 l: cbdklqhnklkmwhpp\n4-5 f: fffff\n1-4 q: qpwq\n9-11 x: vxxgrgxxfvvgx\n3-4 r: rrfd\n3-16 x: sxqxxxxxxxxxxxxwxxx\n5-8 s: ssssssss\n5-13 b: jbbhqqbblbbbbfcb\n17-18 p: ppppppppprpppwppphp\n3-9 p: kdpnpfphppvffpwf\n7-8 r: jvrrrrrnrpqrn\n5-6 m: ncmmdxmm\n10-11 v: vvvvvvvvvdl\n3-8 g: jlsscggg\n1-3 m: gwms\n4-5 f: fpkvnf\n6-15 j: zfdrxjgxtbkbbjjctmsk\n5-6 v: qnnbvvwqvbwqblqd\n2-5 v: kvhvt\n2-6 x: lxrgsxg\n6-8 s: sssxswsjsssxss\n1-6 t: dxwjttnqkt\n13-14 s: sssssssssssvjg\n1-2 t: rtttnf\n9-17 h: hjhhhvhhbhhwhhhhphd\n2-3 m: mmmwcm\n10-12 v: vvvvvvvvvvvv\n6-10 j: flcjtzwhwsnjjjrjrj\n3-8 d: ddmrddddd\n5-8 c: ccccslcrwrc\n13-14 s: hpsqtxvkrsssshh\n11-15 t: hrxtjgtwtlmpqfx\n11-14 l: llbxlsllllqrqlllzlb\n3-4 c: crlcvcxwd\n3-9 s: vjbsxhwwdvshfxstc\n7-8 f: jfhnffbgsfjfwf\n15-16 x: ksxqvfsvnxsgvwgpjzl\n16-19 g: gggggggggggggggdggg\n9-11 h: hhhhhchhzmqhh\n2-6 f: kfdlcfrxftzgq\n2-8 p: ppnppkpwwmgp\n1-2 g: ggfgrg\n1-7 n: nnnnnnqwnnn\n5-6 h: hzjgthqphwhnjh\n10-15 l: xlllllllvlllllll\n3-4 l: lzjb\n4-6 l: ldzlql\n7-12 b: mzlbqzbrqjjbddr\n10-13 t: tttttxtttpttt\n4-6 d: dddddd\n4-5 k: kzvkkrkf\n1-2 h: fqml\n17-18 f: jhtfdfgmnchprbwbfrf\n3-10 s: sjsnsksbdsss\n1-17 j: jgcbjmjjjjjcxsjjjjhj\n15-16 k: tkjkjbkkrkkkklmgkj\n2-13 k: kkkkfkqkkkkkkkkpkk\n4-13 d: ddkkdtddzdnnddjdd\n1-3 v: fxvnxdvxnrbjs\n8-11 s: sssssssssspssssss\n1-8 k: khkkkkkkhk\n8-9 k: kkktmkkzk\n4-9 w: wrqdfwxwnfwwwnmzww\n3-5 l: lgtcvl\n13-18 n: nnnnnnnnnnqnnnvnnnn\n15-17 z: zzzzzzzzzhzzzzqzv\n3-5 q: lxnqsvk\n8-13 g: rzzwdlzgbcmggct\n10-14 d: ddvdbddddnddpd\n6-10 d: dphsszqrvz\n8-16 b: bnbgfbkbbbqlbcbbgqbb\n13-15 d: dddjddmqndddddddvjdd\n8-11 q: tnhgqkqfpjhwqgktq\n1-3 n: nnnbn\n11-12 f: fwfffbfcvflff\n7-8 s: sssglsvssvls\n2-5 z: zzzzzzzzzqjzz\n2-11 m: mhzrwphzxgj\n4-6 k: kkkwkd\n8-15 m: mrmmmmmzmmmmzmgmm\n1-7 w: wwwwwwxwwww\n1-3 l: nllllllll\n4-5 m: zbshj\n3-9 x: xssxqxxxsxkzx\n5-9 d: dnpfhsdmrxbvgxqrs\n7-12 r: vhrgcnxrvksg\n8-10 c: ccccccctzv\n16-17 l: lllllllllllllllhl\n6-15 n: nnnnnnngnnfhwwzsnnn\n6-19 z: zzzzzlzzzzzzzzzzzjv\n7-12 b: bxfbbwsjtbbvbqvbmbpb\n5-14 g: gggcjggggggggmgg\n5-8 d: wdddddkdpd\n4-17 x: xtxxxxxxxxxxxxxxmkx\n4-10 c: cccccccccjcscv\n9-11 v: ttjzfjldvvmswpqt\n5-8 r: tjrwhzgfrkgfq\n5-8 m: cvhsmnzmncsfmbqmm\n7-8 z: krctfpzz\n2-6 w: mlnlswvh\n12-15 t: zrntxzttqlthfdttt\n1-8 r: pkcsjkrrrzxxsfnjw\n5-15 p: xbhpkpnngpqvpcwdppg\n3-4 t: tttt\n8-11 j: wjjjmjjjjjqp\n8-14 h: gjslnklhhkhthxh\n2-6 q: qqlqqq\n3-6 p: pppgppppppppppp\n7-11 l: llhllllrllllxldllll\n11-12 d: lddkdddddddddd\n4-12 z: zwlzzqsvvclw\n15-18 c: tflkxflqccwkmlckck\n8-12 h: hhhhrhhhhfwhhh\n6-9 r: rrrmrjrrvkrrjrdr\n8-19 k: tmdlkcktbkkskvkbkbz\n1-13 s: ssssssssssssss\n4-5 v: xtcfvx\n1-15 s: shtssssswghsfsss\n2-4 k: krkl\n7-12 q: dxqqqhqqqtqj\n1-5 r: rrrvvrrrr\n5-9 t: ttrttftxztnhtvkcmtth\n7-9 g: gngkcggrg\n12-13 z: zzzzzzzzzzzbk\n3-4 l: lslllc\n4-5 t: pqchpk\n5-6 h: hhhhml\n15-16 h: qhhhhhhhhhhhhhhhhhmh\n5-6 d: ddddfs\n1-6 d: dthfdzddfnsdddm\n14-15 d: ddddddddndhtddsdnddh\n5-7 j: qjjjpjplljjc\n2-6 q: qqqqqqq\n3-4 b: bbbb\n5-6 s: ssssjjs\n1-6 b: bbbbfb\n7-13 m: cpvmbbnnrmzwf\n2-4 f: fnhn\n6-7 d: gdktlddrdlvmqdtddpzv\n4-5 w: cwlww\n6-9 z: zmdrgzlmzx\n5-10 m: mjbjmmmlmmwxm\n2-4 s: ssnhht\n8-16 t: tttttttdtttdtttjtt\n3-5 z: wxgzhtswb\n13-14 l: lllllblpllllchll\n4-8 g: vggqzvggnggggggggg\n8-9 h: hhthkshhpnhqhgh\n4-7 l: hglxklpl\n2-3 w: mkmgkwzwmw\n6-7 h: hhhhhrz\n4-6 f: tdffff\n1-7 l: lmfhgqschjqglrvwwnnz\n12-13 g: gggggggggggfhggg\n2-10 m: ntmmmdjmmmsvmm\n1-2 s: sqsssssssssss\n3-8 f: frnthfcfxfft\n7-16 w: wwwwwwmwwnwwwwmk\n11-13 j: jjkjrjjjjjjjjjd\n3-4 s: ssvc\n3-13 t: ddttttnnqpqztzbbdv\n3-5 g: gggggg\n3-4 r: rrrzlklfljqvz\n5-12 m: kdgmmrszqpfsbmz\n11-14 g: gqgvzgxhgbtngxggh\n7-16 k: qqptmhhfkhsgkhbmkx\n5-8 p: ppppptppp\n2-17 x: vxtxpvnxvlctrcpfxxx\n2-3 w: bwmwjwwwwwtnqwxc\n4-5 g: gggggg\n4-7 z: znzwzzzs\n4-5 v: kvvvv\n10-14 r: rrrrrrrrrbrrrx\n4-5 t: xtttt\n4-11 w: wwwwwwdcmkrwx\n6-7 c: ccccccjc\n12-14 w: wkwwwwwwwhwlwwwk\n5-9 s: hmssssssssss\n2-4 r: rrrjsh\n12-13 m: mmmmmmmmmmmmmx\n3-5 s: sshsvssss\n4-5 l: xllwnshchfdfk\n16-18 w: wwpwwpwwwmwwzwwtwwww\n2-9 c: zsfgbrrprlszrr\n10-11 n: thnmfndnnbnnbhnnnn\n13-14 v: vvvcvvvvvvvvvvv\n3-4 n: rwcm\n14-15 t: tttttztttttttpjt\n5-8 j: jjjjwhjdj\n5-8 f: fffffffm\n6-7 z: zzqzzzz\n12-19 j: kjrfnsjnhjdzgqpcjzvh\n3-7 p: pxpdvcqdptxpvk\n1-4 c: wccqj\n1-2 w: pwsqhwjdhcm\n2-3 b: bfbn\n10-19 q: qqqqqqqqqtqqqsqrqqq\n5-8 p: pppppppppppppppp\n10-13 g: rjghslgbggkgjdgvp\n3-9 z: fnnfzzdnf\n9-16 g: tlgtgggggggggggqm\n3-12 r: rrrvccrrvbdbncrrr\n1-4 v: jvhbv\n15-16 v: vvvdvvcvvvpvvvvvvvv\n3-4 g: ggggg\n7-8 w: wwwwwwmpw\n4-9 s: gkxshzkpssc\n2-8 q: pqttnpqqqfwqs\n3-4 n: nlnp\n3-4 h: phhh\n16-18 g: hqsxhmmpfgggmgpqhrjg\n15-17 w: wwgwwwwwwwwwwwmwl\n7-9 x: xxxxxxmxv\n1-7 s: ssssssbs\n1-3 x: dxlxxx\n3-6 p: wpxppq\n5-7 c: clcclcfcbccc\n3-5 s: sbmst\n3-5 d: xqmdn\n5-10 z: zbzwfddkzzc\n18-19 d: dpdddddddddddddddwcv\n3-4 r: rrrr\n8-9 f: ffnffffwd\n12-16 h: hxnlxghzkpcpdhqqvl\n6-7 m: jngclmm\n5-13 m: rmmmpmrmmmmmjmmmm\n1-3 f: dfwfs\n7-8 l: lclllwllll\n16-18 j: jjjjjjjjjjpjnjjgjbj\n12-15 f: fffffffffffnfftfffff\n9-13 w: qbnpmklbwxdbbwkklpwb\n1-3 j: mjkj\n1-3 h: hphwdh\n3-5 x: vrxxb\n1-15 v: mvvvvvvvvvvvvvj\n8-12 v: lvzfvzvvnjtvvvvvv\n4-6 g: gggvggggg\n13-20 l: zdllslnlxslqpnvgwlwm\n2-4 l: lllhll\n3-14 k: kckzddtzbmvkpkhb\n4-5 k: hbkgbxzj\n2-4 p: lppp\n2-11 r: rrrrrbrrbrrrrwrrrwr\n4-7 k: knkqkkk\n7-8 r: njrrrrrrrr\n4-7 h: phhhnhdhc\n4-5 r: nsnrrfktwbbhrrrh\n4-5 q: qqzvzsvkq\n6-7 b: jbbbbbbbkbgb\n5-11 f: ffffcfffffgff\n12-16 v: vfvvqvvdvvvlhvzjvmzv\n9-10 v: vvzvvtvvjgvv\n1-7 v: vvvvvvvv\n3-12 z: zmncxhrdzdmtcbxtlrzq\n3-7 r: rbfbrjrr\n12-13 n: nnnnnnnnpnnldnn\n5-12 j: jjjjjjjjjjjlj\n3-6 s: sssfsgss\n13-17 h: hhhfhhhhhhhhvhhhhhhh\n5-10 g: gggxkggggvg\n1-4 b: hbbtb\n14-15 t: tttttttttttttnb\n7-14 d: ddddddddddddddhdw\n15-16 d: rbdddbdwmjdhmpdd\n7-13 s: kxpdntprmskcs\n3-5 m: gmdmpj\n4-5 j: jjjjjvjsj\n1-14 q: qmqqqntqcqxmqsqkqq\n4-5 q: tqxqq\n8-11 m: mmmxmmmvmmv\n5-8 q: qqqqrqqw\n5-15 d: dqdddddddddddvspd\n1-4 t: tttttt\n12-15 z: dhzzxwfjgnzhzxt\n8-11 v: kdnmfnmqvvdvqdlvk\n16-18 s: sssssssssssssssjstss\n2-5 c: ccwdccc\n3-11 x: nxjxxxxxxxxxxx\n13-15 v: vvvvfvvqvlvvvvvv\n4-11 s: sssszssnqjsbsvs\n9-18 j: gxhjjjjjnjjsjjjrjjjj\n3-4 t: ttfttt\n9-10 s: qzgxjhpsss\n9-10 q: qqqqnrtxrqqsqqq\n6-13 k: kkkrgspkkkpwjshmk\n9-12 h: xhhhmhbhhhdhh\n3-12 g: gfgggggjgggggggggg\n13-15 z: zzfzzszzzzzzvzzzhzw\n5-15 m: mzrmmmmndchfzmmrr\n5-10 z: zzzzzzmzzh\n4-5 s: wscfc\n3-6 m: kmmtmm\n5-7 h: hhhhchh\n4-14 w: wwgwwwqwsfrjvmbwj\n2-9 v: vwvvwvmjcqnxv\n9-10 x: qxxhfxchrx\n17-18 x: xxxxxxxxxxxxxxxxwp\n4-6 v: vvvvvtv\n11-14 l: lcqlclllllrlddxlzll\n1-4 r: trrn\n10-13 h: thnsxphnfgvhvq\n11-12 s: sssssssssshsss\n10-11 z: rqpzszzmzdz\n3-6 q: qrqqqq\n5-6 h: hhkhhhh\n8-16 h: wqhjvhlgwtsgvlpf\n3-6 f: gqfpcfsfhfjgrbqv\n3-5 l: lqllsrjlv\n6-10 v: vvvvvvkvjv\n3-8 t: gtttcbnkxxstttd\n2-6 j: jjjjbjz\n10-15 j: ftwjtfzjmjsvwjj\n8-11 d: ddsksddddzfd\n2-3 b: bhngbb\n14-16 c: mllmcbfxwxrqlcjcw\n7-8 q: qqqqqqdv\n5-6 s: slkzlfvg\n3-5 q: qqqqq\n3-12 p: pddpdpszrppxcpjgv\n11-18 r: rrrrrrrrrrprqrrrrr\n3-4 c: cscckcxcn\n11-13 g: gvgwmlngggqggrtgtkg\n5-7 b: smfpbbb\n7-18 l: sqmlklvkqfjrgtqhzr\n6-7 k: kkpqqfgz\n7-9 n: fnndnnxnndn\n11-12 c: ccccccccccxv\n6-9 d: dddvdddddddvd\n1-4 v: nvjfv\n3-8 g: gggggggg\n6-7 x: xxxxxxx\n4-5 c: scccc\n6-13 q: nchlfqqqlqnqqqtq\n1-8 r: srrrrrrnrrrrrrr\n4-5 p: sdpscpppp\n9-11 r: rrrdrkrrrqg\n3-5 m: xwmscmmm\n7-11 r: rrrfrrrrkfrvrf\n4-6 b: bbfbbb\n4-9 v: rblrvrvpvz\n2-8 w: wwwvwgwww\n8-9 w: wwwwwwwwzw\n16-18 t: ttttttttttttttttttt\n8-9 n: nfrzsdjxr\n1-2 w: wwfwwmfbww\n2-4 s: dfsksft\n1-13 d: dwdrccnddqmndcl\n1-4 f: qfflf\n3-7 d: dnddjpdfgc\n3-5 l: lkglx\n1-13 j: jjjjgbjtpjjhd\n15-16 j: jjjjjjvjjjjjjjjjdjj\n4-13 t: tttxztttttttkt\n4-5 m: mmmmm\n4-6 j: jgjrjwjfjx\n8-11 t: ttbjttvttsttdqtn\n14-15 w: wwwwwwwwwwwwwkjw\n1-6 x: nrscxn\n3-13 w: wwwwwvgnwbwwwwwwv\n12-15 p: pppppppppppjpppppppp\n5-11 v: vvvkvbvwkwnvvvsxvv\n1-11 m: mmmmmmmmmmtm\n10-11 c: cccccccccwc\n5-8 f: xvgstwfxfhxknds\n12-13 t: tttttztttttttt\n4-17 k: vkpkfkkkqnkqnkgkkknk\n9-12 k: kkwpzpdzxhxk\n14-17 v: fgvvvvvvvvvvvjpvv\n7-8 k: kndnqkkk\n7-11 c: gfcbccccjvcskcmrcxc\n4-5 n: ngtnr\n3-12 k: mtcfszkdhkdkd"
+  )
+}
