+++ src/test/scala/io/github/teonistor/adventofcode/_03Test.scala
@@ -1,0 +1,10 @@
+package io.github.teonistor.adventofcode
+
+class _03Test extends AdventOfCodeTestBase {
+  testAndRun(3, _03._1, _03._2,
+    "..##.......\n#...#...#..\n.#....#..#.\n..#.#...#.#\n.#...##..#.\n..#.##.....\n.#.#.#....#\n.#........#\n#.##...#...\n#...##....#\n.#..#...#.#",
+    7L,
+    336L,
+    "...#...#....#....##...###....#.\n#.#...#...#....#.........#..#..\n.#....##..#.#..##..##..........\n.....#.#.............#..#......\n.......#...#.##.#......#..#.#..\n#.#....#......##........##.....\n.....##.#....#..#...#...##...#.\n...#...#..#.......#..#...##...#\n..........#...........##.......\n..#..#..#...................#..\n#..#....#.....##.#.#..........#\n.#.##.......###.....#.#...#....\n.#..##....##....#.......#...###\n#.#..##...#.#..#...............\n.........#....#.......##.#.#...\n...###...##....##...#..##.#..#.\n....#.........#..#...#.......#.\n....................#..#.#.#...\n..#....#..........#...........#\n.#.....#..#.....##........##..#\n#..##..#...##............#..##.\n.#..##....#..........#..#.##.#.\n..#####..#.#............##.....\n...###.#....##..#.#....#.....#.\n.#.......##....#...#.#.....##..\n...#....#...##.#...#..#........\n.####.....#....#.#.#...#.......\n...#....#.....#.......#........\n#..#.#.......#...#............#\n...#.....###.##....#.#.###.#...\n.#.........#.......#.#....##...\n#.#..#...#.#...##......#..#....\n.....#...#..#.#...#..###..#....\n......#.........#...###........\n.....#..##...#..........#.....#\n..#..#.#.##.#...#....#....##..#\n##....#.##...#.##.#..##....#...\n.....#.#.#.#..#....##.#...#.#..\n.....##.......#........#.......\n...#.#.....#...#...##.#......##\n........#..#.#...#.#.....#.#..#\n#..##...#.#...##..##...#.#...##\n.##.#.#..#...#.....#.#.##.#...#\n.#.####.........##.........#..#\n.##..............#....#...#...#\n......#...#..#...#..#..###.#...\n.......##...#.#.#..##..#......#\n.....#....#..##..#.........#...\n.....#..#.#.#........#.#.####..\n#..#.......###....##...........\n#..##..........#.#......#.#....\n.....##........#...#..##.......\n###...#.##.#.#.#.#.##...##.....\n....#...#........##.#.##..##...\n.#..#.#.#......#.......##..#..#\n.#...#.................#....#..\n.##..#..........#..##.......#..\n.#.#.#.....#..#.#.........##..#\n...#......##...#.......#...##..\n##...###....#.###.............#\n#.....#.#..#.#..#........#.#.#.\n.....#.#......##..#.#.....#.##.\n.......#...........#..#.......#\n..#....#.#.#......#.....#...#..\n.....##........#..##..#..##....\n#.#........#...##....#.#..##...\n#......#......#....#..#...#.##.\n....#.#.......#.#.#............\n......####.#.##...#.#.##.....##\n..###.#.#..#.........#.####....\n.#.......#..#.#....#.#..#.#.##.\n#....#....#............##...##.\n....#....#............#....#..#\n..#........#..#....#..#..#...#.\n.#......##....#..........#....#\n#.##.....#..........#.###.#....\n....##...#.....#.#......#.##...\n#.#.....#.......###.###..#..#.#\n..###..##.............#.####.##\n#....#.....#....#..##.......#..\n.....#....#...#.#.#.#..#...#.##\n...#.....#..#....###......#.#.#\n##.........#.#..#..#.#..#.....#\n.#.....#.#....#.........##..#.#\n.#.#..#.###..#..#..........#...\n.##....#.#.#...#......##.....#.\n#.#....#....#...#...##...#..#..\n#...#........#....#....#......#\n#......#...#..#.#.##.....##..#.\n....#...#......##...#..#....#..\n.#......##.##.......#.......#..\n.#...#..####...........#.#.#...\n.........#...#.#.........#.....\n#.##.....#.#..#.#.###...###..#.\n#...##.###......#.###..##.#.##.\n...##.#.....#....#..#......#...\n#....###.#..#...##.....#......#\n........###...#...#............\n........#....#...#...#....#...#\n#....#..#..#....#.#........#.#.\n##...#.....#.#..........#..#..#\n#.#...##.....#........#...#...#\n##.#.#.......#...#..#.###....#.\n.#.......#....##..##...#.....#.\n#....#....#.....#.......#......\n.##.##.##...##...#.#.#..#..#...\n#..#..#.##....#......##....###.\n.......#.#.........#..##.#...##\n.#..##...#....#.....#..........\n..#.#...#......#.#..#..........\n.##....#.#.#.##.......###...#..\n..##.#...#.#.#.#.......#..#....\n#..#.......#...#........#.....#\n.....#.......#......###..#.....\n...##.#.......#.....##.....##..\n##..#.......#.#.....#....#.....\n..#....#.##.##...#...#......#..\n.#..#.###.#....###........#...#\n....##.##...##..#..#.#....#....\n..###...##.....##..............\n#....#...##...#....#..........#\n.##........#......##...##...#.#\n..#.#.##..........#......#.....\n...#...#.........#.##........##\n..#.#..#.#..#...#....#...#.....\n...##...#..#.###.#..#.#...#....\n....###........#..#..##...#....\n#.#....##.......#.#........#...\n.###...#..#.#.#.#..#...#..##.##\n..#.........#####.#......#..#..\n#.....#.....##..#....#...#...#.\n...#..#....##....##.....##.#...\n.........#............#.##.....\n....##.#..#....#.##.......#..##\n.###....#.#..#......#.#.......#\n.###...###.#.........#.#..#...#\n.....#........#..#.#..#.#..##.#\n.###..#....##.........#..##....\n..#.......#..#..##...#.###.#...\n#.......#...........#.#...#.###\n#.##.##...##.#...##..#.....#...\n..#..#........###.#.....##.....\n#.....##....#...##...####..#..#\n....#........#...#...#.........\n......#.#.#.#.......#..#.....##\n..#..#....#.....#.#...##......#\n..#....#...#.###.........#.###.\n...#......##..#.#.....#...#....\n...#.......#...#...#........##.\n............#...#..#....#.....#\n....##......................#..\n#.#.#....#....#..........##....\n#.#.....#.#.##..#...#.##....##.\n...#...#..#...#..#.#.#.......#.\n#.....#..........#.........##.#\n#...##..#..#.#.......###....#..\n.#...#..##....#.....##.......#.\n....#.##.....#.........#.#....#\n........#.#...####..#.......#.#\n.####...#.#......####.....#.##.\n###..#....#..#.......#.#..##..#\n#......#.#....##..#.##.#....#.#\n...###...#...#..##.#..#..#.#...\n...##..##....#..#.....#........\n.....#..............#......#..#\n......#....#......#..#.........\n#..#.....#.##...........##.....\n.#..#.#..................##....\n#.#..#..##...#....#.#......#...\n.##.#.##......#.##...#...#...#.\n..#...#.........#.#..#.#....#..\n.#.####.#..#.#......##.#..#....\n#..#.......#....#..............\n....#............#..#..........\n.....#####.....#.....#..##...##\n#.#....#.#...............#..##.\n.#.#..#...#......#.....#.#.#...\n.#....#.#.#......#.....##....#.\n....#....#.##..#.......###...##\n.....#..#.##...#...#...#..#.#..\n##..#........#.#..#..##......#.\n.#..#..##.......#..#.....#.....\n.#.#.....###..##.#.#...........\n..##..##.####..........#..#....\n..##..#..#...#....#......#.#...\n#...#.#......##.....##.#..###..\n#..#..............#........##.#\n.........#.##..#.#..#..##.##.#.\n#....##....#.#..#.#...##..#....\n.#....#.......#............##..\n.......#.#.......#...#.#......#\n......##...#.......#.#........#\n..###..#.#.##....##...#....##..\n..##.##..........##..###.......\n.#.#.#..#..#.#.......#.#...##..\n..#..##.........#.###..#......#\n....#.#.#...##.#...#...##..###.\n..###..##.........##...#...#..#\n.#..##...#.......#.......#..#.#\n........##....##....#.#.###.#.#\n#.....#.#.................#.#..\n....#.#.#.....##.####.#......#.\n....#.......#.#.##.##..........\n...#...........#...#.##...#.###\n#....#....#..........#.##......\n##..#...........##.....##.##...\n.#.##...##..##....#..#.....####\n#...#...#.##..........##..##...\n....##..#....#.....#.#...#....#\n..#....#..##...###.#.#.........\n#......#.#.#...#...#.........#.\n#............###.#.#.#..##...#.\n.##.....####...##..##..#..##.#.\n#..#........#.....#.#.....#...#\n#............#....#.#.#........\n......##...##.#....#.....#...#.\n..#........##......#.#.....##..\n.#..#..#.....##.......#..#.#..#\n.#....#..#....##.#.#.#..#..#.##\n.####.#..........#...#..##.....\n...###..###...##..#............\n#..#.....##.#...#..##..#.......\n.....##....#...###.##...#......\n...##..#...#..#..##....##....#.\n...###....#.###.#.#.##....#....\n##.#.#.....#....#.#....#..#....\n.......##.....#.#..##...##...#.\n.#....#.#...##.#..#....#.....#.\n..#...#..#...#.##........#...#.\n#....#......##.#....##...#.#..#\n.....#..#..#..#......#...#.#.#.\n..###....#........#...#.......#\n###...#.......#.#.......##.##..\n......##.....#.#........#....#.\n#.##..#.#.#.#..#....#.##.....#.\n..........#.##.#...#...#..#..#.\n..#...##.#..........#..##.###..\n..###..##.##..#.#...##.####..#.\n#.#.#...............##....###.#\n....#.........#.#....#.#....#.#\n..#...#.###...#....###.....#...\n..#..#....#...#............#...\n.#..#....#..##.....##..........\n..#....#.#...#.#.#.#.......##.#\n.........#....##........#.#....\n...#..##.#..#.##...#...#.#....#\n....####...#...####.#....###..#\n......##...#.##.#.......#..#...\n#.#...#.#...#.#...#....#.#.#...\n.#.....##...#.....###.#....#...\n......##.....###...#.#...#.#...\n#..#..##.#.#......#....#..#..#.\n....#.###.....#..#...#.##.....#\n##.##........#......#....#..##.\n##.....##.#.....#.....##.....#.\n.....#.##...#.#..#.#.#.....#...\n.#.##..#...#.#..#.....#.#......\n.....##.......#..#...##..#..#..\n#.....#..#.####......#........#\n.#..#..##.#..##............#..#\n.##..#.#....##.##.....#......#.\n.......##.........#..#.........\n.#...#.......................#.\n#......#.#....##.#.......#..#..\n..##..##......#.......#....#.#.\n##......#......##...##.........\n..#....####....#.#.....##.#.#..\n..........#..#.#.#.....#..#.#..\n##..##...........##.......#....\n##....#.#....#..#......###....#\n...#.#.#..#.......##.......#...\n#....#.......#.......#.........\n...##......##....#...#......#.#\n#......#####.#.........#.....#.\n#..#.............#..#....#...#.\n.......#.##..#..#..#..#....####\n......#.##..##..........###...#\n.#.##....###..#........#....##.\n#......#..#...###.#...#.....#..\n.#.#.......#....##.......#.#...\n..#.##..#..##.....#.........#.#\n#.#...#..#.##....#.......##....\n.#.....###....#.#..#...#.....#.\n#...#..#.......#.#.....##...#.#\n#.#####.........#....##.....#..\n#....#..##...#....#.##.......#.\n.#.#.........##....##....#.....\n...#..##.......#....#.#.#......\n#.###.##...###....#.....#.####.\n.#...#.#.#..##.#..........#....\n#.#.....#.##.#..####.....##.#..\n...###.##..####.......#......##\n.##..#.........#...#.#.....#.##\n..#.....##....###.....#.#...##.\n#....#....#..#....#.##.........\n......###....#.#..#..#....##...\n.#.#................#.......##.\n...#.......#.........#.#.......\n...#..........#...##.....###...\n....#......#...#...............\n.##...#....#.....#.##......#...\n.#.....###...##..##...#.#......\n....##........#.....#...#....#.\n#.........#.#...##...#.#..#....\n...#.#.....#.#........#.#....#.\n.#........#.....#.#.#.#.#..#...\n....#...#.....#.#....#........#\n..###.#....#.#....##...##..#.##\n.#....#.#.####.#.#.....#.......\n.#...#...#.................##.#\n..................##..#..#.#.#.\n.#..#............##....###.....\n.......#....#...........#......\n....#.#.#.....###.........#..##\n...#.#....#.#.##.#.##.....##..#\n.#.##.#...##...#.......#.....##\n.#............#...#..##...#.#.#\n#.##..#.##..#..##.###.#........\n..............##....#...#..#.#.\n.#.#...#.#....#....###........#\n.#....#.#....#......###........\n..#.......##......#.##.....#...\n.....#......#..#...#.#.....#..."
+  )
+}
