+++ src/test/scala/io/github/teonistor/adventofcode/_05Test.scala
@@ -1,0 +1,11 @@
+package io.github.teonistor.adventofcode
+
+class _05Test extends AdventOfCodeTestBase {
+  testAndRun(5, _05._1, _05._2,
+    // No example for this one, back-filling with actual results
+    "FFBFFBFLLR\nBFBFBFBRLR\nFBFFBFBRRL\nBBFFFBFLLR\nBFFFFFBLRL\nBFBBFBBRLR\nFBFBFBBRRL\nFBBBFBFLRL\nBFFFFBBRRL\nFBBBBBFLLR\nFBFFFBFLLR\nFFBBBFBLRR\nBFFFBBFLRL\nFFBBBFBLLL\nFBFBBFFLRR\nFBFBBBBLLR\nFFBBBFFRLR\nFBFBBFFRRL\nFBBBFFFLRR\nFFBBFFBLLR\nBFBFBFBLLR\nBFBBFFFLLL\nFBFBBFBRRL\nFFBBFFBLRR\nFFFBFBBLRR\nFFFFBBBRRL\nBFBBBBFRLL\nFBFBBFFLLR\nFFFBFBBLLR\nFFBFBBBRRL\nBFFBBFFLLR\nBFFFBBFLLR\nFBBFBBFRRR\nFFFBBFFRLL\nBFBFBBFRRR\nFBBFFBBRLR\nFBFFBBFRRL\nBFFFFBFLLR\nBBFFBFBLRL\nFBFBBBBLRL\nFBFBFBFRLL\nFFBFBFFLLR\nBFBFBBBRLR\nBFBBFFBLRR\nFFBFBFFLLL\nFFFFBBBLLR\nFFBBBBBLRR\nFFBFBBFLLR\nBFFFBBBRLL\nBFBFBFFLRR\nFBBBBFFRLR\nBFBFFBBRLL\nFBFBBFBRRR\nBFBFFBBLLL\nBFBFBBFLRL\nBBFFFFBLLR\nFFFBFFFRLL\nFBFFFBBRLL\nFFFBBBBLRR\nFBBBFBBLRL\nBBFFBFFRRR\nBFFBFFFLRR\nFBBBBBBLLR\nBFBBFFBLLL\nBFFBBBFRLR\nFBBBFFFLRL\nBBFFFBFRRL\nBBFFBFFRRL\nFFBBFFFLRR\nBBFFBFBRLR\nBBFFFFBRLR\nFFFFBBFLRL\nFFBBBFBRRR\nFFBFBBFLRR\nFFFBBFBRRR\nFBFFBFBRRR\nBFFFFFBRRR\nBFBFFBFLRL\nFFBBBBBLLL\nBFFFFBBRRR\nFBFBFFBRLR\nFBFFFFBLLR\nFFFBFFFLLL\nBFFFFFFLLR\nBBFFFFBRRL\nFBBFBBBRRL\nBFFFFFFLLL\nFBFBFFBLRR\nFFFBBFFRLR\nFBBFFBBLLL\nBFBBBFFLRL\nBFBFBBBRLL\nFBFBBFFLLL\nFFBBFFBRLL\nBFFFFBBLLR\nFFBBFBFLLL\nFFFFBBBRRR\nBFFFBBBLLL\nBFFBBFBRLR\nFFBBBFFLLR\nFBBBFFBRLR\nFFBFBFBRLL\nFFFBFFFLRR\nFFBBFBBLRL\nBFFBFBFLRR\nFFBFFFFRLR\nBFBFFFFLLL\nBFBBFBFRRR\nBFBFBFFRLL\nBFFFBFBRLL\nFFBBFBFLRR\nFFBFFFFRRL\nBFBBBBFLLL\nBFBFBFBRRR\nFFFFBBFLRR\nFBBBBFFLRL\nBFBFFFBLRL\nBFFBBFBRLL\nFBFFBBFLLL\nBFBBBBBLRL\nFFBFBBBLLR\nFBBFFBBLRR\nFFBFFFBRRR\nFFFBBBBRRR\nFBBBBBBLRL\nFFFBBFFLLL\nBBFFBFBLRR\nBFFBFFBRLR\nFFBBFBBRLL\nBFFBFBBRLR\nFBFBFBFRRL\nBFBBBFBRLL\nFFFFBBBLRR\nBBFFFFFLLR\nFFBFFFFLLL\nBFBBFBFRLR\nFFBBBBFRLR\nBFBFBBBLRL\nFBBFFFFRLL\nBFFFBBBLRL\nFBBBBFBRLL\nFFBFFFFRRR\nBFFFBFBRLR\nFBBBFBBLLL\nFFBFFBBLLL\nFBFFFBFRLL\nFFFFBBBRLR\nBBFFFBBLLL\nBBFFFFBLRL\nFBFBBBFRRL\nBFBFFBBRLR\nFBBFFBBLLR\nBFFBBFFLLL\nFFBFBBBLLL\nBFFBFFFRRL\nBFBBBFBRLR\nFBFBBBFLLL\nBFFBBBBLRR\nBFFBFBFLLL\nFFFBBBBLLR\nBFFFBFFLLR\nFFFBFBBRLL\nFFBBBBFRRL\nFBFBFBBLLR\nBBFFBFFLRL\nFBFBFBFLLR\nFBFBFFFLLR\nBFFFBFFLRR\nBFFBFFBRRR\nFFFFBBBLLL\nFBBFBFFLLR\nFFBBFBFRRL\nBFBFBBBLRR\nBFFFFFBRLR\nBFBBBBFRRL\nFFFBBBFLRR\nFBBFBFBRLL\nFFFBFFBRLR\nBFFFFFFRRL\nBFFBFBFRRL\nBFBBBBBRRL\nFBBFBFFRLL\nFFBFBBBRRR\nBBFFFBFLRL\nFBBBFBFRRR\nFBFFFFFLRL\nFFFBBBBRLR\nFBFFFFBLRR\nFBBFFBBLRL\nBFFFBBFRRL\nFFFBFBFRRR\nBFFBFFFRLL\nBFBFBFFLLR\nBFFBFBFRLL\nFFBFFBBLRR\nBFBBBBBLLR\nFBBBFFBRRL\nFFBBBFFRLL\nBFFBFBFLLR\nFFBBBBFRLL\nBFBFFFBRLR\nFBFFFFBLRL\nFFFBBFBRLR\nFFFBBFFLLR\nBFBBBBFLRR\nFFBFBFFRRR\nBFFBBBBRLR\nFBFFFFFRRL\nBFFFFFBRLL\nFFBBBBFLRL\nBFFFBFBRRL\nFFBFBFFLRL\nFFFBFFBLLR\nFBBBBFBRRL\nFFFFBBFLLR\nBFBFFBFRLR\nFBFFBFBRLL\nBFBBFBFLRL\nFBBFBBFLLL\nBFFFFFFRLL\nFBBBBBBRLR\nBFFFFBBRLL\nFBFBBFFLRL\nFBFFBBFLLR\nFBFBFBBLLL\nFFBBBFBLRL\nFFBBBBFRRR\nFFFFBBBLRL\nBFBBBBFRLR\nBFBFBBFLRR\nFFFBBFFRRL\nFFFBBFFRRR\nFBBBBFBRRR\nFFBBBBBRLL\nFBBBFFBRLL\nFBFBBFBLRR\nBFFFBBBRRR\nBFFBBBBLLR\nBFFFFBFRLR\nFFBFFFFLRR\nBFBBBFBLRR\nFBFFFBBRLR\nFFBFFBFLRR\nFFFBFBBLLL\nBFFBFFBRLL\nFBFFFBFLRR\nFBBBFBFLLL\nFBBFFBFLRR\nFFBFBFBRLR\nFFBBBBBRRR\nBFFBFBBRRL\nBFBFFBBRRR\nBFFBBFBLRR\nFBBFFFBRLL\nFBBBBFFLLL\nFBBBBBFLRR\nFBFFBBBLLL\nBFFBFFBLRR\nBBFFFFBRRR\nFBFFFBBLLL\nFFFFBFBRRL\nFBBFBBBRLR\nFBFBFBFRLR\nFBFFBFFRRR\nBBFFFBFRLL\nBFBBBFFRLR\nFBBFFFFRRL\nBBFFFFFLLL\nFBBFFFBRLR\nFFBFBFFRLR\nBFBFBBFRRL\nBFBBBBFLLR\nBFFFBBFLRR\nFBFBFBBLRR\nFBBFFFBLRR\nBFBFFBFLLR\nBFBFBFBRRL\nFBFBFFBLLR\nFBBFBFBLRL\nBFBFFFFRLR\nBFFBBFBLLR\nBFBBFBBRLL\nBFBFBFFRRR\nFFBBFBBLLR\nFFBFBFFLRR\nFFFBFFBLRR\nBFBFFFFLLR\nFBFBFFBRRL\nFBFBFFFRRL\nFBFFFFBRRR\nFFBBFFFRRL\nFFBFFBBRLR\nBFBBBFFLRR\nBBFFBFBRLL\nFBBBBBFRRL\nBBFFFFBRLL\nFFBBBBFLLR\nFBBBBFBRLR\nFFBFFBFRLR\nFFFFBFBRLR\nFBBFBFBLRR\nBFBBFBFLLR\nFFBFBBFRRR\nFFBBFBFRRR\nFFBBBBBLLR\nBFBBFFBLRL\nBFFBBFFRLR\nFFFBBFBRRL\nBFFFFFBRRL\nBFBBFFBRRL\nBFFBFBBLLR\nFBBFFFBRRR\nFBBBFFFLLR\nBFFFFFFRLR\nFBBFFFFLRL\nBFFFBBBLLR\nBFFFBFBLLR\nBFFFBFBLLL\nFFFBFBFLRL\nFBFBBBBRLR\nFBFFFFFRLL\nBFFBFFBLLL\nFFBBFBBLLL\nFBFFBFFRLL\nFBFFBFBLLR\nFFBBFBBLRR\nFFBFFFFLLR\nFBBFBFBRLR\nFFBBBFBRRL\nFBFBBFBRLL\nFBFFBFFLLR\nBFFBFBFRRR\nFBBFBBBLLR\nBFBBFFBRLR\nBFFBBFBRRR\nFBBBBFFLRR\nBBFFFFFLRR\nFBFFFFFRRR\nFFBBFFFLLR\nFFBFFBBRRL\nFBBFBBFRRL\nBFFFFFBLLL\nFFFBBBFLLL\nFBBBBFBLRR\nFFFBBFFLRL\nBFBBBFBLRL\nFBBFBBFLRL\nFBBFFBBRRR\nFFBFBBFRLL\nFBFFFFFLLL\nFFFBBBFRLL\nBFBBFBBRRR\nFFFBBFBRLL\nFFBBBFBRLL\nFBBBBBBLRR\nFBFBBBBRLL\nFFBBFFFRRR\nFFFBFFBLRL\nFBBFBFFRLR\nBFFFBFFRRR\nBFFFBFFRLL\nFBBBFFFRLR\nFFBBFBBRRL\nFFBBFFFRLL\nFFBBFBFRLL\nFBBFFBFLLL\nFBFFBBBRRL\nFBFBFFBRRR\nFFBFBFFRLL\nFFFBFFFRRR\nFFBFFFBLRR\nFFBFFFFRLL\nFBFFFFBLLL\nBFBBFFFLLR\nBFBFBFBLRR\nFBBFFFFLLR\nFBBBBFBLLR\nFBBBFBFRRL\nBFBBBBBRLR\nFBFFBFBLLL\nFFBBFFFLLL\nFFFFBBFRRR\nFBFBBBFLRR\nBBFFBBFLLL\nFFBBBFBLLR\nFBBFFFFLRR\nFBBFBFBLLR\nFFBFFFBLRL\nFBFBBBFLLR\nBFBBBFFLLR\nBFFFFBBRLR\nFBBFBFBLLL\nFBBFFBFRLR\nBFBBFFBRLL\nFBBBBFFLLR\nBFFBBBBLRL\nFBBFBFBRRL\nFBBFFBFLRL\nFBBBFFFRRL\nBFBFBFFLRL\nBFBFFFFLRR\nBBFFBFBLLL\nBBFFBFFLLR\nFBFBBFFRLL\nBFFBBBFLLR\nFFBFBBFRRL\nFBFFBBBLRR\nFBBBFBBLLR\nFBFFFBBRRL\nFFBBBBFLRR\nBFBFFBFLRR\nFFFBBBFLRL\nFBFFBFFRRL\nFBBFBFFLRL\nFBBFFFBLRL\nFFBBBFFRRR\nBBFFFBFLLL\nFBFFBFBLRR\nBBFFBFBLLR\nFBFFFFBRLR\nBFFFFFFLRL\nFBBFBBBLLL\nBFBFFBBLRR\nFFBBFFBRRL\nBFFFBFBLRL\nBFBBBBBRLL\nFBFFFBBLRR\nBFBBFBBLLR\nFBBBFFBLLL\nFBFFFBFRLR\nFFBBBFFRRL\nFFBFBBFLRL\nFFBFBFBLLR\nFBBFFBFRLL\nFFBBBFFLLL\nBFBBBBBLLL\nFBBFFFBLLL\nBFBFFFBLRR\nFFBFFBFLRL\nBFFFBBBRRL\nBFBFBBBLLR\nBFBFFFFLRL\nBFBFBFFRLR\nFBBFBBBRLL\nFFBFBBBLRL\nFBBBFFFRRR\nFBBFBFFRRR\nFFFFBBBRLL\nFBFFBBBLRL\nFBBBFBBRRL\nFBFFBBBLLR\nFFFBBBFRLR\nBFBFBFBLLL\nBBFFFFFRRR\nBFBBFBFRLL\nFBBBFBBLRR\nBFFFFBFLRL\nFBBBFFBLRL\nBBFFBFFLRR\nFFBFBFFRRL\nFBFFFBBLRL\nFFFFBBFRLL\nFFBFFFBLLR\nFBBBFBFLRR\nFFFBFFBRRR\nBFFBBBFLRR\nFBFFBFBLRL\nFFBFFFBRLR\nFFBFFFBRRL\nFFBBFFBLRL\nFBFBFBBRLR\nFFFBBBBLLL\nBFBBBFFLLL\nFFBBFBFLRL\nBFBFFFBRLL\nBFFBFFBRRL\nFBBFFFFLLL\nFBFFFFBRRL\nFFFBFFFLLR\nFBFFFBFLRL\nFFBFBFBLRL\nFBBBFFBLRR\nBFBFFBBRRL\nBBFFFFFLRL\nBFFBBBFRRR\nBFFFBBFRLR\nBFBBBBBLRR\nBFFFBFBLRR\nFBBBBFFRLL\nBFBBFFFLRR\nBBFFFBFLRR\nFBFFBBFRLL\nFBFFFFFLRR\nFBBBBBFLLL\nFBFBFBBRRR\nFBFBBFBLRL\nBFBBFFBRRR\nBBFFBFFLLL\nFBBFBBFLLR\nBFFBBFFRLL\nFBFBBBBRRL\nFBBFBBFRLL\nFBBFBBFLRR\nFBBFFBBRRL\nFFFBBFFLRR\nBFBBFBFLRR\nBFBFBFFRRL\nFFBBFFFRLR\nFBFBBBBLRR\nBBFFFFBLLL\nBBFFFFFRLR\nFBFBFFBLRL\nFBBFBBBLRR\nBFBFFBFRRL\nBFFFFFFLRR\nFFFBBBFLLR\nFBFFBFFLLL\nBFFBFFFRLR\nFBBFFBFRRR\nFFFBFFFRRL\nFFFFBBFLLL\nBFBFBBFRLL\nBFFFFBBLRL\nFFFBFBFLLR\nFBBBFBFRLR\nBFBBBFBLLR\nBFFBFBFLRL\nBFBFFFBLLL\nFFBFFBBLRL\nBFBBFFFRRR\nFFFBFFBLLL\nBBFFFFBLRR\nFFBBFBFLLR\nFBFBBBBRRR\nFBFBBBFLRL\nBFFBFFFLLR\nFBFBFFFRLL\nFBBBBFBLLL\nFBBBBBBRRR\nFBBBBBFLRL\nBFBBFFFRLL\nFFFBFFFRLR\nFFBFFBBLLR\nFFBFBBBRLL\nBBFFFFFRLL\nFFFBFBBRLR\nFFBFBBBRLR\nBFFFFBFLRR\nBBFFFBBLRL\nFBBBBFFRRL\nFBBFBBBRRR\nBFBFFFFRLL\nFBBBBBBRRL\nBFFBFFBLRL\nBFFBBFBRRL\nFBFFFBBRRR\nBFFBBFBLLL\nBFBFBFBRLL\nBBFFFBBRLL\nFBFBBBBLLL\nBBFFFBBLRR\nBFFFFBFRRL\nBFBBFBFLLL\nFBBBFBBRLR\nBFBBFBBLRL\nBFBBBFBRRL\nFBFFBBBRLL\nFBBBFBBRLL\nFFBBBBBRRL\nFBFBBBFRLL\nBFFFFFBLLR\nBFFBBBBRRR\nBFBBFBBLLL\nFFBFBBBLRR\nFFBFBFBRRR\nBFBFFFBRRL\nFBBBFFBRRR\nBFBFBBBLLL\nBBFFFBBRLR\nBBFFFBBRRL\nFBBBBFBLRL\nFBBBBFFRRR\nFBBBFFBLLR\nBFFFFBFLLL\nBFBFBFFLLL\nBFFFFBFRLL\nBFFBFBBLRL\nFFBFFBBRLL\nFBFBBFFRLR\nBFBFFFBRRR\nFBFFBFBRLR\nFBBFBFFLLL\nFBFFBBFLRR\nFFBFBFBLRR\nFFBBFFBRRR\nFBBFFFBLLR\nFBFFFBFRRR\nBFBFFFFRRL\nFBBBBBFRRR\nFFFFBFBRRR\nFFBBBBBLRL\nFBFBFFFLRR\nFBBBFFFLLL\nBFFFBFFRLR\nFFBFFBFRRL\nFBFBFFFRRR\nFFBFBFBLLL\nBFBBBFFRRL\nBFBFBFBLRL\nBFFFBFFLRL\nFBBBBBFRLR\nFBFBBFBLLL\nFFFBBBBLRL\nFBBFBFFLRR\nBFBBBFBRRR\nFBBBFBBRRR\nFFFBFBFRLL\nFFBBBBFLLL\nFBFFBFFRLR\nFBFBFBFLLL\nFBFFFFBRLL\nFBBBBBBRLL\nFBBBBBFRLL\nFBBFFBFLLR\nBFFBBBFRRL\nFBFFFBFLLL\nFBBBFFFRLL\nBFFFBBBRLR\nFBFBBBFRRR\nBFFBFBBRLL\nFBBFFFBRRL\nFBFFBBBRLR\nBFBFFBBLRL\nBFFBBFFLRL\nFBFBBBFRLR\nBBFFFBBLLR\nFFFBFFBRLL\nBFBFBBFLLL\nBFFBBFBLRL\nFFFBBFBLRL\nBFBBBFFRRR\nFBFBFFFRLR\nBFFBBFFLRR\nBFFBBBFLLL\nFBFFBBBRRR\nFBFBFFFLLL\nFFBFBBFLLL\nFBFFFFFLLR\nBFBBFBFRRL\nBFBBBBBRRR\nBFFBFBBLLL\nBFFFBFBRRR\nFFFBFBFLLL\nFFBFFFBRLL\nFBBFBFBRRR\nFBBBBBBLLL\nFBFBFFFLRL\nFFFBFFBRRL\nFBFBBFFRRR\nFFBBFBFRLR\nBFFFFBBLLL\nFBBFBFFRRL\nBFFBFFFRRR\nBFFBFFFLRL\nBFBBBBFRRR\nBFBFBBBRRL\nFFFBBFBLLL\nFBBFFBBRLL\nBFBBBFFRLL\nBFFFBBFRRR\nFBFBFBBRLL\nBFFFBFFRRL\nFFFBFBBLRL\nFBFFFBBLLR\nBFFFBBFLLL\nBFFBFFFLLL\nBFFBBBBLLL\nBFFFFFFRRR\nFFBFFBFRLL\nFBBFFBFRRL\nBBFFFFFRRL\nFFFBBFBLRR\nBFBFFBFLLL\nFBFBBFBRLR\nBFBBFBBLRR\nFBFFBFFLRL\nBFBFFFBLLR\nFFFBBBFRRR\nFFBBBFFLRR\nFBFBFFBLLL\nBFBFBBFRLR\nFBFBFBFRRR\nFFBBFBBRRR\nFBBFFFFRLR\nFBFBFBBLRL\nFBBFFFFRRR\nFFBBFFBRLR\nBFBFBBFLLR\nBFFBFBFRLR\nBBFFBFFRLL\nBFFBBFFRRR\nBFFFFBFRRR\nFBFBFBFLRL\nFFBFFBFRRR\nFBBFBBBLRL\nFFFBBBFRRL\nFBFFBFFLRR\nFFBFBBFRLR\nBBFFBFBRRL\nFBFFBBFLRL\nFFFFBBFRRL\nBFBFFBFRLL\nBFBBFBBRRL\nFFFBBFBLLR\nFFBFFBFLLL\nFFBBBBBRLR\nBFFFFFBLRR\nBFFBBBFLRL\nBFBBFFBLLR\nFBFFFFFRLR\nFFFFBBFRLR\nFBFFFBFRRL\nBFFBBBBRRL\nFFBFFFFLRL\nBFBBFFFRRL\nFFFBBBBRLL\nBFBFBBBRRR\nBBFFFBFRLR\nBFFBFFBLLR\nFBFBBFBLLR\nBFFBFBBRRR\nBBFFFBFRRR\nBFFBBFFRRL\nFBBBFBFRLL\nFFBFFFBLLL\nBFFBBBFRLL\nFBFFBBFRRR\nFFFBFBFLRR\nBFFBFBBLRR\nFBFFBBFRLR\nFFFBFBFRRL\nFFFBBBBRRL\nBFBFFBBLLR\nBFBFFFFRRR\nFFBBBFFLRL\nFBBFBBFRLR\nFFFBFFFLRL\nBFFBBBBRLL\nBFBBFFFRLR\nFFBFFBBRRR\nFBFBFFBRLL\nBFFFBFFLLL\nFFBFBFBRRL\nFFFBFBBRRL\nBFBBBBFLRL\nBBFFBFFRLR\nBFFFBBFRLL\nBFFFBBBLRR\nBFBFFBFRRR\nFBBBFBFLLR\nBFBBFFFLRL\nFFFBFBFRLR\nFBFBFBFLRR\nBBFFBFBRRR\nBBFFFBBRRR\nFFBBFFBLLL\nBFBBBFBLLL\nFFBBFFFLRL\nFFFBFBBRRR\nFFBBBFBRLR\nFFBBFBBRLR",
+    816,
+    539,
+    "FFBFFBFLLR\nBFBFBFBRLR\nFBFFBFBRRL\nBBFFFBFLLR\nBFFFFFBLRL\nBFBBFBBRLR\nFBFBFBBRRL\nFBBBFBFLRL\nBFFFFBBRRL\nFBBBBBFLLR\nFBFFFBFLLR\nFFBBBFBLRR\nBFFFBBFLRL\nFFBBBFBLLL\nFBFBBFFLRR\nFBFBBBBLLR\nFFBBBFFRLR\nFBFBBFFRRL\nFBBBFFFLRR\nFFBBFFBLLR\nBFBFBFBLLR\nBFBBFFFLLL\nFBFBBFBRRL\nFFBBFFBLRR\nFFFBFBBLRR\nFFFFBBBRRL\nBFBBBBFRLL\nFBFBBFFLLR\nFFFBFBBLLR\nFFBFBBBRRL\nBFFBBFFLLR\nBFFFBBFLLR\nFBBFBBFRRR\nFFFBBFFRLL\nBFBFBBFRRR\nFBBFFBBRLR\nFBFFBBFRRL\nBFFFFBFLLR\nBBFFBFBLRL\nFBFBBBBLRL\nFBFBFBFRLL\nFFBFBFFLLR\nBFBFBBBRLR\nBFBBFFBLRR\nFFBFBFFLLL\nFFFFBBBLLR\nFFBBBBBLRR\nFFBFBBFLLR\nBFFFBBBRLL\nBFBFBFFLRR\nFBBBBFFRLR\nBFBFFBBRLL\nFBFBBFBRRR\nBFBFFBBLLL\nBFBFBBFLRL\nBBFFFFBLLR\nFFFBFFFRLL\nFBFFFBBRLL\nFFFBBBBLRR\nFBBBFBBLRL\nBBFFBFFRRR\nBFFBFFFLRR\nFBBBBBBLLR\nBFBBFFBLLL\nBFFBBBFRLR\nFBBBFFFLRL\nBBFFFBFRRL\nBBFFBFFRRL\nFFBBFFFLRR\nBBFFBFBRLR\nBBFFFFBRLR\nFFFFBBFLRL\nFFBBBFBRRR\nFFBFBBFLRR\nFFFBBFBRRR\nFBFFBFBRRR\nBFFFFFBRRR\nBFBFFBFLRL\nFFBBBBBLLL\nBFFFFBBRRR\nFBFBFFBRLR\nFBFFFFBLLR\nFFFBFFFLLL\nBFFFFFFLLR\nBBFFFFBRRL\nFBBFBBBRRL\nBFFFFFFLLL\nFBFBFFBLRR\nFFFBBFFRLR\nFBBFFBBLLL\nBFBBBFFLRL\nBFBFBBBRLL\nFBFBBFFLLL\nFFBBFFBRLL\nBFFFFBBLLR\nFFBBFBFLLL\nFFFFBBBRRR\nBFFFBBBLLL\nBFFBBFBRLR\nFFBBBFFLLR\nFBBBFFBRLR\nFFBFBFBRLL\nFFFBFFFLRR\nFFBBFBBLRL\nBFFBFBFLRR\nFFBFFFFRLR\nBFBFFFFLLL\nBFBBFBFRRR\nBFBFBFFRLL\nBFFFBFBRLL\nFFBBFBFLRR\nFFBFFFFRRL\nBFBBBBFLLL\nBFBFBFBRRR\nFFFFBBFLRR\nFBBBBFFLRL\nBFBFFFBLRL\nBFFBBFBRLL\nFBFFBBFLLL\nBFBBBBBLRL\nFFBFBBBLLR\nFBBFFBBLRR\nFFBFFFBRRR\nFFFBBBBRRR\nFBBBBBBLRL\nFFFBBFFLLL\nBBFFBFBLRR\nBFFBFFBRLR\nFFBBFBBRLL\nBFFBFBBRLR\nFBFBFBFRRL\nBFBBBFBRLL\nFFFFBBBLRR\nBBFFFFFLLR\nFFBFFFFLLL\nBFBBFBFRLR\nFFBBBBFRLR\nBFBFBBBLRL\nFBBFFFFRLL\nBFFFBBBLRL\nFBBBBFBRLL\nFFBFFFFRRR\nBFFFBFBRLR\nFBBBFBBLLL\nFFBFFBBLLL\nFBFFFBFRLL\nFFFFBBBRLR\nBBFFFBBLLL\nBBFFFFBLRL\nFBFBBBFRRL\nBFBFFBBRLR\nFBBFFBBLLR\nBFFBBFFLLL\nFFBFBBBLLL\nBFFBFFFRRL\nBFBBBFBRLR\nFBFBBBFLLL\nBFFBBBBLRR\nBFFBFBFLLL\nFFFBBBBLLR\nBFFFBFFLLR\nFFFBFBBRLL\nFFBBBBFRRL\nFBFBFBBLLR\nBBFFBFFLRL\nFBFBFBFLLR\nFBFBFFFLLR\nBFFFBFFLRR\nBFFBFFBRRR\nFFFFBBBLLL\nFBBFBFFLLR\nFFBBFBFRRL\nBFBFBBBLRR\nBFFFFFBRLR\nBFBBBBFRRL\nFFFBBBFLRR\nFBBFBFBRLL\nFFFBFFBRLR\nBFFFFFFRRL\nBFFBFBFRRL\nBFBBBBBRRL\nFBBFBFFRLL\nFFBFBBBRRR\nBBFFFBFLRL\nFBBBFBFRRR\nFBFFFFFLRL\nFFFBBBBRLR\nFBFFFFBLRR\nFBBFFBBLRL\nBFFFBBFRRL\nFFFBFBFRRR\nBFFBFFFRLL\nBFBFBFFLLR\nBFFBFBFRLL\nFFBFFBBLRR\nBFBBBBBLLR\nFBBBFFBRRL\nFFBBBFFRLL\nBFFBFBFLLR\nFFBBBBFRLL\nBFBFFFBRLR\nFBFFFFBLRL\nFFFBBFBRLR\nFFFBBFFLLR\nBFBBBBFLRR\nFFBFBFFRRR\nBFFBBBBRLR\nFBFFFFFRRL\nBFFFFFBRLL\nFFBBBBFLRL\nBFFFBFBRRL\nFFBFBFFLRL\nFFFBFFBLLR\nFBBBBFBRRL\nFFFFBBFLLR\nBFBFFBFRLR\nFBFFBFBRLL\nBFBBFBFLRL\nFBBFBBFLLL\nBFFFFFFRLL\nFBBBBBBRLR\nBFFFFBBRLL\nFBFBBFFLRL\nFBFFBBFLLR\nFBFBFBBLLL\nFFBBBFBLRL\nFFBBBBFRRR\nFFFFBBBLRL\nBFBBBBFRLR\nBFBFBBFLRR\nFFFBBFFRRL\nFFFBBFFRRR\nFBBBBFBRRR\nFFBBBBBRLL\nFBBBFFBRLL\nFBFBBFBLRR\nBFFFBBBRRR\nBFFBBBBLLR\nBFFFFBFRLR\nFFBFFFFLRR\nBFBBBFBLRR\nFBFFFBBRLR\nFFBFFBFLRR\nFFFBFBBLLL\nBFFBFFBRLL\nFBFFFBFLRR\nFBBBFBFLLL\nFBBFFBFLRR\nFFBFBFBRLR\nFFBBBBBRRR\nBFFBFBBRRL\nBFBFFBBRRR\nBFFBBFBLRR\nFBBFFFBRLL\nFBBBBFFLLL\nFBBBBBFLRR\nFBFFBBBLLL\nBFFBFFBLRR\nBBFFFFBRRR\nFBFFFBBLLL\nFFFFBFBRRL\nFBBFBBBRLR\nFBFBFBFRLR\nFBFFBFFRRR\nBBFFFBFRLL\nBFBBBFFRLR\nFBBFFFFRRL\nBBFFFFFLLL\nFBBFFFBRLR\nFFBFBFFRLR\nBFBFBBFRRL\nBFBBBBFLLR\nBFFFBBFLRR\nFBFBFBBLRR\nFBBFFFBLRR\nBFBFFBFLLR\nBFBFBFBRRL\nFBFBFFBLLR\nFBBFBFBLRL\nBFBFFFFRLR\nBFFBBFBLLR\nBFBBFBBRLL\nBFBFBFFRRR\nFFBBFBBLLR\nFFBFBFFLRR\nFFFBFFBLRR\nBFBFFFFLLR\nFBFBFFBRRL\nFBFBFFFRRL\nFBFFFFBRRR\nFFBBFFFRRL\nFFBFFBBRLR\nBFBBBFFLRR\nBBFFBFBRLL\nFBBBBBFRRL\nBBFFFFBRLL\nFFBBBBFLLR\nFBBBBFBRLR\nFFBFFBFRLR\nFFFFBFBRLR\nFBBFBFBLRR\nBFBBFBFLLR\nFFBFBBFRRR\nFFBBFBFRRR\nFFBBBBBLLR\nBFBBFFBLRL\nBFFBBFFRLR\nFFFBBFBRRL\nBFFFFFBRRL\nBFBBFFBRRL\nBFFBFBBLLR\nFBBFFFBRRR\nFBBBFFFLLR\nBFFFFFFRLR\nFBBFFFFLRL\nBFFFBBBLLR\nBFFFBFBLLR\nBFFFBFBLLL\nFFFBFBFLRL\nFBFBBBBRLR\nFBFFFFFRLL\nBFFBFFBLLL\nFFBBFBBLLL\nFBFFBFFRLL\nFBFFBFBLLR\nFFBBFBBLRR\nFFBFFFFLLR\nFBBFBFBRLR\nFFBBBFBRRL\nFBFBBFBRLL\nFBFFBFFLLR\nBFFBFBFRRR\nFBBFBBBLLR\nBFBBFFBRLR\nBFFBBFBRRR\nFBBBBFFLRR\nBBFFFFFLRR\nFBFFFFFRRR\nFFBBFFFLLR\nFFBFFBBRRL\nFBBFBBFRRL\nBFFFFFBLLL\nFFFBBBFLLL\nFBBBBFBLRR\nFFFBBFFLRL\nBFBBBFBLRL\nFBBFBBFLRL\nFBBFFBBRRR\nFFBFBBFRLL\nFBFFFFFLLL\nFFFBBBFRLL\nBFBBFBBRRR\nFFFBBFBRLL\nFFBBBFBRLL\nFBBBBBBLRR\nFBFBBBBRLL\nFFBBFFFRRR\nFFFBFFBLRL\nFBBFBFFRLR\nBFFFBFFRRR\nBFFFBFFRLL\nFBBBFFFRLR\nFFBBFBBRRL\nFFBBFFFRLL\nFFBBFBFRLL\nFBBFFBFLLL\nFBFFBBBRRL\nFBFBFFBRRR\nFFBFBFFRLL\nFFFBFFFRRR\nFFBFFFBLRR\nFFBFFFFRLL\nFBFFFFBLLL\nBFBBFFFLLR\nBFBFBFBLRR\nFBBFFFFLLR\nFBBBBFBLLR\nFBBBFBFRRL\nBFBBBBBRLR\nFBFFBFBLLL\nFFBBFFFLLL\nFFFFBBFRRR\nFBFBBBFLRR\nBBFFBBFLLL\nFFBBBFBLLR\nFBBFFFFLRR\nFBBFBFBLLR\nFFBFFFBLRL\nFBFBBBFLLR\nBFBBBFFLLR\nBFFFFBBRLR\nFBBFBFBLLL\nFBBFFBFRLR\nBFBBFFBRLL\nFBBBBFFLLR\nBFFBBBBLRL\nFBBFBFBRRL\nFBBFFBFLRL\nFBBBFFFRRL\nBFBFBFFLRL\nBFBFFFFLRR\nBBFFBFBLLL\nBBFFBFFLLR\nFBFBBFFRLL\nBFFBBBFLLR\nFFBFBBFRRL\nFBFFBBBLRR\nFBBBFBBLLR\nFBFFFBBRRL\nFFBBBBFLRR\nBFBFFBFLRR\nFFFBBBFLRL\nFBFFBFFRRL\nFBBFBFFLRL\nFBBFFFBLRL\nFFBBBFFRRR\nBBFFFBFLLL\nFBFFBFBLRR\nBBFFBFBLLR\nFBFFFFBRLR\nBFFFFFFLRL\nFBBFBBBLLL\nBFBFFBBLRR\nFFBBFFBRRL\nBFFFBFBLRL\nBFBBBBBRLL\nFBFFFBBLRR\nBFBBFBBLLR\nFBBBFFBLLL\nFBFFFBFRLR\nFFBBBFFRRL\nFFBFBBFLRL\nFFBFBFBLLR\nFBBFFBFRLL\nFFBBBFFLLL\nBFBBBBBLLL\nFBBFFFBLLL\nBFBFFFBLRR\nFFBFFBFLRL\nBFFFBBBRRL\nBFBFBBBLLR\nBFBFFFFLRL\nBFBFBFFRLR\nFBBFBBBRLL\nFFBFBBBLRL\nFBBBFFFRRR\nFBBFBFFRRR\nFFFFBBBRLL\nFBFFBBBLRL\nFBBBFBBRRL\nFBFFBBBLLR\nFFFBBBFRLR\nBFBFBFBLLL\nBBFFFFFRRR\nBFBBFBFRLL\nFBBBFBBLRR\nBFFFFBFLRL\nFBBBFFBLRL\nBBFFBFFLRR\nFFBFBFFRRL\nFBFFFBBLRL\nFFFFBBFRLL\nFFBFFFBLLR\nFBBBFBFLRR\nFFFBFFBRRR\nBFFBBBFLRR\nFBFFBFBLRL\nFFBFFFBRLR\nFFBFFFBRRL\nFFBBFFBLRL\nFBFBFBBRLR\nFFFBBBBLLL\nBFBBBFFLLL\nFFBBFBFLRL\nBFBFFFBRLL\nBFFBFFBRRL\nFBBFFFFLLL\nFBFFFFBRRL\nFFFBFFFLLR\nFBFFFBFLRL\nFFBFBFBLRL\nFBBBFFBLRR\nBFBFFBBRRL\nBBFFFFFLRL\nBFFBBBFRRR\nBFFFBBFRLR\nBFBBBBBLRR\nBFFFBFBLRR\nFBBBBFFRLL\nBFBBFFFLRR\nBBFFFBFLRR\nFBFFBBFRLL\nFBFFFFFLRR\nFBBBBBFLLL\nFBFBFBBRRR\nFBFBBFBLRL\nBFBBFFBRRR\nBBFFBFFLLL\nFBBFBBFLLR\nBFFBBFFRLL\nFBFBBBBRRL\nFBBFBBFRLL\nFBBFBBFLRR\nFBBFFBBRRL\nFFFBBFFLRR\nBFBBFBFLRR\nBFBFBFFRRL\nFFBBFFFRLR\nFBFBBBBLRR\nBBFFFFBLLL\nBBFFFFFRLR\nFBFBFFBLRL\nFBBFBBBLRR\nBFBFFBFRRL\nBFFFFFFLRR\nFFFBBBFLLR\nFBFFBFFLLL\nBFFBFFFRLR\nFBBFFBFRRR\nFFFBFFFRRL\nFFFFBBFLLL\nBFBFBBFRLL\nBFFFFBBLRL\nFFFBFBFLLR\nFBBBFBFRLR\nBFBBBFBLLR\nBFFBFBFLRL\nBFBFFFBLLL\nFFBFFBBLRL\nBFBBFFFRRR\nFFFBFFBLLL\nBBFFFFBLRR\nFFBBFBFLLR\nFBFBBBBRRR\nFBFBBBFLRL\nBFFBFFFLLR\nFBFBFFFRLL\nFBBBBFBLLL\nFBBBBBBRRR\nFBBBBBFLRL\nBFBBFFFRLL\nFFFBFFFRLR\nFFBFFBBLLR\nFFBFBBBRLL\nBBFFFFFRLL\nFFFBFBBRLR\nFFBFBBBRLR\nBFFFFBFLRR\nBBFFFBBLRL\nFBBBBFFRRL\nFBBFBBBRRR\nBFBFFFFRLL\nFBBBBBBRRL\nBFFBFFBLRL\nBFFBBFBRRL\nFBFFFBBRRR\nBFFBBFBLLL\nBFBFBFBRLL\nBBFFFBBRLL\nFBFBBBBLLL\nBBFFFBBLRR\nBFFFFBFRRL\nBFBBFBFLLL\nFBBBFBBRLR\nBFBBFBBLRL\nBFBBBFBRRL\nFBFFBBBRLL\nFBBBFBBRLL\nFFBBBBBRRL\nFBFBBBFRLL\nBFFFFFBLLR\nBFFBBBBRRR\nBFBBFBBLLL\nFFBFBBBLRR\nFFBFBFBRRR\nBFBFFFBRRL\nFBBBFFBRRR\nBFBFBBBLLL\nBBFFFBBRLR\nBBFFFBBRRL\nFBBBBFBLRL\nFBBBBFFRRR\nFBBBFFBLLR\nBFFFFBFLLL\nBFBFBFFLLL\nBFFFFBFRLL\nBFFBFBBLRL\nFFBFFBBRLL\nFBFBBFFRLR\nBFBFFFBRRR\nFBFFBFBRLR\nFBBFBFFLLL\nFBFFBBFLRR\nFFBFBFBLRR\nFFBBFFBRRR\nFBBFFFBLLR\nFBFFFBFRRR\nBFBFFFFRRL\nFBBBBBFRRR\nFFFFBFBRRR\nFFBBBBBLRL\nFBFBFFFLRR\nFBBBFFFLLL\nBFFFBFFRLR\nFFBFFBFRRL\nFBFBFFFRRR\nFFBFBFBLLL\nBFBBBFFRRL\nBFBFBFBLRL\nBFFFBFFLRL\nFBBBBBFRLR\nFBFBBFBLLL\nFFFBBBBLRL\nFBBFBFFLRR\nBFBBBFBRRR\nFBBBFBBRRR\nFFFBFBFRLL\nFFBBBBFLLL\nFBFFBFFRLR\nFBFBFBFLLL\nFBFFFFBRLL\nFBBBBBBRLL\nFBBBBBFRLL\nFBBFFBFLLR\nBFFBBBFRRL\nFBFFFBFLLL\nFBBBFFFRLL\nBFFFBBBRLR\nFBFBBBFRRR\nBFFBFBBRLL\nFBBFFFBRRL\nFBFFBBBRLR\nBFBFFBBLRL\nBFFBBFFLRL\nFBFBBBFRLR\nBBFFFBBLLR\nFFFBFFBRLL\nBFBFBBFLLL\nBFFBBFBLRL\nFFFBBFBLRL\nBFBBBFFRRR\nFBFBFFFRLR\nBFFBBFFLRR\nBFFBBBFLLL\nFBFFBBBRRR\nFBFBFFFLLL\nFFBFBBFLLL\nFBFFFFFLLR\nBFBBFBFRRL\nBFBBBBBRRR\nBFFBFBBLLL\nBFFFBFBRRR\nFFFBFBFLLL\nFFBFFFBRLL\nFBBFBFBRRR\nFBBBBBBLLL\nFBFBFFFLRL\nFFFBFFBRRL\nFBFBBFFRRR\nFFBBFBFRLR\nBFFFFBBLLL\nFBBFBFFRRL\nBFFBFFFRRR\nBFFBFFFLRL\nBFBBBBFRRR\nBFBFBBBRRL\nFFFBBFBLLL\nFBBFFBBRLL\nBFBBBFFRLL\nBFFFBBFRRR\nFBFBFBBRLL\nBFFFBFFRRL\nFFFBFBBLRL\nFBFFFBBLLR\nBFFFBBFLLL\nBFFBFFFLLL\nBFFBBBBLLL\nBFFFFFFRRR\nFFBFFBFRLL\nFBBFFBFRRL\nBBFFFFFRRL\nFFFBBFBLRR\nBFBFFBFLLL\nFBFBBFBRLR\nBFBBFBBLRR\nFBFFBFFLRL\nBFBFFFBLLR\nFFFBBBFRRR\nFFBBBFFLRR\nFBFBFFBLLL\nBFBFBBFRLR\nFBFBFBFRRR\nFFBBFBBRRR\nFBBFFFFRLR\nFBFBFBBLRL\nFBBFFFFRRR\nFFBBFFBRLR\nBFBFBBFLLR\nBFFBFBFRLR\nBBFFBFFRLL\nBFFBBFFRRR\nBFFFFBFRRR\nFBFBFBFLRL\nFFBFFBFRRR\nFBBFBBBLRL\nFFFBBBFRRL\nFBFFBFFLRR\nFFBFBBFRLR\nBBFFBFBRRL\nFBFFBBFLRL\nFFFFBBFRRL\nBFBFFBFRLL\nBFBBFBBRRL\nFFFBBFBLLR\nFFBFFBFLLL\nFFBBBBBRLR\nBFFFFFBLRR\nBFFBBBFLRL\nBFBBFFBLLR\nFBFFFFFRLR\nFFFFBBFRLR\nFBFFFBFRRL\nBFFBBBBRRL\nFFBFFFFLRL\nBFBBFFFRRL\nFFFBBBBRLL\nBFBFBBBRRR\nBBFFFBFRLR\nBFFBFFBLLR\nFBFBBFBLLR\nBFFBFBBRRR\nBBFFFBFRRR\nBFFBBFFRRL\nFBBBFBFRLL\nFFBFFFBLLL\nBFFBBBFRLL\nFBFFBBFRRR\nFFFBFBFLRR\nBFFBFBBLRR\nFBFFBBFRLR\nFFFBFBFRRL\nFFFBBBBRRL\nBFBFFBBLLR\nBFBFFFFRRR\nFFBBBFFLRL\nFBBFBBFRLR\nFFFBFFFLRL\nBFFBBBBRLL\nBFBBFFFRLR\nFFBFFBBRRR\nFBFBFFBRLL\nBFFFBFFLLL\nFFBFBFBRRL\nFFFBFBBRRL\nBFBBBBFLRL\nBBFFBFFRLR\nBFFFBBFRLL\nBFFFBBBLRR\nBFBFFBFRRR\nFBBBFBFLLR\nBFBBFFFLRL\nFFFBFBFRLR\nFBFBFBFLRR\nBBFFBFBRRR\nBBFFFBBRRR\nFFBBFFBLLL\nBFBBBFBLLL\nFFBBFFFLRL\nFFFBFBBRRR\nFFBBBFBRLR\nFFBBFBBRLR"
+  )
+}
